+++ src/spherical/pipeline/run_trap.py
@@ -12,6 +12,7 @@ seamless integration into automated data processing workflows.
 """
 
 import os
+import time
 from copy import deepcopy
 from pathlib import Path
 from typing import Union
@@ -26,6 +27,12 @@ from trap.reduction_wrapper import run_complete_reduction
 from spherical.database.ifs_observation import IFSObservation
 from spherical.database.irdis_observation import IRDISObservation
 from spherical.pipeline import ifs_reduction
+from spherical.pipeline.logging_utils import (
+    PipelineLoggerAdapter,
+    get_pipeline_log_context,
+    get_pipeline_logger,
+    remove_queue_listener,
+)
 from spherical.pipeline.pipeline_config import IFSReductionConfig
 from spherical.pipeline.toolbox import make_target_folder_string
 
@@ -107,121 +114,187 @@ def run_trap_on_observation(
     name_mode_date = make_target_folder_string(observation)
     result_folder = os.path.join(str(reduction_config.directories.reduction_directory), 'IFS/trap', name_mode_date)
     
-    trap_parameters = trap_config.get_reduction_parameters()
-    trap_parameters.result_folder = result_folder
+    # Create TRAP result folder
+    os.makedirs(result_folder, exist_ok=True)
 
-    if continuous_satellite_spots:
-        file_identifier = "center"
-    else:
-        file_identifier = "coro"
-
-    wavelengths = (
-        fits.getdata(os.path.join(data_directory, "wavelengths.fits")) * u.nm
-    ).to(u.micron)
-    used_instrument.wavelengths = wavelengths
-
-    pa = pd.read_csv(
-        os.path.join(data_directory, f"frames_info_{file_identifier}.csv")
-    )['DEROT ANGLE'].values
-
-    data_full = fits.getdata(
-        os.path.join(data_directory, f"{file_identifier}_cube.fits")
+    # Initialize logging for TRAP session with trap_ prefix for log files
+    context = get_pipeline_log_context(observation)
+    logger = get_pipeline_logger(
+        f"trap_{name_mode_date}", 
+        Path(result_folder), 
+        verbose=trap_config.processing.verbose,
+        log_prefix="trap_reduction"
     )
+    logger = PipelineLoggerAdapter(logger, context)
+
+    start_time = time.time()
+    logger.info("TRAP session started", extra={"step": "trap_session_start", "status": "started"})
+
     try:
-        flux_psf_full = fits.getdata(
-            os.path.join(data_directory, "psf_cube_for_postprocessing.fits")
+        # Early exit logging
+        if not reduction_config.steps.run_trap_reduction and not reduction_config.steps.run_trap_detection:
+            logger.info("No TRAP processing steps enabled", extra={"step": "trap_session", "status": "skipped"})
+            return
+
+        # Log observation details
+        logger.info(f"Starting TRAP processing for {name_mode_date}")
+        logger.info(f"TRAP steps enabled - Reduction: {reduction_config.steps.run_trap_reduction}, Detection: {reduction_config.steps.run_trap_detection}")
+        logger.debug(f"TRAP result folder: {result_folder}")
+    
+        trap_parameters = trap_config.get_reduction_parameters()
+        trap_parameters.result_folder = result_folder
+
+        if continuous_satellite_spots:
+            file_identifier = "center"
+        else:
+            file_identifier = "coro"
+
+        logger.debug(f"File identifier: {file_identifier}")
+        logger.debug(f"Temporal components fraction: {trap_config.processing.temporal_components_fraction}")
+
+        wavelengths = (
+            fits.getdata(os.path.join(data_directory, "wavelengths.fits")) * u.nm
+        ).to(u.micron)
[…]
