### 2025-08-06 86a0628
Fix trap logging
