### 2025-08-06 28a6623
Fix try-except logic in run_trap
