+++ src/spherical/pipeline/run_trap.py
@@ -14,7 +14,9 @@ seamless integration into automated data processing workflows.
 import copy
 import os
 import time
+import traceback
 from copy import deepcopy
+from datetime import datetime
 from pathlib import Path
 from typing import Union
 
@@ -227,65 +229,137 @@ def run_trap_on_observation(
             logger.info("TRAP reduction completed", extra={"step": "trap_reduction", "status": "success"})
 
         if reduction_config.steps.run_trap_detection:
-            logger.info("Starting TRAP detection", extra={"step": "trap_detection", "status": "started"})
-            logger.debug(f"Detection threshold: {trap_config.detection.detection_threshold}")
-            logger.debug(f"Candidate threshold: {trap_config.detection.candidate_threshold}")
-            logger.debug(f"Search radius: {trap_config.detection.search_radius} pixels")
-            
-            analysis = DetectionAnalysis(
-                reduction_parameters=None,
-                instrument=None,
-            )
-            
-            logger.debug("Reading TRAP reduction outputs")
-            analysis.read_output(
-                trap_config.processing.temporal_components_fraction[0],
-                result_folder=trap_parameters.result_folder,
-                reduction_type="temporal",
-                correlated_residuals=False,
-                read_parameters=True,
-                read_instrument=True,
-            )
+            try:
+                logger.info("Starting TRAP detection", extra={"step": "trap_detection", "status": "started"})
+                
+                # Fix B: Enhanced diagnostic logging for TRAP parameters
+                logger.debug("TRAP detection parameters:")
+                logger.debug(f"  - Detection threshold: {trap_config.detection.detection_threshold}")
+                logger.debug(f"  - Candidate threshold: {trap_config.detection.candidate_threshold}")
+                logger.debug(f"  - Search radius: {trap_config.detection.search_radius} pixels")
+                logger.debug(f"  - Use spectral correlation: {trap_config.detection.use_spectral_correlation}")
+                logger.debug(f"  - Inner mask radius: {trap_config.detection.inner_mask_radius}")
+                logger.debug(f"  - Good fraction threshold: {trap_config.detection.good_fraction_threshold}")
+                logger.debug(f"  - Theta deviation threshold: {trap_config.detection.theta_deviation_threshold}")
+                logger.debug(f"  - YX FWHM ratio threshold: {trap_config.detection.yx_fwhm_ratio_threshold}")
+                
+                # Stellar parameters diagnostic
+                stellar_params = trap_config.get_stellar_parameters()
+                logger.debug("Stellar parameters:")
+                logger.debug(f"  - Effective temperature: {stellar_params.get('teff', 'Not set')} K")
+                logger.debug(f"  - Surface gravity: {stellar_params.get('logg', 'Not set')}")
+                logger.debug(f"  - Metallicity: {stellar_params.get('feh', 'Not set')}")
+                
+                # Processing parameters diagnostic
+                logger.debug("Processing parameters:")
+                logger.debug(f"  - Temporal components fraction: {trap_config.processing.temporal_components_fraction}")
+                logger.debug(f"  - Species database directory: {species_database_directory}")
+                
+                analysis = DetectionAnalysis(
+                    reduction_parameters=None,
+                    instrument=None,
+                )
+                
+                logger.debug("Reading TRAP reduction outputs")
+                analysis.read_output(
+                    trap_config.processing.temporal_components_fraction[0],
+                    result_folder=trap_parameters.result_folder,
+                    reduction_type="temporal",
+                    correlated_residuals=False,
+                    read_parameters=True,
+                    read_instrument=True,
+                )
 
-            # Update result folder in case data was copied after reduction phase
-            # The saved parameters don't know about potential folder moves
[…]
