### 2025-08-06 067752b
Full journal names in .bib / Shortened paper
