+++ joss/paper.md
@@ -23,63 +23,42 @@ bibliography: paper.bib
 
 ### Summary
 
-The Spectro-Polarimetric High-contrast Exoplanet REsearch instrument (SPHERE; @Beuzit:2019) at the Very Large Telescope (VLT) is among the most advanced ground-based instruments for coronagraphic imaging of exoplanets and circumstellar disks in the optical and near-infrared. Over the last decade, SPHERE has been extraordinarily productive, contributing to over 400 [publications](https://telbib.eso.org/) and leading one of the largest ground-based surveys of exoplanets and circumstellar disks (SHINE; see e.g., @Chauvin:2017; @Chomez:2025). It has played a pivotal role in probing the outer regions of planetary systems, revealing both planetary-mass companions and intricate disk structures around nearby stars.
+The Spectro-Polarimetric High-contrast Exoplanet REsearch instrument (SPHERE; @Beuzit:2019) at the Very Large Telescope (VLT) is among the most advanced ground-based instruments for coronagraphic imaging of exoplanets and circumstellar disks in the optical and near-infrared. Over the last decade, SPHERE has contributed to over 400 [publications](https://telbib.eso.org/) and led major surveys of exoplanets and disks (SHINE; @Chauvin:2017; @Chomez:2025), revealing planetary companions and disk structures around nearby stars.
 
 The spherical software package offers both a ready-to-use, searchable database of all SPHERE observations from the ESO archive—available via [Zenodo](https://doi.org/10.5281/zenodo.15147730)—and a framework for generating and updating this database using the latest archive contents, allowing users to explore, download, and analyze high-contrast imaging datasets efficiently.
 The SPHERE instrument consists of three sub-instruments. The InfraRed Dual-band Imager and Spectrograph (IRDIS; @Dohlen:2008; @Vigan:2010), the Integral Field Spectrograph (IFS; @Claudi:2008; @Mesa:2015), and the optical Zurich Imaging Polarimeter (ZIMPOL; @Schmid:2018).
-As of May 2025, the **spherical** database includes ~6000 IRDIS observations in the dual-band imaging (DBI) mode, ~1000 in the IRDIS dual-beam polarimetric imaging (DPI; @deBoer:2020; @vanHolstein:2020) mode, and ~4500 in the IFS mode. As discussed in Section **Future work** below, the ZIMPOl instrument, IRDIS long-slit spectroscopy (LSS; @Vigan:2008) mode, and Sparse Aperture Masking (SAM; @Cheetham:2016) are not yet supported, but are planned to be included in future updates.
+As of May 2025, the spherical database includes ~6000 IRDIS observations in the dual-band imaging (DBI) mode, ~1000 in the IRDIS dual-beam polarimetric imaging (DPI; @deBoer:2020; @vanHolstein:2020) mode, and ~4500 in the IFS mode. The ZIMPOl instrument, IRDIS long-slit spectroscopy (LSS; @Vigan:2008) mode, and Sparse Aperture Masking (SAM; @Cheetham:2016) are not yet supported, but are planned to be included in future updates.
 
-Unlike the ESO archive interface, **spherical** consolidates observational metadata (instrument setup, total exposure time, total parallactic angle rotation, quality flags), target star properties, and observing conditions into a single table, significantly simplifying the identification of optimal datasets for scientific analysis. It integrates an parallelized *end-to-end data reduction pipeline* specifically designed for SPHERE’s IFS, enabling efficient spectral cube extraction, calibration, post-processing, and exoplanet characterization. For IRDIS, **spherical** supports the discovery and download of dual-band imaging and polarimetric sequences. By offering an intuitive path from archival data to calibrated products, **spherical** empowers researchers to maximize the scientific yield of SPHERE observations.
[…]
