+++ joss/paper.bib
@@ -33,7 +33,7 @@ archivePrefix = {ascl},
 @ARTICLE{Holstein:2020,
        author = {{van Holstein}, R.~G. and {Girard}, J.~H. and {de Boer}, J. and {Snik}, F. and {Milli}, J. and {Stam}, D.~M. and {Ginski}, C. and {Mouillet}, D. and {Wahhaj}, Z. and {Schmid}, H.~M. and {Keller}, C.~U. and {Langlois}, M. and {Dohlen}, K. and {Vigan}, A. and {Pohl}, A. and {Carbillet}, M. and {Fantinel}, D. and {Maurel}, D. and {Orign{\'e}}, A. and {Petit}, C. and {Ramos}, J. and {Rigal}, F. and {Sevin}, A. and {Boccaletti}, A. and {Le Coroller}, H. and {Dominik}, C. and {Henning}, T. and {Lagadec}, E. and {M{\'e}nard}, F. and {Turatto}, M. and {Udry}, S. and {Chauvin}, G. and {Feldt}, M. and {Beuzit}, J. -L.},
         title = "{Polarimetric imaging mode of VLT/SPHERE/IRDIS. II. Characterization and correction of instrumental polarization effects}",
-      journal = {\aap},
+      journal = {Astronomy \& Astrophysics},
      keywords = {polarization, techniques: polarimetric, techniques: high angular resolution, techniques: image processing, methods: observational, protoplanetary disks, Astrophysics - Instrumentation and Methods for Astrophysics, Astrophysics - Earth and Planetary Astrophysics},
          year = 2020,
         month = jan,
@@ -62,7 +62,7 @@ archivePrefix = {arXiv},
 @ARTICLE{Gonzales:2017,
        author = {{Gomez Gonzalez}, Carlos Alberto and {Wertz}, Olivier and {Absil}, Olivier and {Christiaens}, Valentin and {Defr{\`e}re}, Denis and {Mawet}, Dimitri and {Milli}, Julien and {Absil}, Pierre-Antoine and {Van Droogenbroeck}, Marc and {Cantalloube}, Faustine and {Hinz}, Philip M. and {Skemer}, Andrew J. and {Karlsson}, Mikael and {Surdej}, Jean},
         title = "{VIP: Vortex Image Processing Package for High-contrast Direct Imaging}",
-      journal = {\aj},
+      journal = {The Astronomical Journal},
      keywords = {methods: data analysis, planetary systems, planets and satellites: detection, techniques: high angular resolution, techniques: image processing, Astrophysics - Instrumentation and Methods for Astrophysics},
          year = 2017,
         month = jul,
@@ -108,7 +108,7 @@ archivePrefix = {arXiv},
 @ARTICLE{Samland:2021,
        author = {{Samland}, M. and {Bouwman}, J. and {Hogg}, D.~W. and {Brandner}, W. and {Henning}, T. and {Janson}, M.},
         title = "{TRAP: a temporal systematics model for improved direct detection of exoplanets at small angular separations}",
-      journal = {\aap},
+      journal = {Astronomy \& Astrophysics},
      keywords = {planets and satellites: detection, methods: data analysis, techniques: high angular resolution, techniques: image processing, Astrophysics - Earth and Planetary Astrophysics, Astrophysics - Instrumentation and Methods for Astrophysics},
          year = 2021,
         month = feb,
@@ -126,7 +126,7 @@ archivePrefix = {arXiv},
[…]
