### 2025-08-07 6a84fa0
Fix try/except in run_trap
