### 2025-08-07 168978c
Savely shutdown ray when TRAP fails
