### 2025-08-07 35bc7f7
Add additional saveguards to trap
