+++ README.md
@@ -8,38 +8,68 @@
 [![License](https://img.shields.io/badge/License-BSD--3-blue.svg)](https://opensource.org/licenses/BSD-3-Clause)
 [![CI](https://github.com/m-samland/spherical/actions/workflows/ci.yml/badge.svg)](https://github.com/m-samland/spherical/actions/workflows/ci.yml)
 
-**spherical** provides comprehensive tools for exploring and analyzing data from the ESO VLT/SPHERE instrument, specifically tailored for high-contrast imaging observations with the IRDIS and IFS instruments.
+**spherical** provides tools for exploring and analyzing data from the ESO VLT/SPHERE instrument, with a focus on high-contrast imaging observations using **IRDIS** and **IFS**.
 
 ---
 
 ## Overview
 
-**spherical** features a specially curated observation database, uniquely built by analyzing headers from all available SPHERE data in the ESO archive. Unlike the official ESO archive, this database systematically discovers every SPHERE observation sequence and provides detailed summary statistics such as observing modes, target properties, integration times, and observing conditions.
+**spherical** ships with a **curated observation database**, constructed by parsing headers from all available SPHERE data in the ESO archive and cross-matching with external catalogs. This database systematically identifies SPHERE observation sequences and summarizes observing modes, target properties, integration times, and observing conditions.
 
-Users can easily explore and filter the database to identify observation sequences of interest and seamlessly feed these sequences into an end-to-end data analysis workflow for SPHERE IFS. This pipeline automates data discovery, downloading, calibration, planet detection, and spectral characterization. For the IRDIS instrument, data discovery and downloading of dual-band imaging (DBI) observations and polarimetric imaging observations is supported.
+From exploration to execution, you can filter sequences of interest and feed them into an **end-to-end IFS pipeline** that automates discovery, download, calibration, post-processing, and spectral characterization. For **IRDIS**, the package supports discovery and download of dual-band imaging (DBI) and polarimetric imaging observations.
+
+> ⚠️ **Note:** The database contains **metadata only**. It does **not** include already reduced data products.
+
+---
+
+## Usage Model
+
+**spherical** is both:
+
+1. **A curated observation database** — metadata about VLT/SPHERE observations in the ESO archive (not the reduced data).
+2. **A Python-based IFS reduction and analysis pipeline** — intended to be run from a **Python script** so that you can tune parameters and control each step.
+
+You can use **spherical** in two complementary ways:
+
+- **As a library**  
+  Import in Python or Jupyter to query/filter the database and to call pipeline components programmatically.
+
+- **As a script-driven workflow**  
+  Start from the provided example script (`ifs_reduction_template.py`) to configure and run a complete IFS reduction with sensible defaults and many tunable parameters.
+
+> ❗ **Why not a one-click CLI pipeline?**  
+> The IFS pipeline exposes many parameters (calibration strategy, post-processing choices, masking, spectral extraction options, etc.). A single command-line entry point would hide crucial configuration and often lead to suboptimal results. Instead, we provide a ready-to-run **Python template script** that you adapt to your dataset.
 
 ---
 
 ## Key Features
 
-- **Comprehensive Database**: Automatically generated by cross-referencing all available SPHERE data headers with the Gaia catalog to ensure accuracy and completeness.
[…]
