+++ joss/paper.md
@@ -1,5 +1,5 @@
 ---
-title: 'spherical: A Comprehensive Database and Automated Pipeline for VLT/SPHERE High-Contrast Imaging'
+title: 'spherical: A Curated Database and Automated IFS Analysis Pipeline for VLT/SPHERE High-Contrast Imaging'
 tags:
   - Python
   - astronomy
@@ -17,64 +17,95 @@ affiliations:
   - name: Max-Planck-Institut für Astronomie (MPIA), Königstuhl 17, 69117 Heidelberg, Germany
     index: 1
 
-date: 23 July 2025
+date: 09 August 2025
 bibliography: paper.bib
 ---
 
 ### Summary
 
-The Spectro-Polarimetric High-contrast Exoplanet REsearch instrument (SPHERE; @Beuzit:2019) at the Very Large Telescope (VLT) is among the most advanced ground-based instruments for coronagraphic imaging of exoplanets and circumstellar disks in the optical and near-infrared. Over the last decade, SPHERE has contributed to over 400 [publications](https://telbib.eso.org/) and led major surveys of exoplanets and disks (SHINE; @Chauvin:2017; @Chomez:2025), revealing planetary companions and disk structures around nearby stars.
+The Spectro-Polarimetric High-contrast Exoplanet REsearch instrument (SPHERE; @Beuzit:2019) at the Very Large Telescope (VLT) is a leading facility for coronagraphic imaging of exoplanets and circumstellar disks in the optical and near-infrared. Over the last decade, SPHERE has contributed to hundreds of publications and major legacy surveys (e.g., SHINE; @Chauvin:2017; @Chomez:2025).
 
-The spherical software package offers both a ready-to-use, searchable database of all SPHERE observations from the ESO archive—available via [Zenodo](https://doi.org/10.5281/zenodo.15147730)—and a framework for generating and updating this database using the latest archive contents, allowing users to explore, download, and analyze high-contrast imaging datasets efficiently.
-The SPHERE instrument consists of three sub-instruments. The InfraRed Dual-band Imager and Spectrograph (IRDIS; @Dohlen:2008; @Vigan:2010), the Integral Field Spectrograph (IFS; @Claudi:2008; @Mesa:2015), and the optical Zurich Imaging Polarimeter (ZIMPOL; @Schmid:2018).
-As of May 2025, the spherical database includes ~6000 IRDIS observations in the dual-band imaging (DBI) mode, ~1000 in the IRDIS dual-beam polarimetric imaging (DPI; @deBoer:2020; @vanHolstein:2020) mode, and ~4500 in the IFS mode. The ZIMPOl instrument, IRDIS long-slit spectroscopy (LSS; @Vigan:2008) mode, and Sparse Aperture Masking (SAM; @Cheetham:2016) are not yet supported, but are planned to be included in future updates.
+`spherical` is both a *curated, searchable database* listing all VLT/SPHERE observations and a *Python-based, automated analysis pipeline* for SPHERE’s *Integral Field Spectrograph (IFS)*. The database—archived on Zenodo and regenerable from the ESO archive—consolidates observational metadata, stellar properties, and observing conditions into a single, analysis-ready table. The pipeline takes users from raw IFS frames to calibrated spectral cubes and exoplanet characterization within a script-driven, configurable workflow.
 
-Unlike the ESO archive, spherical consolidates observational metadata, target star properties, and observing conditions into one table, simplifying dataset selection. It integrates an automated *end-to-end data reduction pipeline* for SPHERE’s IFS, enabling spectral cube extraction, calibration, and exoplanet characterization. For IRDIS, spherical supports downloading DBI and polarimetric sequences. By providing an intuitive path from archival data to calibrated products, spherical maximizes SPHERE's scientific yield.
+**What `spherical` can do**
[…]
