### 2025-08-10 7c5a861
Ruff linting
