+++ src/spherical/pipeline/run_trap.py
@@ -16,7 +16,6 @@ import os
 import time
 import traceback
 from copy import deepcopy
-from datetime import datetime
 from pathlib import Path
 from typing import Union
 
