### 2025-09-09 622f3af
Improve intermediate file cleanup
