### 2025-11-27 ba5a296
Closes #94
