### 2025-12-12 713585f
Fix newly introduced format string issue
