### 2026-02-01 #8690 @warrickball
Hi again @m-samland, and sorry about the long wait. We recently [announced changes to our requirements](https://blog.joss.theoj.org/2026/01/preparing-joss-for-a-generative-ai-future), which now appear in our docs on [what the paper should contain](https://joss.readthedocs.io/en/latest/paper.html#what-should-my-paper-contain), the [example paper](https://joss.readthedocs.io/en/latest/example_paper.html), and [scope definition](https://joss.readthedocs.io/en/latest/submitting.html#scope-and-significance).

Could you update the paper to match the new format, include the newly-required sections? This may be by re-arranging
existing material or writing new material. The maximum word count has increased from 1000 words to 1750 to accommodate the extra requirements. Once you have updated your paper, please comment here to let us know. At that point, we will assess your software against our updated scope criteria.

We want to be transparent that under these new criteria, some submissions that were previously in scope may now be deemed out of scope, and we may choose not to proceed with review. We recognise that this will be frustrating, especially for authors (like you) who have already been waiting many months for an editor to start handling a submission. The editorial board discussed this issue for some time and our consensus was that this choice will give JOSS's volunteer editors and reviewers the most clarity about what requirements to apply and support our commitment to ensuring JOSS publications recognise genuine scholarly contributions.

I'll also note that I don't think I regenerated the paper after your previous changes and there are a number of formatting things that have gone awry. I'll highlight three that I think should be resolved when you rework the paper:

1. The sections headings start at level 3 (e.g. `### Summary`) rather than level 1 (e.g. `# Summary`).
2. Starting from the "Design and Implementation", there is an excessive use of numbered or itemised lists. While lists are in principle fine (many JOSS papers have them), those sections have basically not leading or following text and I find very difficult to read.
3. The "Design and Implementation" section doesn't actually seem to describe either, and instead lists two workflows. It's not clear that these are relevant to the paper, which should principally be about scientific and software context. Those items can probably be kept in the documentation. Note that we do now require a "Software Design" section but our [docs suggest](https://joss.readthedocs.io/en/latest/paper.html#what-should-my-paper-contain) it should contain

   > * **Software design**: An explanation of the trade-offs you weighed, the design/architecture you chose, and why it matters for your research application. This should demonstrate meaningful design thinking beyond a superficial code structure description.
