### 2026-02-16 0fb2f91
Ruff linting
