### 2026-02-20 b70a0a6
Update paper to new JOSS requirements
