+++ joss/paper.md
@@ -21,88 +21,51 @@ date: 09 August 2025
 bibliography: paper.bib
 ---
 
-### Summary
+# Summary
 
-The Spectro-Polarimetric High-contrast Exoplanet REsearch instrument (SPHERE; @Beuzit:2019) at the Very Large Telescope (VLT) is a leading facility for coronagraphic imaging of exoplanets and circumstellar disks in the optical and near-infrared. Over the last decade, SPHERE has contributed to hundreds of publications and major legacy surveys (e.g., SHINE; @Chauvin:2017; @Chomez:2025).
+The Spectro-Polarimetric High-contrast Exoplanet REsearch instrument [SPHERE; @Beuzit:2019] at the Very Large Telescope (VLT) is a leading facility for coronagraphic imaging of exoplanets and circumstellar disks in the optical and near-infrared. Over the last decade, SPHERE has contributed to hundreds of publications and major legacy surveys [e.g., SHINE; @Chauvin:2017; @Chomez:2025].
 
-`spherical` is both a *curated, searchable database* listing all VLT/SPHERE observations and a *Python-based, automated analysis pipeline* for SPHERE’s *Integral Field Spectrograph (IFS)*. The database—archived on Zenodo and regenerable from the ESO archive—consolidates observational metadata, stellar properties, and observing conditions into a single, analysis-ready table. The pipeline takes users from raw IFS frames to calibrated spectral cubes and exoplanet characterization within a script-driven, configurable workflow.
+`spherical` streamlines the end-to-end analysis of SPHERE data by integrating a curated, searchable database with an automated Python-based pipeline. The software enables users to filter the complete observation history by target properties or observing conditions, automatically download raw datasets from the ESO archive, and reduce Integral Field Spectrograph (IFS) data using the adapted CHARIS pipeline. Beyond basic reduction, the workflow handles astrometric calibration and post-processing via TRAP for companion detection, while also facilitating the discovery and retrieval of IRDIS dual-band and polarimetric imaging data for analysis with community tools.
 
-**What `spherical` can do**
-- Search and filter the complete SPHERE observation history by target, stellar properties, observing mode, date, and observing conditions  
-- Download selected raw datasets (and associated calibrations) directly from the ESO archive  
-- Automatically reduce *IFS* data from raw frames to calibrated spectral cubes (using the adapted CHARIS pipeline)  
-- Perform astrometric/photometric calibration and *post-processing with TRAP* for companion detection and spectral extraction  
-- Provide *IRDIS* (DBI and DPI) dataset discovery and download for subsequent analysis with community tools
+As of May 2025, the database includes approximately 6000 IRDIS dual-band imaging (DBI) observations, ~1000 IRDIS dual-beam polarimetric imaging [DPI; @deBoer:2020; @vanHolstein:2020] observations, and ~4500 IFS observations. SPHERE is undergoing the *SPHERE+* upgrade [@Boccaletti:2020; @Boccaletti:2022], including a second-stage adaptive optics system [SAXO+; @Stadler:2022], ensuring long-term scientific relevance and providing a pathfinder for the ELT’s Planetary Camera and Spectrograph [PCS; @Kasper:2021].
 
[…]
