+++ joss/paper.bib
@@ -571,3 +571,61 @@ archivePrefix = {arXiv},
        adsurl = {https://ui.adsabs.harvard.edu/abs/2021Msngr.182...38K},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
+
+@ARTICLE{Franson:2023,
+       author = {{Franson}, Kyle and {Bowler}, Brendan P. and {Bonavita}, Mariangela and {Brandt}, Timothy D. and {Chen}, Minghan and {Samland}, Matthias and {Zhang}, Zhoujian and {Lueber}, Anna and {Heng}, Kevin and {Kitzmann}, Daniel and {Wolf}, Trevor and {Jones}, Brandon A. and {Tran}, Quang H. and {Bardalez Gagliuffi}, Daniella C. and {Biller}, Beth and {Chilcote}, Jeffrey and {Crepp}, Justin R. and {Dupuy}, Trent J. and {Faherty}, Jacqueline and {Fontanive}, Cl{\'e}mence and {Groff}, Tyler D. and {Gratton}, Raffaele and {Guyon}, Olivier and {Jensen-Clem}, Rebecca and {Jovanovic}, Nemanja and {Kasdin}, N. Jeremy and {Lozi}, Julien and {Magnier}, Eugene A. and {Mu{\v{z}}i{\'c}}, Koraljka and {Sanghi}, Aniket and {Theissen}, Christopher A.},
+        title = "{Astrometric Accelerations as Dynamical Beacons: Discovery and Characterization of HIP 21152 B, the First T-dwarf Companion in the Hyades}",
+      journal = {\aj},
+     keywords = {Brown dwarfs, Direct imaging, T dwarfs, Astrometry, Orbit determination, Atmospheric clouds, 185, 387, 1679, 80, 1175, 2180, Astrophysics - Solar and Stellar Astrophysics, Astrophysics - Earth and Planetary Astrophysics},
+         year = 2023,
+        month = feb,
+       volume = {165},
+       number = {2},
+          eid = {39},
+        pages = {39},
+          doi = {10.3847/1538-3881/aca408},
+archivePrefix = {arXiv},
+       eprint = {2211.09840},
+ primaryClass = {astro-ph.SR},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2023AJ....165...39F},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
+@ARTICLE{Hammond:2025,
+       author = {{Hammond}, Iain and {Christiaens}, Valentin and {Price}, Daniel J. and {Blakely}, Dori and {Trevascus}, David and {Bonse}, Markus J. and {Cantalloube}, Faustine and {Marleau}, Gabriel-Dominique and {Pinte}, Christophe and {Juillard}, Sandrine and {Samland}, Matthias and {Thompson}, William and {Wallace}, Alex},
+        title = "{Keplerian motion of a compact source orbiting the inner disc of PDS 70: a third protoplanet in resonance with b and c?}",
+      journal = {\mnras},
+     keywords = {planet─disc interactions, protoplanetary discs, stars: individual: PDS 70, infrared: planetary systems, Earth and Planetary Astrophysics},
+         year = 2025,
+        month = may,
+       volume = {539},
+       number = {2},
+        pages = {1613-1627},
+          doi = {10.1093/mnras/staf586},
+archivePrefix = {arXiv},
+       eprint = {2504.11127},
[…]
