### 2026-02-21 #8690 @warrickball
Hi @m-samland, and thanks for adding the paper. On the basis of the design explained in the paper, the demonstrated research impact and the (limited) collaborative use of issues and PRs, this appears to satisfy [our new scope and significance requirements](https://joss.readthedocs.io/en/latest/submitting.html#scope-and-significance). The bad news is that it remains in the submission queue (though it's placed hasn't changed) and all our editors are still busy.

I did notice that the software apparently depends on the [curated metadata tables on Zenodo](https://zenodo.org/records/15459438). Is it possible to regenerate those tables if necessary and, if so, is the process documented somewhere? I only ask because it otherwise creates the risk that the software remains usable but the archive is not maintained.
