### 2026-02-23 86e1614
Fix citation brackets; add that spherical can produce and update the observation table
