+++ joss/paper.md
@@ -23,11 +23,11 @@ bibliography: paper.bib
 
 # Summary
 
-The Spectro-Polarimetric High-contrast Exoplanet REsearch instrument [SPHERE; @Beuzit:2019] at the Very Large Telescope (VLT) is a leading facility for coronagraphic imaging of exoplanets and circumstellar disks in the optical and near-infrared. Over the last decade, SPHERE has contributed to hundreds of publications and major legacy surveys [e.g., SHINE; @Chauvin:2017; @Chomez:2025].
+The Spectro-Polarimetric High-contrast Exoplanet REsearch instrument (SPHERE; @Beuzit:2019) at the Very Large Telescope (VLT) is a leading facility for coronagraphic imaging of exoplanets and circumstellar disks in the optical and near-infrared. Over the last decade, SPHERE has contributed to hundreds of publications and major legacy surveys (e.g., SHINE; @Chauvin:2017; @Chomez:2025).
 
 `spherical` streamlines the end-to-end analysis of SPHERE data by integrating a curated, searchable database with an automated Python-based pipeline. The software enables users to filter the complete observation history by target properties or observing conditions, automatically download raw datasets from the ESO archive, and reduce Integral Field Spectrograph (IFS) data using the adapted CHARIS pipeline. Beyond basic reduction, the workflow handles astrometric calibration and post-processing via TRAP for companion detection, while also facilitating the discovery and retrieval of IRDIS dual-band and polarimetric imaging data for analysis with community tools.
 
-As of May 2025, the database includes approximately 6000 IRDIS dual-band imaging (DBI) observations, ~1000 IRDIS dual-beam polarimetric imaging [DPI; @deBoer:2020; @vanHolstein:2020] observations, and ~4500 IFS observations. SPHERE is undergoing the *SPHERE+* upgrade [@Boccaletti:2020; @Boccaletti:2022], including a second-stage adaptive optics system [SAXO+; @Stadler:2022], ensuring long-term scientific relevance and providing a pathfinder for the ELT’s Planetary Camera and Spectrograph [PCS; @Kasper:2021].
+As of May 2025, the database includes approximately 6000 IRDIS dual-band imaging (DBI) observations, ~1000 IRDIS dual-beam polarimetric imaging (DPI; @deBoer:2020; @vanHolstein:2020) observations, and ~4500 IFS observations. SPHERE is undergoing the *SPHERE+* upgrade [@Boccaletti:2020; @Boccaletti:2022], including a second-stage adaptive optics system (SAXO+; @Stadler:2022), ensuring long-term scientific relevance and providing a pathfinder for the ELT’s Planetary Camera and Spectrograph (PCS; @Kasper:2021).
 
 # Statement of Need
 
@@ -43,11 +43,11 @@ Several existing tools address isolated components of the SPHERE data workflow. 
 
 The architecture of `spherical` is designed as a high-level abstraction layer over the ESO raw archive, providing a systematic interface that currently does not exist for high-contrast imaging data. A central design choice was the decoupling of metadata curation from data reduction. This is achieved through a multi-stage process that ingests science headers from the ESO archive and cross-matches them with the Gaia catalog based on telescope pointings. To handle the inherent noise in archival metadata, the system employs a proximity, brightness, and object type-based resolution logic to resolve target ambiguities, accounting for epoch of observation of proper motion of stellar targets, ensuring the resulting database is both clean and astronomically accurate.
 
[…]
