### 2026-02-23 #8690 @m-samland
Thank you @warrickball! The software does not depend on the table uploaded on Zenodo. It is fully capable of producing the table from scratch, it just takes quite some time given the amount of data in the archive. It is advice to download the table as a starting point, and if needed it can be updated with a diff of only the newer data than the last release of the Zenodo table (which is much quicker). Thanks for pointing this out, and I have updated the paper to reflect this.

Is there a rough estimate for the waitlist? I can ask around for potential reviewers, if not needing to search for a referee would speed things up.
