### 2026-02-23 #8690 @warrickball
Thanks for the update.

> Is there a rough estimate for the waitlist? I can ask around for potential reviewers, if not needing to search for a referee would speed things up.

My estimates have mostly turned out to be optimistic but I'd say there's at least a few months to wait still. I appreciate the offer of reviewers but really the limiting factor is editors, who are all handling as much as they can. Sorry!
