### 2026-02-24 05fb9be
Removed try-except and return None instead
