+++ src/spherical/database/sphere_database.py
@@ -415,12 +415,13 @@ class SphereDatabase(object):
         Returns
         -------
         astropy.table.Table
-            Table of matching observations.
+            Table of matching observations. Returns None if there are
+            no observations of the ``target_name``.
 
         Raises
         ------
         ValueError
-            If the target cannot be found locally or via SIMBAD.
+            If the target cannot be resolved by SIMBAD.
         """
         def _try_local_lookup(name: str) -> Optional[Table]:
             """Try to find target in local database by normalized name."""
@@ -455,28 +456,20 @@ class SphereDatabase(object):
         # 3. SIMBAD query fallback
         warnings.warn(f"Target '{target_name}' not found in local ID columns. Trying SIMBAD...")
         
-        try:
-            result = Simbad.query_object(target_name)
-            if result is not None and len(result) > 0:
-                simbad_main_id = _normalize_name(result['main_id'][0])
-                main_id_col = self.table_of_observations['MAIN_ID'].astype(str)
-                main_id_norm = np.char.lower(np.char.strip(main_id_col))
-                main_id_norm = np.char.replace(main_id_norm, " ", "")
-                main_id_norm = np.char.replace(main_id_norm, "_", "")
-                mask = main_id_norm == simbad_main_id
-                if np.any(mask):
-                    return self.table_of_observations[mask]
-                else:
-                    warnings.warn(f"SIMBAD resolved '{target_name}' to '{result['main_id'][0]}', but this MAIN_ID is not in the observation table.")
+        result = Simbad.query_object(target_name)
+        if result is not None and len(result) > 0:
+            simbad_main_id = _normalize_name(result['main_id'][0])
+            main_id_col = self.table_of_observations['MAIN_ID'].astype(str)
+            main_id_norm = np.char.lower(np.char.strip(main_id_col))
+            main_id_norm = np.char.replace(main_id_norm, " ", "")
+            main_id_norm = np.char.replace(main_id_norm, "_", "")
+            mask = main_id_norm == simbad_main_id
+            if np.any(mask):
+                return self.table_of_observations[mask]
             else:
-                warnings.warn(f"SIMBAD could not resolve '{target_name}'.")
-        except Exception as e:
-            warnings.warn(f"SIMBAD query failed for '{target_name}': {e}")
-
-        # 4. Failure
-        raise ValueError(
-            f"Target '{target_name}' could not be found in local ID columns or via SIMBAD."
-        )
+                warnings.warn(f"SIMBAD resolved '{target_name}' to '{result['main_id'][0]}', but this MAIN_ID is not in the observation table.")
+        else:
+            raise ValueError(f"SIMBAD could not resolve '{target_name}'.")
 
     def observations_from_name_SIMBAD(
         self,
@@ -502,7 +495,8 @@ class SphereDatabase(object):
         Returns
         -------
         astropy.table.Table
-            Table of matching observations (possibly summarized).
+            Table of matching observations (possibly summarized). Returns
+            None if there are no observations of the ``target_name``.
         """
         if usable_only:
             table_of_observations = self.table_of_observations[
@@ -514,21 +508,22 @@ class SphereDatabase(object):
         # Use robust ID-based matching with SIMBAD fallback
         matching_observations = self._find_observations_by_id_or_simbad(target_name)
 
-        # Only keep those in the current filtered table_of_observations
-        # (in case usable_only is True)
-        mask = np.isin(matching_observations["MAIN_ID"], table_of_observations["MAIN_ID"])
-        matching_observations = matching_observations[mask]
+        if matching_observations is not None:
+            # Only keep those in the current filtered table_of_observations
+            # (in case usable_only is True)
[…]
