### 2026-02-25 5eab3ab
Change ValueError to warning if not resolved by SIMBAD
