+++ CHANGELOG.md
@@ -13,6 +13,18 @@ This project follows [Semantic Versioning](https://semver.org/) and the [Keep a 
 ### Changed
 
 ### Fixed
+- **Fixed IFS pipeline step ordering causing crash at cube header update** – Reordered pipeline steps so that `run_frame_info_computation` executes before `run_cube_header_update`, resolving `FileNotFoundError` for `frames_info_*.csv` files. Also added defensive handling for missing frame-info CSVs, consistent use of `converted_dir` in bundling, and glob escaping for target names with special characters ([@m-samland](https://github.com/m-samland)) ([#97](https://github.com/m-samland/spherical/issues/97)).
+
+---
+
+## [2.1.3] - 2026-02-19
+
+### ✨ Added
+- **Intermediate file cleanup** – Added functionality to clean up intermediate pipeline files after successful reduction. The clean-up can be triggered using the methods outlined in the example reduction script ([@m-samland](https://github.com/m-samland)).
+
+### Fixed
+- **Fixed ESO data download crash when keyring is unavailable** – Fixed crash when downloading proprietary ESO data on machines without an installed keyring. The `store_password` default is now `False`, and keyring interactions are wrapped in try/except for robustness ([@m-samland](https://github.com/m-samland)) ([#94](https://github.com/m-samland/spherical/issues/94)).
+- **Fixed SIMBAD target name resolving** – Fixed SIMBAD query failures caused by using the outdated column name `MAIN_ID` instead of `main_id`, which could cause reductions to be skipped for certain targets ([@m-samland](https://github.com/m-samland)) ([#95](https://github.com/m-samland/spherical/issues/95)).
 
 ---
 
@@ -147,7 +159,8 @@ This project follows [Semantic Versioning](https://semver.org/) and the [Keep a 
 ### Fixed
 - No known issues.
 
-[Unreleased]: https://github.com/m-samland/spherical/compare/v2.1.2...HEAD  
+[Unreleased]: https://github.com/m-samland/spherical/compare/v2.1.3...HEAD  
+[2.1.3]: https://github.com/m-samland/spherical/compare/v2.1.2...v2.1.3  
 [2.1.2]: https://github.com/m-samland/spherical/compare/v2.1.1...v2.1.2  
 [2.1.1]: https://github.com/m-samland/spherical/compare/v2.1.0...v2.1.1  
 [2.1.0]: https://github.com/m-samland/spherical/compare/v2.0.0...v2.1.0  
