### 2026-03-05 ac8c330
Update .gitignore for pixi
