### 2026-03-06 b30ed48
Update readme and pixi.toml
