+++ README.md
@@ -7,6 +7,7 @@
 [![Python Version](https://img.shields.io/badge/Python-3.11%20%7C%203.12%20%7C%203.13-brightgreen.svg)](https://github.com/m-samland/spherical)
 [![License](https://img.shields.io/badge/License-BSD--3-blue.svg)](https://opensource.org/licenses/BSD-3-Clause)
 [![CI](https://github.com/m-samland/spherical/actions/workflows/ci.yml/badge.svg)](https://github.com/m-samland/spherical/actions/workflows/ci.yml)
+[![DOI](https://img.shields.io/badge/DOI-10.48550%2FarXiv.2509.08044-blue.svg)](https://doi.org/10.48550/arXiv.2509.08044)
 
 **spherical** provides tools for exploring and analyzing data from the ESO VLT/SPHERE instrument, with a focus on high-contrast imaging observations using **IRDIS** and **IFS**.
 
@@ -14,7 +15,7 @@
 
 ## Overview
 
-**spherical** ships with a **curated observation database**, constructed by parsing headers from all available SPHERE data in the ESO archive and cross-matching with external catalogs. This database systematically identifies SPHERE observation sequences and summarizes observing modes, target properties, integration times, and observing conditions.
+**spherical** provides a **curated observation database** for VLT/SPHERE, constructed by parsing headers from all available SPHERE data in the ESO archive and cross-matching with external catalogs. This database systematically identifies SPHERE observation sequences and summarizes observing modes, target properties, integration times, and observing conditions. Pre-built database tables are available for download from [Zenodo](https://doi.org/10.5281/zenodo.15147730) and can be updated with newer data using **spherical** itself (see [Database Tables](#database-tables)).
 
 From exploration to execution, you can filter sequences of interest and feed them into an **end-to-end IFS pipeline** that automates discovery, download, calibration, post-processing, and spectral characterization. For **IRDIS**, the package supports discovery and download of dual-band imaging (DBI) and polarimetric imaging observations.
 
@@ -36,9 +37,6 @@ You can use **spherical** in two complementary ways:
 
 - **As a script-driven workflow**  
   Start from the provided example script (`ifs_reduction_template.py`) to configure and run a complete IFS reduction with sensible defaults and many tunable parameters.
-
-> ❗ **Why not a one-click CLI pipeline?**  
-> The IFS pipeline exposes many parameters (calibration strategy, post-processing choices, masking, spectral extraction options, etc.). A single command-line entry point would hide crucial configuration and often lead to suboptimal results. Instead, we provide a ready-to-run **Python template script** that you adapt to your dataset.
 
 ---
 
@@ -65,69 +63,100 @@ You can use **spherical** in two complementary ways:
 
 ## Installation
 
-**Python ≥ 3.11** is required. We recommend a dedicated environment (e.g., via Miniconda or Micromamba):
+**Python ≥ 3.11** is required.
+
+### Option A: Using pip
+
+We recommend a dedicated environment (e.g., via Miniconda or Micromamba):
 
 ```bash
-mamba create -n spherical_env python=3.12
+mamba create -n spherical_env python=3.13
 mamba activate spherical_env
 ```
 
-### Base installation (database exploration only)
+**Base install (database exploration only):**
 
 ```bash
 pip install git+https://github.com/m-samland/spherical.git
 ```
 
-> **Database tables required**  
-> The SPHERE database tables must be downloaded separately from **Zenodo**: [10.5281/zenodo.15147730](https://doi.org/10.5281/zenodo.15147730).  
[…]
