+++ pixi.toml
@@ -2,7 +2,7 @@
 name = "spherical"
 description = "VLT/SPHERE Instrument Calibration Pipeline and Database"
 channels = ["conda-forge"]
-platforms = ["linux-64", "osx-arm64"]
+platforms = ["linux-64", "linux-aarch64", "osx-arm64", "osx-64"]
 
 # ──────────────────────────────────────────────
 # Core dependencies (mirrors [project.dependencies] in pyproject.toml)
