+++ src/spherical/database/database_utils.py
@@ -17,7 +17,7 @@ from astropy.io.ascii.core import InconsistentTableError
 from astropy.table import Column, Table, TableMergeError, hstack, join, unique, vstack
 from astroquery.gaia import Gaia
 
-from spherical.utils.progress import tqdm
+from tqdm.auto import tqdm
 
 
 def convert_table_to_little_endian(table):
