### 2026-03-12 d20b987
Closes #105
