+++ src/spherical/database/file_table.py
@@ -208,7 +208,8 @@ def make_file_table(output_dir,
                     cache=False,
                     existing_table_path=None,
                     batch_size=100,
-                    date_batch_months=1):
+                    date_batch_months=1,
+                    resume=True):
     """
     Create a file table of SPHERE science and calibration observations 
     by retrieving and parsing ESO archive headers.
@@ -252,6 +253,12 @@ def make_file_table(output_dir,
     date_batch_months : int or None, optional
         If provided, the full date range will be split into monthly batches of this size 
         to avoid timeouts. Set to None to query the entire range at once.
+
+    resume : bool, optional
+        If True (default), automatically resumes an interrupted run by detecting a
+        partial file (``*_partial.csv``) and/or the existing output file. Already
+        downloaded DP.IDs from both files are skipped. New data is written to the
+        partial file and merged into the final output only upon successful completion.
 
     Returns
     -------
@@ -313,36 +320,67 @@ def make_file_table(output_dir,
     - Date fields are enhanced with a `NIGHT_START` column for grouping by observing night.
     - Observations from both calibration and science categories are included.
     - Use `existing_table_path` to incrementally update a file table over time.
+    - When ``resume=True``, interrupted runs are automatically resumed. New data
+      is written to a ``*_partial.csv`` file; the final output is only updated
+      upon successful completion.
     """
 
     logger.info(f"Starting file table generation with date range: {start_date} to {end_date}")
 
     instrument = instrument.lower()
 
-    previous_file_table = None
-    existing_entries = set()
-    if existing_table_path is not None:
-        logger.info(f"Loading existing file table from: {existing_table_path}")
-        try:
-            previous_file_table = pd.read_csv(existing_table_path)
-            existing_entries = set(previous_file_table["DP.ID"].values)
-            logger.info(f"Found {len(existing_entries)} existing entries")
-        except Exception as e:
-            logger.error(f"Error loading existing file table: {e}")
-            raise
-
     logger.info("Initializing ESO query interface")
     eso = Eso()
     eso.ROW_LIMIT = 10000000
 
-    # Prepare output path
+    # Prepare output and partial file paths
     output_path = os.path.join(output_dir, f"table_of_files_{instrument}{output_suffix}.csv".lower())
+    partial_path = output_path.replace(".csv", "_partial.csv")
+
+    # Collect known DP.IDs from all available sources to skip already-downloaded files
+    existing_entries = set()
+
+    # Load completed output file if it exists
     if os.path.exists(output_path):
-        timestamp = time.strftime("%Y%m%d_%H%M%S")
-        backup_path = f"{output_path}_{timestamp}.bak"
-        os.rename(output_path, backup_path)
-        logger.info(f"Existing file backed up to {backup_path}")
-    first_batch_file = True
+        try:
+            previous_output = pd.read_csv(output_path)
+            existing_entries.update(previous_output["DP.ID"].values)
+            logger.info(f"Found {len(previous_output)} entries in completed output file")
+        except Exception as e:
+            logger.error(f"Error reading existing output file: {e}")
+            raise
+
+    # Load partial file from a previously interrupted run (resume mode)
+    if resume and os.path.exists(partial_path):
+        try:
+            partial_df = pd.read_csv(partial_path)
+        except Exception:
+            logger.warning("Partial file could not be read normally, attempting recovery with on_bad_lines='skip'")
+            partial_df = pd.read_csv(partial_path, on_bad_lines='skip')
+        existing_entries.update(partial_df["DP.ID"].values)
+        logger.info(f"Resuming: found {len(partial_df)} entries in partial file")
+    elif not resume and os.path.exists(partial_path):
[…]

[…]
