### 2026-03-12 aaecb75
Closes #107
