+++ tests/test_mocadb_matching.py
@@ -1,0 +1,186 @@
+"""Tests for MOCAdb cross-matching module (spherical.database.mocadb_matching).
+
+These tests verify the module against the live MOCAdb MySQL endpoint using
+a small inline target table with known Gaia DR3 identifiers.
+"""
+
+import numpy as np
+import pytest
+from astropy.table import Table
+
+from spherical.database.mocadb_matching import (
+    _clean_gaia_id,
+    query_mocadb_for_targets,
+)
+
+# ---------------------------------------------------------------------------
+# Inline test fixture: 4 targets from the SPHERE archive
+# ---------------------------------------------------------------------------
+# 3 Cen A is a known Sco-Cen member, 51 Eri is a Beta Pic member;
+# 4 Sgr and 5 Vul are field stars.
+
+@pytest.fixture()
+def sample_target_table():
+    """Minimal target table with 4 stars, all with valid Gaia DR3 IDs.
+
+    Only ``MAIN_ID`` and ``ID_GAIA_DR3`` are required by
+    `query_mocadb_for_targets`; the matching is purely based on
+    Gaia DR3 source IDs (no coordinate cross-match).
+    """
+    return Table(
+        {
+            "MAIN_ID": ["*   3 Cen A", "*   4 Sgr", "*   5 Vul", "51 Eri"],
+            "ID_GAIA_DR3": [
+                "Gaia DR3 6170485544575679104",
+                "Gaia DR3 4069112871640783360",
+                "Gaia DR3 4515892996332248960",
+                "Gaia DR3 3205095125321700480",
+            ],
+        }
+    )
+
+
+# ---------------------------------------------------------------------------
+# Unit tests for _clean_gaia_id
+# ---------------------------------------------------------------------------
+
+class TestCleanGaiaId:
+    def test_string_with_prefix(self):
+        assert _clean_gaia_id("Gaia DR3 6126376695498244736") == 6126376695498244736
+
+    def test_plain_integer_string(self):
+        assert _clean_gaia_id("6126376695498244736") == 6126376695498244736
+
+    def test_integer_input(self):
+        assert _clean_gaia_id(6126376695498244736) == 6126376695498244736
+
+    def test_numpy_int(self):
+        assert _clean_gaia_id(np.int64(6126376695498244736)) == 6126376695498244736
+
+    def test_none(self):
+        assert _clean_gaia_id(None) is None
+
+    def test_nan(self):
+        assert _clean_gaia_id(float("nan")) is None
+
+    def test_empty_string(self):
+        assert _clean_gaia_id("") is None
+
+    def test_dash(self):
+        assert _clean_gaia_id("--") is None
+
+    def test_short_number_rejected(self):
+        # source_ids are long; very short numbers should be rejected
+        assert _clean_gaia_id("123") is None
+
+
+# ---------------------------------------------------------------------------
+# Integration tests against the live MOCAdb endpoint
+# ---------------------------------------------------------------------------
+
+@pytest.mark.remote_data
+class TestQueryMOCAdb:
+    """Tests that connect to the public MOCAdb MySQL server.
+
+    Marked with ``@pytest.mark.remote_data`` so they can be skipped in
+    offline / CI environments via ``pytest -m "not remote_data"``.
+    """
+
+    def test_basic_enrichment(self, sample_target_table):
+        """All MOCA_ columns are appended and the table length is unchanged."""
[…]
