+++ tests/test_resume.py
@@ -1,0 +1,190 @@
+"""Tests for the resume functionality in file table generation.
+
+These tests simulate interrupted downloads by running make_file_table for
+a short time window, then extending the date range on a second call. The
+resume logic should detect DP.IDs in the existing output and only download
+new ones.
+
+NOTE: These tests query the live ESO archive and require network access.
+They use very short 1-day date windows to minimize download time.
+"""
+
+import pandas as pd
+import pytest
+
+from spherical.database.file_table import make_file_table
+
+INSTRUMENT = "ifs"
+
+
+@pytest.fixture(scope="module")
+def resume_table_path(tmp_path_factory):
+    return tmp_path_factory.mktemp("resume_test")
+
+
+class TestResumeFiletable:
+    """Test suite for the resume/incremental file table building."""
+
+    def test_fresh_run_creates_output_no_partial(self, resume_table_path):
+        """A complete fresh run should produce an output file and no partial file."""
+        suffix = "_fresh"
+        table = make_file_table(
+            output_dir=resume_table_path,
+            instrument=INSTRUMENT,
+            start_date="2016-09-15",
+            end_date="2016-09-16",
+            output_suffix=suffix,
+            cache=False,
+            batch_size=100,
+            date_batch_months=1,
+        )
+
+        output_path = resume_table_path / f"table_of_files_{INSTRUMENT}{suffix}.csv"
+        partial_path = resume_table_path / f"table_of_files_{INSTRUMENT}{suffix}_partial.csv"
+
+        assert output_path.exists(), "Output file should exist after fresh run"
+        assert not partial_path.exists(), "Partial file should be cleaned up after successful run"
+        assert len(table) > 0, "Table should have entries for this date range"
+
+    def test_incremental_extends_date_range(self, resume_table_path):
+        """Running again with a wider date range should add new entries via resume."""
+        suffix = "_incr"
+
+        # First run: one day
+        table1 = make_file_table(
+            output_dir=resume_table_path,
+            instrument=INSTRUMENT,
+            start_date="2016-09-15",
+            end_date="2016-09-16",
+            output_suffix=suffix,
+            cache=False,
+            batch_size=100,
+            date_batch_months=1,
+        )
+        n_first = len(table1)
+        assert n_first > 0
+
+        # Second run: extended by one day (resume=True by default)
+        table2 = make_file_table(
+            output_dir=resume_table_path,
+            instrument=INSTRUMENT,
+            start_date="2016-09-15",
+            end_date="2016-09-17",
+            output_suffix=suffix,
+            cache=False,
+            batch_size=100,
+            date_batch_months=1,
+        )
+
+        assert len(table2) >= n_first, "Extended run should have at least as many entries"
+        # Check no duplicates
+        dp_ids = list(table2["DP.ID"])
+        assert len(dp_ids) == len(set(dp_ids)), "No duplicate DP.IDs in merged table"
+
+    def test_resume_with_simulated_partial_file(self, resume_table_path):
+        """Simulates an interrupted run by manually creating a partial file,
+        then verifies that resume picks it up and merges correctly."""
+        suffix = "_simpartial"
+
+        # First run to get real data
+        table1 = make_file_table(
+            output_dir=resume_table_path,
+            instrument=INSTRUMENT,
+            start_date="2016-09-15",
+            end_date="2016-09-16",
+            output_suffix=suffix,
+            cache=False,
+            batch_size=100,
+            date_batch_months=1,
+        )
+        n_first = len(table1)
+        assert n_first > 0
+
+        output_path = resume_table_path / f"table_of_files_{INSTRUMENT}{suffix}.csv"
[…]
