+++ src/spherical/database/mocadb_matching.py
@@ -1,0 +1,483 @@
+#!/usr/bin/env python
+# -*- coding: utf-8 -*-
+"""MOCAdb cross-matching for young-association membership and stellar ages.
+
+This module queries the Montreal Open Clusters and Associations database
+(MOCAdb; Gagné et al. 2026) to enrich a spherical target table with young
+stellar association membership, ages, and ancillary youth indicators.
+
+The cross-match is performed via Gaia DR3 ``source_id`` using a direct
+MySQL connection to the public MOCAdb endpoint.  Target IDs are uploaded
+as a temporary table and joined server-side for efficiency.
+
+The query is split into three steps to avoid performance issues with
+the ``summary_all_members`` VIEW:
+
+1. **Resolve** Gaia DR3 source_id → ``moca_oid`` via ``cat_gaiadr3``
+   (indexed temp-table JOIN).
+2. **Tier-1** query: ``summary_all_objects`` + association metadata +
+   adopted age, queried via ``WHERE moca_oid IN (...)``.
+3. **Tier-2** query: ``summary_all_members`` (activity, kinematics),
+   also queried via ``WHERE moca_oid IN (...)``.
+
+References
+----------
+- Gagné et al. (2026), ApJS, arXiv:2602.15695
+- MOCAdb website: https://mocadb.ca/
+- MOCAdb schema: https://mocadb.ca/en/schema
+"""
+
+from __future__ import annotations
+
+import logging
+import re
+import warnings
+
+import numpy as np
+import pandas as pd
+from astropy.table import Column, Table
+
+logger = logging.getLogger(__name__)
+
+__all__ = ["query_mocadb_for_targets"]
+
+# ---------------------------------------------------------------------------
+# MOCAdb public connection parameters
+# ---------------------------------------------------------------------------
+MOCADB_HOST = "104.248.106.21"
+MOCADB_PORT = 3306
+MOCADB_USER = "public"
+MOCADB_PASSWORD = "z@nUg_2h7_%?31y88"
+MOCADB_DATABASE = "mocadb"
+
+
+# ---------------------------------------------------------------------------
+# Helper: parse Gaia DR3 IDs
+# ---------------------------------------------------------------------------
+
+def _clean_gaia_id(value) -> int | None:
+    """Return an integer Gaia DR3 source_id, or ``None`` if not parseable."""
+    if value is None or (isinstance(value, float) and np.isnan(value)):
+        return None
+    if isinstance(value, (int, np.integer)):
+        return int(value)
+    s = str(value).strip()
+    if s.lower() in ("", "nan", "none", "--"):
+        return None
+    m = re.search(r"(?:Gaia\s*DR3\s*)?(\d{5,})$", s, re.I)
+    return int(m.group(1)) if m else None
+
+
+def _execute_query(conn, sql: str, debug: bool = False) -> pd.DataFrame:
+    """Execute a SQL query via pymysql cursor and return a DataFrame.
+
+    Uses the cursor interface directly (rather than ``pd.read_sql``) to
+    avoid spurious SQLAlchemy compatibility warnings.
+    """
+    if debug:
+        preview = sql.strip()[:500]
+        logger.debug("SQL query:\n%s", preview)
+    cursor = conn.cursor()
+    cursor.execute(sql)
+    columns = [desc[0] for desc in cursor.description]
+    rows = cursor.fetchall()
+    cursor.close()
+    return pd.DataFrame(rows, columns=columns)
+
+
+# ---------------------------------------------------------------------------
+# SQL query templates
[…]
