+++ src/spherical/database/gaia_astrophysical_params.py
@@ -1,0 +1,264 @@
+#!/usr/bin/env python
+# -*- coding: utf-8 -*-
+"""Gaia DR3 astrophysical parameter enrichment for spherical target tables.
+
+This module queries the ESA Gaia DR3 archive to enrich a spherical target
+table with GSP-Phot astrophysical parameters: effective temperature,
+surface gravity, metallicity, and extinction.
+
+The cross-match is performed via Gaia DR3 ``source_id`` using a TAP
+upload-join against ``gaiadr3.astrophysical_parameters``.  This is a single
+ADQL query regardless of list size.
+
+References
+----------
+- Andrae et al. (2023), A&A, 674, A27 (GSP-Phot)
+- Gaia DR3 documentation: https://gea.esac.esa.int/archive/documentation/GDR3/
+"""
+
+from __future__ import annotations
+
+import logging
+import re
+import warnings
+
+import numpy as np
+from astropy.table import Column, Table
+
+logger = logging.getLogger(__name__)
+
+__all__ = ["query_gaia_astrophysical_params"]
+
+# ---------------------------------------------------------------------------
+# Gaia TAP endpoint
+# ---------------------------------------------------------------------------
+GAIA_TAP_URL = "https://gea.esac.esa.int/tap-server/tap"
+
+# ---------------------------------------------------------------------------
+# ADQL query template
+# ---------------------------------------------------------------------------
+_ADQL_QUERY = """
+SELECT u.source_id,
+       ap.teff_gspphot, ap.teff_gspphot_lower, ap.teff_gspphot_upper,
+       ap.logg_gspphot, ap.logg_gspphot_lower, ap.logg_gspphot_upper,
+       ap.mh_gspphot,   ap.mh_gspphot_lower,   ap.mh_gspphot_upper,
+       ap.ag_gspphot,   ap.ag_gspphot_lower,    ap.ag_gspphot_upper
+FROM TAP_UPLOAD.targets AS u
+JOIN gaiadr3.astrophysical_parameters AS ap
+    ON ap.source_id = u.source_id
+"""
+
+# ---------------------------------------------------------------------------
+# Output column definitions (all float)
+# ---------------------------------------------------------------------------
+_GAIA_COLUMNS = {
+    "GAIA_TEFF":       "teff_gspphot",
+    "GAIA_TEFF_LOWER": "teff_gspphot_lower",
+    "GAIA_TEFF_UPPER": "teff_gspphot_upper",
+    "GAIA_LOGG":       "logg_gspphot",
+    "GAIA_LOGG_LOWER": "logg_gspphot_lower",
+    "GAIA_LOGG_UPPER": "logg_gspphot_upper",
+    "GAIA_MH":         "mh_gspphot",
+    "GAIA_MH_LOWER":   "mh_gspphot_lower",
+    "GAIA_MH_UPPER":   "mh_gspphot_upper",
+    "GAIA_AG":         "ag_gspphot",
+    "GAIA_AG_LOWER":   "ag_gspphot_lower",
+    "GAIA_AG_UPPER":   "ag_gspphot_upper",
+}
+
+# ---------------------------------------------------------------------------
+# Helper: parse Gaia DR3 IDs
+# ---------------------------------------------------------------------------
+
+def _clean_gaia_id(value) -> int | None:
+    """Return an integer Gaia DR3 source_id, or ``None`` if not parseable."""
+    if value is None or (isinstance(value, float) and np.isnan(value)):
+        return None
+    if isinstance(value, (int, np.integer)):
+        return int(value)
+    s = str(value).strip()
+    if s.lower() in ("", "nan", "none", "--"):
+        return None
+    m = re.search(r"(?:Gaia\s*DR3\s*)?(\d{5,})$", s, re.I)
+    return int(m.group(1)) if m else None
+
[…]
