+++ tests/test_gaia_astrophysical_params.py
@@ -1,0 +1,152 @@
+"""Tests for Gaia DR3 astrophysical parameter enrichment.
+
+These tests verify the module against the live ESA Gaia TAP archive using
+a small inline target table with known Gaia DR3 identifiers (same fixture
+as test_mocadb_matching.py).
+"""
+
+import numpy as np
+import pytest
+from astropy.table import Table
+
+from spherical.database.gaia_astrophysical_params import (
+    _clean_gaia_id,
+    query_gaia_astrophysical_params,
+)
+
+# ---------------------------------------------------------------------------
+# Inline test fixture: 4 targets from the SPHERE archive
+# ---------------------------------------------------------------------------
+# 3 Cen A (B5III, ~17000 K), 4 Sgr (A0, ~10000 K), 5 Vul (A0V, ~10000 K),
+# 51 Eri (F0IV, ~7000 K)
+
+@pytest.fixture()
+def sample_target_table():
+    """Minimal target table with 4 stars, all with valid Gaia DR3 IDs."""
+    return Table(
+        {
+            "MAIN_ID": ["*   3 Cen A", "*   4 Sgr", "*   5 Vul", "51 Eri"],
+            "ID_GAIA_DR3": [
+                "Gaia DR3 6170485544575679104",
+                "Gaia DR3 4069112871640783360",
+                "Gaia DR3 4515892996332248960",
+                "Gaia DR3 3205095125321700480",
+            ],
+        }
+    )
+
+
+# ---------------------------------------------------------------------------
+# Unit tests for _clean_gaia_id
+# ---------------------------------------------------------------------------
+
+class TestCleanGaiaId:
+    def test_string_with_prefix(self):
+        assert _clean_gaia_id("Gaia DR3 6126376695498244736") == 6126376695498244736
+
+    def test_plain_integer_string(self):
+        assert _clean_gaia_id("6126376695498244736") == 6126376695498244736
+
+    def test_integer_input(self):
+        assert _clean_gaia_id(6126376695498244736) == 6126376695498244736
+
+    def test_numpy_int(self):
+        assert _clean_gaia_id(np.int64(6126376695498244736)) == 6126376695498244736
+
+    def test_none(self):
+        assert _clean_gaia_id(None) is None
+
+    def test_nan(self):
+        assert _clean_gaia_id(float("nan")) is None
+
+    def test_empty_string(self):
+        assert _clean_gaia_id("") is None
+
+    def test_dash(self):
+        assert _clean_gaia_id("--") is None
+
+    def test_short_number_rejected(self):
+        assert _clean_gaia_id("123") is None
+
+
+# ---------------------------------------------------------------------------
+# Integration tests against the live Gaia TAP archive
+# ---------------------------------------------------------------------------
+
+@pytest.mark.remote_data
+class TestQueryGaiaAstrophysicalParams:
+
+    def test_basic_enrichment(self, sample_target_table):
+        """All 12 GAIA_ columns should be present after enrichment."""
+        enriched = query_gaia_astrophysical_params(sample_target_table)
+
+        assert len(enriched) == len(sample_target_table)
+
+        expected_cols = [
+            "GAIA_TEFF", "GAIA_TEFF_LOWER", "GAIA_TEFF_UPPER",
+            "GAIA_LOGG", "GAIA_LOGG_LOWER", "GAIA_LOGG_UPPER",
+            "GAIA_MH", "GAIA_MH_LOWER", "GAIA_MH_UPPER",
+            "GAIA_AG", "GAIA_AG_LOWER", "GAIA_AG_UPPER",
+        ]
+        for col in expected_cols:
+            assert col in enriched.colnames, f"Missing column: {col}"
+
[…]
