### 2026-03-18 ae744d4
Replace greek letters with ascii
