### 2026-03-18 a9d4b9a
Implement #110
