+++ src/spherical/database/target_table.py
@@ -441,6 +441,7 @@ def make_target_list_with_SIMBAD(
     table_of_files,
     instrument,
     polarimetry: bool = False,
+    sparse_aperture_masking: bool = False,
     search_radius: float = 0.5,
     parallax_limit: float = 1e-3,
     J_mag_limit: float = 15.0,
@@ -471,6 +472,9 @@ def make_target_list_with_SIMBAD(
         Instrument identifier ('irdis' or 'ifs').
     polarimetry : bool
         If True, include only frames with 'DPR_TECH' containing 'POLARIMETRY'.
+        If False, exclude such frames.
+    sparse_aperture_masking : bool
+        If True, include only frames with 'DPR_TECH' containing 'SAM'.
         If False, exclude such frames.
     search_radius : float
         SIMBAD search radius in arcseconds.
@@ -511,6 +515,7 @@ def make_target_list_with_SIMBAD(
         table_of_files=table_of_files,
         instrument=instrument,
         polarimetry=polarimetry,
+        sparse_aperture_masking=sparse_aperture_masking,
         remove_fillers=remove_fillers,
     )
 
