### 2026-03-24 9b941bc
add Zenodo sync script for database tables
