+++ src/spherical/scripts/sync_zenodo_tables.py
@@ -1,0 +1,382 @@
+from __future__ import annotations
+
+import argparse
+import hashlib
+import json
+import re
+import shutil
+import sys
+import tempfile
+import time
+import warnings
+from pathlib import Path
+from typing import Any
+from urllib.error import HTTPError, URLError
+from urllib.parse import urlparse
+from urllib.request import Request, urlopen
+
+
+DEFAULT_DOI = "10.5281/zenodo.15147730"
+DEFAULT_TIMEOUT = 120
+MANIFEST_NAME = ".zenodo_manifest.json"
+
+
+def _http_get_json(url: str, timeout: int = DEFAULT_TIMEOUT) -> dict[str, Any]:
+    req = Request(
+        url,
+        headers={
+            "Accept": "application/json",
+            "User-Agent": "spherical-zenodo-sync/1.0",
+        },
+    )
+    with urlopen(req, timeout=timeout) as response:
+        return json.load(response)
+
+
+def _resolve_zenodo_record_id(doi_or_record: str, timeout: int = DEFAULT_TIMEOUT) -> str:
+    """
+    Accepts:
+      - DOI: 10.5281/zenodo.15147730
+      - DOI URL: https://doi.org/10.5281/zenodo.15147730
+      - record id: 15459438
+      - Zenodo record URL: https://zenodo.org/records/15459438
+    Returns the concrete Zenodo record id as a string.
+    """
+    doi_or_record = doi_or_record.strip()
+
+    if doi_or_record.isdigit():
+        return doi_or_record
+
+    record_match = re.search(r"zenodo\.org/records/(\d+)", doi_or_record)
+    if record_match:
+        return record_match.group(1)
+
+    doi = doi_or_record
+    if doi.startswith("https://doi.org/"):
+        doi = doi.removeprefix("https://doi.org/")
+    elif doi.startswith("http://doi.org/"):
+        doi = doi.removeprefix("http://doi.org/")
+
+    # Resolve DOI redirect to latest Zenodo landing page.
+    doi_url = f"https://doi.org/{doi}"
+    req = Request(
+        doi_url,
+        headers={"User-Agent": "spherical-zenodo-sync/1.0"},
+    )
+    with urlopen(req, timeout=timeout) as response:
+        final_url = response.geturl()
+
+    record_match = re.search(r"zenodo\.org/records/(\d+)", final_url)
+    if not record_match:
+        raise RuntimeError(
+            f"Could not resolve DOI '{doi}' to a Zenodo record URL. Final URL was: {final_url}"
+        )
+    return record_match.group(1)
+
+
+def _load_record(record_id: str, timeout: int = DEFAULT_TIMEOUT) -> dict[str, Any]:
+    return _http_get_json(f"https://zenodo.org/api/records/{record_id}", timeout=timeout)
+
+
+def _extract_version(record: dict[str, Any]) -> str | None:
+    metadata = record.get("metadata", {})
+    version = metadata.get("version")
+    if version:
+        return str(version)
+
+    # Fall back to publication date if version not set.
+    for key in ("publication_date", "created", "updated", "modified"):
+        value = record.get(key) or metadata.get(key)
+        if value:
+            return str(value)
+    return None
+
+
+def _normalize_file_entry(file_entry: dict[str, Any]) -> dict[str, Any]:
+    """
+    Zenodo file metadata has changed over time. Handle common shapes defensively.
+    """
+    key = (
+        file_entry.get("key")
+        or file_entry.get("filename")
+        or file_entry.get("name")
+    )
+    if not key:
+        raise RuntimeError(f"Could not determine file name from entry: {file_entry}")
+
+    links = file_entry.get("links", {})
+    download_url = (
+        links.get("self")
[…]
