+++ CHANGELOG.md
@@ -9,15 +9,18 @@ This project follows [Semantic Versioning](https://semver.org/) and the [Keep a 
 ## [Unreleased]
 
 ### ✨ Added
+- **`dev-git` pixi environment** – Added a self-contained development environment that installs `charis` and `trap` from git (non-editable) instead of from local sibling directories. Use `pixi install -e dev-git` when working on `spherical` alone; the existing `dev` environment continues to install `charis`/`trap` editable from `../charis-dep/` and `../trap/` for cross-package development ([@m-samland](https://github.com/m-samland)).
 - **MOCAdb integration for stellar ages and young association membership** – New `mocadb_matching` module that cross-matches targets against the MOCAdb database (Gagné et al. 2026) via Gaia DR3 source IDs. Enriches the target table with 27 `MOCA_`-prefixed columns including association membership, BANYAN Σ probabilities, adopted ages, and optionally kinematics and activity indicators. Uses `pymysql` as a lightweight optional dependency. Integrated into the database generation pipeline as an optional enrichment step between target table and observation table ([@m-samland](https://github.com/m-samland)) ([#107](https://github.com/m-samland/spherical/issues/107)).
 - **Pixi package manager support** – Added `pixi.toml` with feature-based environments (`pipeline`, `notebook`, `docs`, `test`, `dev`) for managing conda and PyPI dependencies. The standard `pyproject.toml` remains fully functional for pip-based workflows ([@m-samland](https://github.com/m-samland)) ([#102](https://github.com/m-samland/spherical/issues/102)).
 - **Resume support for file table generation** – `make_file_table()` now automatically resumes interrupted runs. New data is written to a `*_partial.csv` file during processing; the final output is only updated on successful completion. On restart, already-downloaded DP.IDs are detected and skipped, avoiding redundant header retrieval from the ESO archive. Added `resume` parameter (default `True`) ([@m-samland](https://github.com/m-samland)) ([#105](https://github.com/m-samland/spherical/issues/105)).
 
 ### Changed
+- **Notebook dependency switched to JupyterLab** – The `notebook` optional dependency group (and the `notebook` pixi feature) now install `jupyterlab` instead of the classic `notebook` package, the modern Jupyter interface. Affects both `pyproject.toml` and `pixi.toml` ([@m-samland](https://github.com/m-samland)).
 - **README.md improvements** – Restructured installation instructions into clear pip and Pixi options. Consolidated scattered database table information into a dedicated section. Merged "Documentation and Examples" into Quick Start with correct `examples/` paths. Added spherical publication to citation list ([@m-samland](https://github.com/m-samland)).
 - **`python-json-logger` compatibility** – Made `logging_utils.py` compatible with both v2 (conda-forge) and v3 (PyPI) of `python-json-logger` via a try/except import ([@m-samland](https://github.com/m-samland)).
 
 ### Removed
+- **Dropped unused notebook dependencies `ipympl` and `ipydatagrid`** – Neither was imported by the package or any example notebook (`show_in_browser()` uses astropy's `jsviewer`, not `ipydatagrid`). The `notebook` group is now `ipython`, `jupyterlab`, `ipywidgets`, `seaborn`; `ipywidgets` is kept because `tqdm.auto` uses it for widget progress bars in JupyterLab ([@m-samland](https://github.com/m-samland)).
 - **Removed custom `utils.progress` module** – Replaced the custom tqdm environment-detection wrapper with `tqdm.auto`, which provides more robust notebook vs. console detection, proper `ipywidgets` fallback, and async support out of the box ([@m-samland](https://github.com/m-samland)) ([#104](https://github.com/m-samland/spherical/issues/104)).
 
 ### Fixed
