+++ tests/test_gaia_stellar_params_trap.py
@@ -1,0 +1,159 @@
+"""Tests for _apply_gaia_stellar_params helper in run_trap.py."""
+
+import logging
+from types import SimpleNamespace
+
+import numpy as np
+import pytest
+from astropy.table import Table
+
+# trap is not installed in the CI `test` env (pipeline deps are validated
+# locally only); skip the whole module there instead of erroring at collection.
+pytest.importorskip("trap")
+
+from trap.parameters import trap_config_for_ifs
+
+
+# ---------------------------------------------------------------------------
+# Fixtures
+# ---------------------------------------------------------------------------
+
+def _make_observation(gaia_teff=None, gaia_logg=None, gaia_mh=None, include_columns=True):
+    """Create a minimal mock observation with optional Gaia columns."""
+    data = {
+        "MAIN_ID": ["HD 12345"],
+        "FILTER": ["OBS_YJ"],
+        "NIGHT_START": ["2024-01-01"],
+        "WAFFLE_MODE": ["CORO"],
+    }
+    if include_columns:
+        data["GAIA_TEFF"] = [gaia_teff if gaia_teff is not None else np.nan]
+        data["GAIA_LOGG"] = [gaia_logg if gaia_logg is not None else np.nan]
+        data["GAIA_MH"] = [gaia_mh if gaia_mh is not None else np.nan]
+
+    obs_table = Table(data)
+    return SimpleNamespace(observation=obs_table)
+
+
+@pytest.fixture
+def default_config():
+    """Return a default IFS TrapConfig."""
+    return trap_config_for_ifs()
+
+
+@pytest.fixture
+def logger():
+    return logging.getLogger("test_gaia_stellar_params")
+
+
+# ---------------------------------------------------------------------------
+# Import the helper
+# ---------------------------------------------------------------------------
+
+from spherical.pipeline.run_trap import _apply_gaia_stellar_params
+
+
+# ---------------------------------------------------------------------------
+# Tests: basic override
+# ---------------------------------------------------------------------------
+
+class TestGaiaOverrideApplied:
+    """Gaia values in the observation table should override stellar params."""
+
+    def test_teff_override(self, default_config, logger):
+        obs = _make_observation(gaia_teff=6500.0)
+        result = _apply_gaia_stellar_params(obs, default_config, logger)
+        assert result.detection.stellar_parameters.teff == 6500.0
+
+    def test_logg_override(self, default_config, logger):
+        obs = _make_observation(gaia_logg=4.35)
+        result = _apply_gaia_stellar_params(obs, default_config, logger)
+        assert result.detection.stellar_parameters.logg == pytest.approx(4.35)
+
+    def test_feh_override(self, default_config, logger):
+        obs = _make_observation(gaia_mh=-0.12)
+        result = _apply_gaia_stellar_params(obs, default_config, logger)
+        assert result.detection.stellar_parameters.feh == pytest.approx(-0.12)
+
+    def test_all_three_overridden(self, default_config, logger):
+        obs = _make_observation(gaia_teff=5200.0, gaia_logg=3.8, gaia_mh=0.15)
+        result = _apply_gaia_stellar_params(obs, default_config, logger)
+        sp = result.detection.stellar_parameters
+        assert sp.teff == 5200.0
+        assert sp.logg == pytest.approx(3.8)
+        assert sp.feh == pytest.approx(0.15)
+
+    def test_original_config_not_mutated(self, default_config, logger):
+        original_teff = default_config.detection.stellar_parameters.teff
+        obs = _make_observation(gaia_teff=3400.0)
+        _apply_gaia_stellar_params(obs, default_config, logger)
+        assert default_config.detection.stellar_parameters.teff == original_teff
+
+
+# ---------------------------------------------------------------------------
+# Tests: fallback when Gaia values are NaN / missing
[…]
