+++ CHANGELOG.md
@@ -9,6 +9,7 @@ This project follows [Semantic Versioning](https://semver.org/) and the [Keep a 
 ## [Unreleased]
 
 ### ✨ Added
+- **Gaia stellar parameters as TRAP defaults** – TRAP detection now auto-populates the host star's effective temperature, surface gravity, and metallicity for template matching from the `GAIA_TEFF` / `GAIA_LOGG` / `GAIA_MH` columns of the observation table, per observation. NaN or missing columns fall back to the configured (or default) `StellarParameters` values. Controlled by `trap_config.detection.use_gaia_stellar_parameters` (default `True`); set it to `False` to keep a fixed value for all targets ([@m-samland](https://github.com/m-samland)).
 - **`dev-git` pixi environment** – Added a self-contained development environment that installs `charis` and `trap` from git (non-editable) instead of from local sibling directories. Use `pixi install -e dev-git` when working on `spherical` alone; the existing `dev` environment continues to install `charis`/`trap` editable from `../charis-dep/` and `../trap/` for cross-package development ([@m-samland](https://github.com/m-samland)).
 - **MOCAdb integration for stellar ages and young association membership** – New `mocadb_matching` module that cross-matches targets against the MOCAdb database (Gagné et al. 2026) via Gaia DR3 source IDs. Enriches the target table with 27 `MOCA_`-prefixed columns including association membership, BANYAN Σ probabilities, adopted ages, and optionally kinematics and activity indicators. Uses `pymysql` as a lightweight optional dependency. Integrated into the database generation pipeline as an optional enrichment step between target table and observation table ([@m-samland](https://github.com/m-samland)) ([#107](https://github.com/m-samland/spherical/issues/107)).
 - **Pixi package manager support** – Added `pixi.toml` with feature-based environments (`pipeline`, `notebook`, `docs`, `test`, `dev`) for managing conda and PyPI dependencies. The standard `pyproject.toml` remains fully functional for pip-based workflows ([@m-samland](https://github.com/m-samland)) ([#102](https://github.com/m-samland/spherical/issues/102)).
