### 2026-06-26 5b7ef1a
Refactor Gaia stellar parameter handling for trap reduction with fallback
