+++ tests/test_gaia_stellar_params_trap.py
@@ -1,4 +1,4 @@
-"""Tests for _apply_gaia_stellar_params helper in run_trap.py."""
+"""Tests for the TRAP stellar-parameter resolution in run_trap.py."""
 
 import logging
 from types import SimpleNamespace
@@ -13,26 +13,27 @@ pytest.importorskip("trap")
 
 from trap.parameters import trap_config_for_ifs
 
+from spherical.pipeline.pipeline_config import IFSReductionConfig
+from spherical.pipeline.run_trap import (
+    _apply_stellar_params,
+    _teff_from_spectral_type,
+)
 
-# ---------------------------------------------------------------------------
-# Fixtures
-# ---------------------------------------------------------------------------
 
-def _make_observation(gaia_teff=None, gaia_logg=None, gaia_mh=None, include_columns=True):
-    """Create a minimal mock observation with optional Gaia columns."""
+def _make_observation(gaia_teff=None, gaia_logg=None, gaia_mh=None, sp_type=None, include_gaia=True):
+    """Create a minimal mock observation with optional Gaia and SP_TYPE columns."""
     data = {
         "MAIN_ID": ["HD 12345"],
         "FILTER": ["OBS_YJ"],
-        "NIGHT_START": ["2024-01-01"],
-        "WAFFLE_MODE": ["CORO"],
     }
-    if include_columns:
+    if include_gaia:
         data["GAIA_TEFF"] = [gaia_teff if gaia_teff is not None else np.nan]
         data["GAIA_LOGG"] = [gaia_logg if gaia_logg is not None else np.nan]
         data["GAIA_MH"] = [gaia_mh if gaia_mh is not None else np.nan]
+    if sp_type is not None:
+        data["SP_TYPE"] = [sp_type]
 
-    obs_table = Table(data)
-    return SimpleNamespace(observation=obs_table)
+    return SimpleNamespace(observation=Table(data))
 
 
 @pytest.fixture
@@ -43,117 +44,128 @@ def default_config():
 
 @pytest.fixture
 def logger():
-    return logging.getLogger("test_gaia_stellar_params")
+    return logging.getLogger("test_stellar_params")
 
 
-# ---------------------------------------------------------------------------
-# Import the helper
-# ---------------------------------------------------------------------------
+# --- Tier 1: Gaia DR3 ------------------------------------------------------
 
-from spherical.pipeline.run_trap import _apply_gaia_stellar_params
+class TestGaiaTier:
+    """Finite Gaia values take priority and populate teff/logg/feh."""
 
-
-# ---------------------------------------------------------------------------
-# Tests: basic override
-# ---------------------------------------------------------------------------
-
-class TestGaiaOverrideApplied:
-    """Gaia values in the observation table should override stellar params."""
-
-    def test_teff_override(self, default_config, logger):
-        obs = _make_observation(gaia_teff=6500.0)
-        result = _apply_gaia_stellar_params(obs, default_config, logger)
-        assert result.detection.stellar_parameters.teff == 6500.0
-
-    def test_logg_override(self, default_config, logger):
-        obs = _make_observation(gaia_logg=4.35)
-        result = _apply_gaia_stellar_params(obs, default_config, logger)
-        assert result.detection.stellar_parameters.logg == pytest.approx(4.35)
-
-    def test_feh_override(self, default_config, logger):
-        obs = _make_observation(gaia_mh=-0.12)
-        result = _apply_gaia_stellar_params(obs, default_config, logger)
-        assert result.detection.stellar_parameters.feh == pytest.approx(-0.12)
-
-    def test_all_three_overridden(self, default_config, logger):
+    def test_all_three_from_gaia(self, default_config, logger):
         obs = _make_observation(gaia_teff=5200.0, gaia_logg=3.8, gaia_mh=0.15)
[…]
