+++ src/spherical/pipeline/run_trap.py
@@ -13,11 +13,12 @@ seamless integration into automated data processing workflows.
 
 import copy
 import os
+import re
 import time
 import traceback
 from copy import deepcopy
 from pathlib import Path
-from typing import Union
+from typing import Optional, Union
 
 import numpy as np
 import pandas as pd
@@ -39,75 +40,137 @@ from spherical.pipeline.logging_utils import (
 from spherical.pipeline.pipeline_config import IFSReductionConfig
 from spherical.pipeline.toolbox import make_target_folder_string
 
+_SPECTRAL_TYPE_TEFF_PATH = Path(__file__).parent / "spectral_type_teff.csv"
 
-def _apply_gaia_stellar_params(
+# Leading main spectral type + integer subtype (e.g. "G2" from "G2V", "F8" from
+# "F8/G0V"). Luminosity class, peculiarities, and fractional subtypes are
+# ignored; anything not matching falls through to the configured default.
+_SPECTRAL_TYPE_RE = re.compile(r"\s*[(~]?([OBAFGKM])(\d)")
+
+
+def _load_spectral_type_teff(path: Path = _SPECTRAL_TYPE_TEFF_PATH) -> dict[str, float]:
+    """Load the spectral-type -> Teff lookup table from the vendored CSV."""
+    table: dict[str, float] = {}
+    for line in path.read_text().splitlines():
+        line = line.strip()
+        if not line or line.startswith("#") or line.startswith("SpType"):
+            continue
+        sptype, teff = line.split(",")
+        table[sptype] = float(teff)
+    return table
+
+
+_SPECTRAL_TYPE_TEFF = _load_spectral_type_teff()
+
+
+def _finite_float(table, column: str) -> Optional[float]:
+    """Return the first row of *column* as a finite float, else ``None``."""
+    if column not in table.colnames:
+        return None
+    val = table[column][0]
+    if val is None or np.ma.is_masked(val):
+        return None
+    try:
+        fval = float(val)
+    except (TypeError, ValueError):
+        return None
+    return fval if np.isfinite(fval) else None
+
+
+def _nonempty_str(table, column: str) -> Optional[str]:
+    """Return the first row of *column* as a stripped string, else ``None``."""
+    if column not in table.colnames:
+        return None
+    val = table[column][0]
+    if val is None or np.ma.is_masked(val):
+        return None
+    if isinstance(val, bytes):
+        val = val.decode("utf-8", "ignore")
+    text = str(val).strip()
+    return text or None
+
+
+def _teff_from_spectral_type(sp_type: Optional[str]) -> Optional[float]:
+    """Map a spectral-type string to a main-sequence Teff, else ``None``.
+
+    Only the leading main type and integer subtype are used (fractional subtypes
+    are floored); non-OBAFGKM or unparseable types return ``None``.
+    """
+    if not sp_type:
+        return None
+    match = _SPECTRAL_TYPE_RE.match(sp_type)
+    if match is None:
+        return None
+    return _SPECTRAL_TYPE_TEFF.get(match.group(1) + match.group(2))
+
+
+def _apply_stellar_params(
     observation: Union[IFSObservation, IRDISObservation],
     trap_config,
     logger,
 ):
-    """Return a copy of *trap_config* with stellar parameters updated from Gaia columns.
+    """Return a copy of *trap_config* with stellar parameters resolved from the table.
 
-    If ``GAIA_TEFF``, ``GAIA_LOGG``, or ``GAIA_MH`` are present and finite in
-    ``observation.observation``, they override the corresponding values in
-    ``trap_config.detection.stellar_parameters``.  When a Gaia value is NaN or
-    the column is missing the user-provided (or default) value is preserved.
[…]
