### 2026-06-27 2d7dcab
Fix missing package-data so runtime data files ship in wheels
simbad_tap_query.adql and the IRDIS/CPI filter curves (data/*.txt) plus ifu_mask.fits are loaded via __file__ at runtime but were absent from [tool.setuptools.package-data], so non-editable installs raised FileNotFoundError. Editable installs were unaffected
