+++ CHANGELOG.md
@@ -25,6 +25,7 @@ This project follows [Semantic Versioning](https://semver.org/) and the [Keep a 
 - **Removed custom `utils.progress` module** – Replaced the custom tqdm environment-detection wrapper with `tqdm.auto`, which provides more robust notebook vs. console detection, proper `ipywidgets` fallback, and async support out of the box ([@m-samland](https://github.com/m-samland)) ([#104](https://github.com/m-samland/spherical/issues/104)).
 
 ### Fixed
+- **Package data files now ship in built wheels** – `simbad_tap_query.adql` (`spherical.database`) and the IRDIS/CPI filter-curve `data/*.txt` files plus `ifu_mask.fits` (`spherical.pipeline`) were loaded at runtime via `__file__` paths but were missing from `[tool.setuptools.package-data]`, so they were absent from non-editable (wheel) installs and raised `FileNotFoundError`. Added them to `package-data`. Editable installs were unaffected ([@m-samland](https://github.com/m-samland)).
 - **Fixed spurious "No usable science frames" warnings during observation table generation** – Removed `SKY` from the science-frame filter in `filter_for_science_frames()`. SKY frames were included in the observation-matching pool but never handled by `select_primary_science_frames()`, causing ghost observation groups that always failed with warnings ([@m-samland](https://github.com/m-samland)).
 - **Fixed `run_cube_header_update` crash outside git repo** – Fixed `ValueError` in `spherical_populate_fits_header()` when the pipeline runs from a working directory that is not inside a git repository (e.g. network filesystems, HPC). Git metadata collection now targets the spherical source tree directly and falls back to `"unknown"` when git is unavailable ([@m-samland](https://github.com/m-samland)) ([#101](https://github.com/m-samland/spherical/issues/101)).
 - **Fixed IFS pipeline step ordering causing crash at cube header update** – Reordered pipeline steps so that `run_frame_info_computation` executes before `run_cube_header_update`, resolving `FileNotFoundError` for `frames_info_*.csv` files. Also added defensive handling for missing frame-info CSVs, consistent use of `converted_dir` in bundling, and glob escaping for target names with special characters ([@m-samland](https://github.com/m-samland)) ([#97](https://github.com/m-samland/spherical/issues/97)).
