### 2026-06-28 c277f91
Demonstrate sparse_aperture_masking in db example; fix stellar-params comment
- generate_database.py: thread sparse_aperture_masking through the target
  and observation table builders
- ifs_reduction_template.py: document per-target stellar-parameter
  resolution and correct the use_gaia_stellar_parameters config path
