### 2026-07-02 a24952f
feat(database): add resolve_mode_name for Zenodo-style filenames
