### 2026-07-02 46fe1d8
feat(database): add provenance dataclass and JSON I/O
