### 2026-07-02 1bc94a4
fix(database): harden read_provenance; drop stray __init__.py
- Wrap record reconstruction in try-except to catch TypeError/KeyError/AttributeError
  when records are missing required fields (instrument/mode), treating the file as corrupt
- Add test_read_corrupt_record_returns_empty() to verify behavior
- Remove src/spherical/database/__init__.py to restore namespace package convention
- Narrow spherical_version() except clause from Exception to ImportError
- Annotate provenance_path(dest: "str | Path") parameter type
