### 2026-07-02 c0d8a81
feat(database): embed/extract provenance in table FITS meta
