### 2026-07-02 cfce8ba
fix(sync): unverifiable-checksum handling; add --dry-run/--list
