### 2026-07-02 1a8cfe8
feat(sync): write provenance, add tqdm progress and console entry point
