### 2026-07-02 c30f92d
feat(database): add build.build_tables with provenance
