### 2026-07-02 106c98f
feat(database): add build.enrich_tables for standalone re-enrichment
