### 2026-07-02 d4f69c7
feat(database): add update_database orchestration with resume + SAM
