+++ src/spherical/database/build.py
@@ -6,6 +6,7 @@ are delegated to the existing database modules.
 """
 from __future__ import annotations
 
+import datetime
 import logging
 from pathlib import Path
 
@@ -185,3 +186,107 @@ def enrich_tables(
     obs_tbl.write(dest / f"table_of_observations_{mode}{suffix}.fits", format="fits", overwrite=True)
     prov.write_provenance(dest, {mode: record})
     return record
+
+
+def derive_start_date(dest, instrument, overlap_days, suffix=""):
+    """Return the resume start date (YYYY-MM-DD) for an incremental update.
+
+    Prefers the standard-mode provenance ``eso_coverage_end``; falls back to the
+    max NIGHT_START in the local file-table CSV. Backs off by ``overlap_days``.
+    Returns None if no baseline exists.
+    """
+    dest = Path(dest)
+    instrument = instrument.lower()
+    coverage_end = None
+    record = prov.read_provenance(dest).get(instrument)
+    if record is not None:
+        coverage_end = record.eso_coverage_end
+    if coverage_end is None:
+        csv_path = dest / f"table_of_files_{instrument}{suffix}.csv"
+        if csv_path.exists():
+            coverage_end = _max_night_start(Table.read(csv_path, format="csv"))
+    if coverage_end is None:
+        return None
+    end_dt = datetime.datetime.strptime(coverage_end, "%Y-%m-%d")
+    return (end_dt - datetime.timedelta(days=overlap_days)).strftime("%Y-%m-%d")
+
+
+def modes_for_instrument(instrument, skip_sam=False):
+    """Return the list of modes to build for an instrument, as kwargs dicts."""
+    instrument = instrument.lower()
+    modes = [{"polarimetry": False, "sparse_aperture_masking": False}]
+    if instrument == "irdis":
+        modes.append({"polarimetry": True, "sparse_aperture_masking": False})
+    if not skip_sam:
+        modes.append({"polarimetry": False, "sparse_aperture_masking": True})
+    return modes
+
+
+def update_database(
+    dest,
+    instrument,
+    *,
+    start_date=None,
+    end_date=None,
+    overlap_days=7,
+    suffix="",
+    skip_sam=False,
+    enrich=True,
+    parallax_limit=1e-3,
+    J_mag_limit=14.0,
+    search_radius=3.0,
+    cone_size_science=15.0,
+    batch_size=150,
+):
+    """Extend the file table to today, then rebuild every mode's tables.
+
+    ``start_date`` defaults to the provenance-derived resume date minus
+    ``overlap_days``; ``end_date`` defaults to today. Raises ``ValueError`` if
+    there is no baseline and no explicit ``start_date``.
+    """
+    dest = Path(dest)
+    dest.mkdir(parents=True, exist_ok=True)
+    instrument = instrument.lower()
+
+    if end_date is None:
+        end_date = datetime.date.today().strftime("%Y-%m-%d")
+    if start_date is None:
+        start_date = derive_start_date(dest, instrument, overlap_days, suffix)
+    if start_date is None:
+        raise ValueError(
+            "No baseline tables found and no --start-date given. For a build from "
+            "scratch, pass an explicit start date (e.g. 2014-05-01)."
+        )
+
+    existing_csv = dest / f"table_of_files_{instrument}{suffix}.csv"
+    table_of_files = file_table.make_file_table(
+        output_dir=dest,
+        instrument=instrument,
+        start_date=start_date,
+        end_date=end_date,
+        output_suffix=suffix,
+        existing_table_path=existing_csv if existing_csv.exists() else None,
+        batch_size=batch_size,
+        cache=False,
+    )
+
+    records = {}
+    for mode_kw in modes_for_instrument(instrument, skip_sam=skip_sam):
[…]
