### 2026-07-02 3c53224
feat(cli): add spherical-update-database with --enrich-only
