+++ pyproject.toml
@@ -70,6 +70,7 @@ crash_reports = "spherical.scripts.aggregate_crash_reports:main"
 reduction_status = "spherical.scripts.aggregate_reduction_status:main"
 spherical-sync-tables = "spherical.scripts.sync_zenodo_tables:main"
 spherical-sync-zenodo-tables = "spherical.scripts.sync_zenodo_tables:main"
+spherical-update-database = "spherical.scripts.update_database:main"
 
 [project.urls]
 "Bug Tracker" = "https://github.com/m-samland/spherical/issues"
