### 2026-07-02 cb84848
refactor(examples): generate_database uses build orchestration
