+++ examples/generate_database.py
@@ -1,222 +1,38 @@
-"""Generate SPHERE database tables (file table, target table, observation table).
+"""Get an up-to-date SPHERE database: download the Zenodo baseline, then update.
 
-Usage Modes
------------
-1. **Build from scratch**:
-   Set ``build_file_table = True`` and choose a start/end date range.
-   The file table is created by querying the ESO archive for all SPHERE files
-   in that range, retrieving their FITS headers, and saving the result as a CSV.
-   This can take hours for large date ranges (e.g. the full archive 2014–today).
+Two steps, both also available as console commands:
+  1. spherical-sync-tables --dest <dir>            (download latest Zenodo tables)
+  2. spherical-update-database --dest <dir>         (extend to today + rebuild)
 
-2. **Extend an existing table** (e.g. add new observations):
-   Keep ``build_file_table = True`` and widen the date range (e.g. update
-   ``end_date`` to today). The function automatically detects the existing output
-   file, loads all DP.IDs that were already downloaded, and only retrieves
-   headers for new files. This is the recommended way to keep a database
-   up to date.
-
-3. **Resume an interrupted build**:
-   Simply re-run the script with the same parameters. If a previous run was
-   interrupted, a ``*_partial.csv`` file will exist alongside the output.
-   On the next run, already-downloaded entries from both the output file and the
-   partial file are skipped. Once all batches complete, the partial file is
-   merged into the final output and deleted. No manual intervention needed.
-
-4. **Skip file table generation** (use existing):
-   Set ``build_file_table = False`` to load the file table from disk and only
-   rebuild the target and/or observation tables.
-
-Notes
------
-- The ``existing_table_path`` parameter is kept for backward compatibility. When
-  set, its entries are also used to skip known files. In most cases you can leave
-  it pointed at the output file or remove it entirely — the auto-detection
-  handles everything.
-- Set ``resume=False`` in ``make_file_table()`` to force a clean download,
-  discarding any partial progress from a previous interrupted run.
+This script shows the equivalent library calls. `update_in_place` no longer
+exists: update mode always writes the canonical Zenodo-style filenames; pass a
+`suffix` only for experimental/scratch builds. A build from scratch is just an
+update with an explicit early start date and no baseline tables present.
 """
-
-import warnings
 from pathlib import Path
 
-import numpy as np
-from astropy.table import Table, vstack
+from spherical.database import build
 
-from spherical.database import file_table, observation_table, target_table
-from spherical.database.gaia_astrophysical_params import query_gaia_astrophysical_params
-from spherical.database.mocadb_matching import query_mocadb_for_targets
-from spherical.database.sphere_database import SphereDatabase
-
-warnings.filterwarnings("ignore")
-
-# Define path to store the database
+# Where the database tables live.
 table_path = Path.home() / "data/sphere/database"
 table_path.mkdir(parents=True, exist_ok=True)
 
-# Choose instrument ('ifs', 'irdis')
-instrument = "ifs"
-polarimetry = False
-sparse_aperture_masking = False
+# Step 1 (optional): download the latest pre-built tables from Zenodo.
+# Skip this if you already have up-to-date files in `table_path`.
+download_from_zenodo = True
+if download_from_zenodo:
+    build.sync_from_zenodo(table_path, instrument="all", include_polarimetry=True)
 
-# Set file name ending for the database files, e.g. "_test"
-output_suffix = ''
-
-# If True, ignore output_suffix and overwrite the canonical tables in place
[…]
