### 2026-07-02 7074f7f
docs(readme): document sync/update commands; fix table name typo
