### 2026-07-02 8dffe8a
chore: remove dead imports; align FITS coverage-start with JSON provenance
