+++ tests/test_provenance.py
@@ -1,5 +1,3 @@
-from pathlib import Path
-
 from spherical.database import provenance as prov
 
 
