### 2026-07-02 f8aafa1
Harden ESO header retrieval against network hangs; silence cosmetic warnings
make_file_table() issued ESO requests with no timeout, so a stalled
connection blocked indefinitely instead of failing and being skipped/
retried. Configure the ESO session with a (30s connect, 120s read)
timeout and mount urllib3 retries with backoff so transient disconnects
self-heal within the run; combined with the existing *_partial.csv
resume, an interrupted update continues by re-running the same command.

Also silence two cosmetic warnings that flooded long updates: the
VerifyWarning about DP.ID needing a HIER
[…]
