+++ CHANGELOG.md
@@ -9,6 +9,8 @@ This project follows [Semantic Versioning](https://semver.org/) and the [Keep a 
 ## [Unreleased]
 
 ### ✨ Added
+- **Automated database build & update workflow** – New two-step automation for getting an up-to-date SPHERE database, exposed as console commands and as library functions in `spherical.database.build`. Step 1, `spherical-sync-tables`, downloads the latest pre-built tables from Zenodo (with md5 verification, resumable chunked downloads, `--dry-run`/`--list`, and a `.zenodo_manifest.json`). Step 2, `spherical-update-database`, incrementally extends the file table from the last recorded coverage date to today against the ESO archive, then rebuilds the target and observation tables from scratch for every mode — including the newly generated sparse-aperture-masking (SAM) tables for IFS and IRDIS (`ifs`, `ifs_sam`, `irdis`, `irdis_polarimetry`, `irdis_sam`) — and runs Gaia DR3 and MOCAdb enrichment. Output files use the canonical Zenodo-style names (no `True`/`False` in filenames). A `--enrich-only` path re-runs Gaia/MOCA enrichment on existing tables and updates the enrichment status without re-querying ESO or rebuilding. `examples/generate_database.py` demonstrates the equivalent library calls ([@m-samland](https://github.com/m-samland)).
+- **Database provenance tracking** – Each build records provenance both in a human-readable `database_provenance.json` and embedded in the FITS table metadata (compact `SPH*` header keywords). Tracked per mode: spherical version, data source (`zenodo` vs `eso-extend`), ESO query date and coverage range, Gaia data release, Gaia/MOCA query dates, per-enrichment status (`ok`/`skipped`/`failed`), and the build parameters used (cone size, magnitude/parallax limits, search radius). This makes any table set self-describing and ready to publish as a future Zenodo release ([@m-samland](https://github.com/m-samland)).
 - **Per-target stellar parameters for TRAP template matching** – TRAP detection now resolves the host star's effective temperature (and, from Gaia, surface gravity and metallicity) per observation instead of using a single value for every target. Teff is resolved in priority order: Gaia DR3 (`GAIA_TEFF`/`GAIA_LOGG`/`GAIA_MH`), then a spectral-type estimate from `SP_TYPE` (mapped via a vendored Mamajek (2022) main-sequence Teff table; `logg`/`feh` keep their defaults), otherwise the values configured on `trap_config`. The active source is logged per target. Controlled by `IFSReductionConfig.use_gaia_stellar_parameters` (default `True`); set it to `False` to keep the configured values for all targets. The lookup lives entirely in `spherical` — `trap` only exposes generic stellar parameters and is not coupled to Gaia ([@m-samland](https://github.com/m-samland)) ([#109](https://github.com/m-samland/spherical/issues/109)).
 - **`dev-git` pixi environment** – Added a self-contained development environment that installs `charis` and `trap` from git (non-editable) instead of from local sibling directories. Use `pixi install -e dev-git` when working on `spherical` alone; the existing `dev` environment continues to install `charis`/`trap` editable from `../charis-dep/` and `../trap/` for cross-package development ([@m-samland](https://github.com/m-samland)).
[…]
