### 2026-07-02 9a9939d
Fix IndexError in filter_for_science_frames for IRDIS non-polarimetry/SAM
tech_col was computed once before the polarimetry filter, but that
filter shrinks t_science. The subsequent SAM filter applied the stale,
full-length boolean mask to the shortened table, raising 'IndexError:
boolean index did not match indexed array' and blocking target-table
generation for any IRDIS mode where the polarimetry filter removed rows.
Recompute tech_col from the current t_science before each filter step.
