+++ CHANGELOG.md
@@ -27,6 +27,7 @@ This project follows [Semantic Versioning](https://semver.org/) and the [Keep a 
 - **Removed custom `utils.progress` module** – Replaced the custom tqdm environment-detection wrapper with `tqdm.auto`, which provides more robust notebook vs. console detection, proper `ipywidgets` fallback, and async support out of the box ([@m-samland](https://github.com/m-samland)) ([#104](https://github.com/m-samland/spherical/issues/104)).
 
 ### Fixed
+- **Fixed `IndexError` in `filter_for_science_frames()` for IRDIS non-polarimetry / SAM modes** – The `tech_col` array (used to match `DPR_TECH` substrings) was computed once before the polarimetry filter, but that filter shrinks `t_science`. The subsequent SAM filter then applied the stale, full-length boolean mask to the shortened table, raising `IndexError: boolean index did not match indexed array`. This blocked target-table generation for any IRDIS mode where the polarimetry filter removed rows (e.g. plain `irdis`). `tech_col` is now recomputed from the current `t_science` before each filtering step ([@m-samland](https://github.com/m-samland)).
 - **Hardened ESO header retrieval against network hangs during file-table updates** – `make_file_table()` previously issued ESO archive requests with no timeout, so a stalled connection on a flaky link would block indefinitely (the raw error showed `connect timeout=None`) instead of failing and being skipped/retried. The ESO session now sets a (30 s connect, 120 s read) timeout and mounts urllib3 retries with backoff, so transient disconnects (`RemoteDisconnected`, timeouts, 5xx) self-heal within the run. Combined with the existing `*_partial.csv` resume mechanism, an interrupted update can be continued simply by re-running the same command — already-downloaded DP.IDs are skipped and unfetched sub-batches are retried ([@m-samland](https://github.com/m-samland)).
 - **Silenced cosmetic warnings during file-table generation** – Suppressed the astropy `VerifyWarning` about `DP.ID` requiring a HIERARCH card (emitted while building a throwaway `Header` solely to compute file size) and the `MergeConflictWarning` for `ID`/`name` meta keys (emitted when astroquery vstacks per-product header tables; the `.meta` is discarded at `.to_pandas()`). Both were harmless but flooded the logs during long updates ([@m-samland](https://github.com/m-samland)).
 - **Package data files now ship in built wheels** – `simbad_tap_query.adql` (`spherical.database`) and the IRDIS/CPI filter-curve `data/*.txt` files plus `ifu_mask.fits` (`spherical.pipeline`) were loaded at runtime via `__file__` paths but were missing from `[tool.setuptools.package-data]`, so they were absent from non-editable (wheel) installs and raised `FileNotFoundError`. Added them to `package-data`. Editable installs were unaffected ([@m-samland](https://github.com/m-samland)).
