### 2026-07-02 1c7a643
Make MOCA connection failure honest in provenance; fix eso_coverage_start
Two provenance-integrity fixes surfaced by the first full live update run:

1. query_mocadb_for_targets() swallowed a failed MySQL connection and
   returned empty MOCA_ columns, so _enrich recorded moca='ok' despite no
   MOCA data being written (all 27 MOCA_ columns empty across every mode).
   It now raises MocadbConnectionError on connection failure, which
   _enrich catches and records as 'failed'; a genuine no-match result
   still returns empty columns without raising. Affected tables can be
   repopulated later v
[…]
