+++ CHANGELOG.md
@@ -27,6 +27,8 @@ This project follows [Semantic Versioning](https://semver.org/) and the [Keep a 
 - **Removed custom `utils.progress` module** – Replaced the custom tqdm environment-detection wrapper with `tqdm.auto`, which provides more robust notebook vs. console detection, proper `ipywidgets` fallback, and async support out of the box ([@m-samland](https://github.com/m-samland)) ([#104](https://github.com/m-samland/spherical/issues/104)).
 
 ### Fixed
+- **MOCAdb connection failures are no longer silently reported as successful enrichment** – `query_mocadb_for_targets()` swallowed a failed MySQL connection (logging a warning and returning empty `MOCA_` columns), so the build recorded `enrichment.moca = "ok"` in the provenance even though no MOCA data was written. It now raises a dedicated `MocadbConnectionError` on connection failure, which the build catches and records as `"failed"` — a genuine "connected but no matches" result still returns empty columns without raising. Tables affected by a connection outage can be repopulated afterwards with `spherical-update-database --enrich-only` once MOCAdb is reachable ([@m-samland](https://github.com/m-samland)).
+- **`eso_coverage_start` in provenance now reflects the true data span** – When an incremental update extended a Zenodo baseline, `eso_coverage_start` (and the embedded `SPHCOV0` FITS keyword) recorded the incremental *resume date* instead of the table's earliest observing night, understating the real coverage by years. Coverage start is now derived from the file table's minimum `NIGHT_START`, mirroring how `eso_coverage_end` is already computed, so an extended baseline honestly reports its full coverage ([@m-samland](https://github.com/m-samland)).
 - **Fixed `IndexError` in `filter_for_science_frames()` for IRDIS non-polarimetry / SAM modes** – The `tech_col` array (used to match `DPR_TECH` substrings) was computed once before the polarimetry filter, but that filter shrinks `t_science`. The subsequent SAM filter then applied the stale, full-length boolean mask to the shortened table, raising `IndexError: boolean index did not match indexed array`. This blocked target-table generation for any IRDIS mode where the polarimetry filter removed rows (e.g. plain `irdis`). `tech_col` is now recomputed from the current `t_science` before each filtering step ([@m-samland](https://github.com/m-samland)).
 - **Hardened ESO header retrieval against network hangs during file-table updates** – `make_file_table()` previously issued ESO archive requests with no timeout, so a stalled connection on a flaky link would block indefinitely (the raw error showed `connect timeout=None`) instead of failing and being skipped/retried. The ESO session now sets a (30 s connect, 120 s read) timeout and mounts urllib3 retries with backoff, so transient disconnects (`RemoteDisconnected`, timeouts, 5xx) self-heal within the run. Combined with the existing `*_partial.csv` resume mechanism, an interrupted update can be continued simply by re-running the same command — already-downloaded DP.IDs are skipped and unfetched sub-batches are retried ([@m-samland](https://github.com/m-samland)).
 - **Silenced cosmetic warnings during file-table generation** – Suppressed the astropy `VerifyWarning` about `DP.ID` requiring a HIERARCH card (emitted while building a throwaway `Header` solely to compute file size) and the `MergeConflictWarning` for `ID`/`name` meta keys (emitted when astroquery vstacks per-product header tables; the `.meta` is discarded at `.to_pandas()`). Both were harmless but flooded the logs during long updates ([@m-samland](https://github.com/m-samland)).
