### 2026-07-02 73d5bb9
Use mocadb.ca hostname instead of hardcoded IP for MOCAdb
The hardcoded IP 104.248.106.21 no longer serves MOCAdb (connections
time out); the project's db_access page now publishes the hostname
mocadb.ca (currently 173.209.56.106). Using the hostname is also robust
to future server moves. Verified: connection and Gaia->moca_oid
resolution now succeed (352/381 targets resolved on a live subset).
