### 2026-07-02 a644c82
Adapt MOCAdb tier-1 query to new summary_all_objects schema
MOCAdb restructured summary_all_objects: it dropped moca_aid, moca_mtid
and banyan_prob, and replaced the single association pointer with
moca_aid_banyan (BANYAN Sigma) and moca_aid_literature. The old tier-1
query referenced sam.moca_aid and failed with 'Unknown column
sam.moca_aid'.

Rewrite tier-1 to adopt the BANYAN association where present, falling
back to the literature one (COALESCE), record the choice in a new
MOCA_AID_SOURCE column, and source moca_mtid/banyan_prob from the
per-membership table summary_all_members joined on 
[…]
