+++ CHANGELOG.md
@@ -27,6 +27,7 @@ This project follows [Semantic Versioning](https://semver.org/) and the [Keep a 
 - **Removed custom `utils.progress` module** – Replaced the custom tqdm environment-detection wrapper with `tqdm.auto`, which provides more robust notebook vs. console detection, proper `ipywidgets` fallback, and async support out of the box ([@m-samland](https://github.com/m-samland)) ([#104](https://github.com/m-samland/spherical/issues/104)).
 
 ### Fixed
+- **Restored MOCAdb enrichment after a server migration and schema change** – Two changes broke MOCAdb cross-matching: (1) the connection used a hardcoded IP (`104.248.106.21`) that no longer serves the database, and (2) MOCAdb restructured `summary_all_objects`, removing `moca_aid`/`moca_mtid`/`banyan_prob` and replacing the single association pointer with `moca_aid_banyan` (BANYAN Σ) and `moca_aid_literature`. The host is now the stable `mocadb.ca` hostname, and the tier-1 query adopts the BANYAN association where available, falling back to the literature one, sourcing `moca_mtid`/`banyan_prob` from `summary_all_members`. A new `MOCA_AID_SOURCE` column records which pointer (`banyan`/`literature`) supplied each match. Verified live: 429/500 sampled targets now receive associations and ages (e.g. TW Hya, Sco-Cen, Ursa Major) ([@m-samland](https://github.com/m-samland)).
 - **MOCAdb connection failures are no longer silently reported as successful enrichment** – `query_mocadb_for_targets()` swallowed a failed MySQL connection (logging a warning and returning empty `MOCA_` columns), so the build recorded `enrichment.moca = "ok"` in the provenance even though no MOCA data was written. It now raises a dedicated `MocadbConnectionError` on connection failure, which the build catches and records as `"failed"` — a genuine "connected but no matches" result still returns empty columns without raising. Tables affected by a connection outage can be repopulated afterwards with `spherical-update-database --enrich-only` once MOCAdb is reachable ([@m-samland](https://github.com/m-samland)).
 - **`eso_coverage_start` in provenance now reflects the true data span** – When an incremental update extended a Zenodo baseline, `eso_coverage_start` (and the embedded `SPHCOV0` FITS keyword) recorded the incremental *resume date* instead of the table's earliest observing night, understating the real coverage by years. Coverage start is now derived from the file table's minimum `NIGHT_START`, mirroring how `eso_coverage_end` is already computed, so an extended baseline honestly reports its full coverage ([@m-samland](https://github.com/m-samland)).
 - **Fixed `IndexError` in `filter_for_science_frames()` for IRDIS non-polarimetry / SAM modes** – The `tech_col` array (used to match `DPR_TECH` substrings) was computed once before the polarimetry filter, but that filter shrinks `t_science`. The subsequent SAM filter then applied the stale, full-length boolean mask to the shortened table, raising `IndexError: boolean index did not match indexed array`. This blocked target-table generation for any IRDIS mode where the polarimetry filter removed rows (e.g. plain `irdis`). `tech_col` is now recomputed from the current `t_science` before each filtering step ([@m-samland](https://github.com/m-samland)).
