### 2026-07-02 e7d9438
Make Gaia TAP failures honest and non-destructive; refresh coverage on enrich-only
Validation of the re-enriched tables showed GAIA_ columns wiped to 0%
after --enrich-only, while provenance still said gaia='ok'. Cause: a
transient Gaia TAP failure was swallowed (empty columns returned, no
raise), and because the query strips existing GAIA_ columns before
merging, the failure overwrote previously-good data.

- query_gaia_astrophysical_params now raises GaiaTapError on a TAP query
  failure (0-match still returns empty without raising). _enrich catches
  it -> records 'failed', and since the ra
[…]
