+++ CHANGELOG.md
@@ -28,8 +28,8 @@ This project follows [Semantic Versioning](https://semver.org/) and the [Keep a 
 
 ### Fixed
 - **Restored MOCAdb enrichment after a server migration and schema change** – Two changes broke MOCAdb cross-matching: (1) the connection used a hardcoded IP (`104.248.106.21`) that no longer serves the database, and (2) MOCAdb restructured `summary_all_objects`, removing `moca_aid`/`moca_mtid`/`banyan_prob` and replacing the single association pointer with `moca_aid_banyan` (BANYAN Σ) and `moca_aid_literature`. The host is now the stable `mocadb.ca` hostname, and the tier-1 query adopts the BANYAN association where available, falling back to the literature one, sourcing `moca_mtid`/`banyan_prob` from `summary_all_members`. A new `MOCA_AID_SOURCE` column records which pointer (`banyan`/`literature`) supplied each match. Verified live: 429/500 sampled targets now receive associations and ages (e.g. TW Hya, Sco-Cen, Ursa Major) ([@m-samland](https://github.com/m-samland)).
-- **MOCAdb connection failures are no longer silently reported as successful enrichment** – `query_mocadb_for_targets()` swallowed a failed MySQL connection (logging a warning and returning empty `MOCA_` columns), so the build recorded `enrichment.moca = "ok"` in the provenance even though no MOCA data was written. It now raises a dedicated `MocadbConnectionError` on connection failure, which the build catches and records as `"failed"` — a genuine "connected but no matches" result still returns empty columns without raising. Tables affected by a connection outage can be repopulated afterwards with `spherical-update-database --enrich-only` once MOCAdb is reachable ([@m-samland](https://github.com/m-samland)).
-- **`eso_coverage_start` in provenance now reflects the true data span** – When an incremental update extended a Zenodo baseline, `eso_coverage_start` (and the embedded `SPHCOV0` FITS keyword) recorded the incremental *resume date* instead of the table's earliest observing night, understating the real coverage by years. Coverage start is now derived from the file table's minimum `NIGHT_START`, mirroring how `eso_coverage_end` is already computed, so an extended baseline honestly reports its full coverage ([@m-samland](https://github.com/m-samland)).
+- **Enrichment connection failures are no longer silently reported as successful** – Both `query_mocadb_for_targets()` and `query_gaia_astrophysical_params()` swallowed a failed connection/query (logging a warning and returning empty columns), so the build recorded `enrichment.moca`/`enrichment.gaia = "ok"` even though no data was written. They now raise dedicated exceptions (`MocadbConnectionError`, `GaiaTapError`) on infrastructure failure, which the build catches and records as `"failed"`; a genuine "connected but no matches" result still returns empty columns without raising. For Gaia this also prevents a transient TAP outage during `--enrich-only` from **overwriting existing good `GAIA_` columns with empties** — the query now raises before the table is modified, preserving prior data. Affected tables can be repopulated with `spherical-update-database --enrich-only` once the services are reachable ([@m-samland](https://github.com/m-samland)).
+- **`eso_coverage_start` in provenance now reflects the true data span** – When an incremental update extended a Zenodo baseline, `eso_coverage_start` (and the embedded `SPHCOV0` FITS keyword) recorded the incremental *resume date* instead of the table's earliest observing night, understating the real coverage by years. Coverage start is now derived from the file table's minimum `NIGHT_START`, mirroring how `eso_coverage_end` is already computed, so an extended baseline honestly reports its full coverage. The `--enrich-only` path also refreshes both coverage bounds from the file table (repairing any stale value from an older build) without touching `eso_query_utc` ([@m-samland](https://github.com/m-samland)).
[…]
