### 2026-07-02 30e68f7
Add --mode flag for single-mode re-enrichment
spherical-update-database --enrich-only previously re-ran every mode of
an instrument. Add --mode {ifs,ifs_sam,irdis,irdis_polarimetry,irdis_sam}
to refresh just one mode, so a single mode can be retried after a
transient Gaia/MOCA outage without redoing the others. --mode requires
--enrich-only; using it without errors out.
