### 2026-07-02 c43d173
Round Gaia parameters and store as float32
GAIA_* columns carried spurious precision (14636.6279296875 K) from
float32->float64 promotion. Round GAIA_TEFF and bounds to whole Kelvin,
and logg/[M/H]/A_G and bounds to two decimals, matching GSP-Phot
precision. Store all twelve as float32 (the archive's native precision),
halving their on-disk footprint and keeping tables readable.
