+++ CHANGELOG.md
@@ -19,6 +19,7 @@ This project follows [Semantic Versioning](https://semver.org/) and the [Keep a 
 - **Resume support for file table generation** – `make_file_table()` now automatically resumes interrupted runs. New data is written to a `*_partial.csv` file during processing; the final output is only updated on successful completion. On restart, already-downloaded DP.IDs are detected and skipped, avoiding redundant header retrieval from the ESO archive. Added `resume` parameter (default `True`) ([@m-samland](https://github.com/m-samland)) ([#105](https://github.com/m-samland/spherical/issues/105)).
 
 ### Changed
+- **Gaia astrophysical parameters rounded and stored as `float32`** – The `GAIA_*` columns previously carried spurious precision (e.g. `14636.6279296875` K) from float32→float64 promotion. `GAIA_TEFF` and its bounds are now rounded to whole Kelvin, and `GAIA_LOGG`/`GAIA_MH`/`GAIA_AG` (and bounds) to two decimals, matching GSP-Phot's real precision. All twelve columns are stored as `float32` — the archive's native precision — halving their on-disk footprint and keeping the tables readable ([@m-samland](https://github.com/m-samland)).
 - **Notebook dependency switched to JupyterLab** – The `notebook` optional dependency group (and the `notebook` pixi feature) now install `jupyterlab` instead of the classic `notebook` package, the modern Jupyter interface. Affects both `pyproject.toml` and `pixi.toml` ([@m-samland](https://github.com/m-samland)).
 - **README.md improvements** – Restructured installation instructions into clear pip and Pixi options. Consolidated scattered database table information into a dedicated section. Merged "Documentation and Examples" into Quick Start with correct `examples/` paths. Added spherical publication to citation list ([@m-samland](https://github.com/m-samland)).
 - **`python-json-logger` compatibility** – Made `logging_utils.py` compatible with both v2 (conda-forge) and v3 (PyPI) of `python-json-logger` via a try/except import ([@m-samland](https://github.com/m-samland)).
