### 2026-07-02 69e8527
Update explore_database.ipynb for new mode naming, SAM and enrichment
The notebook predated the enrichment and SAM work and still built table
filenames with the removed _pol_{True,False} scheme, so it could not load
the current tables.

- Derive the canonical filename via resolve_mode_name; add an
  instrument/polarimetry/sparse_aperture_masking selector so the SAM
  tables are reachable.
- Add an enrichment section demonstrating the GAIA_* and MOCA_* columns
  with a young-moving-group (MOCA age < 150 Myr) filter example.
- Refresh stale intro text (J<14, dynamic coverage, modes, provenance).
[…]
