### 2026-07-02 8d54622
Support a list of names in observations_from_name_SIMBAD
The method was documented and implemented for a single string, but was
(historically) called with a list, raising 'AttributeError: list object
has no attribute strip' deep in name normalization.

Accept either a single name or a list/tuple: resolve each name (local ID
lookup with SIMBAD fallback), collect the matched MAIN_IDs, and select
the union once so a target reached through multiple names appears only
once. As a side effect of selecting from the (optionally usable-filtered)
table, usable_only=True now correctly returns only usable 
[…]
