+++ CHANGELOG.md
@@ -19,6 +19,7 @@ This project follows [Semantic Versioning](https://semver.org/) and the [Keep a 
 - **Resume support for file table generation** – `make_file_table()` now automatically resumes interrupted runs. New data is written to a `*_partial.csv` file during processing; the final output is only updated on successful completion. On restart, already-downloaded DP.IDs are detected and skipped, avoiding redundant header retrieval from the ESO archive. Added `resume` parameter (default `True`) ([@m-samland](https://github.com/m-samland)) ([#105](https://github.com/m-samland/spherical/issues/105)).
 
 ### Changed
+- **`SphereDatabase.observations_from_name_SIMBAD` accepts one or more names** – The method now takes either a single target name or a list/tuple of names, resolving each (local ID lookup with SIMBAD fallback) and returning the combined observations, with targets reached through multiple names included only once. Previously passing a list raised `AttributeError: 'list' object has no attribute 'strip'`. As part of the rewrite, `usable_only=True` now correctly returns only usable observations (it previously returned *all* observations of any target that had at least one usable observation) ([@m-samland](https://github.com/m-samland)).
 - **Updated `examples/explore_database.ipynb` for the new mode naming, SAM and enrichment** – The notebook still built table filenames with the removed `_pol_{True,False}` scheme; it now derives the canonical filename via `resolve_mode_name` and exposes an `instrument`/`polarimetry`/`sparse_aperture_masking` selector (so the SAM tables are reachable). Added a section demonstrating the Gaia DR3 (`GAIA_*`) and MOCAdb (`MOCA_*`) enrichment columns, including a young-moving-group filter example, refreshed the stale intro text, and cleared stale committed outputs ([@m-samland](https://github.com/m-samland)).
 - **Gaia astrophysical parameters rounded and stored as `float32`** – The `GAIA_*` columns previously carried spurious precision (e.g. `14636.6279296875` K) from float32→float64 promotion. `GAIA_TEFF` and its bounds are now rounded to whole Kelvin, and `GAIA_LOGG`/`GAIA_MH`/`GAIA_AG` (and bounds) to two decimals, matching GSP-Phot's real precision. All twelve columns are stored as `float32` — the archive's native precision — halving their on-disk footprint and keeping the tables readable ([@m-samland](https://github.com/m-samland)).
 - **Notebook dependency switched to JupyterLab** – The `notebook` optional dependency group (and the `notebook` pixi feature) now install `jupyterlab` instead of the classic `notebook` package, the modern Jupyter interface. Affects both `pyproject.toml` and `pixi.toml` ([@m-samland](https://github.com/m-samland)).
