### 2026-07-02 c7e65d5
Fix stale field-star assertion in MOCAdb test
BANYAN classifies field stars into the 'Field stars' association with an
old placeholder age (~11.2 Gyr) rather than returning NaN/empty. Update
the test to assert that field stars are not flagged as a young moving
group, and add a transient-server skip guard matching the sibling tests.
