### 2026-07-03 23ef751
Apply ruff import-sort and formatting to database build/enrichment modules
Run ruff check --fix (import ordering) and ruff format on the source
files touched by this branch. Scoped to branch-modified files only; the
rest of the repo is intentionally left untouched.
