+++ src/spherical/database/database_utils.py
@@ -1,9 +1,7 @@
 #!/usr/bin/env python
 # -*- coding: utf-8 -*-
 
-__author__ = (
-    "M. Samland @ MPIA (Heidelberg, Germany), J. Kemmer @ MPIA (Heidelberg, Germany)"
-)
+__author__ = "M. Samland @ MPIA (Heidelberg, Germany), J. Kemmer @ MPIA (Heidelberg, Germany)"
 
 import glob
 import itertools
@@ -16,7 +14,6 @@ import pandas as pd
 from astropy.io.ascii.core import InconsistentTableError
 from astropy.table import Column, Table, TableMergeError, hstack, join, unique, vstack
 from astroquery.gaia import Gaia
-
 from tqdm.auto import tqdm
 
 
@@ -27,7 +24,7 @@ def convert_table_to_little_endian(table):
     for colname in table.colnames:
         col = table[colname]
         if np.issubdtype(col.dtype, np.number):
-            table[colname] = col.astype(col.dtype.newbyteorder('='))
+            table[colname] = col.astype(col.dtype.newbyteorder("="))
     return table
 
 
@@ -45,7 +42,7 @@ def normalize_shutter_column(table):
     table : same type as input
         Modified table with 'SHUTTER' column coerced to bool or None.
     """
-    if 'SHUTTER' not in table.columns:
+    if "SHUTTER" not in table.columns:
         return table  # nothing to do
 
     def convert_to_bool(val):
@@ -64,10 +61,10 @@ def normalize_shutter_column(table):
         return None  # fallback for unexpected values
 
     if isinstance(table, pd.DataFrame):
-        table['SHUTTER'] = table['SHUTTER'].apply(convert_to_bool)
+        table["SHUTTER"] = table["SHUTTER"].apply(convert_to_bool)
     else:
         # Assume astropy Table
-        table['SHUTTER'] = [convert_to_bool(v) for v in table['SHUTTER']]
+        table["SHUTTER"] = [convert_to_bool(v) for v in table["SHUTTER"]]
 
     return table
 
@@ -88,8 +85,8 @@ def normalize_shutter_column(table):
 
 def add_night_start_date(df: pd.DataFrame, key="DATE_OBS") -> pd.DataFrame:
     if "NIGHT_START" not in df.columns:
-        times = pd.to_datetime(df[key], errors='coerce', utc=True)
-        night_start = (times - pd.to_timedelta((times.dt.hour < 12).astype(int), unit='D')).dt.date.astype(str)
+        times = pd.to_datetime(df[key], errors="coerce", utc=True)
+        night_start = (times - pd.to_timedelta((times.dt.hour < 12).astype(int), unit="D")).dt.date.astype(str)
         # Replace invalid parsed dates with a warning string
         night_start = night_start.where(times.notna(), "INVALID_DATE")
         df["NIGHT_START"] = night_start
@@ -219,9 +216,7 @@ def find_names_of_duplicates(
             available_values.append(fluxes)
         print(available_values)
         index_of_entry_to_retain = np.argmax(available_values)
-        duplicate_indices = np.where(np.arange(group_size) != index_of_entry_to_retain)[
-            0
-        ]
+        duplicate_indices = np.where(np.arange(group_size) != index_of_entry_to_retain)[0]
         print("indices of duplicate: {}".format(duplicate_indices))
         print("argmax: {}".format(np.argmax(available_values)))
         names_of_duplicates.append(group["MAIN_ID"][duplicate_indices].data)
@@ -240,26 +235,28 @@ def remove_spaces(name):
     splitname = " ".join(name.split())
     return splitname.replace(" ", "_")
 
+
 def convert_bytestring_columns(df):
     """
     Convert all bytestring columns in the DataFrame to normal strings.
-    
+
[…]
