+++ src/spherical/database/mocadb_matching.py
@@ -29,10 +29,10 @@ References
 
 from __future__ import annotations
 
+import html.entities
 import logging
 import re
 import warnings
-import html.entities
 
 import numpy as np
 import pandas as pd
@@ -65,6 +65,7 @@ class MocadbConnectionError(RuntimeError):
 # ---------------------------------------------------------------------------
 # Helper: parse Gaia DR3 IDs
 # ---------------------------------------------------------------------------
+
 
 def _clean_gaia_id(value) -> int | None:
     """Return an integer Gaia DR3 source_id, or ``None`` if not parseable."""
@@ -178,6 +179,7 @@ WHERE samem.moca_oid IN ({oid_list})
 # Core public function
 # ---------------------------------------------------------------------------
 
+
 def query_mocadb_for_targets(
     target_table: Table,
     *,
@@ -233,17 +235,13 @@ def query_mocadb_for_targets(
         import pymysql
     except ImportError:
         warnings.warn(
-            "pymysql is not installed – skipping MOCAdb enrichment. "
-            "Install with:  pip install pymysql",
+            "pymysql is not installed – skipping MOCAdb enrichment. Install with:  pip install pymysql",
             stacklevel=2,
         )
         return _attach_empty_columns(target_table, include_tier2)
 
     if gaia_id_column not in target_table.colnames:
-        raise ValueError(
-            f"Column '{gaia_id_column}' not found in target table. "
-            f"Available columns: {target_table.colnames}"
-        )
+        raise ValueError(f"Column '{gaia_id_column}' not found in target table. Available columns: {target_table.colnames}")
 
     # ----- Parse Gaia IDs -----
     raw_ids = list(target_table[gaia_id_column])
@@ -252,9 +250,9 @@ def query_mocadb_for_targets(
 
     if not valid_ids:
         logger.warning(
-            "MOCAdb: No valid Gaia DR3 IDs found in column '%s' "
-            "(checked %d rows) – returning table with empty MOCA columns.",
-            gaia_id_column, len(raw_ids),
+            "MOCAdb: No valid Gaia DR3 IDs found in column '%s' (checked %d rows) – returning table with empty MOCA columns.",
+            gaia_id_column,
+            len(raw_ids),
         )
         return _attach_empty_columns(target_table, include_tier2)
 
@@ -281,9 +279,7 @@ def query_mocadb_for_targets(
             charset="utf8mb4",
         )
     except Exception as e:
-        raise MocadbConnectionError(
-            f"Could not connect to MOCAdb at {MOCADB_HOST}:{MOCADB_PORT}: {e}"
-        ) from e
+        raise MocadbConnectionError(f"Could not connect to MOCAdb at {MOCADB_HOST}:{MOCADB_PORT}: {e}") from e
 
     try:
         # -- Step 0: Upload Gaia IDs as indexed temporary table --------
@@ -295,8 +291,7 @@ def query_mocadb_for_targets(
 
         if len(df_resolve) == 0:
             logger.warning(
-                "MOCAdb: None of the %d valid Gaia DR3 IDs were found in MOCAdb. "
-                "Returning table with empty MOCA columns.",
+                "MOCAdb: None of the %d valid Gaia DR3 IDs were found in MOCAdb. Returning table with empty MOCA columns.",
[…]

[…]
