### 2026-07-03 a488e6d
Fix import ordering in target_table and observation_table
Resolve the two ruff I001 (isort) errors that fail the CI lint job
(ruff check .). Import-ordering only; no other changes. Repo-wide
formatting is handled separately on the chore/ruff-format-repo branch.
