### 2026-07-03 d73554c
Migrate TRAP post-processing to immutable TrapReductionConfig
trap froze TrapReductionConfig, so spherical's in-place mutation of
result_folder on the config returned by read_output raised
FrozenInstanceError, and the deprecated get_reduction_parameters()
round-trip (new -> legacy Reduction_parameters -> new) emitted three
DeprecationWarnings per run. Build and adjust the config via
trap_config.reduction.merge(...) instead, and pass the frozen
TrapReductionConfig straight to trap, which accepts it zero-copy.

Also update examples/ifs_reduction_template.py to the same .merge()
pattern (direct f
[…]
