+++ CHANGELOG.md
@@ -31,6 +31,8 @@ This project follows [Semantic Versioning](https://semver.org/) and the [Keep a 
 - **Removed custom `utils.progress` module** – Replaced the custom tqdm environment-detection wrapper with `tqdm.auto`, which provides more robust notebook vs. console detection, proper `ipywidgets` fallback, and async support out of the box ([@m-samland](https://github.com/m-samland)) ([#104](https://github.com/m-samland/spherical/issues/104)).
 
 ### Fixed
+- **TRAP post-processing migrated to the immutable `TrapReductionConfig` (no more legacy round-trip)** – `run_trap.py` passed `trap` a legacy `Reduction_parameters` built via the deprecated `TrapConfig.get_reduction_parameters()`, and then *mutated* the config's `result_folder` in place. After `trap` made its config a frozen dataclass, the detection path's `analysis.reduction_parameters.result_folder = …` (on the frozen config returned by `read_output`) raised `FrozenInstanceError`, and every reduction emitted three `DeprecationWarning`s from the `get_reduction_parameters()` → legacy → `_to_reduction_config()` round-trip (new → legacy → new). Both the reduction and detection steps now build/adjust the config with `trap_config.reduction.merge(result_folder=…)` and pass the `TrapReductionConfig` straight to `trap` (which accepts it zero-copy), eliminating the round-trip, the deprecation warnings, and the crash. No behavioural change to reduction settings ([@m-samland](https://github.com/m-samland)) ([#115](https://github.com/m-samland/spherical/issues/115)).
+- **`examples/ifs_reduction_template.py` updated for the now-frozen TRAP config** – After `trap` made `TrapReductionConfig` an immutable (`frozen=True`) dataclass, the template's direct field assignments (`trap_config.reduction.search_region_outer_bound = 65`, …) raised `FrozenInstanceError`. The TRAP-parameter block now updates each sub-config by reassigning `.merge(...)` — the same pattern already used above for `config.steps`/`config.preprocessing` — which returns a copy with only the named fields overridden and preserves the `ncpus` set by `apply_trap_resources`. Also repaired a second staleness in the commented stellar-parameter example: the opt-out flag moved to `config.use_gaia_stellar_parameters` (the referenced `trap_config.detection.use_gaia_stellar_parameters` no longer exists), and the paired line now uses `stellar_parameters.merge(teff=...)`. The `run_trap.py` pipeline path is migrated to the same immutable config in the companion entry above ([@m-samland](https://github.com/m-samland)).
 - **Restored MOCAdb enrichment after a server migration and schema change** – Two changes broke MOCAdb cross-matching: (1) the connection used a hardcoded IP (`104.248.106.21`) that no longer serves the database, and (2) MOCAdb restructured `summary_all_objects`, removing `moca_aid`/`moca_mtid`/`banyan_prob` and replacing the single association pointer with `moca_aid_banyan` (BANYAN Σ) and `moca_aid_literature`. The host is now the stable `mocadb.ca` hostname, and the tier-1 query adopts the BANYAN association where available, falling back to the literature one, sourcing `moca_mtid`/`banyan_prob` from `summary_all_members`. A new `MOCA_AID_SOURCE` column records which pointer (`banyan`/`literature`) supplied each match. Verified live: 429/500 sampled targets now receive associations and ages (e.g. TW Hya, Sco-Cen, Ursa Major) ([@m-samland](https://github.com/m-samland)).
[…]
