### 2026-07-05 649fd8d
Adapt TRAP stellar-param handling and require trap >= 1.3.0
Two changes tied to the trap 1.3.0 release the IFS pipeline now targets:

Fix template-matching crash on sub-solar metallicity. trap builds the
stellar template from the solar-only bt-nextgen grid, so passing a real
Gaia [Fe/H] (e.g. beta Pic's -0.37) raised ValueError from species'
get_model and aborted detection. _apply_stellar_params now resolves only
Teff (Gaia -> spectral type) and log g (Gaia) and keeps [Fe/H] at the
configured solar default: at IFS resolution metallicity negligibly
changes the broadband template shape and the g
[…]
