### 2026-07-05 068da10
docs: add observation-table filtering & presentation design spec
Design for a single db.filter() entry point (validated equality/membership
kwargs + numpy mask escape hatch for comparisons and cross-column arithmetic),
a standalone view() for column-set presentation, exclude-missing semantics,
consolidation of usable_only into a single usable_mask(), removal of the
deprecated FAILED_SEQ path, and a ROTATION -10000 -> np.nan source fix so the
observation table's missing-value representation is uniform.
