### 2026-07-05 8bf5e44
feat(database): add usable_mask() and USABLE_MIN_EXPTIME_SCI
