### 2026-07-05 cb36cde
feat(database): add SUMMARY_COLUMNS and standalone view()
