### 2026-07-05 3ffd3af
feat(database): add SphereDatabase.filter() with validated criteria and mask escape hatch
