+++ tests/test_filter.py
@@ -4,15 +4,18 @@ These construct small astropy Tables directly, or a lightweight SphereDatabase
 via ``__new__`` (bypassing the heavy, network-dependent ``__init__``), so no
 real data or network access is needed.
 """
+import numpy as np
+import numpy.ma as ma
+import pytest
 from astropy.table import Table
 
 from spherical.database.sphere_database import (
-    USABLE_MIN_EXPTIME_SCI,
     SUMMARY_COLUMNS,
+    USABLE_MIN_EXPTIME_SCI,
+    SphereDatabase,
     usable_mask,
     view,
 )
-import pytest
 
 
 def _usable_table():
@@ -74,3 +77,89 @@ def test_view_silently_drops_absent_columns():
 def test_view_unknown_summary_raises():
     with pytest.raises(KeyError):
         view(_obs_table(), "NOPE")
+
+
+def _filter_table():
+    return Table(
+        {
+            "MAIN_ID": ["a", "b", "c", "d"],
+            "MEAN_FWHM": [0.8, 1.5, 0.9, 0.7],
+            "FLUX_J": [6.0, 7.0, 9.0, 5.0],
+            "DISTANCE": [10.0, 50.0, 20.0, 30.0],
+            "MOCA_ASSOCIATION_NAME": ma.MaskedArray(
+                [b"bpmg", b"other", b"bpmg", b"zzz"], mask=[False, False, False, True]
+            ),
+            "OBS_PROG_ID": [b"095.C-0298", b"111.24QL.001", b"110.240D.001", b"095.C-0298"],
+            "MOCA_AGE_MYR": ma.MaskedArray([20.0, 30.0, 99.0, 10.0], mask=[False, True, False, False]),
+        }
+    )
+
+
+def _lightweight_db(table):
+    db = SphereDatabase.__new__(SphereDatabase)  # bypass heavy __init__
+    db.table_of_observations = table
+    return db
+
+
+def test_filter_equality_criterion():
+    db = _lightweight_db(_filter_table())
+    result = db.filter(MOCA_ASSOCIATION_NAME="bpmg")
+    assert set(np.asarray(result["MAIN_ID"]).astype(str)) == {"a", "c"}
+
+
+def test_filter_membership_list():
+    db = _lightweight_db(_filter_table())
+    result = db.filter(OBS_PROG_ID=["110.240D.001", "111.24QL.001"])
+    assert set(np.asarray(result["MAIN_ID"]).astype(str)) == {"b", "c"}
+
+
+def test_filter_not_in_exclusion():
+    db = _lightweight_db(_filter_table())
+    result = db.filter(OBS_PROG_ID=("not in", ["095.C-0298"]))
+    assert set(np.asarray(result["MAIN_ID"]).astype(str)) == {"b", "c"}
+
+
+def test_filter_mask_and_criteria_compose():
+    db = _lightweight_db(_filter_table())
+    result = db.filter(lambda t: t["MEAN_FWHM"] < 1.0, MOCA_ASSOCIATION_NAME="bpmg")
+    assert set(np.asarray(result["MAIN_ID"]).astype(str)) == {"a", "c"}
+
+
+def test_filter_cross_column_mask():
+    db = _lightweight_db(_filter_table())
+    # absolute J mag = FLUX_J - 5*log10(DISTANCE/10); keep intrinsically bright.
+    # a: 6-0=6.0, b: 7-3.49=3.51, c: 9-1.51=7.49, d: 5-2.39=2.61  -> < 5 keeps {b, d}
+    result = db.filter(lambda t: t["FLUX_J"] - 5 * np.log10(t["DISTANCE"] / 10) < 5)
+    assert set(np.asarray(result["MAIN_ID"]).astype(str)) == {"b", "d"}
+
+
[…]
