### 2026-07-05 f4b3c52
feat(database): support target_list resolution in filter()
