### 2026-07-05 974d991
refactor(database): consolidate usable_only into usable_mask, drop FAILED_SEQ, route summaries through view()
