+++ src/spherical/database/sphere_database.py
@@ -245,99 +245,6 @@ class SphereDatabase(object):
         if isinstance(self.table_of_files["SHUTTER"][0], str):
             self.table_of_files["SHUTTER"] = [s.lower() in ("true", "t", "1") for s in self.table_of_files["SHUTTER"]]
 
-        # ---------- 5) flag column name (HCI_READY ↔ ~FAILED_SEQ) ---------------
-        self._ready_flag = "HCI_READY" if "HCI_READY" in self.table_of_observations.colnames else "FAILED_SEQ"
-
-        self._not_usable_observations_mask = self._mask_not_usable_observations(5.0)
-
-        # ---------- 6) build lists of "summary" keys ---------------------------
-        def _keys(base):
-            """Replace placeholders with actual column names that exist."""
-            out = []
-            for item in base:
-                if item == "{FILTER}":
-                    out.append(self._filter_keyword)
-                elif item == "{READY}":
-                    out.append(self._ready_flag)
-                else:
-                    out.append(item)
-            # silently drop keys not present in the table
-            return [k for k in out if k in self.table_of_observations.colnames]
-
-        base_head = ["MAIN_ID"]
-        base_tail = ["NIGHT_START", "{FILTER}", "WAFFLE_MODE", "{READY}", "DEROTATOR_MODE", "PRIMARY_SCIENCE"]
-
-        self._keys_for_summary = _keys(
-            base_head
-            + [
-                "ID_GAIA_DR3",
-                "ID_HIP",
-                "RA",
-                "DEC",
-                "OTYPE",
-                "SP_TYPE",
-                "FLUX_H",
-                "STARS_IN_CONE",
-            ]
-            + base_tail
-            + ["TOTAL_EXPTIME_SCI", "ROTATION", "MEAN_FWHM", "OBS_PROG_ID", "TOTAL_FILE_SIZE_MB"]
-        )
-
-        self._keys_for_obslog_summary = _keys(
-            base_head
-            + base_tail
-            + [
-                "DIT",
-                "NDIT",
-                "NCUBES",
-                "TOTAL_EXPTIME_SCI",
-                "TOTAL_EXPTIME_FLUX",
-                "ROTATION",
-                "MEAN_FWHM",
-                "MEAN_TAU",
-                "OBS_PROG_ID",
-            ]
-        )
-
-        self._keys_for_short_summary = _keys(
-            base_head
-            + base_tail
-            + [
-                "GAIA_TEFF",
-                "MOCA_AID",
-                "MOCA_BANYAN_PROB",
-                "MOCA_AGE_MYR",
-                "TOTAL_EXPTIME_SCI",
-                "ROTATION",
-                "MEAN_FWHM",
-                "OBS_PROG_ID",
-            ]
-        )
-
-        self._keys_for_medium_summary = _keys(
-            base_head
-            + base_tail
-            + [
-                "GAIA_TEFF",
-                "GAIA_LOGG",
-                "GAIA_MH",
-                "MOCA_AID",
-                "MOCA_BANYAN_PROB",
-                "MOCA_ASSOCIATION_NAME",
-                "MOCA_ASSOCIATION_TYPE",
-                "MOCA_AGE_MYR",
-                "MOCA_AGE_MYR_UNC",
-                "FLUX_H",
-                "DIT",
-                "NDIT",
-                "TOTAL_EXPTIME_SCI",
-                "ROTATION",
-                "MEAN_FWHM",
-                "STDDEV_FWHM",
-                "OBS_PROG_ID",
-            ]
-        )
-
         if table_of_files is not None:
             self._non_instrument_mask = self._mask_non_instrument_files()
             self._non_science_file_mask = self._mask_non_science_files()
[…]
