### 2026-07-05 f015dca
fix(database): use np.nan instead of -10000 sentinel for missing ROTATION
