+++ CHANGELOG.md
@@ -41,6 +41,9 @@ This project follows [Semantic Versioning](https://semver.org/) and the [Keep a 
   `HCI_READY`).
 
 ### Fixed
+- `ROTATION` now stores `np.nan` (not the `-10000` sentinel) when derotation is
+  not applicable/failed, so it round-trips to a masked column and is treated as
+  missing by filtering. Takes effect on the next database rebuild.
 - **TRAP template matching no longer crashes on sub-solar metallicity** – TRAP builds its stellar template from the solar-only `bt-nextgen` grid, so feeding it a real Gaia `[Fe/H]` (e.g. β Pic's `-0.37`) raised `ValueError: 'feh' … smaller than the lower boundary of the model grid` and aborted detection. `_apply_stellar_params` now resolves only Teff (Gaia → spectral-type) and log g (Gaia) from the archive and leaves `[Fe/H]` at the configured solar default — at IFS resolution metallicity negligibly changes the broadband template shape, and the grid is solar-only regardless. log g is still passed through; `trap` separately clamps any out-of-grid Teff/log g/`[Fe/H]` to the nearest grid boundary as a safety net (requires the companion `trap` change) ([@m-samland](https://github.com/m-samland)).
 - **TRAP post-processing migrated to the immutable `TrapReductionConfig` (no more legacy round-trip)** – `run_trap.py` passed `trap` a legacy `Reduction_parameters` built via the deprecated `TrapConfig.get_reduction_parameters()`, and then *mutated* the config's `result_folder` in place. After `trap` made its config a frozen dataclass, the detection path's `analysis.reduction_parameters.result_folder = …` (on the frozen config returned by `read_output`) raised `FrozenInstanceError`, and every reduction emitted three `DeprecationWarning`s from the `get_reduction_parameters()` → legacy → `_to_reduction_config()` round-trip (new → legacy → new). Both the reduction and detection steps now build/adjust the config with `trap_config.reduction.merge(result_folder=…)` and pass the `TrapReductionConfig` straight to `trap` (which accepts it zero-copy), eliminating the round-trip, the deprecation warnings, and the crash. No behavioural change to reduction settings ([@m-samland](https://github.com/m-samland)) ([#115](https://github.com/m-samland/spherical/issues/115)).
 - **`examples/ifs_reduction_template.py` updated for the now-frozen TRAP config** – After `trap` made `TrapReductionConfig` an immutable (`frozen=True`) dataclass, the template's direct field assignments (`trap_config.reduction.search_region_outer_bound = 65`, …) raised `FrozenInstanceError`. The TRAP-parameter block now updates each sub-config by reassigning `.merge(...)` — the same pattern already used above for `config.steps`/`config.preprocessing` — which returns a copy with only the named fields overridden and preserves the `ncpus` set by `apply_trap_resources`. Also repaired a second staleness in the commented stellar-parameter example: the opt-out flag moved to `config.use_gaia_stellar_parameters` (the referenced `trap_config.detection.use_gaia_stellar_parameters` no longer exists), and the paired line now uses `stellar_parameters.merge(teff=...)`. The `run_trap.py` pipeline path is migrated to the same immutable config in the companion entry above ([@m-samland](https://github.com/m-samland)).
