### 2026-07-05 2d449ad
docs(examples): use database.filter(target_list=...) in IFS reduction template
