### 2026-07-05 c9be41f
fix(database): preserve mask on directly-passed filter() masks; enrich unknown-column error
