### 2026-07-06 8303a9e
feat(database): add comparison, membership, and substring operators to filter()
filter() previously only supported equality/list-membership keywords plus
callable masks for anything else, forcing simple range/threshold cuts
(e.g. DISTANCE < 300) into ad-hoc lambdas. Add a (op, value) tuple form
covering >, <, >=, <=, ==, !=, 'in'/'not in', and 'contains', so most
criteria are expressible as keywords; drop callable-mask support in favor
of pre-computed boolean arrays as the sole escape hatch. Update the
example notebook and IFS reduction template to the new keyword style and
document DEROTATOR_
[…]
