+++ src/spherical/database/sphere_database.py
@@ -1,3 +1,4 @@
+import operator
 import warnings
 from typing import Dict, List, Optional, Sequence, Union
 
@@ -163,21 +164,62 @@ def _isin(column, values) -> np.ndarray:
     return np.isin(data, list(values))
 
 
+_COMPARE = {
+    ">": operator.gt,
+    "<": operator.lt,
+    ">=": operator.ge,
+    "<=": operator.le,
+    "==": operator.eq,
+    "!=": operator.ne,
+}
+
+
+def _contains(column, needle) -> np.ndarray:
+    """Boolean mask: string-column entries containing ``needle`` as a substring.
+
+    Byte (``|S``) and unicode (``|U``) columns are both handled by stringifying
+    the data, mirroring :func:`_isin`. ``needle`` may be ``str`` or ``bytes``.
+    """
+    data = np.asarray(column)
+    if data.dtype.kind not in ("S", "U"):
+        raise TypeError("'contains' requires a string column.")
+    if isinstance(needle, bytes):
+        needle = needle.decode()
+    return np.char.find(data.astype(str), str(needle)) >= 0
+
+
 def _criterion_mask(column, cond) -> np.ndarray:
     """Boolean mask for one filter criterion on a single column.
 
-    ``cond`` is a scalar (equality), a list/set/ndarray (membership), or a
-    ``('not in', sequence)`` tuple (exclusion). Missing-value handling is done
-    by the caller via the column mask, so masked positions may take an
-    arbitrary value here.
+    ``cond`` may be:
+
+    * a **scalar** -> equality (``==``);
+    * a **list/set/ndarray** -> membership (``in``);
+    * a **``(op, value)`` tuple**, where ``op`` is one of ``>``, ``<``, ``>=``,
+      ``<=``, ``==``, ``!=`` (comparison against ``value``); ``'in'`` /
+      ``'not in'`` (membership / exclusion against a sequence ``value``); or
+      ``'contains'`` (substring test on a string column).
+
+    Missing-value handling is done by the caller via the column mask, so masked
+    positions may take an arbitrary value here.
     """
     if isinstance(cond, tuple):
-        if len(cond) == 2 and cond[0] == "not in":
-            values = cond[1]
-            if isinstance(values, (str, bytes)) or not hasattr(values, "__iter__"):
-                raise TypeError("('not in', ...) requires a sequence of values.")
-            return ~_isin(column, values)
-        raise ValueError("Tuple criteria must be of the form ('not in', sequence).")
+        if len(cond) != 2:
+            raise ValueError("Tuple criteria must be of the form (op, value).")
+        op, value = cond
+        if op in _COMPARE:
+            return np.asarray(_COMPARE[op](column, value))
+        if op in ("in", "not in"):
+            if isinstance(value, (str, bytes)) or not hasattr(value, "__iter__"):
+                raise TypeError(f"({op!r}, ...) requires a sequence of values.")
+            result = _isin(column, value)
+            return ~result if op == "not in" else result
+        if op == "contains":
+            return _contains(column, value)
+        raise ValueError(
+            f"Unknown operator {op!r}. Use one of: "
+            ">, <, >=, <=, ==, !=, 'in', 'not in', 'contains'."
+        )
     if isinstance(cond, (list, set, np.ndarray)):
         return _isin(column, cond)
     return np.asarray(column == cond)
@@ -382,19 +424,23 @@ class SphereDatabase(object):
 
         Parameters
         ----------
-        *masks : numpy.ndarray or callable
-            Boolean arrays, or callables ``f(table) -> bool array``, combined
-            with logical AND. Use these for comparisons and cross-column
-            arithmetic (e.g. absolute magnitude, colours).
+        *masks : numpy.ndarray
[…]
