### 2026-07-06 d1d038a
feat(database): support list-valued 'not contains' (and 'contains') in filter()
'contains' only took a single substring, so excluding several patterns at
once (e.g. multiple program-ID prefixes) required chaining filter() calls
or falling back to a boolean mask. Let both 'contains' and 'not contains'
take a substring or a sequence of substrings, matched as OR ("contains any
of" / "contains none of").
