### 2026-07-06 69e72f4
feat(database): add exclude_targets to filter()
target_list only let you keep a set of named targets; excluding a few
targets from an otherwise broad query (e.g. drop known binaries from a
survey) required post-filtering the result table by MAIN_ID by hand. Add
exclude_targets, resolved the same way as target_list (local ID lookup
with SIMBAD fallback), applied after the target_list restriction. Names
that don't resolve to any observations are silently ignored, matching
target_list's per-name resolution behavior.
