### 2026-07-07 5bb5b75
docs: remove internal design and planning documents from repo
