+++ docs/superpowers/plans/2026-07-05-observation-filtering.md
@@ -1,927 +1,0 @@
-# Observation Table Filtering & Presentation Implementation Plan
-
-> **For agentic workers:** REQUIRED SUB-SKILL: Use superpowers:subagent-driven-development (recommended) or superpowers:executing-plans to implement this plan task-by-task. Steps use checkbox (`- [ ]`) syntax for tracking.
-
-**Goal:** Give `SphereDatabase` a single `filter()` entry point for the observation table (validated equality/membership kwargs + numpy mask escape hatch), a standalone `view()` for column-set presentation, uniform exclude-missing semantics, and consolidate the `usable_only` machinery — replacing the hand-rolled `np.logical_and.reduce` masks duplicated across the notebook and reduction template.
-
-**Architecture:** All new code lives in `src/spherical/database/sphere_database.py` as module-level functions (`usable_mask`, `view`, `SUMMARY_COLUMNS`, criterion helpers) plus a `filter()` method and `columns` property. `filter()` returns a filtered astropy `Table` (the existing interchange object). Missing values are excluded uniformly: `~np.ma.getmaskarray(col)` per criterion, `np.ma.filled(mask, False)` for masks. The deprecated `FAILED_SEQ`/`_ready_flag` path and the precomputed usable-mask subsystem are removed; a one-line source fix in `observation_table.py` makes the last non-masked sentinel (`ROTATION = -10000`) round-trip to a mask.
-
-**Tech Stack:** Python, numpy (incl. `numpy.ma`), astropy Tables, pytest, pixi.
-
-## Global Constraints
-
-- Work on branch `feature/observation-filtering` (already created off `develop`).
-- **Run everything through the pixi `dev` environment.** Use `pixi run -e dev <cmd>` for all commands. Concretely: tests = `pixi run -e dev pytest tests/... -v`; lint = `pixi run -e dev ruff check src/...`; ad-hoc Python = `pixi run -e dev python ...`. Never use bare `python`/`pip`.
-- **Commit once per task**, at the task's final step — not per sub-step. Plain commit messages, **no attribution footers** (no `Co-authored-by`, no "Generated with…" trailer).
-- The `spherical.database` half MUST import without the `pipeline` extra — no `scipy`/`charis`/`trap`/`photutils` imports in this code.
-- Style: ruff line-length **127**, **double quotes**, Google-style docstrings. Lint touched files before committing.
-- Comment only *why*, never *what* — name things well (per CLAUDE.md).
-- Tests live in `tests/`. Every behavior change lands or updates a test.
-- CHANGELOG discipline: add bullets under the right `[Unreleased]` subsection in `CHANGELOG.md`, Keep-a-Changelog format.
-- Never `git add -A`/`.`; stage explicit paths only.
-- Missing-value semantics are **fixed** (exclude-missing), not a flag.
-
----
-
-## File Structure
-
[…]
