+++ docs/superpowers/specs/2026-07-05-observation-filtering-design.md
@@ -1,328 +1,0 @@
-# Observation table filtering & presentation — design
-
-Date: 2026-07-05
-Status: Approved (brainstorming), pending spec review
-Repo: `spherical`, target branch: feature branch off `develop`
-
-## Problem
-
-Filtering the observation table is currently hand-rolled and duplicated. Two
-entry points both build a raw `np.logical_and.reduce([...])` mask:
-
-- `explore_database.ipynb` masks `table_of_observations` directly, bypassing the
-  `SphereDatabase` object entirely, then `table[mask].show_in_browser()`.
-- `examples/ifs_reduction_template.py` calls
-  `target_list_to_observation_table(target_list)` → Table → a second
-  hand-written mask → `retrieve_observation_metadata(table)`.
-
-Pain points:
-
-1. The filtering mechanism (`logical_and.reduce`) is duplicated and verbose in
-   two places.
-2. List membership (`OBS_PROG_ID` include/exclude, `MOCA_ASSOCIATION_NAME`) is
-   awkward with raw numpy (`~np.isin(...)`).
-3. "Usability" filtering (`usable_only`, a precomputed mask on the DB) and
-   ad-hoc filtering are two unrelated mechanisms that don't compose.
-4. Presentation (the `NORMAL`/`SHORT`/`MEDIUM`/`OBSLOG` column-sets) is welded
-   into `show_in_browser` and the SIMBAD lookups, so a filtered table cannot
-   reuse a named summary without copying the column list.
-5. **Latent correctness bug:** many columns are `MaskedColumn`s with real
-   missing values (`MOCA_AGE_MYR`, `GAIA_TEFF`, `MOCA_ASSOCIATION_NAME`, …). A
-   naive comparison such as `table['MOCA_AGE_MYR'] < 50` returns a *masked*
-   boolean; when used to index, astropy falls back to the underlying fill value,
-   so missing-value rows are included or excluded by an arbitrary fill rather
-   than by the criterion. The result is effectively undefined and silent.
-
-The motivating new capability is include/exclude filtering by lists of
-`OBS_PROG_ID`. The guiding constraint is minimalism: solve the real issues
-without inventing a query DSL or bloating the class.
-
-## Current state (findings)
-
-- Interchange object is an astropy `Table` in both flows;
-  `retrieve_observation_metadata(table)` consumes a `Table`.
-- `SphereDatabase` already has: a precomputed `_not_usable_observations_mask`
-  (HCI_READY + not FIELD + min exposure) exposed via `usable_only=`; the
-  summary column-sets built in `__init__`; robust name resolution
-  (`_find_observations_by_id_or_simbad`, with binary-naming and SIMBAD
-  fallback) used by `observations_from_name_SIMBAD` /
-  `get_observation_SIMBAD` / `retrieve_observation`.
-- Missing values in the shipped table are encoded as **masked entries in
-  `MaskedColumn`s**. `np.ma.getmaskarray(col)` reads the mask uniformly
-  (all-present for unmasked columns like `FLUX_J`, `DISTANCE`).
-- The build machinery writes **plain `Column`s with raw placeholders**, not
-  masked columns: Gaia and MOCA floats use `np.nan`, MOCA strings and target IDs
-  use `""` / `np.nan`. The astropy **FITS write→read cycle silently converts
-  `np.nan` and `""` to masks**, so on read these all become proper masked
-  columns — verified on the shipped table (`GAIA_*`, `MOCA_*`, `ID_HD`,
-  `ID_HIP` all masked). The one exception is a numeric sentinel that does *not*
-  round-trip to a mask: `ROTATION = -10000` (see the source-fix decision).
-- pandas is a base dependency, but the design deliberately stays in
-  numpy/astropy — no query-string path (rejected, see below).
-
-## Decisions
-
[…]
