### 2026-07-07 532d635
feat: migrate visualize_trap_results into spherical.pipeline.visualize
Add the plot_trap_mosaics console script and its mosaic-building module,
migrated from the standalone visualize_trap_results package.

- src/spherical/pipeline/visualize/mosaic.py: per-template detection/
  spectrum/combined mosaics from a TRAP results tree (self-contained;
  numpy/matplotlib/pandas/astropy only, GIF/imageio feature dropped).
  visualize/ is a namespace package (no __init__.py), matching pipeline/
  steps/ and pipeline/fits/
- src/spherical/scripts/plot_trap_mosaics.py: argparse CLI mirroring the
  existing
[…]
