+++ src/spherical/scripts/plot_trap_mosaics.py
@@ -1,0 +1,206 @@
+"""Command-line entry point for creating mosaics of TRAP results."""
+from __future__ import annotations
+
+import argparse
+import logging
+import sys
+from pathlib import Path
+from typing import Callable, Optional
+
+import matplotlib.pyplot as plt
+
+from spherical.pipeline.visualize import mosaic
+
+logger = logging.getLogger(__name__)
+
+TEMPLATE_TYPES = ["flat", "L-type", "T-type"]
+CONTENT_CHOICES = ["combined", "detection", "spectrum"]
+OBS_TABLE_FILENAME = "table_of_observations_ifs.fits"
+
+# content -> (single function name, batched function name, supports auto_scale)
+_PLOT_DISPATCH = {
+    "combined": ("plot_combined_mosaic", "plot_combined_mosaic_batched", True),
+    "detection": ("plot_detection_mosaic", "plot_detection_mosaic_batched", True),
+    "spectrum": ("plot_spectrum_mosaic", "plot_spectrum_mosaic_batched", False),
+}
+
+
+def build_parser() -> argparse.ArgumentParser:
+    parser = argparse.ArgumentParser(
+        description="Create mosaic summaries (PDF by default) of TRAP detection results."
+    )
+    parser.add_argument(
+        "base_path",
+        type=Path,
+        help="TRAP results root (contains target/mode/date/template_matching/...).",
+    )
+    parser.add_argument(
+        "--database-dir",
+        type=Path,
+        default=None,
+        help=f"Directory holding {OBS_TABLE_FILENAME}. Optional; without it, titles "
+        "omit exposure-time/rotation metadata.",
+    )
+    parser.add_argument(
+        "--output",
+        type=Path,
+        default=None,
+        help="Output directory (default: BASE_PATH/mosaics/).",
+    )
+    parser.add_argument(
+        "--suffix",
+        type=str,
+        default=None,
+        help="Filename suffix so subset runs do not overwrite "
+        "(e.g. combined_mosaic_flat_young.pdf).",
+    )
+    parser.add_argument(
+        "--templates",
+        nargs="+",
+        choices=TEMPLATE_TYPES,
+        default=list(TEMPLATE_TYPES),
+        help="Template subset to plot (default: all three).",
+    )
+    parser.add_argument(
+        "--content",
+        choices=CONTENT_CHOICES,
+        default="combined",
+        help="What to plot (default: combined = detection maps + spectra).",
+    )
+    parser.add_argument(
+        "--format",
+        choices=["pdf", "png"],
+        default="pdf",
+        help="Output format (default: pdf).",
+    )
+    parser.add_argument(
+        "--batch-size",
+        type=int,
+        default=None,
+        help="Split into files of N observations each (default: single file). "
+        "Batched filenames add a _batch_NN_of_MM marker and honor --format/--suffix.",
+    )
+    parser.add_argument(
+        "--snr-min",
+        type=float,
+        default=None,
+        help="Drop candidates with SNR below this value.",
+    )
+    parser.add_argument(
+        "--snr-max",
+        type=float,
+        default=None,
+        help="Drop candidates with SNR above this value.",
+    )
+    parser.add_argument(
+        "--dpi",
+        type=int,
+        default=300,
+        help="Output DPI (default: 300).",
+    )
+    parser.add_argument(
+        "--auto-scale",
+        action="store_true",
+        help="Enable global auto color-scaling (default: off, fixed sigma limits). "
+        "Ignored for --content spectrum.",
+    )
+    return parser
+
+
+def resolve_output_dir(base_path: Path, output: Optional[Path]) -> Path:
+    return output if output is not None else base_path / "mosaics"
+
+
+def build_output_path(
+    output_dir: Path, content: str, template: str, suffix: Optional[str], fmt: str
[…]
