### 2026-07-08 b8962e4
Default IFS coronagraph transmission for TRAP runs
IFS TRAP reductions now default trap_config.reduction.coronagraph_transmission
to the packaged N_ALC_JYH_S IFS curve so contrasts near the coronagraph are not
underestimated. Applies to both OBS_YJ and OBS_H. Gated by the new
IFSReductionConfig.apply_coronagraph_transmission toggle (default True); an
explicit table set on the trap config always wins, and an older trap without
the field is warned-and-skipped rather than crashing.

The two transmission tables ship as .txt to match the existing data/*.txt
package-data glob (no pyproject/MANIFEST 
[…]
