### 2026-07-08 cd1c869
Require trap >= 1.3.1 for default coronagraph transmission
coronagraph_transmission landed in trap 1.3.1 (now released), so bump the
runtime version floor from 1.3.0 to 1.3.1 and update the matching notes in
pyproject.toml, pixi.toml, and the CHANGELOG.
