+++ CHANGELOG.md
@@ -9,7 +9,7 @@ This project follows [Semantic Versioning](https://semver.org/) and the [Keep a 
 ## [Unreleased]
 
 ### ✨ Added
-- **Default IFS coronagraph transmission** – IFS TRAP runs now default `trap_config.reduction.coronagraph_transmission` to the packaged `N_ALC_JYH_S` IFS curve so contrasts near the coronagraph are not underestimated. Applied to both `OBS_YJ` and `OBS_H`. Toggle with `IFSReductionConfig.apply_coronagraph_transmission` (default `True`); an explicit table set on the trap config always wins. Requires a `trap` build with `coronagraph_transmission` support (older builds warn and skip) ([@m-samland](https://github.com/m-samland)).
+- **Default IFS coronagraph transmission** – IFS TRAP runs now default `trap_config.reduction.coronagraph_transmission` to the packaged `N_ALC_JYH_S` IFS curve so contrasts near the coronagraph are not underestimated. Applied to both `OBS_YJ` and `OBS_H`. Toggle with `IFSReductionConfig.apply_coronagraph_transmission` (default `True`); an explicit table set on the trap config always wins. Requires `trap >= 1.3.1`, where `coronagraph_transmission` was introduced ([@m-samland](https://github.com/m-samland)).
 - `SphereDatabase.filter()` for composable, validated observation-table filtering,
   plus `view()` and a `columns` property. Missing values are excluded per criterion.
   Per-column keyword criteria support equality (scalar), membership (list), and a
@@ -30,7 +30,7 @@ This project follows [Semantic Versioning](https://semver.org/) and the [Keep a 
 - **Resume support for file table generation** – `make_file_table()` now automatically resumes interrupted runs. New data is written to a `*_partial.csv` file during processing; the final output is only updated on successful completion. On restart, already-downloaded DP.IDs are detected and skipped, avoiding redundant header retrieval from the ESO archive. Added `resume` parameter (default `True`) ([@m-samland](https://github.com/m-samland)) ([#105](https://github.com/m-samland/spherical/issues/105)).
 
 ### Changed
-- **Requires `trap >= 1.3.0`** – The IFS post-processing now depends on `trap`'s dataclass configuration API (`TrapConfig` / immutable `TrapReductionConfig`), introduced in trap 1.3.0. Because `trap` is a git-URL dependency (which cannot carry a PEP 508 version floor), the minimum is enforced at runtime: importing `spherical.pipeline.run_trap` with an older `trap` now raises a clear `ImportError` telling you to upgrade, rather than failing later with a cryptic `AttributeError`. The floor is also documented next to the `trap` entries in `pyproject.toml` and `pixi.toml` ([@m-samland](https://github.com/m-samland)).
+- **Requires `trap >= 1.3.1`** – The IFS post-processing depends on `trap`'s dataclass configuration API (`TrapConfig` / immutable `TrapReductionConfig`, introduced in trap 1.3.0) and now also on the `coronagraph_transmission` reduction parameter it injects by default (introduced in trap 1.3.1), so the floor is 1.3.1. Because `trap` is a git-URL dependency (which cannot carry a PEP 508 version floor), the minimum is enforced at runtime: importing `spherical.pipeline.run_trap` with an older `trap` now raises a clear `ImportError` telling you to upgrade, rather than failing later with a cryptic `AttributeError`. The floor is also documented next to the `trap` entries in `pyproject.toml` and `pixi.toml` ([@m-samland](https://github.com/m-samland)).
[…]
