### 2026-07-08 1181aca
Remove dead trap-lacks-coronagraph fallback
The trap >= 1.3.1 floor guarantees the coronagraph_transmission field is
present, so the hasattr guard in _resolve_coronagraph_transmission, the
elif warn-and-skip branch in the wiring, and the now-unreachable
field-unsupported test are dead code. Drop them.
