### 2026-07-08 9b8ae85
feat: resolve plot_trap_mosaics database dir from $SPHERICAL_DATABASE_DIR
--database-dir still wins when given; otherwise fall back to the
$SPHERICAL_DATABASE_DIR environment variable, then to no metadata.
Adds resolve_database_dir() with tests for the three-way precedence.
