### 2026-07-09 9618c03
feat(pipeline): add step registry as single source of truth for step outputs/order
