### 2026-07-09 991d46b
feat(pipeline): add should_run resume gate emitting skipped_complete
