### 2026-07-09 2473106
feat(config): add force field to PipelineStepsConfig
Introduces the force resume/skip control (bool | set[str]) used by the
new step_registry.should_run logic, without yet removing the legacy
overwrite_* flags so existing readers keep working until Task 7.
