### 2026-07-09 ef6089c
feat(pipeline): gate IFS steps through should_run; drop per-step overwrite params
Wires ifs_reduction.py's IFS step call sites to should_run() so a resumed
run skips steps whose outputs already exist, with steps.force cascading
recomputation from a chosen step onward. This requires each step to no
longer decide overwrite behavior itself (that decision now belongs to
should_run/_forced), so the per-step overwrite/overwrite_bundle/
overwrite_preprocessing parameters are removed and their fits.writeto
calls hardcode overwrite=True. extract_cubes now writes a parent-side
completion marker since pe
[…]
