+++ src/spherical/pipeline/steps/flux_psf_calibration.py
@@ -5,8 +5,6 @@ Parameters
 ----------
 converted_dir : str
     Directory where the output files are stored and written.
-overwrite_preprocessing : bool
-    Whether to overwrite existing files.
 reduction_parameters : dict
     Reduction parameters dict, must contain 'flux_combination_method', 'exclude_first_flux_frame', and 'exclude_first_flux_frame_all'.
 """
@@ -28,7 +26,6 @@ from spherical.pipeline.steps.find_star import guess_position_psf, star_centers_
 @optional_logger
 def run_flux_psf_calibration(
     converted_dir: str,
-    overwrite_preprocessing: bool,
     reduction_parameters: Dict[str, str | bool],
     logger,
 ) -> None:
@@ -90,8 +87,6 @@ def run_flux_psf_calibration(
     ----------
     converted_dir : str
         Directory containing the input files and where outputs will be written.
-    overwrite_preprocessing : bool
-        Whether to overwrite existing output files.
     reduction_parameters : dict
         Reduction parameters dict, must contain:
         - flux_combination_method: str
@@ -161,7 +156,6 @@ def run_flux_psf_calibration(
     --------
     >>> run_flux_psf_calibration(
     ...     converted_dir="/path/to/converted",
-    ...     overwrite_preprocessing=True,
     ...     reduction_parameters={
     ...         "flux_combination_method": "mean",
     ...         "exclude_first_flux_frame": True,
@@ -263,20 +257,20 @@ def run_flux_psf_calibration(
     flux_centers = np.expand_dims(np.swapaxes(np.array(flux_centers), 0, 1), axis=2)
     flux_amplitudes = np.swapaxes(np.array(flux_amplitudes), 0, 1)
     logger.debug(f"Extracted flux_centers shape: {flux_centers.shape}, flux_amplitudes shape: {flux_amplitudes.shape}")
-    fits.writeto(additional_outputs_dir / 'flux_centers.fits', flux_centers, overwrite=overwrite_preprocessing)
-    fits.writeto(additional_outputs_dir / 'flux_gauss_amplitudes.fits', flux_amplitudes, overwrite=overwrite_preprocessing)
+    fits.writeto(additional_outputs_dir / 'flux_centers.fits', flux_centers, overwrite=True)
+    fits.writeto(additional_outputs_dir / 'flux_gauss_amplitudes.fits', flux_amplitudes, overwrite=True)
     flux_stamps = toolbox.extract_satellite_spot_stamps(
         flux_cube, flux_centers, stamp_size=57, shift_order=3, plot=False)
     logger.debug(f"Extracted flux_stamps shape: {flux_stamps.shape}")
     fits.writeto(additional_outputs_dir / 'flux_stamps_uncalibrated.fits',
-                 flux_stamps.astype('float32'), overwrite=overwrite_preprocessing)
+                 flux_stamps.astype('float32'), overwrite=True)
     if len(frames_info['FLUX']['INS4 FILT2 NAME'].unique()) > 1:
         logger.warning('Non-unique ND filters in sequence.', extra={"step": "flux_psf_calibration", "status": "failed"})
         raise ValueError('Non-unique ND filters in sequence.')
     else:
         ND = frames_info['FLUX']['INS4 FILT2 NAME'].unique()[0]
     _, attenuation = transmission.transmission_nd(ND, wave=wavelengths)
-    fits.writeto(additional_outputs_dir / 'nd_attenuation.fits', attenuation, overwrite=overwrite_preprocessing)
+    fits.writeto(additional_outputs_dir / 'nd_attenuation.fits', attenuation, overwrite=True)
[…]
