+++ src/spherical/pipeline/ifs_reduction.py
@@ -99,6 +99,7 @@ matplotlib.use(backend='Agg')  # Must be set before any matplotlib imports
 
 # Local imports
 from spherical.pipeline.pipeline_config import IFSReductionConfig, defaultIFSReduction
+from spherical.pipeline.step_registry import StepDirs, _forced, should_run, validate_force
 from spherical.pipeline.steps.bundle_output import run_bundle_output
 from spherical.pipeline.steps.cube_header_update import run_cube_header_update
 from spherical.pipeline.steps.download_data import download_data_for_observation, update_observation_file_paths
@@ -310,7 +311,8 @@ def execute_target(
     
     # Extract step configuration for cleaner code
     steps = config.steps
-    
+    validate_force(steps.force)
+
     # Get directory paths from config
     raw_directory = config.directories.raw_directory
     reduction_directory = config.directories.reduction_directory
@@ -395,7 +397,13 @@ def execute_target(
 
         calibration_time_name = str(observation.frames['WAVECAL']['DP.ID'][0][6:])  # type: ignore
         wavecal_outputdir = os.path.join(str(reduction_directory), 'IFS/calibration', obs_band, calibration_time_name)
-        
+
+        dirs = StepDirs(
+            converted_dir=Path(converted_dir),
+            cube_outputdir=Path(cube_outputdir),
+            wavecal_outputdir=Path(wavecal_outputdir),
+        )
+
         if not os.path.exists(outputdir):
             os.makedirs(outputdir)
 
@@ -406,11 +414,11 @@ def execute_target(
                 instrument=instrument,
                 calibration_parameters=calibration_parameters,
                 wavecal_outputdir=wavecal_outputdir,
-                overwrite_calibration=steps.overwrite_calibration,
+                overwrite_calibration=_forced("reduce_calibration", steps.force),
                 logger=logger,
             )
 
-        if steps.extract_cubes:
+        if should_run("extract_cubes", steps.extract_cubes, dirs, steps.force, logger):
             extract_cubes_with_multiprocessing(
                 observation=observation,
                 frame_types_to_extract=['CORO', 'CENTER', 'FLUX'],
@@ -420,10 +428,9 @@ def execute_target(
                 cube_outputdir=cube_outputdir,
                 logger=logger,
                 non_least_square_methods=non_least_square_methods,
-                extract_cubes=steps.extract_cubes,
             )
 
-        if steps.bundle_output:
+        if should_run("bundle_output", steps.bundle_output, dirs, steps.force, logger):
             run_bundle_output(
                 frame_types_to_extract=frame_types_to_extract,
                 cube_outputdir=cube_outputdir,
@@ -431,16 +438,15 @@ def execute_target(
                 extraction_parameters=extraction_parameters,
                 instrument=instrument,
                 non_least_square_methods=non_least_square_methods,
-                overwrite_bundle=steps.overwrite_bundle,
                 bundle_hexagons=steps.bundle_hexagons,
                 bundle_residuals=steps.bundle_residuals,
                 logger=logger,
             )
 
-        if steps.compute_frames_info:
+        if should_run("compute_frames_info", steps.compute_frames_info, dirs, steps.force, logger):
             run_frame_info_computation(observation, converted_dir, logger=logger)
 
-        if steps.cube_header_update:
+        if should_run("cube_header_update", steps.cube_header_update, dirs, steps.force, logger):
             run_cube_header_update(
                 frame_types_to_extract=frame_types_to_extract,
                 converted_dir=converted_dir,
@@ -449,34 +455,32 @@ def execute_target(
                 logger=logger,
             )
 
-        if steps.find_centers:
+        if should_run("find_centers", steps.find_centers, dirs, steps.force, logger):
[…]
