### 2026-07-09 cd5668c
feat(pipeline): drive TRAP overwrite from force; resume-gate TRAP detection
TRAP reduction's overwrite flag now follows the pipeline-wide force
mechanism instead of a hardcoded trap_config setting, so it participates
in the same cascade as the rest of the pipeline. TRAP detection is now
resume-gated via should_run/StepDirs and writes a completion marker on
success, matching the resume behavior already wired into ifs_reduction.py.
