### 2026-07-09 fd394a1
refactor(config): remove dead overwrite_* fields, superseded by force
Tasks 4-5 migrated every step caller off overwrite_calibration/bundle/
preprocessing/trap onto the force/resume mechanism; these fields had
no remaining readers, so drop them to avoid config drift.
