### 2026-07-09 24ec4a6
refactor(pipeline): derive check_output from the step registry (fixes additional_outputs path)
