### 2026-07-09 32d9f8b
docs: document resume/force model in template and CHANGELOG
The Tasks 1-8 refactor made the IFS pipeline resume by default and
replaced the four overwrite_* flags with a single force field that
cascades to downstream steps. Document the new behavior for users:
a commented RESUME/FORCE example in the reduction template, and
Unreleased CHANGELOG entries for the change and the removed flags.
