### 2026-07-09 de15d60
test(config): update merge test for force field replacing overwrite_*
Replace removed overwrite_calibration parameter with the new force field
in test_pipeline_steps_merge to match the PipelineStepsConfig refactoring.
