### 2026-07-09 2fb658c
docs: fix stale check_output docstring to point at step registry
The check_output function was reimplemented to derive completeness from
STEP_REGISTRY rather than hardcoded filenames. Update the docstring Notes
section to reflect this, removing the stale hardcoded file list and instead
pointing readers to the registry as the source of truth for required outputs.
