### 2026-07-09 a569a44
docs: link resume/force CHANGELOG entries to PR #121
