### 2026-07-09 49bcad1
Quiet trap logging and add progress bar in run_trap_on_observations
Batch TRAP runs lower the trap logger to WARNING (INFO when verbose) so recent
trap releases (which log via the standard logging module) no longer flood the
console, and wrap the observation loop in a tqdm progress bar. Harmless no-op on
older trap that still prints. The trap version floor is left unchanged to avoid
referencing an as-yet-untagged release.
