+++ CHANGELOG.md
@@ -30,6 +30,7 @@ This project follows [Semantic Versioning](https://semver.org/) and the [Keep a 
 - **Resume support for file table generation** – `make_file_table()` now automatically resumes interrupted runs. New data is written to a `*_partial.csv` file during processing; the final output is only updated on successful completion. On restart, already-downloaded DP.IDs are detected and skipped, avoiding redundant header retrieval from the ESO archive. Added `resume` parameter (default `True`) ([@m-samland](https://github.com/m-samland)) ([#105](https://github.com/m-samland/spherical/issues/105)).
 
 ### Changed
+- **Quieter TRAP batch runs with a progress bar** – `run_trap_on_observations` now lowers the `trap` logger to `WARNING` (or `INFO` when `trap_config.processing.verbose`) before iterating, so recent `trap` releases (which log via the standard `logging` module instead of `print`) no longer flood the console during multi-target runs, and wraps the observation loop in a `tqdm` progress bar. On older `trap` versions that still `print()` the level change is a harmless no-op. The `trap` floor is deliberately not raised yet; bump `_MIN_TRAP_VERSION` once the trap logging release is tagged to make the suppression a hard guarantee ([@m-samland](https://github.com/m-samland)).
 - **Requires `trap >= 1.3.1`** – The IFS post-processing depends on `trap`'s dataclass configuration API (`TrapConfig` / immutable `TrapReductionConfig`, introduced in trap 1.3.0) and now also on the `coronagraph_transmission` reduction parameter it injects by default (introduced in trap 1.3.1), so the floor is 1.3.1. Because `trap` is a git-URL dependency (which cannot carry a PEP 508 version floor), the minimum is enforced at runtime: importing `spherical.pipeline.run_trap` with an older `trap` now raises a clear `ImportError` telling you to upgrade, rather than failing later with a cryptic `AttributeError`. The floor is also documented next to the `trap` entries in `pyproject.toml` and `pixi.toml` ([@m-samland](https://github.com/m-samland)).
 - **`SphereDatabase.observations_from_name_SIMBAD` accepts one or more names** – The method now takes either a single target name or a list/tuple of names, resolving each (local ID lookup with SIMBAD fallback) and returning the combined observations, with targets reached through multiple names included only once. Previously passing a list raised `AttributeError: 'list' object has no attribute 'strip'`. As part of the rewrite, `usable_only=True` now correctly returns only usable observations (it previously returned *all* observations of any target that had at least one usable observation) ([@m-samland](https://github.com/m-samland)).
 - **Updated `examples/explore_database.ipynb` for the new mode naming, SAM and enrichment** – The notebook still built table filenames with the removed `_pol_{True,False}` scheme; it now derives the canonical filename via `resolve_mode_name` and exposes an `instrument`/`polarimetry`/`sparse_aperture_masking` selector (so the SAM tables are reachable). Added a section demonstrating the Gaia DR3 (`GAIA_*`) and MOCAdb (`MOCA_*`) enrichment columns, including a young-moving-group filter example, refreshed the stale intro text, and cleared stale committed outputs ([@m-samland](https://github.com/m-samland)).
