### 2026-07-10 c8084c4
Add quantitative enrichment health checks for Gaia/MOCA
The enrichment status was binary (ok/failed) and set only by whether the
query raised, so a query that succeeded while returning far fewer rows than
a known-good run — the flaky-partial failure mode — was still tagged "ok",
and update runs always exited 0. This makes a silently degraded enrichment
invisible.

Record a per-source match fraction (matched / valid-Gaia-ID rows) in
provenance and evaluate it against absolute floors plus a >10% relative drop
vs the previous run (SPHERE grows only incrementally and Gaia DR3/MOCAdb are
frozen unt
[…]
