+++ src/spherical/scripts/update_database.py
@@ -6,16 +6,27 @@ target and observation tables for every mode (including SAM). With
 querying ESO. ``--mode`` narrows an ``--enrich-only`` run to a single mode
 (e.g. ``irdis``), which is handy for retrying one mode after a transient
 Gaia/MOCA outage without redoing the others.
+
+After the tables are written, a per-mode Gaia/MOCA enrichment health summary is
+printed. Each source is compared against absolute match-fraction floors and the
+previous run's fractions (see ``spherical.database.enrichment_health``); the CLI
+exits non-zero if any mode's enrichment failed or degraded, so unattended runs
+surface a likely-failed enrichment rather than silently succeeding.
 """
 
 from __future__ import annotations
 
 import argparse
+import logging
 import sys
 from pathlib import Path
 
 from spherical.database import build
+from spherical.database import enrichment_health as eh
+from spherical.database import provenance as prov
 from spherical.database.database_utils import resolve_mode_name
+
+logger = logging.getLogger("spherical.update")
 
 MODE_CHOICES = ("ifs", "ifs_sam", "irdis", "irdis_polarimetry", "irdis_sam")
 
@@ -54,8 +65,63 @@ def _resolve_single_mode(mode_name):
     raise ValueError(f"Unknown mode: {mode_name}")
 
 
+def _modes_processed(args):
+    """Canonical mode names the run (re)built, used to scope the health summary."""
+    if args.mode is not None:
+        return [args.mode]
+    modes = []
+    for instrument in _instruments(args.instrument):
+        for mode_kw in build.modes_for_instrument(instrument, skip_sam=args.skip_sam):
+            modes.append(resolve_mode_name(instrument, mode_kw["polarimetry"], mode_kw["sparse_aperture_masking"]))
+    return modes
+
+
+def _summarize_enrichment(dest, prev, modes) -> int:
+    """Print a per-mode Gaia/MOCA health summary; return the number of breaches.
+
+    Each (re)built mode's stored metrics are compared against the pre-run
+    provenance (``prev``) using the floor + regression policy in
+    ``enrichment_health``. A breach is a failed/degraded source.
+    """
+    new = prov.read_provenance(dest)
+    rows = []
+    breaches = 0
+    for mode in modes:
+        record = new.get(mode)
+        if record is None or not isinstance(record.enrichment, dict):
+            continue
+        prev_enrich = prev[mode].enrichment if mode in prev else {}
+        for source in ("gaia", "moca"):
+            metrics = record.enrichment.get(source)
+            if not isinstance(metrics, dict):
+                continue
+            prev_metrics = prev_enrich.get(source) if isinstance(prev_enrich, dict) else None
+            verdict = eh.evaluate(source, metrics, prev_metrics)
+            status = metrics.get("status", "?")
+            frac = metrics.get("frac")
+            frac_s = f"{frac * 100:4.0f}%" if isinstance(frac, (int, float)) else "  --"
+            if status == "skipped":
+                tag = "SKIP"
+            elif verdict.passed:
+                tag = "OK"
+            else:
+                tag = "WARN"
+                breaches += 1
+                for reason in verdict.reasons:
+                    logger.warning("enrichment health: %s/%s: %s", mode, source, reason)
+            rows.append((mode, source, frac_s, tag, verdict.reasons))
+
+    if rows:
+        print("\n=== enrichment health summary ===", file=sys.stderr)
+        for mode, source, frac_s, tag, reasons in rows:
+            detail = f"  ({'; '.join(reasons)})" if reasons else ""
+            print(f"  {mode:<20} {source:<5} {frac_s}  {tag}{detail}", file=sys.stderr)
+    return breaches
+
+
[…]
