+++ CHANGELOG.md
@@ -9,6 +9,7 @@ This project follows [Semantic Versioning](https://semver.org/) and the [Keep a 
 ## [Unreleased]
 
 ### ✨ Added
+- **Quantitative enrichment health checks** – `spherical-update-database` now records a per-source Gaia/MOCA *match fraction* (matched / valid-Gaia-ID rows) alongside the query status, and evaluates it against absolute floors and the previous run's fractions to catch a silently degraded or partially failed enrichment (a query that succeeds but returns far fewer rows than a known-good run). Previously the only signal was a binary `ok`/`failed` status set solely by whether the query raised, so a zero-match or partial result was still tagged `ok`. Provenance `enrichment` is now nested per source (`status`, `n_total`, `n_valid_ids`, `n_matched`, `frac`, and `moca.tier2_ok`; legacy string values are still read). After the tables are written the CLI prints a per-mode health summary, logs a warning per breach, and **exits non-zero** if any mode's enrichment failed or dropped >10% relative to the prior run, so unattended runs no longer succeed silently on a bad enrichment. Thresholds live in the new `spherical.database.enrichment_health` module ([@m-samland](https://github.com/m-samland)).
 - **Default IFS coronagraph transmission** – IFS TRAP runs now default `trap_config.reduction.coronagraph_transmission` to the packaged `N_ALC_JYH_S` IFS curve so contrasts near the coronagraph are not underestimated. Applied to both `OBS_YJ` and `OBS_H`. Toggle with `IFSReductionConfig.apply_coronagraph_transmission` (default `True`); an explicit table set on the trap config always wins. Requires `trap >= 1.3.1`, where `coronagraph_transmission` was introduced ([@m-samland](https://github.com/m-samland)).
 - `SphereDatabase.filter()` for composable, validated observation-table filtering,
   plus `view()` and a `columns` property. Missing values are excluded per criterion.
