### 2026-07-17 e8f5843
feat(irdis): add minimal IRDISReductionConfig for download-only skeleton
