### 2026-07-17 f420ba3
feat(irdis): add execute_irdis_target orchestrator wired for download_data
