### 2026-07-17 4bcae20
feat(irdis): dispatch execute_targets to IRDIS orchestrator by instrument
