### 2026-07-17 adcf3c8
docs(irdis): add Phase 1 template and CHANGELOG entry
