+++ CHANGELOG.md
@@ -9,6 +9,7 @@ This project follows [Semantic Versioning](https://semver.org/) and the [Keep a 
 ## [Unreleased]
 
 ### ✨ Added
+- **IRDIS reduction — Phase 1 download-only skeleton** – `execute_targets()` now dispatches by instrument: IRDIS observations route to a new `execute_irdis_target()` in `spherical.pipeline.irdis_reduction` that wires the shared `download_data` step, avoiding the IFS orchestrator's unconditional WAVECAL access that crashes for IRDIS. Adds `IRDISReductionConfig` / `defaultIRDISReduction()` (composite dataclass reusing `DirectoryConfig`, `Resources`, `PreprocConfig`, `PipelineStepsConfig`) and `examples/irdis_reduction_template.py` for pulling a reference DBI dataset. Calibration, preprocessing, TRAP integration, and shared-step generalization land in later phases per the IRDIS design spec ([@m-samland](https://github.com/m-samland)).
 - **`plot_trap_mosaics` console script for visualizing TRAP results** – Migrated the standalone `visualize_trap_results` package into `spherical.pipeline.visualize` (module `mosaic`) with a `plot_trap_mosaics` command-line front end (`spherical.scripts.plot_trap_mosaics`). It builds per-template (`flat`/`L-type`/`T-type`) detection-map and spectrum mosaics from a TRAP results tree (`target/mode/date/template_matching/...`), optionally annotating candidates with SNR-based filtering and exposure/rotation metadata from the observation table. Supports `combined`/`detection`/`spectrum` content, single-file or `--batch-size`-split output (filenames honor `--format` and `--suffix`), PDF (searchable) or PNG, and fixed-sigma or `--auto-scale` color limits. The observation-table directory is resolved from `--database-dir`, then the `$SPHERICAL_DATABASE_DIR` environment variable, otherwise plots are drawn without exposure/rotation metadata. Depends only on the existing base stack (numpy/matplotlib/pandas/astropy); the old GIF feature (imageio) was dropped ([@m-samland](https://github.com/m-samland)).
 - **Default IFS coronagraph transmission** – IFS TRAP runs now default `trap_config.reduction.coronagraph_transmission` to the packaged `N_ALC_JYH_S` IFS curve so contrasts near the coronagraph are not underestimated. Applied to both `OBS_YJ` and `OBS_H`. Toggle with `IFSReductionConfig.apply_coronagraph_transmission` (default `True`); an explicit table set on the trap config always wins. Requires `trap >= 1.3.1`, where `coronagraph_transmission` was introduced ([@m-samland](https://github.com/m-samland)).
 - `SphereDatabase.filter()` for composable, validated observation-table filtering,
