### 2026-07-17 c77b08c
fix(irdis): widen execute_targets config type to accept IRDISReductionConfig
