### 2026-07-17 8e5af1f
feat(irdis): add IRDIS calibration/preprocess configs + step registry surface
