+++ src/spherical/pipeline/pipeline_config.py
@@ -170,6 +170,10 @@ class PipelineStepsConfig:
     extract_cubes: bool = True
     bundle_output: bool = True
     cube_header_update: bool = True
+
+    # IRDIS-only step flags
+    irdis_calibration: bool = True
+    preprocess_irdis: bool = True
     
     # Bundle options
     bundle_hexagons: bool = False
@@ -196,7 +200,7 @@ class PipelineStepsConfig:
     # Class-level list of all IFS pipeline steps (excludes TRAP and overwrite settings)
     _IFS_STEPS = [
         'download_data',
-        'reduce_calibration', 
+        'reduce_calibration',
         'extract_cubes',
         'bundle_output',
         'cube_header_update',
@@ -210,23 +214,49 @@ class PipelineStepsConfig:
         'calibrate_flux_psf',
         'spot_to_flux',
     ]
-    
+
+    # Class-level list of all IRDIS pipeline steps (excludes TRAP)
+    _IRDIS_STEPS = [
+        'download_data',
+        'irdis_calibration',
+        'preprocess_irdis',
+        'cube_header_update',
+        'compute_frames_info',
+        'find_centers',
+        'plot_image_center_evolution',
+        'process_extracted_centers',
+        'calibrate_spot_photometry',
+        'calibrate_flux_psf',
+        'spot_to_flux',
+    ]
+
     def merge(self, **kw) -> "PipelineStepsConfig":
         """Return a copy with selected fields overridden."""
         return replace(self, **kw)
-    
+
     def all_steps_disabled(self) -> bool:
-        """Check if all pipeline steps are disabled."""
-        return not any(getattr(self, step) for step in self._IFS_STEPS)
-    
+        """Check if all pipeline steps are disabled (IFS + IRDIS union)."""
+        all_steps = set(self._IFS_STEPS) | set(self._IRDIS_STEPS)
+        return not any(getattr(self, step) for step in all_steps)
+
     def enable_all_ifs_steps(self):
         """Enable all IFS pipeline steps (excludes TRAP and overwrite settings)."""
         for step in self._IFS_STEPS:
             setattr(self, step, True)
-    
+
     def disable_all_ifs_steps(self):
         """Disable all IFS pipeline steps (excludes TRAP and overwrite settings)."""
         for step in self._IFS_STEPS:
+            setattr(self, step, False)
+
+    def enable_all_irdis_steps(self):
+        """Enable all IRDIS pipeline steps (excludes TRAP)."""
+        for step in self._IRDIS_STEPS:
+            setattr(self, step, True)
+
+    def disable_all_irdis_steps(self):
+        """Disable all IRDIS pipeline steps (excludes TRAP)."""
+        for step in self._IRDIS_STEPS:
             setattr(self, step, False)
 
 # --- Composite reduction config --------------------------------------------
@@ -277,38 +307,69 @@ def defaultIFSReduction() -> IFSReductionConfig:
     return IFSReductionConfig()
 
 
-# --- IRDIS composite reduction config (Phase 1 skeleton) ------------------
-# Minimal composite for the download-only IRDIS path. Calibration and
-# preprocessing sub-configs will be added in later phases; only the fields
-# actually consumed by download_data + orchestration are present here.
+# -------- IRDIS-specific calibration & preprocess sub-configs --------------
+
+@dataclass(slots=True)
+class IRDISCalibrationConfig:
+    """Master-calibration parameters for the IRDIS calibration step.
+
+    Controls the construction of the master background, master flat, and
+    bad-pixel map from archive FLAT and BG_SCIENCE frames.
+    """
+    combination_method: str = "median"
[…]
