### 2026-07-17 682316a
feat(irdis): add IRDIS step registry and order-parametrize resume/force guards
