### 2026-07-17 173cb67
feat(irdis): add instrument-aware check_output for IRDIS reduction outputs
