### 2026-07-17 3e2e4ae
docs(irdis): CHANGELOG entry for Phase 2 config + step registry
