+++ CHANGELOG.md
@@ -10,6 +10,7 @@ This project follows [Semantic Versioning](https://semver.org/) and the [Keep a 
 
 ### ✨ Added
 - **IRDIS reduction — Phase 1 download-only skeleton** – `execute_targets()` now dispatches by instrument: IRDIS observations route to a new `execute_irdis_target()` in `spherical.pipeline.irdis_reduction` that wires the shared `download_data` step, avoiding the IFS orchestrator's unconditional WAVECAL access that crashes for IRDIS. Adds `IRDISReductionConfig` / `defaultIRDISReduction()` (composite dataclass reusing `DirectoryConfig`, `Resources`, `PreprocConfig`, `PipelineStepsConfig`) and `examples/irdis_reduction_template.py` for pulling a reference DBI dataset. Calibration, preprocessing, TRAP integration, and shared-step generalization land in later phases per the IRDIS design spec ([@m-samland](https://github.com/m-samland)).
+- **IRDIS reduction — Phase 2 config surface + step registry** – Adds `IRDISCalibrationConfig` (master flat / background / bad-pixel-map parameters) and `IRDISPreprocessConfig` (crop, anamorphism factor, gain, read noise, star-mask radius, badpix replacement) to the composite `IRDISReductionConfig`. `PipelineStepsConfig` gains `irdis_calibration` and `preprocess_irdis` step flags, a `_IRDIS_STEPS` class list, and `enable/disable_all_irdis_steps()` helpers; `all_steps_disabled()` now checks the IFS ∪ IRDIS union. `step_registry` gains `IRDIS_STEP_REGISTRY` / `IRDIS_STEP_ORDER` (reusing shared `StepSpec` entries so `reduction_status` aggregates across instruments) and `StepDirs.irdis_calibration_dir`; `_forced`, `should_run`, `validate_force`, and `expected_outputs` gain optional `step_order` / `registry` kwargs (default to IFS — zero churn at IFS call sites). `irdis_reduction.check_output` verifies IRDIS reductions off the IRDIS registry. Scaffolding only — no calibration/preprocess step runs yet (Phase 3+) ([@m-samland](https://github.com/m-samland)).
 - **`plot_trap_mosaics` console script for visualizing TRAP results** – Migrated the standalone `visualize_trap_results` package into `spherical.pipeline.visualize` (module `mosaic`) with a `plot_trap_mosaics` command-line front end (`spherical.scripts.plot_trap_mosaics`). It builds per-template (`flat`/`L-type`/`T-type`) detection-map and spectrum mosaics from a TRAP results tree (`target/mode/date/template_matching/...`), optionally annotating candidates with SNR-based filtering and exposure/rotation metadata from the observation table. Supports `combined`/`detection`/`spectrum` content, single-file or `--batch-size`-split output (filenames honor `--format` and `--suffix`), PDF (searchable) or PNG, and fixed-sigma or `--auto-scale` color limits. The observation-table directory is resolved from `--database-dir`, then the `$SPHERICAL_DATABASE_DIR` environment variable, otherwise plots are drawn without exposure/rotation metadata. Depends only on the existing base stack (numpy/matplotlib/pandas/astropy); the old GIF feature (imageio) was dropped ([@m-samland](https://github.com/m-samland)).
[…]
