### 2026-07-18 9dbb807
feat(irdis): add detector-half split helpers and dead-region constants
