+++ src/spherical/pipeline/steps/irdis_calibration.py
@@ -1,0 +1,98 @@
+"""VLT/SPHERE IRDIS master calibration step (Phase 3).
+
+Builds per-observation master background, master flat, and bad-pixel map from
+archive ``FLAT`` and ``BG_SCIENCE`` frames. All products are stored in the
+canonical split layout ``(2, 1024, 1024)`` so downstream science-frame
+application is a straight indexed operation.
+
+Dead detector regions are named constants — the row/column bands that were
+hardcoded in ``simplified_IRDIS_reduction.py`` are declared once here and
+consumed everywhere.
+
+Pixel-validity convention (charis-compatible; see design spec):
+
+- Dead detector regions → NaN in master data, ``False`` in the bad-pixel map
+  (so they are never interpolated). Downstream steps write ``ivar = 0`` at
+  these pixels.
+- Bad pixels flagged by this step → ``True`` in the bad-pixel map.
+  Downstream steps replace the data by interpolation and set ``ivar = 0``.
+
+This step is an internal-guard step (spec §6): if all outputs already exist
+on disk it logs ``skipped_complete`` and returns immediately.
+"""
+from __future__ import annotations
+
+import numpy as np
+
+# --- detector geometry -----------------------------------------------------
+# The IRDIS detector is 1024×2048; the two dual-band halves live at
+# columns [0:1024] (channel 0) and [1024:2048] (channel 1).
+
+DETECTOR_ROWS: int = 1024
+DETECTOR_COLS: int = 2048
+HALF_COLS: int = 1024
+
+# Dead regions in RAW detector coordinates (rows apply to both halves; column
+# slices are RAW-column indices — see dead_region_mask() for the per-half
+# projection).
+DEAD_ROW_SLICE_BOTTOM: slice = slice(0, 15)
+DEAD_ROW_SLICE_TOP: slice = slice(1013, 1024)
+DEAD_COL_SLICE_LEFT_HALF_EDGE: slice = slice(0, 50)
+DEAD_COL_SLICE_MID: slice = slice(941, 1078)
+DEAD_COL_SLICE_RIGHT_HALF_EDGE: slice = slice(1966, 2048)
+
+
+def split_detector_image(img: np.ndarray) -> np.ndarray:
+    """Split a raw ``(1024, 2048)`` IRDIS frame into ``(2, 1024, 1024)``.
+
+    Channel 0 is the left half (raw columns 0:1024); channel 1 is the right
+    half (raw columns 1024:2048).
+    """
+    if img.shape != (DETECTOR_ROWS, DETECTOR_COLS):
+        raise ValueError(
+            f"Expected shape ({DETECTOR_ROWS}, {DETECTOR_COLS}), got {img.shape}"
+        )
+    out = np.empty((2, DETECTOR_ROWS, HALF_COLS), dtype=img.dtype)
+    out[0] = img[:, 0:HALF_COLS]
+    out[1] = img[:, HALF_COLS:DETECTOR_COLS]
+    return out
+
+
+def split_detector_cube(cube: np.ndarray) -> np.ndarray:
+    """Split a raw ``(N, 1024, 2048)`` IRDIS cube into ``(N, 2, 1024, 1024)``."""
+    if cube.ndim != 3 or cube.shape[1:] != (DETECTOR_ROWS, DETECTOR_COLS):
+        raise ValueError(
+            f"Expected shape (N, {DETECTOR_ROWS}, {DETECTOR_COLS}), got {cube.shape}"
+        )
+    n = cube.shape[0]
[…]
