### 2026-07-18 69d6163
feat(irdis): master background via sigma-clipped mean of BG_SCIENCE
