### 2026-07-18 319b3d0
feat(irdis): master flat via DIT-linear fit with block-median illumination detrending
