+++ src/spherical/pipeline/steps/irdis_calibration.py
@@ -163,3 +163,180 @@ def build_master_background(bg_frames_paths, config, logger) -> np.ndarray:
         extra={"step": "irdis_calibration", "status": "background_complete"},
     )
     return master
+
+
+def estimate_smooth_illumination(
+    image: np.ndarray,
+    block_size: int = 64,
+) -> np.ndarray:
+    """Estimate the large-scale illumination pattern of a detector-half image.
+
+    Downsamples ``image`` by taking a ``nanmedian`` in ``block_size``-sized
+    blocks, then upsamples the low-resolution grid back to the input shape
+    via cubic-spline interpolation. Fast (~30 ms per 1024×1024 half) and
+    intrinsically robust to bad pixels (a handful of outliers per block are
+    invisible to the median).
+
+    Used by :func:`build_master_flat` to remove the integrating-sphere
+    illumination gradient before normalizing the flat. Can be reused for
+    per-frame illumination correction in later phases.
+
+    Parameters
+    ----------
+    image : np.ndarray
+        Shape ``(H, W)``. Bad pixels should be NaN so ``nanmedian`` ignores
+        them.
+    block_size : int, optional
+        Side length of the median-downsampling blocks. Default 64 → 16×16
+        grid for a 1024×1024 half.
+
+    Returns
+    -------
+    np.ndarray
+        Same shape as ``image``, float32, containing the smooth illumination
+        model. NaN-only blocks are replaced with the global nanmedian so the
+        interpolation stays finite.
+    """
+    import warnings
+
+    from scipy.interpolate import RegularGridInterpolator
+
+    h, w = image.shape
+    if h % block_size or w % block_size:
+        raise ValueError(
+            f"image shape {image.shape} not divisible by block_size {block_size}"
+        )
+    nb_y = h // block_size
+    nb_x = w // block_size
+
+    blocks = image.reshape(nb_y, block_size, nb_x, block_size).swapaxes(1, 2)
+    with warnings.catch_warnings():
+        warnings.simplefilter("ignore", RuntimeWarning)
+        low = np.nanmedian(blocks, axis=(2, 3)).astype(np.float32)
+
+    if not np.all(np.isfinite(low)):
+        fill = float(np.nanmedian(image))
+        low = np.where(np.isfinite(low), low, fill)
+
+    y_low = (np.arange(nb_y) + 0.5) * block_size
+    x_low = (np.arange(nb_x) + 0.5) * block_size
+    interpolator = RegularGridInterpolator(
+        (y_low, x_low), low, method="cubic", bounds_error=False, fill_value=None,
+    )
+    yy, xx = np.mgrid[:h, :w].astype(np.float32)
+    pts = np.stack([yy.ravel(), xx.ravel()], axis=-1)
+    return interpolator(pts).reshape(h, w).astype(np.float32)
+
+
+def _pixelwise_linear_slope(stack: np.ndarray, dits: np.ndarray) -> np.ndarray:
+    """Per-pixel least-squares slope of ``counts`` vs. ``dit`` along axis 0.
+
+    Fully vectorized closed-form OLS: ``slope = sum((x-x̄)(y-ȳ)) / sum((x-x̄)^2)``.
+
+    Parameters
+    ----------
+    stack : np.ndarray
+        Shape ``(N_frames, 2, 1024, 1024)``, float32.
+    dits : np.ndarray
+        Shape ``(N_frames,)``, DIT of each frame.
+
+    Returns
+    -------
+    np.ndarray
+        Slope per pixel, shape ``(2, 1024, 1024)``, float32.
+    """
+    dits = np.asarray(dits, dtype=np.float64)
+    x = dits - dits.mean()
+    denom = float((x * x).sum())
[…]
