### 2026-07-18 118fa87
feat(irdis): bad-pixel map from union of flat-response and background outliers
