### 2026-07-18 824a522
feat(irdis): wire run_irdis_calibration into execute_irdis_target
