### 2026-07-18 d73563c
fix(irdis): flag bad pixels off detrended flat/bg to avoid vignetting false-positives
The abnormal-response test previously consulted the raw per-pixel slope and
its global mad_std. Because the slope drops smoothly toward the edges of
the illuminated field (integrating-sphere vignetting), legitimate edge
pixels appeared as >5σ low outliers and got flagged en masse: 34% of pixels
in the r=450–550 ring on channel 0 of a K12 dataset.

Move the flat sigma/floor/ceiling test to the illumination-detrended
`master_flat` (median ≈ 1 by construction; spread encodes real pixel-to-
pixel response variati
[…]
