+++ tests/test_irdis_calibration.py
@@ -333,7 +333,12 @@ class TestBuildMasterFlat:
 
 
 class TestBuildBadPixelMap:
-    def _uniform_response(self, value=100.0):
+    """Tests for build_bad_pixel_map. Note: first argument is the NORMALIZED
+    ``master_flat`` (median ≈ 1 per half after illumination detrending),
+    not the raw response — so that vignetting doesn't cause legitimate
+    edge pixels to be flagged as sigma-outliers."""
+
+    def _uniform_flat(self, value=1.0):
         return np.full((2, 1024, 1024), value, dtype=np.float32)
 
     def _uniform_bg(self, value=50.0, n_frames=3):
@@ -353,11 +358,11 @@ class TestBuildBadPixelMap:
 
         from spherical.pipeline.steps.irdis_calibration import build_bad_pixel_map
 
-        response = self._uniform_response()
+        flat = self._uniform_flat()
         bg_stack = self._uniform_bg()
         master_bg = bg_stack.mean(axis=0)
 
-        bpm = build_bad_pixel_map(response, bg_stack, master_bg, self._cfg(), logger=MagicMock())
+        bpm = build_bad_pixel_map(flat, bg_stack, master_bg, self._cfg(), logger=MagicMock())
         assert not bpm[0, 512, 512]
         assert not bpm[1, 512, 512]
 
@@ -366,12 +371,12 @@ class TestBuildBadPixelMap:
 
         from spherical.pipeline.steps.irdis_calibration import build_bad_pixel_map
 
-        response = self._uniform_response(100.0)
-        response[0, 500, 500] = 10.0  # 10% of median → below hard floor.
+        flat = self._uniform_flat(1.0)
+        flat[0, 500, 500] = 0.1  # 10% of nominal → below hard floor.
         bg_stack = self._uniform_bg()
         master_bg = bg_stack.mean(axis=0)
 
-        bpm = build_bad_pixel_map(response, bg_stack, master_bg, self._cfg(), logger=MagicMock())
+        bpm = build_bad_pixel_map(flat, bg_stack, master_bg, self._cfg(), logger=MagicMock())
         assert bpm[0, 500, 500]
 
     def test_hot_pixel_in_background_flagged(self):
@@ -379,12 +384,12 @@ class TestBuildBadPixelMap:
 
         from spherical.pipeline.steps.irdis_calibration import build_bad_pixel_map
 
-        response = self._uniform_response(100.0)
+        flat = self._uniform_flat(1.0)
         master_bg = np.full((2, 1024, 1024), 50.0, dtype=np.float32)
         master_bg[1, 300, 300] = 50000.0
         bg_stack = np.stack([master_bg, master_bg, master_bg])
 
-        bpm = build_bad_pixel_map(response, bg_stack, master_bg, self._cfg(), logger=MagicMock())
+        bpm = build_bad_pixel_map(flat, bg_stack, master_bg, self._cfg(), logger=MagicMock())
         assert bpm[1, 300, 300]
 
     def test_dead_regions_are_false(self):
@@ -395,13 +400,13 @@ class TestBuildBadPixelMap:
             dead_region_mask,
         )
 
-        response = np.full((2, 1024, 1024), np.nan, dtype=np.float32)
-        response[:, 20:1000, 100:900] = 100.0
+        flat = np.full((2, 1024, 1024), np.nan, dtype=np.float32)
+        flat[:, 20:1000, 100:900] = 1.0
         bg_stack = self._uniform_bg()
         master_bg = bg_stack.mean(axis=0)
         master_bg[dead_region_mask()] = np.nan
 
-        bpm = build_bad_pixel_map(response, bg_stack, master_bg, self._cfg(), logger=MagicMock())
[…]
