### 2026-07-18 cfbdd66
docs(irdis): CHANGELOG entry for Phase 3 master-calibration step
