### 2026-07-18 80def38
feat(irdis): nominal per-filter star positions + coarse localization helper
