+++ src/spherical/pipeline/steps/irdis_preprocess.py
@@ -1,0 +1,126 @@
+"""VLT/SPHERE IRDIS science preprocess step (Phase 4).
+
+Replaces the IFS pipeline's ``extract_cubes`` + ``bundle_output`` combo for
+IRDIS. Consumes raw CORO/CENTER/FLUX frames plus the Phase 3 master
+calibration products (background, flat, bpm) and writes the ``converted/``
+folder in a layout byte-compatible with the IFS pipeline so downstream
+shared steps and TRAP work without further generalization.
+
+Pixel-validity convention (charis-compatible; see design spec):
+
+- Dead detector regions → NaN in data, ``ivar = 0``, ``False`` in the
+  propagated bad-pixel map (never interpolated).
+- Bad pixels flagged by calibration → interpolated in data, ``ivar = 0``.
+- Real measurements → finite data, ``ivar > 0``.
+
+The step is gated by :func:`spherical.pipeline.step_registry.should_run`
+against the eight output files listed in ``IRDIS_STEP_REGISTRY``; it is not
+internal-guard.
+"""
+from __future__ import annotations
+
+import numpy as np
+
+from spherical.pipeline.transmission import wavelength_bandwidth_filter
+
+# --- per-filter nominal star positions -------------------------------------
+# Per-half (x, y) coordinates. Values verbatim from
+# ``simplified_IRDIS_reduction.py``'s K/H band guesses (measured empirically
+# across archive datasets). A robust matched-filter refinement is deferred to
+# Phase 5 (find_centers generalization).
+
+NOMINAL_STAR_POSITIONS_H_BAND: tuple[tuple[float, float], tuple[float, float]] = (
+    (485.81, 523.54),
+    (487.95, 514.36),
+)
+NOMINAL_STAR_POSITIONS_K_BAND: tuple[tuple[float, float], tuple[float, float]] = (
+    (480.0, 524.7),
+    (482.5, 511.4),
+)
+
+
+def nominal_star_positions(filter_comb: str) -> np.ndarray:
+    """Return per-channel nominal ``(x, y)`` star positions for a filter.
+
+    Dispatches K-band vs. H-band by peak wavelength (``> 2000 nm`` → K),
+    matching ``simplified_IRDIS_reduction.py``'s convention. Fallback used
+    when a coarse-localization search fails; also used as the search-window
+    center in :func:`coarse_star_position`.
+
+    Parameters
+    ----------
+    filter_comb : str
+        IRDIS filter combination string (e.g. ``"DB_K12"``, ``"DB_H23"``,
+        ``"BB_H"``). Passed to
+        :func:`spherical.pipeline.transmission.wavelength_bandwidth_filter`.
+
+    Returns
+    -------
+    np.ndarray
+        Shape ``(2, 2)`` — rows are channels, columns are ``(x, y)`` in
+        per-half detector coordinates.
+    """
+    wavelengths, _ = wavelength_bandwidth_filter(filter_comb)
+    if np.max(np.asarray(wavelengths, dtype=float)) > 2000.0:
+        return np.array(NOMINAL_STAR_POSITIONS_K_BAND, dtype=np.float64)
+    return np.array(NOMINAL_STAR_POSITIONS_H_BAND, dtype=np.float64)
+
+
+def coarse_star_position(
+    image: np.ndarray,
+    nominal_xy: tuple[float, float],
+    search_radius: int = 100,
+) -> tuple[float, float]:
+    """Locate the brightest peak in a window around ``nominal_xy``.
+
+    Searches inside a square window of half-side ``search_radius`` around
+    ``nominal_xy``; if no peak reaches 5σ above the local ``mad_std`` the
[…]
