### 2026-07-18 8907c6c
fix(irdis): warn when coarse_star_position falls back to nominal
