### 2026-07-18 5d355a2
feat(irdis): scaled-background fit helpers (mask + robust scalar least-squares)
