+++ tests/test_irdis_preprocess.py
@@ -4,11 +4,15 @@ from __future__ import annotations
 import numpy as np
 import pytest
 
+from spherical.pipeline.steps.irdis_calibration import dead_region_mask
 from spherical.pipeline.steps.irdis_preprocess import (
     NOMINAL_STAR_POSITIONS_H_BAND,
     NOMINAL_STAR_POSITIONS_K_BAND,
+    background_region_mask,
     coarse_star_position,
+    fit_background_scale,
     nominal_star_positions,
+    subtract_scaled_background,
 )
 
 
@@ -77,3 +81,104 @@ class TestCoarseStarPosition:
         cx, cy = coarse_star_position(img, nominal_xy=(500.0, 500.0), search_radius=30)
         assert abs(cx - 505.0) < 1.5
         assert abs(cy - 507.0) < 1.5
+
+
+class TestBackgroundRegionMask:
+    def test_shape_and_dtype(self):
+        dm = dead_region_mask()
+        bpm = np.zeros((1024, 1024), dtype=bool)
+        mask = background_region_mask(
+            shape=(1024, 1024),
+            star_xy=(500.0, 500.0),
+            star_mask_radius=40,
+            dead_mask_ch=dm[0],
+            bpm_ch=bpm,
+        )
+        assert mask.shape == (1024, 1024)
+        assert mask.dtype == np.bool_
+
+    def test_star_region_excluded(self):
+        dm = np.zeros((1024, 1024), dtype=bool)
+        bpm = np.zeros((1024, 1024), dtype=bool)
+        mask = background_region_mask(
+            shape=(1024, 1024),
+            star_xy=(500.0, 500.0),
+            star_mask_radius=40,
+            dead_mask_ch=dm,
+            bpm_ch=bpm,
+        )
+        assert not mask[500, 500]
+        assert not mask[500, 460]  # on the circle boundary
+        assert mask[500, 600]  # outside
+
+    def test_dead_and_bad_excluded(self):
+        dm = dead_region_mask()
+        bpm = np.zeros((1024, 1024), dtype=bool)
+        bpm[300, 300] = True
+        mask = background_region_mask(
+            shape=(1024, 1024),
+            star_xy=(500.0, 500.0),
+            star_mask_radius=40,
+            dead_mask_ch=dm[0],
+            bpm_ch=bpm,
+        )
+        assert not mask[0, 500]  # dead row band
+        assert not mask[300, 300]  # bad pixel
+
+
+class TestFitBackgroundScale:
+    def test_recovers_known_scale_without_noise(self):
+        rng = np.random.default_rng(0)
+        bg = rng.uniform(50.0, 200.0, size=(200, 200)).astype(np.float32)
+        s_true = 1.35
+        frame = s_true * bg
+        mask = np.ones_like(bg, dtype=bool)
+        s = fit_background_scale(frame, bg, mask)
+        assert abs(s - s_true) < 1e-4
+
+    def test_ignores_masked_pixels(self):
+        bg = np.full((100, 100), 100.0, dtype=np.float32)
+        frame = 2.0 * bg
+        frame[50, 50] = 1e6  # would blow up an unweighted fit
+        mask = np.ones_like(bg, dtype=bool)
+        mask[50, 50] = False
+        s = fit_background_scale(frame, bg, mask)
+        assert abs(s - 2.0) < 1e-4
+
+    def test_sigma_clips_outliers(self):
+        rng = np.random.default_rng(1)
+        bg = rng.uniform(50.0, 200.0, size=(200, 200)).astype(np.float32)
+        s_true = 1.35
+        frame = s_true * bg + rng.standard_normal(bg.shape).astype(np.float32) * 0.5
[…]
