### 2026-07-18 42fcdb8
feat(irdis): analytic inverse-variance builder (photon + read noise, flat-propagated)
