### 2026-07-18 8515225
feat(irdis): NaN-safe bad-pixel replacement + dead-region-aware sigma clip
