### 2026-07-18 9d6d529
fix(irdis): only pass real bad pixels as fix_badpix targets, correct misleading comment
The neighbor-exclusion bpm previously unioned in the entire dead-region
mask, so imutils.fix_badpix's Python target loop reprocessed ~15% of
the frame as targets on every call, even though the result was always
immediately overwritten with NaN for those pixels. Restrict the bpm
argument to real bad pixels (plus residual NaN outside the dead
region); the unconditional post-call NaN reset still enforces the
dead-region invariant regardless of what imutils computes for
dead-region "good neighbor" pixels. Also 
[…]
