### 2026-07-18 c2b38fe
feat(irdis): anamorphism-correction and crop helpers with offset recording
