### 2026-07-18 2b6583c
feat(irdis): per-frame-type preprocess pipeline (bg + flat + ivar + fix_badpix + crop)
