+++ src/spherical/pipeline/steps/irdis_preprocess.py
@@ -194,8 +194,14 @@ def fit_background_scale(
     for _ in range(max_iter):
         if not mask.any():
             return 0.0
-        num = float((mask * frame_ch * bg_ch).sum())
-        den = float((mask * bg_ch * bg_ch).sum())
+        # `np.where` selects 0.0 at excluded pixels before the sum; a plain
+        # `mask * frame_ch * bg_ch` would multiply first, and `0 * NaN = NaN`
+        # leaks any NaN in frame_ch/bg_ch outside the mask (e.g. dead-region
+        # background pixels) into the sum despite being masked out.
+        frame_masked = np.where(mask, frame_ch, 0.0)
+        bg_masked = np.where(mask, bg_ch, 0.0)
+        num = float((frame_masked * bg_masked).sum())
+        den = float((bg_masked * bg_masked).sum())
         if den <= 0.0:
             return 0.0
         s_new = num / den
@@ -462,3 +468,176 @@ def apply_crop(
         cube_c = cube_ch[y0:y0 + crop_size, x0:x0 + crop_size]
         ivar_c = ivar_ch[y0:y0 + crop_size, x0:x0 + crop_size]
     return cube_c, ivar_c, (x0, y0)
+
+
+def _load_raw_frames_expanded(paths: list[str]) -> np.ndarray:
+    """Read raw IRDIS files and return a ``(N_total, 1024, 2048)`` cube.
+
+    Each file may store one 2-D frame or an NDIT × 2-D cube; NDIT frames
+    are unrolled into the leading axis, preserving intra-file order.
+    """
+    from astropy.io import fits
+
+    chunks = []
+    for path in paths:
+        data = np.asarray(fits.getdata(path), dtype=np.float32)
+        if data.ndim == 2:
+            data = data[np.newaxis, ...]
+        elif data.ndim != 3:
+            raise ValueError(
+                f"Unexpected data shape {data.shape} in {path}; expected 2-D or 3-D."
+            )
+        chunks.append(data)
+    return np.concatenate(chunks, axis=0)
+
+
+def preprocess_frame_type(
+    frame_paths: list[str],
+    master_flat: np.ndarray,
+    master_background: np.ndarray,
+    bpm: np.ndarray,
+    star_positions_xy: np.ndarray,
+    filter_comb: str,
+    is_flux: bool,
+    preprocess_config,
+    logger,
+) -> tuple[np.ndarray, np.ndarray, np.ndarray, np.ndarray | None]:
+    """Full spec §4 preprocess pipeline for one frame type of one observation.
+
+    Runs, per frame per channel:
+        1. Split into halves (already done in-bulk).
+        2. Scaled background subtraction (per-frame scalar fit on star-masked region).
+        3. Flat division.
+        4. Analytic inverse variance (with NaN → 0).
+        5. Dead-region NaN in data + 0 in ivar.
+        6. NaN-safe bad-pixel replacement + ivar=0 at replaced pixels.
+        7. Non-FLUX only: transient sigma-clip, ivar=0 at flagged pixels.
+        8. Optional anamorphism correction (default off).
+        9. Optional crop around the star (default off).
+
+    Returns
+    -------
+    cube : np.ndarray
+        ``(2, n_time, ny, nx)`` float32 — n_wave-first ordering matching
+        the IFS bundle-output convention.
+    ivar : np.ndarray
+        Same shape as ``cube``, float32.
+    bpm_out : np.ndarray
+        ``(2, 1024, 1024)`` uint8 — Phase 3 bpm unioned with per-observation
+        transient bad pixels; never ``True`` in dead regions. Stays at
+        native detector resolution.
+    crop_offsets : np.ndarray or None
+        ``(2, 2)`` int array of ``[[x0_ch0, y0_ch0], [x0_ch1, y0_ch1]]``
[…]
