### 2026-07-18 e1b940a
fix(irdis): preserve un-interpolated data when transient-clip exceeds 1% threshold
