### 2026-07-19 09e2ae8
feat(irdis): wire run_irdis_preprocess into execute_irdis_target (should_run-gated)
