### 2026-07-19 1c283fe
fix(irdis): check_output looks under converted/ subdir; drop dead crop_offsets_by_type
