### 2026-07-19 a80464d
docs(irdis): CHANGELOG entry for Phase 4 preprocess step
