+++ CHANGELOG.md
@@ -11,6 +11,7 @@ This project follows [Semantic Versioning](https://semver.org/) and the [Keep a 
 ### ✨ Added
 - **IRDIS reduction — Phase 1 download-only skeleton** – `execute_targets()` now dispatches by instrument: IRDIS observations route to a new `execute_irdis_target()` in `spherical.pipeline.irdis_reduction` that wires the shared `download_data` step, avoiding the IFS orchestrator's unconditional WAVECAL access that crashes for IRDIS. Adds `IRDISReductionConfig` / `defaultIRDISReduction()` (composite dataclass reusing `DirectoryConfig`, `Resources`, `PreprocConfig`, `PipelineStepsConfig`) and `examples/irdis_reduction_template.py` for pulling a reference DBI dataset. Calibration, preprocessing, TRAP integration, and shared-step generalization land in later phases per the IRDIS design spec ([@m-samland](https://github.com/m-samland)).
 - **IRDIS reduction — Phase 3 master-calibration step** – New `spherical.pipeline.steps.irdis_calibration` module builds per-observation master background (sigma-clipped mean of `BG_SCIENCE`), master flat (per-pixel linear fit of counts vs. DIT when the FLAT sequence has ≥ 2 distinct DITs; background-subtracted median fallback), and bad-pixel map (union of abnormal flat response, hot pixels in the master background, and temporally noisy pixels). All products are stored in split `(2, 1024, 1024)` layout. Dead detector regions live as named module constants (`DEAD_ROW_SLICE_TOP` / `DEAD_ROW_SLICE_BOTTOM` / `DEAD_COL_SLICE_*`) and are NaN in the master data, `False` in the bpm (charis convention — never interpolated). A block-median + cubic-spline `estimate_smooth_illumination` helper removes the integrating-sphere illumination gradient before the flat is normalized and is reused to subtract the smooth thermal profile before the hot-pixel test — otherwise vignetting causes legitimate edge pixels to appear as sigma-outliers. `execute_irdis_target` now hydrates FLAT / BG_SCIENCE file paths via `update_observation_file_paths` and invokes `run_irdis_calibration` (internal-guard: skips when the three outputs already exist) with cascade support through `_forced` on `IRDIS_STEP_ORDER`. Outputs land in `{reduction_dir}/IRDIS/calibration/{filter}/{date}/` ([@m-samland](https://github.com/m-samland)).
[…]
