### 2026-07-19 d74307a
fix(irdis): ivar header + zero ivar in transient warning branch + cleanups
