### 2026-07-19 36948de
fix(irdis): calibrate star-mask radius per frame type (285 CORO/CENTER, 150 FLUX)
