### 2026-07-19 55b2fda
perf(irdis): vectorize fix_badpix_nan_safe (two-pass K-nearest, 5x faster; drops dead-band bias)
