+++ src/spherical/pipeline/steps/irdis_preprocess.py
@@ -283,6 +283,65 @@ def analytic_ivar(
     return ivar.astype(np.float32)
 
 
+def _fix_badpix_vec_one_pass(
+    frame: np.ndarray,
+    target_bpm: np.ndarray,
+    exclusion_mask: np.ndarray,
+    dmax: int,
+    npix: int,
+) -> tuple[np.ndarray, np.ndarray]:
+    """One vectorized pass of the K-nearest-good-neighbor fixer.
+
+    Returns ``(fixed, updated)`` where ``updated`` marks the target pixels
+    that received a new value. Targets with fewer than ``npix`` good
+    neighbors inside a ``(2·dmax+1)²`` window are left at their original
+    value and reported ``False`` in ``updated`` — the caller can drive a
+    second pass over them with a wider window.
+    """
+    h, w = frame.shape
+    win = 2 * dmax + 1
+
+    dy, dx = np.mgrid[-dmax:dmax + 1, -dmax:dmax + 1]
+    dist_win = np.sqrt(dx.astype(np.float32) ** 2 + dy.astype(np.float32) ** 2).ravel()
+    dist_win[dist_win == 0] = 1e-6  # avoid /0 at the target (excluded anyway)
+
+    bp_y, bp_x = np.where(target_bpm)
+    n_bad = bp_y.size
+    updated = np.zeros_like(target_bpm)
+    if n_bad == 0:
+        return frame.copy(), updated
+
+    padded_frame = np.pad(frame, dmax, mode="constant", constant_values=0.0)
+    padded_exc = np.pad(exclusion_mask, dmax, mode="constant", constant_values=True)
+
+    yi = bp_y[:, None, None] + np.arange(win)[None, :, None]
+    xi = bp_x[:, None, None] + np.arange(win)[None, None, :]
+    windows_vals = padded_frame[yi, xi].reshape(n_bad, -1)
+    windows_exc = padded_exc[yi, xi].reshape(n_bad, -1)
+
+    LARGE = np.float32(1e10)
+    dists_masked = np.where(windows_exc, LARGE, dist_win)
+
+    part = np.argpartition(dists_masked, npix, axis=1)[:, :npix]
+    top_vals = np.take_along_axis(windows_vals, part, axis=1)
+    top_dists = np.take_along_axis(dists_masked, part, axis=1)
+    enough = (top_dists < LARGE).all(axis=1)
+
+    val_order = np.argsort(top_vals, axis=1)
+    sorted_vals = np.take_along_axis(top_vals, val_order, axis=1)
+    sorted_dists = np.take_along_axis(top_dists, val_order, axis=1)
+
+    mid_vals = sorted_vals[:, 1:-1]
+    mid_dists = sorted_dists[:, 1:-1]
+    weights = np.float32(1.0) / mid_dists
+    new_vals = (mid_vals * weights).sum(axis=1) / weights.sum(axis=1)
+
+    result = frame.copy()
+    result[bp_y[enough], bp_x[enough]] = new_vals[enough].astype(frame.dtype)
+    updated[bp_y[enough], bp_x[enough]] = True
+    return result, updated
+
+
 def fix_badpix_nan_safe(
     frame_ch: np.ndarray,
     bpm_ch: np.ndarray,
@@ -291,14 +350,21 @@ def fix_badpix_nan_safe(
 ) -> np.ndarray:
     """Interpolate bad pixels without pulling NaN from dead-region neighbors.
 
-    Wraps :func:`spherical.pipeline.imutils.fix_badpix`. Preprocess substitutes
-    NaN pixels with ``0.0`` so the vendored routine's arithmetic mean stays
-    finite; the ``bpm`` argument passed to it is restricted to real bad
-    pixels (plus any residual NaN outside the dead region) so the target
[…]
