### 2026-07-19 e242029
perf(irdis): parallelize preprocess_frame_type via ProcessPoolExecutor (ncpu_preprocess)
