+++ src/spherical/pipeline/steps/irdis_preprocess.py
@@ -554,6 +554,133 @@ def _load_raw_frames_expanded(paths: list[str]) -> np.ndarray:
     return np.concatenate(chunks, axis=0)
 
 
+# Module-level worker state for the ProcessPoolExecutor path. Populated by
+# ``_worker_init``; consumed by ``_process_chunk``. Each worker process gets
+# its own module-level dict on init.
+_WORKER_STATE: dict = {}
+
+
+def _worker_init(
+    master_flat: np.ndarray,
+    master_bg: np.ndarray,
+    bpm_bool: np.ndarray,
+    star_positions_xy: np.ndarray,
+    mask_radius: int,
+    is_flux: bool,
+    gain: float,
+    read_noise: float,
+    fix_badpix: bool,
+    warn_threshold: int,
+) -> None:
+    """Initializer for both ProcessPoolExecutor workers and the serial path.
+
+    Every call fully overwrites ``_WORKER_STATE`` so back-to-back invocations
+    from ``run_irdis_preprocess`` (one per frame type) don't leak state.
+    """
+    from spherical.pipeline.steps.irdis_calibration import dead_region_mask
+
+    _WORKER_STATE.clear()
+    _WORKER_STATE.update(
+        master_flat=master_flat,
+        master_bg=master_bg,
+        bpm_bool=bpm_bool,
+        dead=dead_region_mask(),
+        star_positions=star_positions_xy,
+        mask_radius=mask_radius,
+        is_flux=is_flux,
+        gain=gain,
+        read_noise=read_noise,
+        fix_badpix=fix_badpix,
+        warn_threshold=warn_threshold,
+    )
+
+
+def _process_chunk(
+    chunk_data: tuple[list[int], np.ndarray],
+) -> tuple[np.ndarray, np.ndarray, np.ndarray, list[tuple[int, int, int]]]:
+    """Process one chunk of frames. Runs in worker or main process.
+
+    Parameters
+    ----------
+    chunk_data : (frame_indices, split_slice)
+        ``frame_indices`` is a list of global frame indices for warning
+        messages; ``split_slice`` is ``(n_chunk, 2, 1024, 1024)``.
+
+    Returns
+    -------
+    cube_slice : np.ndarray
+        ``(2, n_chunk, 1024, 1024)`` float32.
+    ivar_slice : np.ndarray
+        Same shape.
+    transient_local : np.ndarray
+        ``(2, 1024, 1024)`` bool — union of transient masks seen in this chunk.
+    warnings : list of (frame_idx, ch, n_transient)
+        Frames whose transient count exceeded ``warn_threshold``. Returned to
+        the main process for logging.
+    """
+    frame_indices, chunk_split = chunk_data
+    n = len(frame_indices)
+
+    mf = _WORKER_STATE["master_flat"]
+    mb = _WORKER_STATE["master_bg"]
+    bpm_bool = _WORKER_STATE["bpm_bool"]
+    dead = _WORKER_STATE["dead"]
+    star_positions = _WORKER_STATE["star_positions"]
+    mask_radius = _WORKER_STATE["mask_radius"]
+    is_flux = _WORKER_STATE["is_flux"]
+    gain = _WORKER_STATE["gain"]
+    read_noise = _WORKER_STATE["read_noise"]
+    do_fix_badpix = _WORKER_STATE["fix_badpix"]
+    warn_threshold = _WORKER_STATE["warn_threshold"]
+
+    cube_slice = np.empty((2, n, 1024, 1024), dtype=np.float32)
+    ivar_slice = np.empty((2, n, 1024, 1024), dtype=np.float32)
+    transient_local = np.zeros((2, 1024, 1024), dtype=bool)
[…]
