### 2026-07-19 7e45b3e
refactor(irdis): drop NaN->0 substitution in fix_badpix_nan_safe (redundant for vec, keeps intent explicit)
