### 2026-07-19 7c022a1
fix(irdis): move template setup into main() so spawn workers don't re-execute it
