+++ examples/irdis_reduction_template.py
@@ -1,10 +1,54 @@
-"""IRDIS reduction — Phase 1 driver (download-only).
+"""IRDIS reduction driver — Phases 1 + 3 + 4.
 
-Downloads the raw science + calibration files (``CORO``, ``CENTER``, ``FLUX``,
-``FLAT``, ``BG_SCIENCE``) for the selected IRDIS observations from the ESO
-archive. Nothing else runs yet: calibration + preprocessing are wired in later
-phases. Use this template to pull the reference dataset that subsequent phases
-will validate against.
+Runs the IRDIS pipeline up to and including preprocessing. What each phase
+produces:
+
+- **Phase 1 — download** (``steps.download_data``): pulls raw ``CORO``,
+  ``CENTER``, ``FLUX``, ``FLAT``, and ``BG_SCIENCE`` FITS files for each
+  observation from the ESO archive into ``raw_directory``. Idempotent — files
+  already present locally are skipped, so re-runs are cheap.
+- **Phase 3 — master calibration** (``steps.irdis_calibration``): from the
+  ``FLAT`` and ``BG_SCIENCE`` frames, builds ``master_flat.fits``,
+  ``master_background.fits``, and ``badpixel_map.fits`` under
+  ``{reduction_directory}/IRDIS/calibration/{filter}/{date}/``. Internal-guard
+  step — skips itself when all three outputs already exist.
+- **Phase 4 — science preprocess** (``steps.preprocess_irdis``): per
+  observation, per frame type (CORO/CENTER/FLUX), scaled-background
+  subtraction + flat division + analytic ivar + NaN-safe bad-pixel replacement
+  (+ transient sigma-clip on non-FLUX). Writes 8 files under
+  ``{reduction_directory}/IRDIS/observation/{target}/{filter}/{date}/converted/``:
+  ``{coro,center,flux}_cube.fits``, matching ``*_ivar_cube.fits``,
+  ``wavelengths.fits``, and ``badpixel_map.fits``. Gated by ``should_run`` —
+  skips when all 8 outputs already exist. Cube axis order matches the IFS
+  pipeline: ``(n_wave=2, n_time, ny, nx)``.
+
+Phases 5 (shared-step generalization: find_centers, process_extracted_centers,
+spot_to_flux) and 6 (TRAP integration) are not wired yet and remain disabled
+below.
+
+Inspecting results after a run:
+
+    ls -1 ~/data/sphere/reduction/IRDIS/calibration/*/*/                     # Phase 3
+    ls -1 ~/data/sphere/reduction/IRDIS/observation/*/*/*/converted/         # Phase 4
+
+Quick FITS inspection:
+
+    python -c "
+    from astropy.io import fits
+    import numpy as np
+    from pathlib import Path
+    for p in sorted(Path.home().glob('data/sphere/reduction/IRDIS/observation/*/*/*/converted/*_cube.fits')):
+        data = fits.getdata(p)
+        print(f'{p.name}: shape={data.shape}, dtype={data.dtype}, med={np.nanmedian(data):.3g}')
+    "
+
+For a persistent validation report with per-channel PNG previews, see the
+plan's Task 7 Step 4 snippet in
+``docs/superpowers/plans/2026-07-18-irdis-reduction-phase4-preprocess-step.md``.
+
+Force recompute (e.g., after editing a step) via
+``config.steps.force = {"preprocess_irdis"}`` — force cascades to every
+downstream step in ``IRDIS_STEP_ORDER``.
 """
 from pathlib import Path
 
@@ -14,64 +58,80 @@ from spherical.database.sphere_database import SphereDatabase
 from spherical.pipeline.ifs_reduction import execute_targets
[…]
