### 2026-07-19 b72862f
fix(irdis): drop transient union from converted/badpixel_map.fits; default 5sigma; log per-type summary
