+++ tests/test_irdis_preprocess.py
@@ -531,17 +531,15 @@ class TestPreprocessFrameType:
         """When >1% of pixels are flagged transient, data is NOT interpolated
         (per the transient-warning branch's log message) but ivar IS still
         zeroed at every flagged pixel — matches the 'no measurement' semantic
-        of ivar=0 in the charis convention."""
+        of ivar=0 in the charis convention.
+
+        Uses ``transient_nsigma=4.0`` to keep the same synthetic geometry
+        (spike grid spacing 8 > box=7) known to yield ~13k flagged pixels;
+        the production default is 5.0.
+        """
         flat, bg, bpm = self._make_calibration()
         from astropy.io import fits
         raw = np.full((1024, 2048), 100.0, dtype=np.float32)
-        # Scattered single-pixel spikes on a grid (spacing 8, > box=7) so
-        # each one is an outlier relative to its own local neighborhood. A
-        # contiguous block of outliers does NOT work here: interior pixels
-        # of a block see only other outliers as neighbors, so the box
-        # filter's leave-one-out mean matches and they are never flagged.
-        # This grid yields ~13300 flagged pixels, comfortably above
-        # warn_threshold = int(0.01 * 1024 * 1024) = 10485.
         ys = np.arange(50, 1000, 8)
         xs = np.arange(50, 1000, 8)
         yy, xx = np.meshgrid(ys, xs, indexing="ij")
@@ -550,9 +548,9 @@ class TestPreprocessFrameType:
         fits.writeto(path, raw[np.newaxis, ...])
 
         star_positions = np.array([[500.0, 500.0], [500.0, 500.0]])
-        cfg = IRDISPreprocessConfig()
+        cfg = IRDISPreprocessConfig(transient_nsigma=4.0)
         logger = MagicMock()
-        cube, ivar, bpm_out, _ = preprocess_frame_type(
+        cube, ivar, _, _ = preprocess_frame_type(
             [str(path)], flat, bg, bpm, star_positions,
             is_flux=False,
             preprocess_config=cfg, logger=logger,
@@ -563,20 +561,94 @@ class TestPreprocessFrameType:
         ]
         assert warning_calls, "expected the >1% transient-warning branch to fire"
 
-        warn_threshold = int(0.01 * 1024 * 1024)
-        # bpm passed in is all-False, so bpm_out[0] is exactly the
-        # transient mask (unioned with an empty bpm, dead regions excluded).
-        transient_mask = bpm_out[0].astype(bool)
-        assert transient_mask.sum() > warn_threshold
-
-        # Every transient-flagged pixel is zeroed in ivar (marked as no
-        # measurement), even though the >1% branch fired and skipped
-        # interpolation.
-        assert (ivar[0, 0][transient_mask] == 0.0).all(), (
-            "expected ivar=0 at every transient-flagged pixel even when >1% branch fired"
+        # Every injected spike position is at ivar=0 because it tripped the
+        # sigma-clip AND the warning branch left the value un-interpolated.
+        # No dependence on bpm_out — transient hits are recorded per-frame in
+        # ivar, not in the persistent bpm.
+        assert (ivar[0, 0, yy, xx] == 0.0).all(), (
+            "expected ivar=0 at every injected transient-flagged spike, "
+            "even when >1% branch fired"
         )
         # A clean pixel away from any spike keeps a real (nonzero) ivar.
         assert ivar[0, 0, 500, 700] > 0.0, "expected ivar>0 at a clean pixel"
+
+    def test_bpm_out_is_calibration_bpm_only(self, tmp_path):
+        """``bpm_out`` echoes the Phase 3 detector bpm dead-region-clamped;
[…]
