+++ src/spherical/pipeline/pipeline_config.py
@@ -358,6 +358,12 @@ class IRDISPreprocessConfig:
     # frames use a smaller radius because the PSF is compact off the coronagraph.
     star_mask_radius: int = 285
     flux_star_mask_radius: int = 150
+    # Per-frame transient sigma-clip threshold (imutils.sigma_filter box=7).
+    # Conservative default: at 5σ the Gaussian tail rate is ~5e-7 per pixel, so a
+    # 1024x1024 half yields ~0.5 statistical false positives/frame; hits above
+    # that are almost certainly cosmic rays or true transient bad pixels. 4σ is
+    # much noisier (~60 false positives/frame). Non-FLUX only; skipped when 0.
+    transient_nsigma: float = 5.0
 
     def merge(self, **kw) -> "IRDISPreprocessConfig":
         return replace(self, **kw)
