+++ src/spherical/pipeline/steps/irdis_preprocess.py
@@ -577,6 +577,7 @@ def _worker_init(
     read_noise: float,
     fix_badpix: bool,
     warn_threshold: int,
+    transient_nsigma: float,
 ) -> None:
     """Initializer for both ProcessPoolExecutor workers and the serial path.
 
@@ -598,19 +599,20 @@ def _worker_init(
         read_noise=read_noise,
         fix_badpix=fix_badpix,
         warn_threshold=warn_threshold,
+        transient_nsigma=transient_nsigma,
     )
 
 
 def _process_chunk(
     chunk_data: tuple[list[int], np.ndarray],
-) -> tuple[np.ndarray, np.ndarray, np.ndarray, list[tuple[int, int, int]]]:
+) -> tuple[np.ndarray, np.ndarray, list[tuple[int, int, int]], list[tuple[int, int, int]]]:
     """Process one chunk of frames. Runs in worker or main process.
 
     Parameters
     ----------
     chunk_data : (frame_indices, split_slice)
-        ``frame_indices`` is a list of global frame indices for warning
-        messages; ``split_slice`` is ``(n_chunk, 2, 1024, 1024)``.
+        ``frame_indices`` is a list of global frame indices for logging;
+        ``split_slice`` is ``(n_chunk, 2, 1024, 1024)``.
 
     Returns
     -------
@@ -618,11 +620,14 @@ def _process_chunk(
         ``(2, n_chunk, 1024, 1024)`` float32.
     ivar_slice : np.ndarray
         Same shape.
-    transient_local : np.ndarray
-        ``(2, 1024, 1024)`` bool — union of transient masks seen in this chunk.
+    transient_counts : list of (frame_idx, ch, n_transient)
+        Per (frame, channel) transient sigma-clip hit counts. Empty when
+        ``is_flux`` (no transient clip runs) or when ``transient_nsigma <= 0``.
+        Consumed by the main process to log a summary; not persisted.
     warnings : list of (frame_idx, ch, n_transient)
-        Frames whose transient count exceeded ``warn_threshold``. Returned to
-        the main process for logging.
+        Subset of ``transient_counts`` where the per-frame count exceeded
+        ``warn_threshold``; separately surfaced so the main process can log
+        one WARNING per excess.
     """
     frame_indices, chunk_split = chunk_data
     n = len(frame_indices)
@@ -638,11 +643,13 @@ def _process_chunk(
     read_noise = _WORKER_STATE["read_noise"]
     do_fix_badpix = _WORKER_STATE["fix_badpix"]
     warn_threshold = _WORKER_STATE["warn_threshold"]
+    transient_nsigma = _WORKER_STATE["transient_nsigma"]
 
     cube_slice = np.empty((2, n, 1024, 1024), dtype=np.float32)
     ivar_slice = np.empty((2, n, 1024, 1024), dtype=np.float32)
-    transient_local = np.zeros((2, 1024, 1024), dtype=bool)
+    transient_counts: list[tuple[int, int, int]] = []
     warnings: list[tuple[int, int, int]] = []
+    do_transient = (not is_flux) and (transient_nsigma > 0)
 
     for i, frame_idx in enumerate(frame_indices):
         for ch in range(2):
@@ -671,11 +678,13 @@ def _process_chunk(
                 divided = fix_badpix_nan_safe(divided, bpm_bool[ch], dead[ch])
                 ivar[bpm_bool[ch]] = 0.0
 
-            if not is_flux:
-                cleaned, transient_mask = sigma_filter_ignore_dead(divided, dead[ch])
+            if do_transient:
+                cleaned, transient_mask = sigma_filter_ignore_dead(
+                    divided, dead[ch], nsigma=transient_nsigma,
+                )
[…]
