### 2026-07-19 d99b5a4
perf(irdis): disable transient sigma-clip by default (~25%% wall time; not scientifically useful on real data)
