### 2026-07-19 9441ec3
docs(irdis): document irdis_preprocessing knobs in the reduction template
