### 2026-07-19 fc735e7
feat(irdis): instrument-thread fit_centers + nominal seed + fix instrument-dispatch else-raise
