+++ tests/test_find_star_irdis.py
@@ -1,0 +1,139 @@
+"""Tests for the IRDIS branch of the shared find_centers step (Phase 5)."""
+from __future__ import annotations
+
+from unittest.mock import MagicMock, patch
+
+import numpy as np
+import pandas as pd
+import pytest
+from astropy.io import fits
+
+
+def _make_center_cube_file(tmp_path, shape=(2, 4, 60, 60), apply_crop=False, crop_offset=(0, 0)):
+    cube = np.zeros(shape, dtype=np.float32)
+    header = fits.Header()
+    header["HIERARCH SPHERICAL CROP APPLIED"] = apply_crop
+    if apply_crop:
+        header["HIERARCH SPHERICAL CROP SIZE"] = shape[-1]
+        header["HIERARCH SPHERICAL CROP X0 CH0"] = int(crop_offset[0])
+        header["HIERARCH SPHERICAL CROP Y0 CH0"] = int(crop_offset[1])
+        header["HIERARCH SPHERICAL CROP X0 CH1"] = int(crop_offset[0])
+        header["HIERARCH SPHERICAL CROP Y0 CH1"] = int(crop_offset[1])
+    path = tmp_path / "center_cube.fits"
+    fits.writeto(path, cube, header, overwrite=True)
+    return path
+
+
+class TestFitCentersInParallelInstrumentDispatch:
+    def test_ifs_path_passes_ifs_and_none_guess(self, tmp_path):
+        from spherical.pipeline.steps import find_star
+
+        _make_center_cube_file(tmp_path, shape=(39, 4, 60, 60))
+        fits.writeto(tmp_path / "wavelengths.fits", np.linspace(1000, 1600, 39), overwrite=True)
+        pd.DataFrame({
+            "OCS WAFFLE ORIENT": ["+"] * 4,
+            "INS COMB IFLT": ["OBS_YJ"] * 4,
+        }).to_csv(tmp_path / "frames_info_center.csv", index=False)
+
+        observation = MagicMock()
+        observation.observation = {"INSTRUMENT": ["IFS"], "FILTER": ["OBS_YJ"]}
+        observation.frames = {"CORO": None}
+
+        with patch.object(find_star, "parallel_map_ordered", return_value=[]) as pm:
+            try:
+                find_star.fit_centers_in_parallel(str(tmp_path), observation, ncpu=1)
+            except Exception:
+                pass
+            args_list = pm.call_args.kwargs.get("args_list") or pm.call_args.args[1]
+            assert all(a[-2] == "IFS" for a in args_list)
+            assert all(a[-1] is None for a in args_list)
+
+    def test_irdis_path_passes_nominal_center_guess(self, tmp_path):
+        from spherical.pipeline.steps import find_star
+        from spherical.pipeline.steps.irdis_preprocess import NOMINAL_STAR_POSITIONS_K_BAND
+
+        _make_center_cube_file(tmp_path)
+        fits.writeto(tmp_path / "wavelengths.fits", np.array([2110.0, 2251.0]), overwrite=True)
+        pd.DataFrame({
+            "OCS WAFFLE ORIENT": ["+"] * 4,
+            "INS COMB IFLT": ["DB_K12"] * 4,
+        }).to_csv(tmp_path / "frames_info_center.csv", index=False)
+
+        observation = MagicMock()
+        observation.observation = {"INSTRUMENT": ["IRDIS"], "FILTER": ["DB_K12"]}
+        observation.frames = {"CORO": None}
+
+        with patch.object(find_star, "parallel_map_ordered", return_value=[]) as pm:
+            try:
+                find_star.fit_centers_in_parallel(str(tmp_path), observation, ncpu=1)
+            except Exception:
+                pass
[…]
