### 2026-07-19 912fb50
feat(irdis): persist empirical cross-channel offset from fitted centers
