### 2026-07-19 bb78721
feat(irdis): instrument dispatch in process_extracted_centers (waffle-CENTER path)
