+++ src/spherical/pipeline/steps/process_centers.py
@@ -1,22 +1,19 @@
-"""
-Process extracted centers: fits, robust fitting, and output of center positions.
+"""Process extracted centers: instrument-dispatched center fitting.
 
-Parameters
-----------
-converted_dir : str
-    Directory where the output files are stored and written.
-extraction_parameters : dict
-    Extraction parameters dict, must contain 'method' and 'linear_wavelength'.
-non_least_square_methods : list
-    List of method names that are not least-square based.
+For IFS: polynomial-across-wavelength two-pass fit with sigma-clipping. For
+IRDIS (waffle-CENTER path): temporal moving-median outlier flagging + local
+median replacement (2 wavelength points make a polynomial across wavelength
+meaningless). For IRDIS (non-waffle, with CORO): DMS-header offset
+propagation from CENTER waffle measurements (Task 4).
 """
-
 import os
+from pathlib import Path
 from typing import Dict, List
 
 import numpy as np
 from astropy.io import fits
-from astropy.stats import sigma_clip
+from astropy.stats import mad_std, sigma_clip
+from scipy.ndimage import median_filter
 
 from spherical.pipeline.logging_utils import optional_logger
 
@@ -24,147 +21,202 @@ from spherical.pipeline.logging_utils import optional_logger
 @optional_logger
 def run_polynomial_center_fit(
     converted_dir: str,
+    observation,
     extraction_parameters: Dict[str, str | bool],
     non_least_square_methods: List[str],
-    logger
+    logger,
 ) -> None:
-    """
-    Fit polynomial models to star center positions across wavelength.
+    """Fit / smooth star center positions for IFS or IRDIS.
 
-    This is the seventh step in the SPHERE/IFS data reduction pipeline. It fits
-    polynomial models to the star center positions as a function of wavelength,
-    using a two-pass approach with robust outlier rejection. The first pass fits
-    all valid points, and the second pass removes outliers based on sigma clipping.
-
-    Required Input Files
-    -------------------
-    From previous step (find_star):
-    - converted_dir/image_centers.fits
-        Raw star center positions from waffle spot fitting
-    - converted_dir/wavelengths.fits
-        Wavelength array for the data cube
-
-    Generated Output Files
-    ---------------------
-    In converted_dir:
-    - image_centers_fitted.fits
-        First-pass polynomial fits to center positions
-    - image_centers_fitted_robust.fits
-        Second-pass robust polynomial fits after outlier rejection
-    - center_plots/
-        Directory for center fitting visualization plots
+    Dispatches on ``observation.observation['INSTRUMENT'][0]``. IFS behavior
+    is byte-identical to the previous implementation; IRDIS gets a per-channel
+    temporal moving-median outlier flag with local-median replacement.
 
     Parameters
     ----------
     converted_dir : str
-        Directory containing the input files and where outputs will be written.
+        Directory containing ``image_centers.fits`` and ``wavelengths.fits``.
+    observation : IFSObservation | IRDISObservation
+        Observation carrying ``observation`` (metadata Table) and ``frames`` (dict).
     extraction_parameters : dict
-        Extraction parameters dict, must contain:
-        - method: str
-            Extraction method used
-        - linear_wavelength: bool
-            Whether linear wavelength sampling was used
-    non_least_square_methods : list
-        List of method names that are not least-square based.
-    logger : logging.Logger
[…]
