### 2026-07-19 6b3f202
feat(irdis): DMS-offset propagation from CENTER to CORO in process_extracted_centers
