+++ tests/test_process_centers_irdis.py
@@ -102,3 +102,51 @@ class TestIRDISWaffleCenterFit:
         )
         assert (tmp_path / "image_centers_fitted.fits").exists()
         assert (tmp_path / "image_centers_fitted_robust.fits").exists()
+
+
+class TestIRDISDmsPropagation:
+    def test_propagates_nearest_center_with_dms_offset(self, tmp_path):
+        import pandas as pd
+
+        from spherical.pipeline.steps.process_centers import run_polynomial_center_fit
+
+        # 3 CENTER frames at MJD 0.0, 0.5, 1.0.
+        center_centers = np.zeros((2, 3, 2), dtype=np.float32)
+        center_centers[0] = [[25.0, 30.0], [25.0, 30.0], [25.0, 30.0]]
+        center_centers[1] = [[28.0, 27.0], [28.0, 27.0], [28.0, 27.0]]
+        fits.writeto(tmp_path / "image_centers.fits", center_centers, overwrite=True)
+        fits.writeto(tmp_path / "wavelengths.fits", np.array([2100.0, 2250.0]), overwrite=True)
+
+        pd.DataFrame({
+            "MJD_OBS": [0.0, 0.5, 1.0],
+            "INS1 PAC X": [0.0, 0.0, 0.0],
+            "INS1 PAC Y": [0.0, 0.0, 0.0],
+        }).to_csv(tmp_path / "frames_info_center.csv", index=False)
+        # 2 CORO frames at MJD 0.1, 0.9 with DMS offsets in µm.
+        pd.DataFrame({
+            "MJD_OBS": [0.1, 0.9],
+            "INS1 PAC X": [18.0, -36.0],  # 1 px, -2 px in x
+            "INS1 PAC Y": [-18.0, 0.0],   # -1 px, 0 px in y
+        }).to_csv(tmp_path / "frames_info_coro.csv", index=False)
+
+        observation = MagicMock()
+        observation.observation = {"INSTRUMENT": ["IRDIS"]}
+        # A non-empty CORO frames table (content doesn't matter; branch dispatches on truthiness).
+        observation.frames = {"CORO": [1, 2]}
+
+        run_polynomial_center_fit(
+            converted_dir=str(tmp_path),
+            observation=observation,
+            extraction_parameters={"method": "optext", "linear_wavelength": True},
+            non_least_square_methods=["optext"],
+        )
+        robust = fits.getdata(str(tmp_path / "image_centers_fitted_robust.fits"))
+        assert robust.shape == (2, 2, 2)
+        # CORO frame 0 (MJD 0.1) nearest to CENTER 0 (MJD 0.0): PAC delta
+        #   (18 - 0, -18 - 0) µm → (1, -1) px. Propagated center = center - delta.
+        assert abs(robust[0, 0, 0] - (25.0 - 1.0)) < 1e-3
+        assert abs(robust[0, 0, 1] - (30.0 - (-1.0))) < 1e-3
+        # CORO frame 1 (MJD 0.9) nearest to CENTER 2 (MJD 1.0): PAC delta
+        #   (-36 - 0, 0 - 0) µm → (-2, 0) px.
+        assert abs(robust[1, 1, 0] - (28.0 - (-2.0))) < 1e-3
+        assert abs(robust[1, 1, 1] - (27.0 - 0.0)) < 1e-3
