### 2026-07-19 649d987
feat(irdis): symlink center_cube→coro_cube for continuous-waffle observations
