### 2026-07-19 244540a
feat(irdis): wire shared downstream steps into execute_irdis_target
