### 2026-07-19 75cef1e
fix(irdis): flux-PSF calibration falls back to full nwave for 2-channel cubes + IRDIS pixel scale + IRDIS process_centers writes image_centers_fitted.fits
