### 2026-07-19 d25888b
docs(irdis): CHANGELOG entry for Phase 5 shared-steps generalization
