+++ CHANGELOG.md
@@ -9,6 +9,7 @@ This project follows [Semantic Versioning](https://semver.org/) and the [Keep a 
 ## [Unreleased]
 
 ### ✨ Added
+- **IRDIS reduction — Phase 5 shared-steps generalization** – `find_centers`, `process_extracted_centers`, and `spot_to_flux` now dispatch on instrument, and `execute_irdis_target` wires the full shared downstream block (`compute_frames_info`, `cube_header_update`, `find_centers`, `process_extracted_centers`, `plot_image_center_evolution`, `calibrate_spot_photometry`, `calibrate_flux_psf`, `spot_to_flux`). IRDIS seeds the waffle fit with the per-band static nominal (crop-aware from Phase 4 header keys) — no local-peak refinement, since the coronagraph masks the direct starlight and an AO-ring center-of-light probe biased by ~7 px in x on the Beta Pic K12 reference dataset. Also fixes a paired-`if` bug in `measure_center_waffle` that raised `ValueError('Only IRDIS and IFS instruments known.')` for IRDIS. Adds a temporal moving-median outlier flag for continuous-waffle CENTER sequences (`image_centers_fitted_robust.fits` + `additional_outputs/center_outlier_frames.fits`), DMS-header offset propagation from CENTER to CORO for non-waffle IRDIS (`INS1 PAC X/Y ÷ 18 µm/px`), and a per-band `spot_to_flux` normalization range (auto-detected: ≥ 3 channels → IFS 1.0–1.3 µm, 2 channels dispatch by max wavelength → IRDIS H 1.5–1.7 µm / K 2.0–2.3 µm). Continuous-waffle observations (no CORO block) are handled by symlinking `center_cube.fits → coro_cube.fits` (+ ivar) before the shared-step dispatch. First-run empirical cross-channel offset persisted to `converted/additional_outputs/cross_channel_offset.fits`. Concrete-failure fixes: `guess_position_psf` no longer collapses to (0, 0) on 2-channel cubes (empty `cube[1:-1]` slice); `run_flux_psf_calibration` picks pixel scale by channel count (12.25 mas/px for IRDIS DBI, 7.46 for IFS); the IRDIS branch of `run_polynomial_center_fit` writes `image_centers_fitted.fits` as the pre-outlier empirical centers so `plot_image_center_evolution` renders. End-to-end validated on the Beta Pic DB_K12 continuous-waffle reference dataset ([@m-samland](https://github.com/m-samland)).
 - **IRDIS reduction — Phase 1 download-only skeleton** – `execute_targets()` now dispatches by instrument: IRDIS observations route to a new `execute_irdis_target()` in `spherical.pipeline.irdis_reduction` that wires the shared `download_data` step, avoiding the IFS orchestrator's unconditional WAVECAL access that crashes for IRDIS. Adds `IRDISReductionConfig` / `defaultIRDISReduction()` (composite dataclass reusing `DirectoryConfig`, `Resources`, `PreprocConfig`, `PipelineStepsConfig`) and `examples/irdis_reduction_template.py` for pulling a reference DBI dataset. Calibration, preprocessing, TRAP integration, and shared-step generalization land in later phases per the IRDIS design spec ([@m-samland](https://github.com/m-samland)).
[…]
