### 2026-07-19 33c2dd3
refactor(irdis): gate nominal_star_positions K/H to filter-name whitelist; Vigan default fallback
