### 2026-07-19 261a712
feat(irdis): instrument-aware paths + coronagraph selection in run_trap
Detects the observation instrument and dispatches result folder + data
directory + coronagraph transmission table per instrument:

- IFS keeps {reduction}/IFS/trap/{name_mode_date}/ and the packaged
  N_ALC_JYH_S-IFS_YJ table.
- IRDIS lands under {reduction}/IRDIS/trap/{name_mode_date}/ (no
  {method} segment per spec §2) and uses the packaged
  N_ALC_JYH_S-IRDIS_H23 table.

Drops the OBS_YJ/OBS_H assertion; widens type unions to accept either
IFSReductionConfig or IRDISReductionConfig. IRDIS non-waffle observations
pass t
[…]
