+++ src/spherical/pipeline/run_trap.py
@@ -24,7 +24,6 @@ from typing import Optional, Union
 
 import numpy as np
 import pandas as pd
-import ray
 from astropy import units as u
 from astropy.io import fits
 from packaging.version import Version
@@ -34,23 +33,25 @@ from trap.reduction_wrapper import run_complete_reduction
 
 from spherical.database.ifs_observation import IFSObservation
 from spherical.database.irdis_observation import IRDISObservation
-from spherical.pipeline import ifs_reduction
+from spherical.pipeline import ifs_reduction, irdis_reduction
 from spherical.pipeline.logging_utils import (
     PipelineLoggerAdapter,
     get_pipeline_log_context,
     get_pipeline_logger,
     remove_queue_listener,
 )
-from spherical.pipeline.pipeline_config import IFSReductionConfig
+from spherical.pipeline.pipeline_config import IFSReductionConfig, IRDISReductionConfig
 from spherical.pipeline.step_registry import StepDirs, _forced, should_run, validate_force, write_marker
 from spherical.pipeline.toolbox import make_target_folder_string
 
 # The default IFS coronagraph transmission this module injects relies on trap's
 # ``coronagraph_transmission`` reduction parameter, which landed in trap 1.3.1
-# (the dataclass config API it also uses landed in 1.3.0). A git URL dependency
-# cannot carry a PEP 508 version floor, so enforce the minimum here with a clear
-# message instead of a cryptic AttributeError.
-_MIN_TRAP_VERSION = "1.3.1"
+# (the dataclass config API it also uses landed in 1.3.0). Phase 6 additionally
+# requires ``trap_config_for_irdis`` and the IRDIS obs-mode dispatch in
+# ``InstrumentConfig.to_instrument`` — both introduced in 1.3.2.dev. A git URL
+# dependency cannot carry a PEP 508 version floor, so enforce the minimum here
+# with a clear message instead of a cryptic AttributeError.
+_MIN_TRAP_VERSION = "1.3.2.dev0"
 
 
 def _require_trap_version(minimum: str = _MIN_TRAP_VERSION) -> None:
@@ -85,16 +86,67 @@ def _load_coronagraph_transmission(instrument: str) -> np.ndarray:
     return np.loadtxt(path)
 
 
-def _resolve_coronagraph_transmission(reduction_config, trap_reduction_config) -> Optional[np.ndarray]:
+def _instrument_of(observation) -> str:
+    """Return the observation's instrument key (``"IFS"`` or ``"IRDIS"``)."""
+    return str(observation.observation["INSTRUMENT"][0]).upper()
+
+
+def _result_folder_for(
+    instrument: str,
+    reduction_directory: str,
+    name_mode_date: str,
+) -> str:
+    """Return the TRAP result folder for *instrument*.
+
+    Both IFS and IRDIS use ``{reduction_directory}/{instrument}/trap/{name_mode_date}``.
+    No ``{method}`` segment — matches the historical IFS path (which also omits
+    it) and the IRDIS layout in the design spec §2.
+    """
+    return os.path.join(reduction_directory, f"{instrument}/trap", name_mode_date)
+
+
+def _data_directory_for(
+    instrument: str,
+    reduction_config,
+    observation,
+) -> str:
+    """Return the ``converted/`` data directory for *observation*.
+
+    IFS uses :func:`spherical.pipeline.ifs_reduction.output_directory_path`
+    (carries the extraction method segment); IRDIS uses
+    :func:`spherical.pipeline.irdis_reduction.output_directory_path`
+    (no method segment).
+    """
+    if instrument == "IFS":
+        return ifs_reduction.output_directory_path(
+            str(reduction_config.directories.reduction_directory),
+            observation,
[…]
