### 2026-07-19 45bc43c
feat(irdis): enable shared steps + TRAP in irdis_reduction_template
Wires Phase 5 (shared downstream steps) and Phase 6 (TRAP reduction +
detection) into the reference driver. Adds trap_config_for_irdis() and
run_trap_on_observations() imports; enables the eight shared-step flags
plus run_trap_reduction / run_trap_detection; and calls TRAP after
execute_targets, pointing at a species database directory next to the
sphere data root.
