### 2026-07-20 467452e
fix(irdis): expand per-file DIT to per-frame in master-flat linear fit
`_load_frames_as_split_stack` unrolls NDIT cubes, so the flat stack is
`(N_total_frames, 2, 1024, 1024)` while `flat_dits` was one DIT per file.
The (5,)-vs-(200,)-frames broadcast crashed `_pixelwise_linear_slope` on
real IRDIS data where FLAT files are multi-frame cubes; single-frame
test flats didn't trigger it. Expand DITs via `NAXIS3` per file.

Test added for multi-frame flat cubes.

Claude-Session: https://claude.ai/code/session_01UzZTnuhwbv4UVBx1FxwjTd
