### 2026-07-20 0640262
fix: harden raw-file resolution against .fits.Z remnants and missing files
Three coupled bugs in the download → file-path resolution → preprocess
handoff:

* `SPHER.*.fits` glob at the callers missed `.fits.Z` remnants left by
  interrupted downloads (`eso.retrieve_data(unzip=True)` skips them if
  the run dies mid-flight). Widened to `SPHER.*.fits*`.
* `update_observation_file_paths` warned but left `<missing>` sentinels
  in the FILE column, which crashed the preprocess loader with a
  cryptic `astropy: Empty filename: ''`. It now raises
  `FileNotFoundError` at the boundary, with the branch
[…]
