### 2026-07-20 db7f95d
feat(irdis): TRAP bad-pixel mask + Vigan-style global-anchor DMS propagation
Two Phase-6 hardening changes required for real IRDIS data:

TRAP hand-over (`run_trap.py`):
* Load `converted/badpixel_map.fits` and pass it as `bad_pixel_mask_full`
  to `run_complete_reduction` (previously the local variable existed but
  was never plumbed through — dead code).
* Assert `xy_image_centers.shape[1] == len(pa)` before calling TRAP so a
  frame-axis mismatch fails loudly instead of silently smearing the
  parallactic-angle assignment.

CORO centering (`process_centers.py`):
* Read `MJD` (per-frame midp
[…]
