### 2026-07-20 ef52d77
docs(changelog): Phase 6 IRDIS TRAP integration + master-flat and raw-file-resolution fixes
Adds an Unreleased Added entry describing the end-to-end Phase 6 TRAP
integration (instrument dispatch in run_trap, IRDIS coronagraph
transmission, bad-pixel-mask wiring, shape assertion, and the Vigan-style
global-anchor DMS propagation validated on 51 Eri DB_K12 2015-09-24) and
two Fixed entries for the master-flat multi-frame crash and the
download → path-hydration → preprocess robustness pass.

Claude-Session: https://claude.ai/code/session_01UzZTnuhwbv4UVBx1FxwjTd
