### 2026-07-20 9aa1378
feat: opt-in TRAP inputs (inverse variance + amplitude modulation)
Adds `pass_inverse_variance_to_trap` and `pass_amplitude_modulation_to_trap`
flags to both `IFSReductionConfig` and `IRDISReductionConfig` (default
`False`, no behavior change). When enabled, `run_trap_on_observation`
loads the matching product from the observation's `converted/` folder and
hands it to `run_complete_reduction`:

* `pass_inverse_variance_to_trap` → `converted/{coro,center}_ivar_cube.fits`
  → TRAP's `inverse_variance_full`. Requires the companion trap change on
  `develop` that renames `include_noise` → `estimat
[…]
