### 2026-07-20 225f0f2
feat: default pass_inverse_variance_to_trap = True
Empirically improves detection on the IRDIS DBI 51 Eri 2015-09-24
reference (norm-detection at the known planet position tightens).
Amplitude-modulation toggle stays False by default — meaningful only
for continuous-waffle observations, so it should stay explicit.

Requires the companion trap change on develop (renamed `include_noise`
→ `estimate_noise_from_data`, explicit ivar cube always wins), landed
in trap commit 63ab58b — without it the flag would be a silent no-op.

Claude-Session: https://claude.ai/code/session_01UzZTnuhwbv4UVBx1FxwjTd
