+++ CHANGELOG.md
@@ -9,7 +9,7 @@ This project follows [Semantic Versioning](https://semver.org/) and the [Keep a 
 ## [Unreleased]
 
 ### ✨ Added
-- **Opt-in TRAP inputs: inverse-variance cube + waffle amplitude modulation** – `IFSReductionConfig` and `IRDISReductionConfig` gain two flags: `pass_inverse_variance_to_trap` (loads `converted/{coro,center}_ivar_cube.fits`) and `pass_amplitude_modulation_to_trap` (loads `converted/spot_amplitude_variation.fits`; continuous-waffle only, no-op otherwise). Both default to `False` — no behavior change for existing users. Requires the companion trap change that renames `include_noise` → `estimate_noise_from_data` and makes an explicit ivar cube always win over the flag (previously ivar was silently discarded when the flag was False) ([@m-samland](https://github.com/m-samland)).
+- **Opt-in TRAP inputs: inverse-variance cube + waffle amplitude modulation** – `IFSReductionConfig` and `IRDISReductionConfig` gain two flags: `pass_inverse_variance_to_trap` (loads `converted/{coro,center}_ivar_cube.fits`, default `True` — empirically improves detection on the IRDIS DBI 51 Eri reference) and `pass_amplitude_modulation_to_trap` (loads `converted/spot_amplitude_variation.fits`; continuous-waffle only, no-op otherwise; default `False`). Requires the companion trap change that renames `include_noise` → `estimate_noise_from_data` and makes an explicit ivar cube always win over the flag (previously ivar was silently discarded when the flag was False) ([@m-samland](https://github.com/m-samland)).
 - **IRDIS reduction — Phase 6 TRAP integration** – End-to-end TRAP post-processing for IRDIS observations. `spherical.pipeline.run_trap` now dispatches on instrument for both directory layout (`{reduction}/IRDIS/trap/{name_mode_date}/` — no `{method}` segment) and reduction inputs: reads `converted/badpixel_map.fits` and passes it as TRAP's `bad_pixel_mask_full` (previously the local variable existed but was never plumbed through); asserts `xy_image_centers.shape[1] == len(pa)` so a frame-axis mismatch fails loudly instead of silently mis-associating parallactic angles; selects the packaged `N_ALC_JYH_S-IRDIS_H23-transmission.txt` coronagraph curve for IRDIS by default. Requires `trap >= 1.3.2.dev`, which adds `trap_config_for_irdis()` (IRDIS DBI/broadband defaults: `pixel_scale=0.01225″/px`, `instrument_type="imaging"`, `wavelength_indices=range(0, 2)`) and accepts IRDIS obs modes in `InstrumentConfig.to_instrument`. Non-continuous-waffle IRDIS centering is derived from CENTER waffle fits via a Vigan-style DMS-anchor propagation: `S₀[ch] = median_k(center_k − PAC_center/18)`, `propagated[ch, i] = S₀[ch] + PAC_coro[i]/18` — mathematically equivalent to the SPHERE-community reference (`ImagingReduction.sph_ird_combine_data`) but with anchor noise reduced by √N over all N CENTER frames (per-CENTER fit scatter ~0.1 px → ~0.035 px on S₀ for the 8-CENTER 51 Eri DB_K12 reference). Falls back to the filter's nominal star position when a channel's CENTER fits are all NaN; logs S₀ and per-CENTER scatter so any temporal drift beyond DMS is visible in the pipeline log. End-to-end validated on the 51 Eridani DB_K12 2015-09-24 dataset (dithered non-continuous-waffle mode) with tight `search_region_inner_bound=31 / outer_bound=43` producing a ~4.5σ recovery of 51 Eri b at `yx_known_companion_posit
[…]
