### 2026-07-20 9b11319
feat(irdis): TRAP template-matching detection support
TRAP's SpectralTemplate.__init__ already contains a working 'photometry'
branch that integrates model spectra through per-channel filter bandpasses
via species.SyntheticPhotometry — the correct treatment for IRDIS dual-band
imaging. What was missing on the spherical side was the SPHERE-specific
translation from IRDIS obs-mode strings to species-registered filter names,
and the wiring that populates Instrument.filters before detection.

- Add spherical.pipeline.irdis_filters: obs-mode → (left, right) tuple of
  SVO filter IDs for the five DB
[…]
