+++ CHANGELOG.md
@@ -9,6 +9,26 @@ This project follows [Semantic Versioning](https://semver.org/) and the [Keep a 
 ## [Unreleased]
 
 ### ✨ Added
+- **IRDIS template-matching detection** – TRAP's spectral template matching now
+  works for IRDIS dual-band imaging (`DB_K12`, `DB_H23`, `DB_H34`, `DB_Y23`,
+  `DB_J23`). A new module `spherical.pipeline.irdis_filters` maps each DBI
+  obs-mode string to a pair of SVO / species filter names
+  (e.g. `DB_K12 → (Paranal/SPHERE.IRDIS_D_K12_1, Paranal/SPHERE.IRDIS_D_K12_2)`);
+  `run_trap` populates `Instrument.filters` accordingly and, for
+  `instrument_type == "photometry"` observations without a mapping (broadband
+  / narrow-band / dual-pol modes — single-channel, where template matching
+  would be degenerate), falls back to
+  `analysis.detection_and_characterization()` so contrast curves, normalized
+  detection map, candidate tables, and extracted candidate spectra are still
+  produced. Any error resolving filters is caught, logged, and downgraded to
+  the fallback path. Also handles a pickled-Instrument edge case: after
+  `read_output(read_instrument=True)` the loaded `analysis.instrument` is
+  patched in-place from `used_instrument` so detection-only re-runs against
+  an older reduction folder pick up the current `instrument_type` and filter
+  names instead of stale on-disk state. End-to-end validated on 51 Eri
+  DB_K12 2015-09-24: T-type template picked correctly, K1 flux > K2 flux
+  (methane absorption), astrometry within 10 mas / 2° of published values
+  ([@m-samland](https://github.com/m-samland)).
 - **Opt-in TRAP inputs: inverse-variance cube + waffle amplitude modulation** – `IFSReductionConfig` and `IRDISReductionConfig` gain two flags: `pass_inverse_variance_to_trap` (loads `converted/{coro,center}_ivar_cube.fits`, default `True` — empirically improves detection on the IRDIS DBI 51 Eri reference) and `pass_amplitude_modulation_to_trap` (loads `converted/spot_amplitude_variation.fits`; continuous-waffle only, no-op otherwise; default `False`). Requires the companion trap change that renames `include_noise` → `estimate_noise_from_data` and makes an explicit ivar cube always win over the flag (previously ivar was silently discarded when the flag was False) ([@m-samland](https://github.com/m-samland)).
[…]
