### 2026-07-20 2196264
feat(flux-psf): BPM-aware calibration + Moffat core repair
Phase 1 — BPM-aware calibration
-------------------------------
Reads flux_ivar_cube.fits (written by preprocess) and derives a per-frame
bad-pixel mask (ivar == 0) that is threaded into:
  * star_centers_from_PSF_img_cube (Gaussian centering fit)
  * flux_calibration.get_aperture_photometry (new bad_pixel_mask kwarg;
    excludes flagged pixels from both source aperture sum and background
    annulus sigma-clip)

Two per-channel diagnostics fire on every observation:
  * Amplitude-delta warning — Gaussian is fit twice per frame (maske
[…]
