+++ src/spherical/pipeline/steps/flux_psf_calibration.py
@@ -20,6 +20,7 @@ from astropy.io import fits
 
 from spherical.pipeline import flux_calibration, toolbox, transmission
 from spherical.pipeline.logging_utils import optional_logger
+from spherical.pipeline.psf_repair import repair_psf_core
 from spherical.pipeline.steps.find_star import guess_position_psf, star_centers_from_PSF_img_cube
 
 
@@ -187,6 +188,31 @@ def run_flux_psf_calibration(
     flux_cube = fits.getdata(flux_cube_path).astype('float64')
     logger.debug(f"Loaded wavelengths shape: {wavelengths.shape}, flux_cube shape: {flux_cube.shape}")
 
+    # Per-frame inverse variance cube written by preprocess: ``ivar == 0``
+    # marks pixels that are either dead detector regions, calibration-flagged
+    # bad pixels, or per-frame sigma-clipped transients. Used below to (a)
+    # keep the centering fit and aperture photometry from being pulled by
+    # interpolated values in the PSF core, and (b) drive the Phase 2 Moffat
+    # repair of core bad pixels (IRDIS K1 in particular has 3-4 permanent bad
+    # pixels at the flux-PSF landing spot).
+    flux_ivar_path = os.path.join(converted_dir, 'flux_ivar_cube.fits')
+    if os.path.exists(flux_ivar_path):
+        flux_ivar_cube = fits.getdata(flux_ivar_path).astype('float64')
+        flux_bpm_cube = flux_ivar_cube == 0
+        logger.debug(
+            f"Loaded flux_ivar_cube for BPM: {int(flux_bpm_cube.sum())} bad "
+            f"pixels across the full cube"
+        )
+    else:
+        flux_ivar_cube = None
+        flux_bpm_cube = None
+        logger.warning(
+            "flux_ivar_cube.fits not found — centering and aperture photometry "
+            "will not mask bad pixels for this observation, and Phase 2 Moffat "
+            "core repair will be skipped.",
+            extra={"step": "flux_psf_calibration", "status": "bpm_missing"},
+        )
+
     # Instrument-aware pixel scale for loD-based Gaussian init. A 2-channel
     # cube marks IRDIS DBI (12.25 mas/px); anything larger is IFS (7.46 mas/px).
     is_irdis = flux_cube.shape[0] == 2
@@ -220,16 +246,27 @@ def run_flux_psf_calibration(
     
     logger.debug(f"Computed guess positions for {len(guess_positions_yx)} frames")
     
-    # Now compute flux centers using pre-computed guess positions
+    # Now compute flux centers using pre-computed guess positions.
+    #
+    # We call ``star_centers_from_PSF_img_cube`` twice: once with the per-frame
+    # bad-pixel mask applied (this becomes the canonical result), and once
+    # without any mask so we can log the flux-amplitude delta and raise a
+    # warning if the interpolated bad-pixel values were biasing the fit by
+    # more than 10% on any channel. Even for well-behaved data this is nearly
+    # free (a single 2D Gaussian fit per frame per channel), and it surfaces
+    # the "permanent bad pixel in PSF core" case (see IRDIS K1 for 51 Eri) as
+    # an explicit log line instead of a silent bias.
     flux_centers = []
     flux_amplitudes = []
+    flux_amplitudes_unmasked = []
     for frame_number in range(flux_cube.shape[1]):
         data = flux_cube[:, frame_number]
+        per_frame_mask = flux_bpm_cube[:, frame_number] if flux_bpm_cube is not None else None
         flux_center, flux_amplitude = star_centers_from_PSF_img_cube(
-            cube=data,
+            cube=data.copy(),
             wave=wavelengths,
             pixel=pixel_scale_mas,
[…]
