### 2026-07-20 fac04db
feat(irdis): center-fit seed-delta log, bad-frames plumbing, warning fix
Three post-Phase-6 follow-ups from validating IRDIS reductions on both
Beta Pic K12 2014-12-07 (continuous-waffle) and 51 Eri K12 2015-09-24
(non-waffle DMS-dithered). Having a second K12 dataset with essentially
zero DMS PAC on both reveals the K-band nominal is a near-perfect match
for the 51 Eri epoch (delta < 0.1 px in both channels/axes) but ~9 px off
in y on Beta Pic — traced to a physical coronagraph realignment between
those two epochs, not a calibration bug. The nominal is only a seed for
the waffle fit box (21 p
[…]
