### 2026-07-20 80c2df1
docs(examples): expand IRDIS reduction template with more knobs
Trims the phase-by-phase docstring at the top (design docs and CHANGELOG
already carry that detail — driver templates should read as drivers, not
documentation). Exposes more of the config surface as commented-out
knobs so users see the parameters they can actually tune without having
to grep the dataclasses:

* Resources: `set_ncpu` (already there) + commented per-step overrides.
* IRDISCalibrationConfig: combination method, badpix sigma thresholds,
  flat response bounds.
* Existing IRDISPreprocessConfig block: unchanged.
* IRDI
[…]
