+++ examples/irdis_reduction_template.py
@@ -1,81 +1,39 @@
-"""IRDIS reduction driver — Phases 1 + 3 + 4 + 5 + 6.
+"""IRDIS reduction driver — end-to-end template.
 
-Runs the IRDIS pipeline end-to-end. What each phase produces:
+Runs download → master calibration → preprocess → shared downstream steps
+(centering, flux PSF, spot-to-flux) → TRAP reduction + detection for the
+observations selected below. Outputs land under
+``{reduction_directory}/IRDIS/{calibration,observation,trap}/…``.
 
-- **Phase 1 — download** (``steps.download_data``): pulls raw ``CORO``,
-  ``CENTER``, ``FLUX``, ``FLAT``, and ``BG_SCIENCE`` FITS files for each
-  observation from the ESO archive into ``raw_directory``. Idempotent — files
-  already present locally are skipped, so re-runs are cheap.
-- **Phase 3 — master calibration** (``steps.irdis_calibration``): from the
-  ``FLAT`` and ``BG_SCIENCE`` frames, builds ``master_flat.fits``,
-  ``master_background.fits``, and ``badpixel_map.fits`` under
-  ``{reduction_directory}/IRDIS/calibration/{filter}/{date}/``. Internal-guard
-  step — skips itself when all three outputs already exist.
-- **Phase 4 — science preprocess** (``steps.preprocess_irdis``): per
-  observation, per frame type (CORO/CENTER/FLUX), scaled-background
-  subtraction + flat division + analytic ivar + NaN-safe bad-pixel
-  replacement. Writes 8 files under
-  ``{reduction_directory}/IRDIS/observation/{target}/{filter}/{date}/converted/``:
-  ``{coro,center,flux}_cube.fits``, matching ``*_ivar_cube.fits``,
-  ``wavelengths.fits``, and ``badpixel_map.fits``. Gated by ``should_run``.
-  Cube axis order: ``(n_wave=2, n_time, ny, nx)``.
-- **Phase 5 — shared downstream steps** (``compute_frames_info``,
-  ``cube_header_update``, ``find_centers``, ``process_extracted_centers``,
-  ``plot_image_center_evolution``, ``calibrate_spot_photometry``,
-  ``calibrate_flux_psf``, ``spot_to_flux``): shared with the IFS pipeline
-  under instrument dispatch. Continuous-waffle observations get a
-  ``center_cube.fits → coro_cube.fits`` symlink so downstream code stays
-  instrument-agnostic. Produces ``image_centers.fits``,
-  ``image_centers_fitted_robust.fits``, ``spot_amplitude_variation.fits``,
-  ``psf_cube_for_postprocessing.fits`` etc.
-- **Phase 6 — TRAP post-processing** (``steps.run_trap_reduction`` /
-  ``steps.run_trap_detection``): reduction + detection via
-  ``run_trap_on_observations``. Outputs land under
-  ``{reduction_directory}/IRDIS/trap/{target}/{filter}/{date}/`` (no
-  ``{method}`` segment for IRDIS).
-
-Inspecting results after a run:
-
-    ls -1 ~/data/sphere/reduction/IRDIS/calibration/*/*/                     # Phase 3
-    ls -1 ~/data/sphere/reduction/IRDIS/observation/*/*/*/converted/         # Phase 4
-
-Quick FITS inspection:
-
-    python -c "
-    from astropy.io import fits
-    import numpy as np
-    from pathlib import Path
-    for p in sorted(Path.home().glob('data/sphere/reduction/IRDIS/observation/*/*/*/converted/*_cube.fits')):
-        data = fits.getdata(p)
[…]
