### 2026-07-20 30a2ff5
feat(database): add `public` filter to SphereDatabase.filter()
Restricts observations to those with `NIGHT_START` more than one year
before today, i.e. out of the ESO proprietary period.

Claude-Session: https://claude.ai/code/session_01UzZTnuhwbv4UVBx1FxwjTd
