### 2026-07-21 3397006
docs(examples): document yx_anamorphism knob in IRDIS reduction template
Now a TRAP-side default via `trap_config_for_irdis()`; the template keeps a
commented override that notes the coupling with `correct_anamorphism` in the
preprocessing block.
