### 2026-07-24 723c72a
test(regression): 51 Eri IRDIS phase-6 astrometry regression vs frozen TRAP baseline
Opt-in (-m regression), runs in the pipeline env. Compares the
template-matched companion table from the current TRAP branch against a
frozen old-code, regressors-off baseline (template_matching_without_sdi),
asserting positions stay within the free-fit-at-threshold budget and the
previously-NaN astrometric sigmas are now finite and self-consistent.
Skips cleanly when no fresh new-schema output is present.
