+++ tests/test_51eri_astrometry_regression.py
@@ -1,0 +1,219 @@
+"""Regression test for the TRAP astrometric-uncertainty change on real data.
+
+Runs (or reads a prior run of) the 51 Eridani IRDIS DB_K12 phase-6 smoke test
+and compares the resulting template-matched companion table against a frozen
+baseline produced by the *old* TRAP code (``feature/edge-fov-reduction``,
+multiwavelength regressors off). See
+``trap/docs/superpowers/specs/2026-07-23-trap-astrometry-uncertainty-design.md``
+(Testing section) for the tolerance rationale.
+
+This is a "no surprises" test, not a correctness proof: the new positions are
+the more principled ones by construction. It asserts that
+
+- positions have not moved more than the free-fit-at-threshold noise budget,
+- the astrometric σ columns that were ``NaN`` in the old output are now finite
+  and self-consistent (positive-definite 2×2 covariance, polar matches
+  Cartesian).
+
+**Heavy + data-dependent — opt in with ``-m regression`` and run in the
+pipeline env** (``pixi run -e dev pytest tests/test_51eri_astrometry_regression.py
+-m regression``), because it needs the TRAP sibling and the 51 Eri reduction
+products on disk.
+
+Workflow:
+
+1. Produce the new output once, with the *current* TRAP branch checked out::
+
+       pixi run -e dev python examples/irdis_reduction_phase6_smoketest.py
+
+   (``MULTIWAVELENGTH_REGRESSORS = None`` in that driver — keep it off; the
+   baseline is regressors-off.) This writes
+   ``<result>/template_matching/overall_validated_companion_detections.csv``.
+   Alternatively set ``SPHERICAL_RUN_SMOKETEST=1`` and this test will run the
+   driver itself as a subprocess before comparing.
+
+2. Run this test. It locates the fresh CSV, reads the frozen baseline in
+   ``tests/data/``, and checks the tolerances below.
+
+Override the search root with ``SPHERICAL_SMOKETEST_RESULT_DIR`` (pointing at
+the smoke test's per-observation result folder, or any parent of it).
+"""
+
+import os
+import subprocess
+from pathlib import Path
+
+import numpy as np
+import pandas as pd
+import pytest
+
+# Position tolerances expressed as multiples of the IRDIS K-band PSF σ
+# (FWHM ≈ 3.4 px → σ_PSF ≈ 1.45 px). At threshold SNR (≈4–6) a free centroid
+# fit is legitimately noisier than the old fixed-width fit; that shift is the
+# expected behaviour, not a regression.
+SIGMA_PSF_PIX = 1.45
+TOL_POSITION_HIGH_SNR_PIX = 0.5 * SIGMA_PSF_PIX
+TOL_POSITION_THRESHOLD_SNR_PIX = 2.0 * SIGMA_PSF_PIX
+HIGH_SNR_CUTOFF = 8.0
+MATCH_RADIUS_PIX = 3.0
+
+BASELINE = (
+    Path(__file__).parent
+    / "data"
+    / "51eri_baseline_overall_validated_companion_detections.csv"
+)
+NEW_CSV_NAME = "overall_validated_companion_detections.csv"
+
+SMOKETEST_DRIVER = "examples/irdis_reduction_phase6_smoketest.py"
+DEFAULT_REDUCTION_ROOT = Path.home() / "data" / "sphere" / "reduction"
+
+
+def _locate_new_csv() -> Path | None:
+    """Find the freshly produced template-matched short table.
+
+    The per-observation result folder name can contain a literal ``*`` (a
+    sanitised target prefix), so we walk real directory names with rglob rather
+    than shell-globbing.
+    """
[…]
