+++ pyproject.toml
@@ -135,4 +135,5 @@ omit = [
 addopts = "--log-format='%(asctime)s %(levelname)s %(message)s' --log-date-format='%Y-%m-%d %H:%M:%S'"
 markers = [
     "remote_data: tests that require a network connection to external services",
+    "regression: heavy end-to-end regression tests against a frozen baseline (opt in with -m regression)",
 ]
