### 2026-07-27 643f0b5
test(regression): re-freeze 51 Eri astrometry baseline to validated per-channel result
The baseline was the old template-collapse position (462.4 mas, no sigma). Trap
now reports per-channel astrometry as primary (454.6 mas, within ~0.7 mas of
GRAVITY), so the baseline is re-frozen to that validated output, turning the test
from a loose "no gross surprise" smoke check into a tight drift guard. Adds the
51 Eri astrometry benchmark reference doc (GRAVITY/Maire numbers, IFS checklist).

Claude-Session: https://claude.ai/code/session_01BNXhnrRQXbMCYsNpsVRTLv
