+++ tests/data/51eri_astrometry_benchmark.md
@@ -1,0 +1,288 @@
+# 51 Eridani b — astrometry benchmark & baseline reference
+
+Reference document for evaluating pipeline astrometry (position + uncertainties) of the
+51 Eri b companion against external ground truth. Written 2026-07-27 from the IRDIS
+DB_K12 phase-6 smoke test; **designed to be reused for the simultaneous IFS dataset**.
+
+> **History (2026-07-27), read this first:** this benchmark went through two corrections.
+> (1) An early draft compared a `flat` "K1-only" detection to GRAVITY using a *stale* CSV
+> that the pipeline erroneously ingested — a real bug, now fixed
+> (`../trap/docs/llm_reference/GITHUB_ISSUE_stale_template_csv_ingestion.md`).
+> (2) After fixing that, a **clean, freshly-computed per-channel measurement** confirmed the
+> underlying point anyway: the per-channel astrometry (channels above threshold, here K1
+> only) lands on GRAVITY (Δρ ≈ −0.7 mas) while the template collapse is +8.7 mas / 2.6σ
+> off. TRAP now **measures per-channel astrometry and reports it as primary**
+> (`astrometry_source` column), collapse as fallback. The numbers below reflect that.
+
+All conversions use the instrument plate scale; see the provenance table for exact values.
+
+---
+
+## 1. Dataset identification
+
+| Property | Value |
+|---|---|
+| Target | 51 Eridani (companion 51 Eri b) |
+| Night (`NIGHT_START`) | 2015-09-24 |
+| Epoch | ≈ 2015.73 (Maire et al. 2019 lists this observation as **2015.74**) |
+| Instrument / mode | VLT/SPHERE **IRDIS** DBI, filter **DB_K12** |
+| Channels | ch0 = K1 (2.110 µm), ch1 = K2 (2.251 µm) |
+| Derotator | PUPIL (ADI) |
+| IRDIS plate scale | **12.25 mas/px** (`0.01225 arcsec/px`, `trap_config_for_irdis()`) |
+| Anamorphism | `yx_anamorphism = [1.0062, 1.0]` |
+| TRAP search annulus | inner 31 px, outer 43 px; `yx_known_companion_position = (-35.95, -8.43)` |
+| `temporal_components_fraction` | `[0.2]` |
+| Multiwavelength regressors | **None** (WP1 baseline; regressors OFF — the baseline condition) |
+| Driver | `examples/irdis_reduction_phase6_smoketest.py` (full) or a detection-only re-run |
+| σ_PSF (K-band) | FWHM ≈ 3.4 px → σ_PSF ≈ **1.44 px ≈ 17.7 mas** |
+
+> **Regressors must be OFF for the baseline.** Use `template_matching/` (regressors off)
+> or the frozen `template_matching_without_sdi/`. The `template_matching_{pool,occluded,
+> sdi}/` folders are WP2 regressor experiments and must **not** be used as an astrometry
+> baseline.
+>
+> **Clear stale per-template CSVs before re-running detection.** A template that finds
+> nothing in a run leaves its previous `companion_table_*.csv` on disk. The combination
+> bug that ingested those has been fixed (it now uses in-memory tables), but stale files
+> on disk are still confusing when inspecting output by hand.
+
+---
+
+## 2. How template matching works here (important for interpretation)
+
+TRAP template matching does **not** detect per channel and combine. For each spectral
+template it **collapses** all wavelength channels into a *single* template-weighted
+detection map, then detects on that one map (`template_matching_detection` →
[…]
