### 2026-07-28 48a5058
feat(pipeline): derive bad-pixel masks from ivar instead of testing for zero
`ivar == 0` was the bad-pixel test everywhere: it drove flux_bpm_cube in
flux_psf_calibration, the Phase-2 Moffat core repair, the mask handed to the PSF
combination, and (via badpixel_map.fits) TRAP's regressor exclusion. On IFS it finds
nothing at all, so all of that was silently inert -- the 51 Eri OBS_H run wrote no
flux_cube_repaired.fits, no flux_psf_repair_log.csv, and passed bad_pixel_mask_full=None
to TRAP.

The premise does not hold for an IFU. charis zeroes ivar on the raw 2048^2 detector
frame before the o
[…]
