### 2026-07-28 5224bcf
docs(benchmark): correct the true-north claim and record the IFS run
Two fixes to 51eri_astrometry_benchmark.md.

Section 7 claimed the reported PA carries no true-north correction and so cannot be
compared to GRAVITY in absolute terms. That is wrong for this pipeline:
database.metadata.compute_angles(true_north=-1.75) folds true north, the pupil offset
and the per-instrument offset (IFS -100.48, IRDIS 0.0) into DEROT ANGLE, and
{coro,center}_parallactic_angles.fits -- what TRAP actually receives -- is exactly that
column. PA is directly comparable; the residual is the RSS of the calibration
c
[…]
