+++ tests/data/51eri_astrometry_benchmark.md
@@ -188,10 +188,15 @@ episode pointed at, now established on freshly-computed clean data.
 - **n = 1 caveat.** One dataset, one GRAVITY point. The *mechanism* is principled and
   dataset-independent; keep validating the per-channel-vs-collapse gap across more
   companions with interferometric truth (IRDIS and IFS).
-- **Residual offset is uncalibrated systematics.** Even the per-channel value carries no
-  true-north correction (IRDIS TN ≈ −1.75°) or epoch plate-scale calibration; separation is
-  calibration-light (plate scale only), which is why it agrees to <1 mas, while PA still
-  needs TN before absolute comparison.
+- **True north *is* already applied — corrected 2026-07-28.** An earlier version of this
+  document claimed the reported PA was uncalibrated. It is not:
+  `spherical.database.metadata.compute_angles(true_north=-1.75)` folds true north, the
+  pupil offset and the per-instrument offset (`IFS = -100.48`, `IRDIS = 0.0`) into
+  `DEROT ANGLE`, and `{coro,center}_parallactic_angles.fits` — what TRAP receives — is
+  exactly that column. PA may therefore be compared to GRAVITY directly. The residual
+  calibration uncertainty is the RSS of the constants: ≈ 0.14° for IRDIS
+  (TN ±0.08, PUPOFF ±0.11) and ≈ 0.17° for IFS (plus the ±0.10 instrument offset).
+  What remains genuinely uncalibrated is the **epoch plate scale**.
 - **The σ magnitude is realistic.** The per-channel σ_ρ ≈ 5.75 mas and σ_PA ≈ 0.31° are in
   the range of published SPHERE measurement errors for this epoch/band (ANDROMEDA 1.4 mas /
   0.06°; TLOCI 4.4 mas / 0.55°). Radial σ (5.75) > tangential σ (2.45) as expected for ADI
@@ -240,10 +245,58 @@ guard against when reusing this document.
 5. **No stale contamination.** Confirm `n_templates_above_threshold` equals the number of
    templates that actually validated this run (the stale-CSV bug is fixed, but verify).
 
+### 8b-bis. Results of the first IFS run (2026-07-28) — read before reusing §8b
+
+The IFS end-to-end run happened; §8b's criteria were applied and three of them needed
+qualification. Full numbers in [`../../llm_docs/decisions.md`](../../llm_docs/decisions.md)
+2026-07-28.
+
+| Measurement | ρ @ 7.46 mas/px, anamorphism applied | Δρ vs GRAVITY | PA | ΔPA |
+|---|---|---|---|---|
+| T-type collapse (now reported) | 454.30 ± 3.85 | −1.07 (0.3σ) | 166.160 ± 0.292 | −0.68 |
+| L-type collapse | 451.99 ± 5.00 | −3.38 | 167.565 ± 0.414 | +0.73 |
+| per-channel (2 of 37 ch, now gated off) | 449.86 ± 2.77 | −5.50 | 167.160 ± 0.259 | +0.32 |
+
+Three things the IRDIS benchmark did not anticipate:
+
+1. **`yx_anamorphism` was `[1, 1]` for IFS.** Now `[1.0059, 1.0011]` in
+   `trap_config_for_ifs()`. Worth +0.30 px / +2.2 mas here (the naive +2.5 mas is diluted
+   to +2.2 by 43° of field rotation, since TRAP applies the distortion per frame).
+2. **The per-channel override made IFS astrometry worse, not better** — see §8c.
+3. **The plate scale is the open item, not the detection.** Once (1) and (2) are handled the
+   T-type collapse sits 1.07 mas (0.3σ) from GRAVITY at the nominal 7.46 mas/px, so this
[…]
