### 2026-07-28 0ddc5ce
docs(ivar): correct why `ivar == 0` fails on IFS -- charis does mask spaxels
The rationale committed in 48a5058 was incomplete and in one respect wrong. It
explained the absence of zeros purely by the detector-plane zeroing that happens
before extraction, and attributed the observed 1e-15 floor to the unilluminated
field. Both need correcting.

charis does mask bad spaxels after extraction: _smoothandmask_hexgeometry rejects
lenslets deviating more than 10 robust sigma in ivar or 5 sigma in flux from their
hexagonal neighbours, separately at each wavelength. The 1e-15 floor is that filter's
se
[…]
