### 2026-07-28 2e72b4a
fix(ivar): rely on exact ivar==0 for IFS bad pixels after the charis resample fix
charis's hexagon->square resample now propagates variance instead of applying
the flux operator to ivar (charis #42), so a lenslet charis flags survives as an
exact ivar==0 in every square it touches. Validated on a fresh 51 Eri OBS_H
extraction: illuminated exact-zeros 0 -> ~7%, positive-ivar scale x24.

Recalibrate the bad-pixel path for this:
- IFS ivar_bad_pixel_ratio_threshold default 0.2 -> 0.0. The exact zeros are
  caught by the ivar<=0 branch; a positive threshold now only flags good moire
  troughs the 
[…]
