### 2026-07-29 53db6be
fix(ivar): gate bad-pixel mask by the data footprint so zero clusters can't leak
bad_pixel_mask_from_ivar inferred "in field" from a local median baseline. Inside
a cluster of exact zeros the baseline collapses to ~0, so the cluster interior
read as out-of-field and escaped the mask -- then reached the TRAP reduction as
an all-zero-weight pixel, making the WLS singular (the pca_regression
divide-by-zero fixed separately in trap).

Add an optional `in_field` argument that states the illuminated field directly,
and have run_trap pass the data-cube footprint (np.isfinite(data)): the
unilluminated
[…]
