### 2026-07-29 37140b0
docs(changelog): record the upstream charis hex bad-lenslet flagging rewrite
charis devel e255469 replaces the two-armed hex flagging with a single
one-sided ivar criterion. spherical has no code change, but the reduction
output does: ~40% fewer spaxels are masked and the mask stops churning from
frame to frame, so the entry records the measurements behind it and notes
that existing reductions need a re-extraction to pick it up.

Claude-Session: https://claude.ai/code/session_01BNXhnrRQXbMCYsNpsVRTLv
