### 2026-07-29 e7f9ca2
docs(changelog): trim the charis flagging entry to changelog depth
The entry carried the full measurement narrative from the analysis session.
Cut it to what a reader needs and point at charis e255469, whose commit
message holds the numbers.

Claude-Session: https://claude.ai/code/session_01BNXhnrRQXbMCYsNpsVRTLv
