### 2026-07-29 ec275eb
docs(benchmark): drop the manual stale-CSV cleanup from the reproduce steps
trap now removes the per-template and overall companion tables at the start of a
detection run, so the `rm template_matching/{companion_table_*,...}` step before
a detection-only re-run is obsolete. Left in place it would read as still
required.

Claude-Session: https://claude.ai/code/session_01BNXhnrRQXbMCYsNpsVRTLv
