### 2026-07-30 f450f29
test(regression): freeze the 51 Eri IFS astrometry baseline
The IRDIS sibling test only pinned DB_K12, so nothing guarded the IFS
conclusions of benchmark §8: that the reported position comes from the
template collapse rather than the gated-off per-channel override, and that
7.46 mas/px needs no revision. Freeze the measured 2026-07-29 run (T-type
collapse 453.30 ± 3.83 mas, 0.53σ from GRAVITY) as two baseline CSVs plus
five checks.

The benchmark's provenance note is corrected: the logged trap version is an
editable install's build stamp, not the code that ran, so the run did carry
the trap d
[…]
