+++ tests/test_51eri_ifs_astrometry_regression.py
@@ -1,0 +1,270 @@
+"""Regression test for the 51 Eridani **IFS** astrometry on real data.
+
+The IRDIS sibling (:mod:`tests.test_51eri_astrometry_regression`) guards the
+DB_K12 result. This module guards the simultaneous IFS YJH reduction of the same
+night, and additionally freezes the *conclusions* reached in
+``tests/data/51eri_astrometry_benchmark.md`` §8, so a code change cannot quietly
+undo them:
+
+- the reported astrometry comes from the **template collapse**, not the
+  per-channel override (only 2 of 37 channels clear ``candidate_threshold``, far
+  below ``per_channel_min_channel_fraction``; the override was 4.4 mas *further*
+  from GRAVITY),
+- the collapse position agrees with the GRAVITY ground truth within its own σ at
+  the nominal 7.46 mas/px plate scale, so this dataset does not demand a scale
+  revision,
+- the σ columns are finite, positive and self-consistent,
+- ``n_templates_above_threshold`` matches the per-template tables actually
+  written by the run (stale-CSV contamination guard).
+
+**Heavy + data-dependent — opt in with ``-m regression`` and run in the pipeline
+env** (``pixi run -e dev pytest tests/test_51eri_ifs_astrometry_regression.py
+-m regression``), because it needs the 51 Eri IFS TRAP products on disk.
+
+Produce them with a driver that has ``run_trap_reduction`` / ``run_trap_detection``
+enabled, e.g. ``examples/ifs_reduction_template.py`` pointed at 51 Eri, then run
+this test. Override the search root with ``SPHERICAL_IFS_RESULT_DIR`` (the
+per-observation TRAP result folder, or any parent of it).
+
+GRAVITY values are **unpublished** — do not redistribute outside this project.
+"""
+
+import os
+from pathlib import Path
+
+import numpy as np
+import pandas as pd
+import pytest
+
+# IFS-specific constants. The plate scale is the single most common mistake when
+# reusing the IRDIS benchmark — it is 7.46 mas/px here, not 12.25.
+IFS_PLATE_SCALE_MAS = 7.46
+N_USABLE_CHANNELS = 37
+# Mean diffraction FWHM over the 37 used YJH channels is 4.66 px (0.94–1.68 µm,
+# 8 m aperture, 7.46 mas/px) → σ_PSF ≈ 1.98 px ≈ 14.8 mas.
+SIGMA_PSF_PIX = 1.98
+
+# GRAVITY interferometric ground truth for epoch 2015.73 (unpublished).
+GRAVITY_RHO_MAS = 455.364
+GRAVITY_RHO_SIGMA_MAS = 0.653
+GRAVITY_PA_DEG = 166.839
+GRAVITY_PA_SIGMA_DEG = 0.086
+
+# Agreement budget against GRAVITY. Separation is checked in units of the
+# combined σ; the PA residual is dominated by the true-north/pupil-offset
+# calibration folded into DEROT ANGLE, so it gets an absolute degree budget
+# rather than an nσ one.
+MAX_SEPARATION_NSIGMA = 2.0
+MAX_POSITION_ANGLE_RESIDUAL_DEG = 1.0
+
+# Drift budget against the frozen baseline. At the ≈6σ SNR of this detection a
+# free centroid fit moves by a sizeable fraction of σ_PSF between runs; anything
+# beyond half a PSF σ is a real change, not fit noise.
+TOL_POSITION_PIX = 0.5 * SIGMA_PSF_PIX
+MATCH_RADIUS_PIX = 3.0
+
+BASELINE_OVERALL = (
+    Path(__file__).parent
+    / "data"
+    / "51eri_ifs_baseline_overall_validated_companion_detections.csv"
+)
[…]
