+++ tests/data/51eri_astrometry_benchmark.md
@@ -245,17 +245,38 @@ guard against when reusing this document.
 5. **No stale contamination.** Confirm `n_templates_above_threshold` equals the number of
    templates that actually validated this run (the stale-CSV bug is fixed, but verify).
 
-### 8b-bis. Results of the first IFS run (2026-07-28) — read before reusing §8b
+### 8b-bis. Frozen IFS result (run 2026-07-29 15:16) — read before reusing §8b
 
-The IFS end-to-end run happened; §8b's criteria were applied and three of them needed
-qualification. Full numbers in [`../../llm_docs/decisions.md`](../../llm_docs/decisions.md)
-2026-07-28.
+This is the **frozen IFS baseline**, measured end-to-end (fixed ivar propagation, ivar-derived
+bad pixels with the data-footprint gate, `yx_anamorphism = [1.0059, 1.0011]`), not extrapolated.
+It is the source of `51eri_ifs_baseline_overall_validated_companion_detections.csv` /
+`51eri_ifs_baseline_per_channel_astrometry.csv` and is guarded by
+[`../test_51eri_ifs_astrometry_regression.py`](../test_51eri_ifs_astrometry_regression.py).
+§8b's criteria were applied and three of them needed qualification.
 
-| Measurement | ρ @ 7.46 mas/px, anamorphism applied | Δρ vs GRAVITY | PA | ΔPA |
+| Measurement | ρ @ 7.46 mas/px | Δρ vs GRAVITY | PA | ΔPA |
 |---|---|---|---|---|
-| T-type collapse (now reported) | 454.30 ± 3.85 | −1.07 (0.3σ) | 166.160 ± 0.292 | −0.68 |
-| L-type collapse | 451.99 ± 5.00 | −3.38 | 167.565 ± 0.414 | +0.73 |
-| per-channel (2 of 37 ch, now gated off) | 449.86 ± 2.77 | −5.50 | 167.160 ± 0.259 | +0.32 |
+| **T-type collapse (reported)** | **453.30 ± 3.83** | **−2.06 (0.53σ)** | **166.174 ± 0.316** | **−0.67** |
+| per-channel (2 of 37 ch, gated off) | 448.86 ± 3.82 | −6.51 (1.68σ) | 167.317 ± 0.357 | +0.48 |
+
+Only T-type validated (peak norm-SNR 6.50); L-type peaks at 4.68 and flat at 3.89, both below
+`candidate_threshold = 4.75`, so `n_templates_above_threshold = 1` and no L-type/flat companion
+tables are written. The pre-fix 2026-07-27 run for comparison: it reported the *per-channel*
+position 447.66 ± 2.77 (Δρ −7.70, −2.71σ), and both L-type and T-type validated.
+
+Implied plate scale to place the collapse exactly on GRAVITY: **7.494 mas/px** (was 7.517
+before the fixes) — 0.24% from the 7.512 waffle transfer, still not a reason to revise 7.46.
+
+Provenance of the frozen numbers: spherical `2.1.4.dev169+g643f0b598`, charis `2.0.1`,
+trap `1.3.2.dev45+g791b3ccfc`, run 2026-07-29 15:16→16:19. **The trap version string in the
+log lags the code that ran** — an editable install stamps `_version.py` at install time, so a
+`git pull` in the checkout does not update it. This run did carry the relevant trap fixes
+(`1e0e709` unconstrained-WLS infinite variance, `404335c` stacked-cube rewrite, `7395902`
+per-run table purge); the two that leave a trace on disk were verified here: the stacked
+`detection_ncomp*.fits` restacks byte-identically from the 37 `detection_lam*.fits`, and only
+the one validated template has a companion table. Everything in trap after those is 2.0.0
[…]
