### 2026-07-30 e28810e
build(deps): require trap v2.0.0
2.0.0 is the first trap release whose astrometry the frozen 51 Eri baselines
were measured against — per-channel astrometry behind the channel-fraction
gate, the SPHERE anamorphism defaults, ivar always honoured, and the
footprint-aware reduction that run_trap feeds `valid_pixel_mask` — and it
removed the deprecated `Reduction_parameters` path. Pin both dependency
declarations to the tag and lift `_MIN_TRAP_VERSION` to match.

The floor error now also tells you to reinstall an editable sibling checkout:
its recorded version is stamped at install time, so a `git
[…]
