### 2026-07-30 f81dc75
docs(examples): note auto_footprint in the IRDIS driver template
trap 2.0.0 derives the reduction footprint from border-connected NaNs and
trap_config_for_irdis() turns it on, which also makes
search_region_outer_bound=None meaningful. Surface both in the knob list so
the template still documents the defaults it runs with.
