### 2026-07-30 449233f
fix(scripts): attribute TRAP crash reports and IRDIS runs correctly
`crash_reports` parsed the dataset identifier only from the reduction crash
report's wording, so every TRAP report -- which opens with a different sentence
-- lost its target/band/night and was listed as "unknown". Both aggregators also
labelled IRDIS reductions `ifs`, because the pipeline type was inferred from the
log file name, which IRDIS shares with IFS.

Instrument is now derived from the observation band (only IFS uses `OBS_*`) and
reported in its own column, and the old TYPE column becomes PIPELINE with values
reductio
[…]
