### 2026-07-30 eb704ae
refactor(irdis): drop the superseded single-script IRDIS prototype
`simplified_IRDIS_reduction` was the prototype the step-based `irdis_reduction`
pipeline replaced. It has no callers and stopped importing on scikit-image >= 0.19,
which removed `skimage.feature.register_translation`; its numeric constants were
already carried over into `irdis_calibration` and `irdis_preprocess`. Those two
comments cited it as the source, so they now state the provenance directly
instead of pointing at a deleted file.

Also sorts a function-local import block flagged by ruff in the same file.
