+++ src/spherical/pipeline/simplified_IRDIS_reduction.py
@@ -1,1305 +1,0 @@
-import collections
-import os
-from os import path
-
-import matplotlib.pyplot as plt
-import numpy as np
-import pandas as pd
-from astropy.io import fits
-from astropy.modeling import fitting, models
-from astropy.table import Column, Table, vstack
-from skimage.feature import register_translation
-from tqdm import tqdm
-from trap.utils import crop_box_from_3D_cube, crop_box_from_image
-
-from spherical.pipeline import imutils, toolbox, transmission
-from spherical.pipeline.toolbox import (
-    # measure_center_waffle,
-    # prepare_dataframe,
-    star_centers_from_waffle_img_cube,
-)
-
-
-def reshape_IRDIS_cube_loop(img):
-    nframes = img.shape[0]
-    nimg = np.zeros((nframes, 2, 1024, 1024))
-    for f in range(len(img)):
-        nimg[f, 0] = img[f, :, 0:1024]
-        nimg[f, 1] = img[f, :, 1024:]
-    return nimg
-
-
-def reshape_IRDIS_cube(cube):
-    nframes = cube.shape[0]
-    cube_reshaped = np.zeros((nframes, 2, 1024, 1024))
-    cube_reshaped[:, 0] = cube[:, :, 0:1024]
-    cube_reshaped[:, 1] = cube[:, :, 1024:]
-    return cube_reshaped
-
-
-def reshape_IRDIS_image(img):
-    nimg = np.zeros((2, 1024, 1024))
-    nimg[0] = img[:, 0:1024]
-    nimg[1] = img[:, 1024:]
-    return nimg
-
-
-def prepare_dataframe(file_table):
-    header_keys = collections.OrderedDict([
-        ('TARG_ALPHA', 'TEL TARG ALPHA'),
-        ('TARG_DELTA', 'TEL TARG DELTA'),
-        ('TARG_PMA', 'TEL TARG PMA'),
-        ('TARG_PMD', 'TEL TARG PMD'),
-        ('DROT2_RA', 'INS4 DROT2 RA'),
-        ('DROT2_DEC', 'INS4 DROT2 DEC'),
-        ('INS4 DROT2 MODE', 'INS4 DROT2 MODE'),
-        ('DROT2_BEGIN', 'INS4 DROT2 BEGIN'),
-        ('DROT2_END', 'INS4 DROT2 END'),
-        ('GEOLAT', 'TEL GEOLAT'),
-        ('GEOLON', 'TEL GEOLON'),
-        ('ALT_BEGIN', 'TEL ALT'),
-        ('EXPTIME', 'EXPTIME'),
-        ('SEQ1_DIT', 'DET SEQ1 DIT'),
-        ('DIT_DELAY', 'DET DITDELAY'),
-        ('NDIT', 'DET NDIT'),
-        ('DB_FILTER', 'INS COMB IFLT'),
-        ('CORO', 'INS COMB ICOR'),
-        ('ND_FILTER', 'INS4 FILT2 NAME'),
-        ('SHUTTER', 'INS4 SHUT ST'),
-        ('IFS_MODE', 'INS2 COMB IFS'),
-        ('PRISM', 'INS2 OPTI2 ID'),
-        ('LAMP', 'INS2 CAL'),
-        ('DPR_TYPE', 'DPR TYPE'),
-        ('DPR_TECH', 'DPR TECH'),
-        ('DET_ID', 'DET ID'),
-        ('SEQ_ARM', 'SEQ ARM'),
-        ('DEROTATOR_MODE', 'INS4 COMB ROT'),
[…]
