### 2026-07-30 779d71f
test: repair the coronagraph-transmission tests and scope collection to tests/
`_resolve_coronagraph_transmission` gained an `observation` argument when the
IRDIS instrument dispatch landed, but the three resolve tests were not updated
and had been failing with a TypeError since. They now pass a minimal observation
stub, plus a new case pinning that the IFS and IRDIS curves are actually
different and each matches its packaged file.

`testpaths` restricts a bare `pytest` to `tests/`; without it collection also
pulled in the driver scripts under `examples/`, which fail at import.
