+++ pyproject.toml
@@ -133,6 +133,9 @@ omit = [
 ]
 
 [tool.pytest.ini_options]
+# Without this, a bare `pytest` also collects driver scripts under examples/ and
+# any scratch test_*.py in the working tree, which fail at import.
+testpaths = ["tests"]
 addopts = "--log-format='%(asctime)s %(levelname)s %(message)s' --log-date-format='%Y-%m-%d %H:%M:%S'"
 markers = [
     "remote_data: tests that require a network connection to external services",
