### 2026-07-30 6d33398
docs(changelog): consolidate the unreleased entries into a 3.0.0 section
The unreleased block had grown to 50 kB across three duplicated sets of
Added/Changed/Removed/Fixed headings, with entries that superseded each other
still listed separately -- the trap floor appeared as both 1.3.1 and 2.0.0, the
ivar bad-pixel threshold as both its introduction at 0.2 and its change to 0.0,
and the IRDIS work as six "Phase N" bullets. Entries are now merged per topic and
stated at their final behaviour, one to two sentences each, with the measurement
detail and internal-branch references left to the comm
[…]
