+++ CHANGELOG.md
@@ -8,281 +8,280 @@ This project follows [Semantic Versioning](https://semver.org/) and the [Keep a 
 
 ## [Unreleased]
 
+---
+
+## [3.0.0] – End-to-End IRDIS Reduction (2026-07-30)
+
+Major release. IRDIS dual-band imaging is now reduced end to end, mirroring the
+IFS workflow, and the database ships with an automated build/update workflow.
+Breaking: the `overwrite_*` step flags are replaced by `force`, `trap >= 2.0.0`
+is required, and the monitoring scripts changed their column and flag names.
+
 ### ✨ Added
-- **51 Eri IFS astrometry regression test + frozen baseline** –
-  `tests/test_51eri_ifs_astrometry_regression.py` (`-m regression`) guards the IFS
-  half of the 2015-09-24 IRDIFS night against
-  `tests/data/51eri_ifs_baseline_{overall_validated_companion_detections,per_channel_astrometry}.csv`.
-  Five checks: drift from the frozen position, agreement with the GRAVITY ground
-  truth at 7.46 mas/px, that the reported astrometry still comes from the template
-  collapse rather than the gated-off per-channel override, σ self-consistency, and
-  that `n_templates_above_threshold` matches the per-template tables on disk.
-  **Why:** the IRDIS sibling test only covered DB_K12, so nothing pinned the IFS
-  conclusions (collapse ≥ per-channel; 7.46 mas/px needs no revision) that
-  `tests/data/51eri_astrometry_benchmark.md` §8 had reached — measured 453.30 ±
-  3.83 mas, 0.53σ from GRAVITY ([@m-samland](https://github.com/m-samland)).
-- **`ivar_bad_pixel_frame_fraction` (both configs, default `0.5`)** – exposes the
-  previously-hardcoded frame collapse in `run_trap`: TRAP's bad-pixel mask is
-  2-D per wavelength, so the per-frame ivar flags are collapsed to "bad in more
-  than this fraction of frames". Default `0.5` is unchanged behaviour. **Why:**
-  since the charis variance-propagating resample ([#42](https://github.com/PrincetonUniversity/charis-dep/issues/42))
-  makes charis's per-frame flags exact `ivar == 0`, measuring them on a fresh
-  51 Eri OBS_H extraction showed the flagging is overwhelmingly transient (~82%
-  of the interior flagged in ≥1 of 256 frames, only ~0.1% in all, median spaxel
-  bad in ~3% of frames), so masking "bad in any frame" would erase ~82% of the
-  regressor pool. `0.5` keeps sigma-clipped cosmics out of the mask while still
-  excluding persistently damaged spaxels; the knob lets a dataset tighten
-  (`0.25` ≈ 2× the mask) or loosen without a code change
+- **End-to-end IRDIS dual-band imaging reduction** – IRDIS DBI observations now run the
+  full chain — download, master calibrations (background, DIT-resolved flat, bad-pixel
+  map), preprocessing into `converted/` cubes with analytic inverse variance, waffle-spot
+  centering, flux-PSF and spot photometry calibration, and TRAP post-processing with
+  spectral template matching — through the same `execute_targets()` entry point and the
+  same resume/force semantics as IFS. Configuration lives in `IRDISReductionConfig`
+  (with `IRDISCalibrationConfig` / `IRDISPreprocessConfig`); `examples/irdis_reduction_template.py`
+  is the reference driver. Star centers for dithered, non-continuous-waffle sequences are
+  propagated from the CENTER waffle fits through the DMS header offsets using a global
+  anchor averaged over all CENTER frames. Modes without a dual-band filter pair
+  (broadband, narrow-band, dual-polarization) fall back to TRAP's contrast-curve and
+  candidate-extraction path instead of template matching. Validated end to end against
[…]
