### 2026-07-30 0b2dd56
docs(readme): document IRDIS as a first-class end-to-end pipeline
The README still described IRDIS as database discovery and download only. It now
covers the full reduction -- calibration, pre-processing, astrometric and
photometric calibration, TRAP post-processing with template matching -- states
what it is based on and that it was validated against the published 51 Eri b
photometry and astrometry, and is explicit about which IRDIS modes are out of
scope.

Also adds the missing `plot_trap_mosaics` script, documents the astrometry
regression tests, and merges the two duplicate Zenodo-download
[…]
