+++ README.md
@@ -2,7 +2,7 @@
   <img src="assets/logos/spherical_banner.png" alt="Spherical - VLT/SPHERE Data Analysis Tools" width="65%">
 </p>
 
-# spherical: VLT/SPHERE Observation Database and IFS Data Analysis Pipeline
+# spherical: VLT/SPHERE Observation Database and IFS + IRDIS Data Analysis Pipeline
 
 [![Python Version](https://img.shields.io/badge/Python-3.11%20%7C%203.12%20%7C%203.13-brightgreen.svg)](https://github.com/m-samland/spherical)
 [![License](https://img.shields.io/badge/License-BSD--3-blue.svg)](https://opensource.org/licenses/BSD-3-Clause)
@@ -17,7 +17,7 @@
 
 **spherical** provides a **curated observation database** for VLT/SPHERE, constructed by parsing headers from all available SPHERE data in the ESO archive and cross-matching with external catalogs. This database systematically identifies SPHERE observation sequences and summarizes observing modes, target properties, integration times, and observing conditions. Pre-built database tables are available for download from [Zenodo](https://doi.org/10.5281/zenodo.15147730) and can be updated with newer data using **spherical** itself (see [Database Tables](#database-tables)).
 
-From exploration to execution, you can filter sequences of interest and feed them into an **end-to-end IFS pipeline** that automates discovery, download, calibration, post-processing, and spectral characterization. For **IRDIS**, the package supports discovery and download of dual-band imaging (DBI) and polarimetric imaging observations.
+From exploration to execution, you can filter sequences of interest and feed them into **end-to-end reduction pipelines for both IFS and IRDIS dual-band imaging (DBI)**. Both automate the same chain — discovery, download, calibration, pre-processing, post-processing, and photometric/astrometric characterization of detected companions — behind one configuration model and one entry point. Polarimetric (DPI) and sparse-aperture-masking observations are covered by the database for discovery and download, but are not reduced by the pipeline.
 
 > ⚠️ **Note:** The database contains **metadata only**. It does **not** include already reduced data products.
 
@@ -28,7 +28,7 @@ From exploration to execution, you can filter sequences of interest and feed the
 **spherical** is both:
 
 1. **A curated observation database** — metadata about VLT/SPHERE observations in the ESO archive (not the reduced data).
-2. **A Python-based IFS reduction and analysis pipeline** — intended to be run from a **Python script** so that you can tune parameters and control each step.
+2. **Python-based IFS and IRDIS reduction and analysis pipelines** — intended to be run from a **Python script** so that you can tune parameters and control each step.
 
 You can use **spherical** in two complementary ways:
 
@@ -36,7 +36,9 @@ You can use **spherical** in two complementary ways:
   Import in Python or Jupyter to query/filter the database and to call pipeline components programmatically.
 
 - **As a script-driven workflow**  
-  Start from the provided example script (`ifs_reduction_template.py`) to configure and run a complete IFS reduction with sensible defaults and many tunable parameters.
+  Start from the provided example scripts (`ifs_reduction_template.py`, `irdis_reduction_template.py`) to configure and run a complete reduction with sensible defaults and many tunable parameters.
+
[…]
