### 2026-07-30 a4a17e1
fix(lint): clear the ruff errors CI sees outside src/
CI runs `ruff check .` while the pixi `lint` task only covered src/, so an
unused assignment in the download tests and a mid-cell import in the database
notebook were never caught locally. Widen the task to src/ + tests/; matching CI
verbatim is not workable while untracked scratch files sit in the tree.
