### 2026-07-30 4dd6ae0
fix(irdis): hand the CPU budget to TRAP
`set_ncpu()` only reached the reduction steps: `IRDISReductionConfig` had no
`apply_trap_resources()`, so TRAP kept its own `TrapReductionConfig.ncpus`
default of 4 however many cores were requested. On an MPCDF batch asking for 46
that is an order of magnitude of wall time.

The call has to precede the `trap_config.reduction.merge(...)` block in the
driver: `TrapResources.apply()` rewrites the whole reduction config, including
`scratch_dir`.
