### 2026-07-30 4571078
fix(logging): rebuild a reused logger whose queue listener has stopped
`name_mode_date` is both the output-folder key and the logger name, so two
observations of the same target, band and night share one cached
`logging.Logger`. `get_pipeline_logger()` archived the log files first and only
then took its "already configured" early return, handing back a logger whose
`QueueHandler` fed a queue that the previous observation's
`remove_queue_listener()` had already stopped.

The second observation therefore wrote nothing *and* its folder was left with no
`reduction.jsonlog` at all, so `aggregate_re
[…]
