### 2026-07-30 d3f0205
fix(trap): isolate each observation in the batch loop
`run_trap_on_observation()` guards its body and writes a crash report, but its
prologue — instrument lookup, data/result path construction, `validate_force`,
logger setup, stellar-parameter resolution — runs before that guard. Anything
raising there took the whole batch down with no crash report and left the
log-queue listener running.

Note this cannot help when the driver dies from a signal (an OOM kill leaves no
exception to catch); it covers the failures that are actually catchable.
