+++ tests/test_run_trap_irdis.py
@@ -122,3 +122,64 @@ class TestCoronagraphResolution:
         assert run_trap._resolve_coronagraph_transmission(
             reduction_config, trap_reduction_config, observation
         ) is None
+
+
+class TestBatchErrorIsolation:
+    """One target dying must not take the rest of the batch with it.
+
+    `run_trap_on_observation` guards its own body, but its prologue (instrument
+    lookup, path construction, `validate_force`, logger setup) runs before that
+    guard, so the batch loop needs its own.
+    """
+
+    @staticmethod
+    def _observation(main_id):
+        observation = MagicMock()
+        observation.observation = {
+            "MAIN_ID": [main_id],
+            "FILTER": ["DB_K12"],
+            "NIGHT_START": ["2015-09-24"],
+        }
+        return observation
+
+    def test_failing_observation_does_not_abort_the_batch(self, monkeypatch):
+        from spherical.pipeline import run_trap
+
+        first, second = self._observation("51 Eri"), self._observation("bet Pic")
+        seen = []
+
+        def fake_run(observation, **_kwargs):
+            seen.append(observation)
+            if observation is first:
+                raise RuntimeError("prologue blew up before the logger existed")
+
+        monkeypatch.setattr(run_trap, "run_trap_on_observation", fake_run)
+
+        run_trap.run_trap_on_observations(
+            observations=[first, second],
+            trap_config=MagicMock(),
+            reduction_config=MagicMock(),
+            species_database_directory="/tmp/species",
+        )
+
+        assert seen == [first, second]
+
+
+class TestDescribeObservation:
+    def test_builds_target_band_night_label(self):
+        from spherical.pipeline.run_trap import _describe_observation
+
+        observation = MagicMock()
+        observation.observation = {
+            "MAIN_ID": ["51 Eri"],
+            "FILTER": ["DB_K12"],
+            "NIGHT_START": ["2015-09-24"],
+        }
+        assert _describe_observation(observation) == "51_Eri/DB_K12/2015-09-24"
+
+    def test_missing_columns_do_not_raise(self):
+        from spherical.pipeline.run_trap import _describe_observation
+
+        observation = MagicMock()
+        observation.observation = {}
+        assert _describe_observation(observation) == "?/?/?"
