### 2026-07-30 ab95cdf
fix(examples): correct the IRDIS template's stellar-parameter assignment
`trap_config.detection.stellar_parameters = StellarParameters(...)` replaces the
object the factory built rather than adjusting it, so with every field commented
out it silently reinstates the library defaults — a no-op for
`trap_config_for_irdis()` today, but it would discard the values
`trap_config_for_beta_pic()` sets. Use `.merge()`, which is a genuine no-op when
nothing is overridden.

Also document `scratch_dir` (unset means /dev/shm, i.e. RAM that a signal-killed
run leaks), state what `verbose` actually does — it 
[…]
