### 2026-07-30 87d35fa
docs(examples): document scratch_dir in the IFS template
Same gap as the IRDIS template: unset means /dev/shm, i.e. RAM that a
signal-killed run leaks with no cleanup.
