### 2026-07-30 08dbe60
docs(examples): expose the optional TRAP inputs in the IFS template
`pass_center_outliers_as_bad_frames_to_trap` and
`pass_amplitude_modulation_to_trap` were reachable only by reading
pipeline_config.py, though the IRDIS template has surfaced them since Phase 6
and the second one's rationale names Beta Pic — the target this template
selects. Both default to False on the config; the template sets them True
because its reference targets are continuous-waffle.
