+++ examples/ifs_reduction_template.py
@@ -78,6 +78,16 @@ table_of_observations = Table.read(
 table_of_files = Table.read(
     database_directory / f"table_of_files_{instrument}.csv")
 
+# ===== OPTIONAL TRAP INPUTS =====
+# Extra per-frame information handed to trap. Each flag is safe to leave on for
+# observations that don't produce the corresponding file (INFO log + no-op
+# fallback), but explicit is better for reproducibility. The two waffle flags
+# default to False on the config; the template turns them on because the
+# reference targets are continuous-waffle.
+config.pass_inverse_variance_to_trap = True                # ivar cube (config default True); improves noise weighting
+config.pass_center_outliers_as_bad_frames_to_trap = True   # union of per-channel waffle-fit outliers → trap bad_frames (continuous-waffle only)
+config.pass_amplitude_modulation_to_trap = True            # continuous-waffle only; loads spot_amplitude_variation.fits
+
 # =================== TRAP CONFIG ===================
 # Create the main TRAP configuration object using the new framework
 trap_config = trap_config_for_ifs()
