### 2026-07-30 1dbe77f
docs(examples): keep the IFS template on the config's TRAP-input defaults
Exposing the two waffle knobs should not also change what the reference driver
does. They stay False, matching pipeline_config, with the opt-in documented.
