### 2026-07-31 70ce6c8
fix(trap): surface TRAP's own logs and forward the candidate-search knobs
Debugging the two IRDIS detection crashes was harder than it needed to be:
HD_140408's trap_reduction.log is 108 lines, 96 of them the traceback. Nothing
here configures the `trap` logger tree, so its records died at
logging.lastResort's WARNING gate — and processing.verbose=True does not help,
since that only adds a stdout handler to the pipeline logger. Bridged for the
lifetime of each target, restoring the tree's level and handlers afterwards.
Library records carry no target/band/night/step/status, so
aggregate_reduct
[…]
