+++ src/spherical/pipeline/logging_utils.py
@@ -8,6 +8,7 @@ import platform
 import socket
 import sys
 import time
+from contextlib import contextmanager
 from datetime import datetime
 from logging.handlers import QueueHandler, QueueListener, RotatingFileHandler
 from multiprocessing import Queue
@@ -20,7 +21,12 @@ except ImportError:
 
     JsonFormatter = jsonlogger.JsonFormatter  # v2
 
-__all__ = ["get_pipeline_logger", "install_queue_listener", "remove_queue_listener"]
+__all__ = [
+    "bridge_library_logger",
+    "get_pipeline_logger",
+    "install_queue_listener",
+    "remove_queue_listener",
+]
 
 # One listener per interpreter
 _listener: QueueListener | None = None
@@ -133,6 +139,55 @@ def get_pipeline_logger(name: str,
         _listener.start()
 
     return logger
+
+
+@contextmanager
+def bridge_library_logger(pipeline_logger, library_name="trap", level=logging.INFO):
+    """Route a third-party library's log records into the per-target log files.
+
+    `trap` logs its progress on its own logger tree, which the pipeline never
+    configures: nothing in this package adds a root handler, so those records
+    die at `logging.lastResort` (level WARNING). The practical effect is that a
+    target that runs for hours writes a `trap_reduction.log` containing only the
+    pipeline's own bookend messages — and, when it crashes, a traceback with no
+    indication of what it was doing. `processing.verbose=True` does not help,
+    because that flag only adds a stdout handler to the *pipeline* logger.
+
+    Parameters
+    ----------
+    pipeline_logger : logging.Logger or logging.LoggerAdapter
+        The per-target logger from `get_pipeline_logger`; its queue handler is
+        borrowed so library records land in the same files.
+    library_name : str
+        Root name of the library's logger tree.
+    level : int
+        Level to force on that tree for the duration.
+
+    Notes
+    -----
+    Library records carry none of the `target`/`band`/`night`/`step`/`status`
+    fields, so `aggregate_reduction_status` ignores them — they enrich the
+    human-readable log without polluting status aggregation.
+    """
+    logger = getattr(pipeline_logger, "logger", pipeline_logger)
+    queue_handlers = [h for h in logger.handlers if isinstance(h, QueueHandler)]
+
+    library_logger = logging.getLogger(library_name)
+    previous_level = library_logger.level
+    added = []
+    try:
+        for handler in queue_handlers:
+            if handler not in library_logger.handlers:
+                library_logger.addHandler(handler)
+                added.append(handler)
+        # NOTSET would inherit root's WARNING and drop exactly the progress
+        # messages this bridge exists to capture.
+        library_logger.setLevel(level)
+        yield library_logger
+    finally:
+        for handler in added:
+            library_logger.removeHandler(handler)
+        library_logger.setLevel(previous_level)
 
 
 def install_queue_listener():  # for tests that run without pipeline
