### 2026-08-05 5fb25f3
build(deps): re-attribute the numpy cap to numba and widen it to <2.5
The pipeline envs pinned numpy<=2.2 with a comment blaming charis. charis
declares numpy>=2,<3 and does not cap at 2.2, so the comment was wrong -- but
the constraint is not vestigial. Dropping it makes conda resolve numpy 2.5.1,
after which uv tries to satisfy numba (numpy<2.5, pulled in transitively by
trap) from PyPI rather than accepting the conda-provided numpy, backtracks into
a numpy sdist that refuses Python >=3.10, and fails the linux-aarch64 solve.
Verified by re-locking from scratch on both manifests.

Widened to 
[…]
