### 2026-08-05 0bfc8e3
docs(changelog): fold the dependency entries into 3.0.0 and flag it unreleased
3.0.0 is still the open release/3.0.0 PR, not a shipped release, so opening a
separate [Unreleased] section above it split one upcoming release across two
headings. The entries move into 3.0.0's Changed section and are cut to the
length of their neighbours.

The 3.0.0 heading now says UNRELEASED and [Unreleased] carries a note pointing
new entries at 3.0.0 until the PR merges, so the same mistake is harder to
repeat.
