+++ CHANGELOG.md
@@ -8,23 +8,13 @@ This project follows [Semantic Versioning](https://semver.org/) and the [Keep a 
 
 ## [Unreleased]
 
-### 🔧 Changed
-- **Corrected the numpy cap in the pipeline envs: `<=2.2` → `<2.5`.** The old bound was
-  attributed to charis, which in fact declares `numpy>=2,<3` — the real constraint comes
-  from **numba**, pulled in transitively by trap, which caps `numpy<2.5`. Removing the cap
-  entirely does not work: conda then resolves numpy 2.5.x, uv tries to satisfy numba from
-  PyPI instead of accepting the conda-provided numpy, and backtracks into a numpy sdist
-  that refuses Python ≥3.10, failing the `linux-aarch64` solve. Widening to numba's actual
-  bound keeps all four platforms solvable and moves the `dev` env from numpy 2.2.0 to
-  2.4.6 ([@m-samland](https://github.com/m-samland)).
-- **`pixi.toml` core dependencies re-synced with `pyproject.toml`.** `python-json-logger`
-  is a runtime dependency in `[project.dependencies]` but was missing from the conda
-  mirror, so it was only ever present as a transitive pip install
-  ([@m-samland](https://github.com/m-samland)).
+> **Nothing goes here yet.** 3.0.0 below is **not released and not merged** — it is still
+> the open `release/3.0.0` PR. Add new entries to the 3.0.0 section, not here, until that
+> PR lands; only then does `[Unreleased]` reopen for 3.1.0 work.
 
 ---
 
-## [3.0.0] – End-to-End IRDIS Reduction (2026-07-30)
+## [3.0.0] – End-to-End IRDIS Reduction (UNRELEASED — pending merge of `release/3.0.0`)
 
 Major release. IRDIS dual-band imaging is now reduced end to end, mirroring the
 IFS workflow, and the database ships with an automated build/update workflow.
@@ -152,6 +142,13 @@ is required, and the monitoring scripts changed their column and flag names.
   ([@m-samland](https://github.com/m-samland)).
 
 ### 🔄 Changed
+- **numpy cap in the pipeline envs widened to `<2.5`** – The old `<=2.2` bound was
+  attributed to charis, which actually declares `numpy>=2,<3`; the real cap is numba's,
+  pulled in transitively by trap. Dropping it entirely fails the `linux-aarch64` solve
+  ([@m-samland](https://github.com/m-samland)).
+- **`python-json-logger` added to `pixi.toml`** – A runtime dependency in
+  `[project.dependencies]` that was missing from the conda mirror
+  ([@m-samland](https://github.com/m-samland)).
 - **The pipeline resumes by default; `force` replaces the `overwrite_*` flags** – Enabled
   steps whose outputs already exist on disk are skipped, so re-running over a growing
   target list is cheap and adding new targets just works. Recomputation is opt-in via the
