### 2026-08-05 7b9e005
fix(tests): guard the pipeline-extra imports so CI can collect again
CI installs `.[test]`, which has neither scipy nor photutils. Five test
modules imported them at import time without a guard, so collection aborted
with `Interrupted: 5 errors` and exit code 2 — meaning the database coverage
CI exists to provide had been silently absent, not merely incomplete.

Use the module-level `pytest.importorskip` that test_coronagraph_transmission.py
already establishes. Verified in a `pip install ".[test]"` venv: 389 tests
collect, 0 errors. The pixi `test` env is not a faithful proxy here — it pulls

[…]
