+++ tests/test_find_star_warning.py
@@ -4,6 +4,11 @@ from __future__ import annotations
 import warnings
 
 import numpy as np
+import pytest
+
+# scipy arrives with the pipeline extra only; skip cleanly in the CI `test` env
+# instead of erroring at collection.
+pytest.importorskip("scipy")
 
 from spherical.pipeline.steps.find_star import guess_position_psf
 
