### 2026-08-05 104f650
test: split the suite by subject into directories, scope cost with markers
A full run took 19m 04s, so it was not run; scoping fell back to a
hand-maintained --ignore list rebuilt from memory each session, which is how
the 15-minute test_resume.py kept landing in supposedly pipeline-only runs.

Subject is now the directory (pipeline / database / regression) and cost is a
marker. A directory choice is visible in review; a missing pytestmark is not,
which is why this replaces the marker-plus-collection-guard approach.

Deriving the network set from imports does not work: test_database.py spends

[…]
