### 2026-08-05 2492d07
test(database): replay the ESO resume tests from cassettes
test_resume.py cost ~15 min against the live archive, so it was deselected by
default, so the resume/incremental logic was covered by nothing anyone ran. Its
failures were also as likely to be archive rate-limiting as regressions.

Recorded with pytest-recording over vcrpy. Five tests now replay offline in
1.85s; test_live_archive_still_responds stays live behind -m remote_data as the
canary for API drift.

The window moved to the night of 2020-03-07: 6 files of a single target,
extending to 8. Header retrieval is one GET per DP.ID, so
[…]
