+++ tests/database/cassettes/README.md
@@ -1,0 +1,121 @@
+# Recorded ESO archive traffic
+
+These cassettes let `test_resume.py` exercise the resume/incremental logic in the default
+offline suite. Before them the tests cost ~15 minutes against the live archive, so they
+were deselected by default — which meant the resume logic was covered by nothing anyone
+actually ran, and its failures were as likely to be rate-limiting as regressions.
+
+Recorded with [`pytest-recording`](https://github.com/kiwicom/pytest-recording) (a thin
+plugin over `vcrpy`). Matching and header-filtering are configured by the `vcr_config`
+fixture in `../conftest.py`.
+
+## Why the window is what it is
+
+`make_file_table` retrieves headers with **one HTTP GET per DP.ID**
+(`archive.eso.org/hdr?DpId=...`), so the number of files in the date range *is* the cost —
+of the tests, and of these files.
+
+The night of **2020-03-07** holds 6 IFS files of a single target (`A124420`); extending to
+2020-03-09 adds 2 `WAVE,LAMP` calibrations. That is 6 and 8 requests, against 171 and 305
+for the 2016-09-15 window this replaced. Six files also halve cleanly for
+`test_resume_with_simulated_partial_file`.
+
+**ESO's `stime`/`etime` are night-based, not calendar-based.** `[2020-03-07, 2020-03-08]`
+selects the night beginning 2020-03-07, whose exposures carry `2020-03-08` `DATE_OBS`
+values. Verified against the live archive: `[2016-09-15, 2016-09-17]` returns 305 files,
+which is `NIGHT_START` 2016-09-15 (171) + 2016-09-16 (134), not any calendar-date sum.
+Keep this in mind when changing the window — picking dates off `DATE_OBS` will select the
+wrong night.
+
+## Re-recording
+
+Needed when the ESO query interface changes, when `make_file_table` changes which requests
+it issues, or when `test_live_archive_still_responds` (the live canary, `-m remote_data`)
+starts failing.
+
+> ### A partially-recorded cassette is silently green
+>
+> This is the trap, and it cost two wasted recordings here. `query_eso_data` catches
+> request failures, logs an ERROR, and returns an empty list. So if the archive resets a
+> connection mid-recording, the run still "succeeds" — it just writes a cassette with
+> interactions missing. On replay those tests then pass against empty results, because
+> most of them only assert `>= n_first` or compare two tables that are both empty.
+>
+> Two consequences:
+>
+> 1. **Nothing else may touch the ESO archive while recording.** Concurrent queries from
+>    another shell are enough to trigger the resets. Record with the archive to yourself.
+> 2. **Verify interaction counts afterwards — do not trust a green run.** Every broken
+>    recording here still reported five passing tests; one of them "passed" in 18.78s with
+>    cassettes that could never have worked on another machine. Counting is the only check
+>    that caught it.
+>
+>    | test | form GET | wdb queries | header GETs | total |
+>    |---|---|---|---|---|
+>    | `test_fresh_run_creates_output_no_partial` | 1 | 2 | 6 | 9 |
+>    | `test_incremental_extends_date_range` | 1 | 4 | 8 | 13 |
+>    | `test_resume_with_simulated_partial_file` | 1 | 4 | 8 | 13 |
+>    | `test_resume_false_ignores_partial` | 1 | 2 | 6 | 9 |
+>    | `test_idempotent_rerun` | 1 | 2 | 6 | 9 |
+>
+>    Two counts that look wrong but are not: the extra GET is astroquery fetching the WDB
+>    query *form* once per session, and `test_idempotent_rerun` has 2 queries rather than 4
+>    because its second `make_file_table` call is served by astroquery's own cache, which
+>    the `_isolate_astroquery_cache` fixture makes cold-per-test and therefore reproducible.
+
+```sh
[…]
