### 2026-08-05 b2b4a0f
fix(database): make cache=False actually disable the archive query cache
make_file_table's `cache` argument reached Eso.get_headers but never
Eso.query_instrument, so astroquery's default cache=True applied to the queries
regardless. A run explicitly asking the archive for fresh data resolved *which
files exist* from a stale on-disk cache while fetching those files' headers
live — meaning files added to an already-queried night stayed invisible to an
update no matter what the caller passed.

Thread `cache` through query_eso_data to both call sites.

This increases the number of queries a full 
[…]
