### 2026-08-05 d0d7327
fix(database): silence MergeConflictWarning from the SIMBAD vstack
Each batched SIMBAD TAP result carries its own 'ID'/'name' in .meta, so
vstack warned once per batch — thousands of lines on a full build, enough
to bury the real log output. file_table.py already guards the equivalent
vstack inside Eso.get_headers; this applies the same idiom to the target
table. Nothing downstream reads those meta keys.
