### 2026-08-05 fac702e
docs(changelog): trim the test-infrastructure entries in 3.0.0
The suite reorg and cassette work had grown their own '### Changed'
section — a duplicate heading — and between them ran to 27 lines, about
twice the entry for the end-to-end IRDIS reduction this release is named
after. Test plumbing should not outweigh the headline feature.

Merges the duplicate section into the existing Changed, condenses the two
entries into one, folds the SIMBAD MergeConflictWarning fix into the
existing 'Cosmetic warnings silenced' entry it duplicated, and moves the
CI-collection fixes below the user-facing on
[…]
