+++ CHANGELOG.md
@@ -134,11 +134,11 @@ is required, and the monitoring scripts changed their column and flag names.
   interrupted run is continued simply by re-running it; already-fetched DP.IDs are
   skipped. New `resume` parameter, default `True`
   ([@m-samland](https://github.com/m-samland)) ([#105](https://github.com/m-samland/spherical/issues/105)).
-- **51 Eridani astrometry regression tests** – `tests/test_51eri_astrometry_regression.py`
-  and `tests/test_51eri_ifs_astrometry_regression.py` (opt in with `-m regression`) pin
-  the reported companion position of the 2015-09-24 IRDIS DB_K12 and IFS OBS_H datasets
-  against frozen baselines in `tests/data/`, and check agreement with the published
-  GRAVITY position. `tests/data/51eri_astrometry_benchmark.md` documents the comparison
+- **51 Eridani astrometry regression tests** – `tests/regression/` (opt in with
+  `-m regression`) pins the reported companion position of the 2015-09-24 IRDIS DB_K12 and
+  IFS OBS_H datasets against frozen baselines in `tests/regression/data/`, and checks
+  agreement with the published GRAVITY position.
+  `tests/regression/data/51eri_astrometry_benchmark.md` documents the comparison
   ([@m-samland](https://github.com/m-samland)).
 
 ### 🔄 Changed
@@ -206,10 +206,16 @@ is required, and the monitoring scripts changed their column and flag names.
 - **README rewritten** – Documents the IRDIS reduction as a first-class workflow alongside
   IFS, restructures installation into pip and pixi options, and consolidates the database
   table information into one section ([@m-samland](https://github.com/m-samland)).
-- **`pytest` collects only `tests/`** – `testpaths` is now set, so a bare `pytest` no
-  longer tries to import the driver scripts under `examples/` (or any scratch `test_*.py`
-  in the working tree) and fail at collection
-  ([@m-samland](https://github.com/m-samland)).
+- **Test suite reorganised by subject** – Tests now live in `tests/pipeline/`,
+  `tests/database/` and `tests/regression/`, with cost expressed as markers (`remote_data`,
+  `regression`, and a new `slow`) instead of a hand-maintained `--ignore` list. `addopts`
+  deselects the expensive markers, so a bare `pytest` is offline and fast (~45s, previously
+  19 minutes), and `testpaths` keeps collection out of `examples/`. New pixi tasks
+  `test-pipeline`, `test-database`, `test-network`, `test-regression` and `test-all` select
+  by path. The ESO resume tests replay from `pytest-recording` cassettes in
+  `tests/database/cassettes/` instead of costing ~15 minutes against the live archive;
+  `test_live_archive_still_responds` stays live behind `-m remote_data` as the API-drift
+  canary ([@m-samland](https://github.com/m-samland)).
 
 ### 🗑️ Removed
 - **`PipelineStepsConfig.overwrite_calibration` / `overwrite_bundle` /
@@ -228,62 +234,14 @@ is required, and the monitoring scripts changed their column and flag names.
   no callers and no longer imported on scikit-image ≥ 0.19
   ([@m-samland](https://github.com/m-samland)).
 
-### 🔧 Changed
-- **The ESO resume tests now replay from cassettes and run in the default suite** –
-  `tests/database/test_resume.py` cost ~15 min against the live archive, so it was
-  deselected, so the resume/incremental logic was covered by nothing anyone actually ran.
-  Recorded with `pytest-recording`/`vcrpy` into `tests/database/cassettes/`; the five tests
-  now replay offline in **1.85s**, and `test_live_archive_still_responds` remains live
-  behind `-m remote_data` as the canary for API drift. The recording window moved from
[…]
