### 2026-08-05 bbf66d4
feat(scripts): resolve the database directory from $SPHERICAL_DATABASE_DIR
The tables live in one directory that every entry point needs, and each was
told separately: --dest on both database CLIs, --database-dir on
plot_trap_mosaics, and a hardcoded ~/data/sphere/database in both reduction
templates. Only plot_trap_mosaics honoured the environment variable, and it
owned the constant.

spherical/database/paths.py is now the single source of truth, with precedence
explicit -> $SPHERICAL_DATABASE_DIR -> caller default -> None. It lives in
database/ and stays stdlib-only because sync_zenodo_table
[…]
