### 2026-08-05 536297b
feat(scripts): add --instrument to crash_reports
reduction_status has had the filter since the monitoring-script fixes;
crash_reports did not, so narrowing a mixed IFS/IRDIS reduction tree to one
instrument meant reading past the other one. The exception tally is computed
after the filter, so the counts describe what is shown rather than the
unfiltered set.

Reports whose dataset identifier could not be parsed carry instrument
"unknown" and are therefore dropped by any explicit filter; the help text
says so.
