### 2026-08-05 d689810
feat(pipeline): track the center-fit quality assessment module
The module and its example driver sat untracked inside the package, so a
wheel built from a clean checkout did not contain them while the working tree
did. Tracking them settles that mismatch before the 3.0.0 tag.

Nothing in src/ imports it yet -- the only consumer is the example -- so the
open question of whether the metric battery is heavier than what is actually
consumed stays open, now against a tracked file.
