### 2026-08-05 2193789
docs: document $SPHERICAL_DATABASE_DIR and the crash_reports filter
The README explained where to get the tables but never how to tell spherical
where they ended up, and the environment variable was mentioned only in a
trailing clause of the mosaic paragraph. A new "Telling spherical where the
tables are" subsection sits under Database Tables, where someone setting up
will look, with a table of what consumes the variable and what each entry
point does without it.
