### 2026-08-11 672096a
fix(database): let the Zenodo sync reach the v3.0.0 tables
The download set was a hardcoded list of five filenames, so the target and
SAM tables the v3.0.0 record ships would have been invisible to the command
the README recommends. Requiring them outright was not an option either: the
published v2.0.0 record has neither, and sync_tables() aborts on any wanted
file the record lacks, so a hard requirement would have broken the command
for everyone between this commit and the upload.

So the set is now split. File and observation tables stay required; target
tables and database_provenance.json a
[…]
