+++ tests/database/test_sync_zenodo.py
@@ -1,3 +1,4 @@
+import json
 from pathlib import Path
 from unittest.mock import patch
 
@@ -22,6 +23,33 @@ def test_wanted_filenames_all_includes_polarimetry():
 def test_wanted_filenames_ifs_only():
     names = sz._wanted_filenames("ifs", include_polarimetry=False)
     assert all("irdis" not in n for n in names)
+
+
+def test_sam_tables_are_opt_in():
+    without = sz._wanted_filenames("all", include_polarimetry=False)
+    assert not any("_sam" in n for n in without)
+
+    with_sam = sz._wanted_filenames("all", include_polarimetry=False, include_sam=True)
+    assert "table_of_observations_ifs_sam.fits" in with_sam
+    assert "table_of_observations_irdis_sam.fits" in with_sam
+
+
+def test_target_tables_are_optional_not_required():
+    """Requiring them would break syncing against the v2.0.0 record, which has none."""
+    required = sz._wanted_filenames("all", include_polarimetry=True)
+    assert not any(n.startswith("table_of_targets_") for n in required)
+
+    optional = sz._optional_filenames("all", include_polarimetry=True)
+    assert "table_of_targets_ifs.fits" in optional
+    assert "table_of_targets_irdis.fits" in optional
+    assert "table_of_targets_irdis_polarimetry.fits" in optional
+    assert sz.PROVENANCE_NAME in optional
+
+
+def test_optional_filenames_follow_instrument_and_sam_selection():
+    optional = sz._optional_filenames("ifs", include_polarimetry=False, include_sam=True)
+    assert "table_of_targets_ifs_sam.fits" in optional
+    assert not any("irdis" in n for n in optional if n.startswith("table_of_targets_"))
 
 
 def test_checksum_missing_file_is_false(tmp_path):
@@ -70,3 +98,85 @@ def test_dry_run_still_works_without_dest_or_env():
     with patch.object(sz, "sync_tables", return_value=0) as sync:
         assert sz.main(["--dry-run"]) == 0
     assert sync.call_args.kwargs["dest"] == Path(".")
+
+
+V2_FILES = [
+    "table_of_files_ifs.csv",
+    "table_of_files_irdis.csv",
+    "table_of_observations_ifs.fits",
+    "table_of_observations_irdis.fits",
+    "table_of_observations_irdis_polarimetry.fits",
+]
+
+V3_FILES = V2_FILES + [
+    "table_of_targets_ifs.fits",
+    "table_of_targets_irdis.fits",
+    "table_of_targets_irdis_polarimetry.fits",
+    sz.PROVENANCE_NAME,
+]
+
+
+def _fake_record(filenames):
+    return {
+        "metadata": {"version": "v3.0.0"},
+        "files": [{"key": name, "links": {"self": f"https://example.invalid/{name}"}} for name in filenames],
+    }
+
+
+def _run_sync(monkeypatch, dest, filenames, **kwargs):
+    """Drive sync_tables against a fabricated record, writing plausible file content."""
+    monkeypatch.setattr(sz, "_resolve_zenodo_record_id", lambda *a, **k: "999")
+    monkeypatch.setattr(sz, "_load_record", lambda *a, **k: _fake_record(filenames))
+
+    downloaded = []
+
+    def fake_download(url, destination, timeout=None):
+        downloaded.append(destination.name)
+        if destination.name.endswith(".csv"):
+            destination.write_text("NIGHT_START\n2014-05-14\n2026-08-09\n")
+        elif destination.name == sz.PROVENANCE_NAME:
[…]
