### 2026-08-11 9e506ba
Update paper for v3.0.0 of spherical
