+++ joss/paper.bib
@@ -30,24 +30,6 @@ archivePrefix = {ascl},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
 
-@ARTICLE{Holstein:2020,
-       author = {{van Holstein}, R.~G. and {Girard}, J.~H. and {de Boer}, J. and {Snik}, F. and {Milli}, J. and {Stam}, D.~M. and {Ginski}, C. and {Mouillet}, D. and {Wahhaj}, Z. and {Schmid}, H.~M. and {Keller}, C.~U. and {Langlois}, M. and {Dohlen}, K. and {Vigan}, A. and {Pohl}, A. and {Carbillet}, M. and {Fantinel}, D. and {Maurel}, D. and {Orign{\'e}}, A. and {Petit}, C. and {Ramos}, J. and {Rigal}, F. and {Sevin}, A. and {Boccaletti}, A. and {Le Coroller}, H. and {Dominik}, C. and {Henning}, T. and {Lagadec}, E. and {M{\'e}nard}, F. and {Turatto}, M. and {Udry}, S. and {Chauvin}, G. and {Feldt}, M. and {Beuzit}, J. -L.},
-        title = "{Polarimetric imaging mode of VLT/SPHERE/IRDIS. II. Characterization and correction of instrumental polarization effects}",
-      journal = {Astronomy \& Astrophysics},
-     keywords = {polarization, techniques: polarimetric, techniques: high angular resolution, techniques: image processing, methods: observational, protoplanetary disks, Astrophysics - Instrumentation and Methods for Astrophysics, Astrophysics - Earth and Planetary Astrophysics},
-         year = 2020,
-        month = jan,
-       volume = {633},
-          eid = {A64},
-        pages = {A64},
-          doi = {10.1051/0004-6361/201834996},
-archivePrefix = {arXiv},
-       eprint = {1909.13108},
- primaryClass = {astro-ph.IM},
-       adsurl = {https://ui.adsabs.harvard.edu/abs/2020A&A...633A..64V},
-      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
-}
-
 @software{Holstein:2020ascl,
        author = {{van Holstein}, R.~G. and {Girard}, J.~H. and {de Boer}, J. and {Snik}, F. and {Milli}, J. and {Stam}, D.~M. and {Ginski}, C. and {Mouillet}, D. and {Wahhaj}, Z. and {Schmid}, H.~M. and {Keller}, C.~U. and {Langlois}, M. and {Dohlen}, K. and {Vigan}, A. and {Pohl}, A. and {Carbillet}, M. and {Fantinel}, D. and {Maurel}, D. and {Orign{\'e}}, A. and {Petit}, C. and {Ramos}, J. and {Rigal}, F. and {Sevin}, A. and {Boccaletti}, A. and {Le Coroller}, H. and {Dominik}, C. and {Henning}, T. and {Lagadec}, E. and {M{\'e}nard}, F. and {Turatto}, M. and {Udry}, S. and {Chauvin}, G. and {Feldt}, M. and {Beuzit}, J. -L.},
         title = "{IRDAP: SPHERE-IRDIS polarimetric data reduction pipeline}",
@@ -629,3 +611,72 @@ archivePrefix = {arXiv},
 
 
 
+
+@ARTICLE{Wenger:2000,
+       author = {{Wenger}, M. and {Ochsenbein}, F. and {Egret}, D. and {Dubois}, P. and {Bonnarel}, F. and {Borde}, S. and {Genova}, F. and {Jasniewicz}, G. and {Lalo{\"e}}, S. and {Lesteven}, S. and {Monier}, R.},
[…]
