+++ joss/paper.md
@@ -10,65 +10,69 @@ tags:
   - direct imaging
 authors:
   - name: Matthias Samland
-    orcid: 0000-0002-0101-8814
+    orcid: 0000-0001-9992-4067
     affiliation: 1
 
 affiliations:
   - name: Max-Planck-Institut für Astronomie (MPIA), Königstuhl 17, 69117 Heidelberg, Germany
     index: 1
 
-date: 09 August 2025
+date: 11 August 2026
 bibliography: paper.bib
 ---
 
 # Summary
 
-The Spectro-Polarimetric High-contrast Exoplanet REsearch instrument (SPHERE; @Beuzit:2019) at the Very Large Telescope (VLT) is a leading facility for coronagraphic imaging of exoplanets and circumstellar disks in the optical and near-infrared. Over the last decade, SPHERE has contributed to hundreds of publications and major legacy surveys (e.g., SHINE; @Chauvin:2017; @Chomez:2025).
+The Spectro-Polarimetric High-contrast Exoplanet REsearch instrument (SPHERE; @Beuzit:2019) at the Very Large Telescope (VLT) is a leading facility for coronagraphic imaging of exoplanets and circumstellar disks in the optical and near-infrared. Over the last decade, SPHERE has contributed to hundreds of publications and major legacy surveys (e.g., the SpHere INfrared survey for Exoplanets, SHINE; @Chauvin:2017; @Chomez:2025).
 
-`spherical` streamlines the end-to-end analysis of SPHERE data by integrating a curated, searchable database with an automated Python-based pipeline. The software enables users to filter the complete observation history by target properties or observing conditions, automatically download raw datasets from the ESO archive, and reduce Integral Field Spectrograph (IFS) data using the adapted CHARIS pipeline. Beyond basic reduction, the workflow handles astrometric calibration and post-processing via TRAP for companion detection, while also facilitating the discovery and retrieval of IRDIS dual-band and polarimetric imaging data for analysis with community tools.
+`spherical` combines a curated, searchable database of the complete SPHERE observation history with automated reduction pipelines for two of its subsystems. Users filter observations by target properties, observing mode, or observing conditions, download the selected raw data and calibrations from the European Southern Observatory (ESO) archive, and reduce them end to end. Integral Field Spectrograph (IFS; @Claudi:2008) data are extracted into spectral cubes by a pipeline adapted from the Coronagraphic High Angular Resolution Imaging Spectrograph (CHARIS; @Brandt:2017; @Samland:2022); data from the Infra-Red Dual-band Imager and Spectrograph (IRDIS; @Dohlen:2008) in dual-band imaging (DBI) mode are calibrated and pre-processed by a pipeline that mirrors the IFS workflow step for step. Both feed post-processing with TRAP (Temporal Reference Analysis of Planets; @Samland:2021), a detection algorithm that models the temporal systematics of the residual starlight instead of subtracting a reference image, and recovers the astrometry and contrast spectra of detected companions.
 
-As of May 2025, the database includes approximately 6000 IRDIS dual-band imaging (DBI) observations, ~1000 IRDIS dual-beam polarimetric imaging (DPI; @deBoer:2020; @vanHolstein:2020) observations, and ~4500 IFS observations. SPHERE is undergoing the *SPHERE+* upgrade [@Boccaletti:2020; @Boccaletti:2022], including a second-stage adaptive optics system (SAXO+; @Stadler:2022), ensuring long-term scientific relevance and providing a pathfinder for the ELT’s Planetary Camera and Spectrograph (PCS; @Kasper:2021).
[…]
