### 2026-08-11 f5d6483
docs: finalize the 3.0.0 changelog and README for release
The changelog still carried the "not released and not merged" banner and no
release date. Reopen an empty [Unreleased], date 3.0.0, and add the missing
[3.0.0] compare link.

Three Fixed entries described bugs in code introduced in this same cycle, so
no user ever saw the broken state: the provenance eso_coverage_start bug, the
enrichment connection-failure honesty fix, and the MOCAdb host/schema
migration. Fold the surviving behaviour into the corresponding Added entries
and drop the rest, per the 2026-07-30 rule that entries state fin
[…]
