+++ CHANGELOG.md
@@ -8,41 +8,30 @@ This project follows [Semantic Versioning](https://semver.org/) and the [Keep a 
 
 ## [Unreleased]
 
-> **Nothing goes here yet.** 3.0.0 below is **not released and not merged** — it is still
-> the open `release/3.0.0` PR. Add new entries to the 3.0.0 section, not here, until that
-> PR lands; only then does `[Unreleased]` reopen for 3.1.0 work.
-
 ---
 
-## [3.0.0] – End-to-End IRDIS Reduction (UNRELEASED — pending merge of `release/3.0.0`)
+## [3.0.0] - 2026-08-11 – End-to-End IRDIS Reduction
 
 Major release. IRDIS dual-band imaging is now reduced end to end, mirroring the
 IFS workflow, and the database ships with an automated build/update workflow.
 Breaking: the `overwrite_*` step flags are replaced by `force`, `trap >= 2.0.0`
-is required, and the monitoring scripts changed their column and flag names.
+is required, database table filenames now name the mode instead of encoding a
+boolean, and the monitoring scripts changed their column and flag names.
 
 ### ✨ Added
 - **`$SPHERICAL_DATABASE_DIR` names the table directory once for the whole install** – The
-  database tables live in one directory that every entry point needs, and each of them used
-  to be told separately. Setting the variable now supplies it to `spherical-sync-tables` and
-  `spherical-update-database` (whose `--dest` is no longer required), to `plot_trap_mosaics`
-  (which already honoured it), and to both reduction templates, which previously hardcoded
-  `~/data/sphere/database`. An explicit flag or argument always wins, so a one-off run
-  against a different copy needs no unsetting. The precedence lives in one place,
-  `spherical.database.paths.resolve_database_dir()`, which is stdlib-only and so usable from
-  a base install ([@m-samland](https://github.com/m-samland)).
-- **`crash_reports --instrument {ifs,irdis,all}`** – Closes the asymmetry with
-  `reduction_status`, which has had the filter since the monitoring-script fixes. The
-  exception tally follows the filter rather than the unfiltered set
-  ([@m-samland](https://github.com/m-samland)).
+  variable supplies the database directory to `spherical-sync-tables` and
+  `spherical-update-database` (whose `--dest` is therefore optional), to `plot_trap_mosaics`,
+  and to both reduction templates. An explicit flag or argument always wins. The precedence
+  lives in `spherical.database.paths.resolve_database_dir()`, which is stdlib-only and so
+  usable from a base install ([@m-samland](https://github.com/m-samland)).
 - **Candidate-search knobs forwarded to TRAP** – `minimum_candidate_separation`,
   `candidate_exclusion_radius` and `max_candidates` from `trap_config.detection` now reach
   `detection_and_characterization_with_template_matching()`, and both reduction templates
-  document them. They exist because `search_region_inner_bound` must stay small (the inner
-  pixels feed the annulus statistics) while the *detection* floor should not: a residual at
-  1 px separation was being promoted to a candidate and then crashing the target. Forwarded
-  only when the installed TRAP defines them, so a `pipeline` env on an older git checkout
-  degrades rather than raising `TypeError` ([@m-samland](https://github.com/m-samland)).
+  document them. They matter because `search_region_inner_bound` must stay small — the inner
+  pixels feed the annulus statistics — while the *detection* floor should not, so a residual
+  a pixel from the star was being promoted to a candidate and crashing the target. Forwarded
+  only when the installed TRAP defines them ([@m-samland](https://github.com/m-samland)).
 - **End-to-end IRDIS dual-band imaging reduction** – IRDIS DBI observations now run the
[…]
