### 2026-08-12 c54226e
fix: anchor configured directories to absolute paths
TRAP's add_default_templates() chdirs into the species database directory
and never restores the cwd (m-samland/trap#39). A relative reduction
directory therefore resolved against the species directory for everything
written after that point, putting the whole template_matching/ tree in the
wrong place while the run still reported success.

DirectoryConfig normalizes in __setattr__, not just __post_init__: both
reduction templates set the layout by assigning onto an already-built
config, which never re-runs __post_init__, so a constructor-on
[…]
