### 2026-08-12 1652005
test: mark the SpectralTemplate smoke test as remote_data
It initializes its species database in a fresh tmp_path, so add_filter
downloads the SVO filter profiles on every run and the skip guard fires
whenever that fetch fails. It was intermittently skipping in the default
offline suite, which is exactly what the marker exists to prevent, and no
import in the file reveals the network use.
