### 2026-08-12 42a609a
docs: correct the stale rationale on the species path absolutization
The chdir this comment described is gone as of trap#39. The call stays
because species writes the path into its config and hands the value back
to h5py verbatim, and because logging the same value the run uses keeps
crash reports honest.
