### 2026-08-14 8f46014
feat: select targets whose epochs can separate companions from background stars
Telling a companion from a background star needs the host to have moved far
enough between epochs that the two hypotheses predict measurably different
positions. select_multi_epoch_targets() keeps the observations where that
holds: at least two epochs, and a proper motion that displaces a stationary
background object by at least one pixel over the span between the first and
the last of them.

Quality cuts stay with the caller. The span is measured across the
observations that survive them, so applying them afterwar
[…]
