### 2026-08-14 0987a6b
refactor: make the TRAP result-folder layout a public surface
Analysis code outside the reduction has to find TRAP output, and until now the
only ways were a private helper in run_trap or re-encoding the layout at the
call site. The second fails quietly: a consumer that tolerates missing TRAP
output, as any resumable survey must, cannot tell a layout change from an
unfinished reduction.

step_registry is where step layout is already defined and it imports only the
standard library, so the helpers are usable from a base install.
run_trap imports trap and toolbox imports scipy and matplotlib, so
[…]
