### 2026-08-18 3c110de
fix: make DirectoryConfig constructible on interpreters without gh-90562
dataclass(slots=True) cannot add __slots__ in place, so it discards the
decorated class and builds a replacement. A zero-argument super() in the
class body keeps a __class__ cell pointing at the discarded original, and
DirectoryConfig.__setattr__ used that form: every construction raised
"TypeError: obj is not an instance or subtype of type". IFSReductionConfig
and IRDISReductionConfig were unconstructible too, since both default to a
DirectoryConfig.

CPython's fix for this (gh-90562) reached 3.14 and late 3.13 patch rel
[…]
