### 2026-08-18 97b09ea
fix: guard the centre-frame split against an empty coronagraphic sequence
IRDISObservation and IFSObservation indexed frames['CORO'][0] unconditionally
whenever WAFFLE_MODE was False, so any sequence with no coronagraphic frames
raised IndexError during construction.

WAFFLE_MODE is False whenever FLUX wins the PRIMARY_SCIENCE exposure-time
contest, and FLUX wins precisely when CORO is short or absent. The condition
that selects the branch is therefore the same condition that empties the table
it indexes, so the failure was deterministic rather than incidental. It fired
inside retrieve_observa
[…]
