### 2026-08-19 #8690 @manunicholasjacob
@samaloney I saw that finding reviewers has been the hard part here. I am happy to take a slot if one is open.

To be straight about what I would and would not cover: I would review the software rather than the astronomy. `spherical` is a database plus an end-to-end reduction pipeline that pulls from the ESO archive, so the question I would spend most time on is whether someone who is not the author can actually run it end to end: install from a clean environment, the test suite, and how far the documented workflow gets on a machine with no existing SPHERE setup. For a pipeline paper that seems like the substantive check, and it is the kind of thing that is easy to miss when the author's environment already works.

The paper does not make performance claims, so I would not be inventing any to measure. The database counts it does state are concrete and checkable, and I would verify those.

For context, I am currently a reviewer on #11170 (Optiland). My background is computer architecture and edge inference rather than astronomy, and I maintain llama-roofline (10.5281/zenodo.21842493) and ml-systems-lab (10.5281/zenodo.21867055), both open with DOIs.

I am in the JOSS reviewer database as @manunicholasjacob. ORCID 0009-0007-6589-6572.
