### 2026-08-20 #8690 @samaloney
Hi @m-samland
> Should the referee be a first time user of the package or can it be someone who has used it before?

No, there's no requirement on this. I think both viewpoints would bring valid information. As Warrickball mentioned above, due to the specific nature of your field, large consortiums and papers with many co-authors, we may be able to relax the CoI constraints - I'll check this for IainHammond.
