### 2026-08-26 2bd1e39
Include regular TRAP detection map for broadband filters in mosaic
