### 2026-08-27 8699478
Selecting correct number of frames for center plot
