### 2026-08-27 3c1a0c1
Fixing the colors
