### 2026-08-28 46e5a80
Restored plot_center_evolution.py
