### 2026-08-28 cab6e09
Fixing the empty flux plots issue
