+++ src/spherical/pipeline/flux_calibration.py
@@ -293,10 +293,15 @@ def plot_flux_normalization_factors(
         savefig=False, savedir=None):
 
     plt.close()
-    mjd_range = np.max(flux_calibration_indices['flux_lst']) - \
-        np.min(flux_calibration_indices['flux_lst'])
-    scaled_values = (flux_calibration_indices['flux_lst'].values -
-                     np.min(flux_calibration_indices['flux_lst'])) / mjd_range
+
+    flux_lst = flux_calibration_indices['flux_lst'].values
+    lst_range = np.ptp(flux_lst)
+
+    if np.isclose(lst_range, 0):
+        scaled_values = np.full(len(flux_lst), 0.5)
+    else:
+        scaled_values = (flux_lst - np.min(flux_lst)) / lst_range
+
     colors = cmap(scaled_values)
 
     if wavelengths is None:
