### 2026-09-01 7dc907c
Fix trap version pin to 2.0.1
