### 2026-09-01 687b026
Always fit 2D background for satellite spots
