### 2026-09-02 #8690 @samaloney
@xiaziyna I'm looking reviewers for this submission, as a recent JOSS author of a somewhat related paper would you be willing to review?

Our check list style review process means you don't have to complete the entire review in one go, it can be tackled in smaller pieces as time permits.

Please be so kind to let me know even if you're not interested, so that I can start looking for other reviewers?
