### 2026-09-03 #11243 @editorialbot
[REVIEW]: spherical: A Comprehensive Database and Automated Pipeline for VLT/SPHERE High-Contrast Imaging
**Submitting author:** @m-samland (<a href="http://orcid.org/0000-0001-9992-4067">Matthias Samland</a>)
**Repository:** https://github.com/m-samland/spherical
**Branch with paper.md** (empty if default branch): paper
**Version:** v2.1.1
