### 2026-09-03 6719703
Update changelog, add missing test for new filter
