### 2026-09-04 #11243 @manunicholasjacob
@samaloney one quick thing before I go further. The review records v2.1.1, but the repository has since released v3.0.0 and v3.0.1, and the paper describes v3.0.0 explicitly. Which would you like me to review against?
