### 2026-09-04 #11243 @m-samland
I would strongly prefer if you could used v3.0.1 (just a bug fix release I put out today, no functional changes). The paper was on the waitlist for over a year and the original submitted version is very outdated and not compatible with the paper. Not sure how this is changed in the review system though.
