### 2026-09-04 c0b3a1a
Use WAFFLE_MODE to check if CENTER or CORO are used for science
