### 2026-09-05 m-samland/spherical#136 @manunicholasjacob
[JOSS Review] Missing pymysql masks the ID_GAIA_DR3 check
Part of the JOSS review at openjournals/joss-reviews#11243.

On a clean v3.0.1 environment installed with `.[pipeline,test]`, `tests/database/test_mocadb_matching.py` gives 1 failed, 3 passed, 3 skipped under `-m remote_data`. The failure is `test_no_gaia_column_raises`.

The test drops `ID_GAIA_DR3` and expects `ValueError`. In `query_mocadb_for_targets` the `pymysql` import sits at `mocadb_matching.py:235` and warns and returns at 238, while the column check is at 244. Without the extra installed the function returns before it validates, so nothing is raised.

Installing `.[mocadb]` takes the same file to 7 passed, so this is the ordering rather than the test.

The wider effect is that a missing optional dependency silences an input error: pass a table with the wrong column name and you get a warning about MOCAdb instead. I can open a PR moving the column check above the import guard so validation happens either way.
