### 2026-09-05 m-samland/spherical#137 @manunicholasjacob
[JOSS Review] explore_database.ipynb hardcodes the table path
Part of the JOSS review at openjournals/joss-reviews#11243.

The README says to set `$SPHERICAL_DATABASE_DIR` once and no entry point needs telling again. The notebook does not read it. Code cell 3 has `table_path = Path.home() / "data/sphere/database"`, and the variable appears nowhere in the file.

This is Quick Start step 1, so someone who follows the README and puts their tables elsewhere fails on the first thing they run. With the tables at the hardcoded path the notebook runs clean, 23 cells in 20.6 s, so the path is the only problem.

`resolve_database_dir()` already returns the env-var path when set. I can open a PR switching the notebook to it.
