### 2026-09-05 m-samland/spherical#138 @manunicholasjacob
[JOSS Review] Install fails on Windows while the package declares OS Independent
Part of the JOSS review at openjournals/joss-reviews#11243.

`pip install .` on Windows 11 with Python 3.11 fails after 94 s. `healpy` has no Windows wheel, so pip builds it from source, looks for cfitsio, falls back to pykg-config and exits with `error: [WinError 193] %1 is not a valid Win32 application`. `spherical` itself builds fine, healpy is the only failure. The same commit installs on Linux in 6 s.

`pyproject.toml:11` declares `Operating System :: OS Independent`, and CI runs ubuntu-only, so nothing exercises the claim. Narrowing the classifier or a line in the README would settle it either way.
