### 2026-09-05 #11243 @manunicholasjacob
Thanks @m-samland and @samaloney for sorting the version. Reviewed against v3.0.1.

Clean installs on Linux (Python 3.12) and Windows (3.11). Full suite 531 passed, 25 deselected, 64 s. `ruff check .` clean. Pulled the published tables with `spherical-sync-tables`, 7 files and 262 MB, md5-verified, 47 s. Ran `examples/explore_database.ipynb` end to end, exercised the template selection path against the real IRDIS table, and ran the `remote_data` and `regression` suites.

The database numbers in the paper are exact: 6101 IRDIS DBI, 4733 IFS, 1205 DPI, 355 SAM as 203 plus 152, coverage 2014-05 to 2026-08. The Zenodo tables are past 640 downloads against the paper's "more than 400". Selection is quick, the IRDIS table of 6101 observations and 277,755 files loads in 6.4 s and filters to 17 observations for 51 Eridani in 1.1 s.

Three issues filed, the first two of which I can attach a PR to:

* m-samland/spherical#136: a missing `pymysql` masks the `ID_GAIA_DR3` check
* m-samland/spherical#137: `explore_database.ipynb` hardcodes the table path
* m-samland/spherical#138: install fails on Windows while the package declares `OS Independent`

One thing I would rather raise here than file. v3.0.1 moved the `trap` dependency from `@v2.0.0` to `@main`. The old comment said the tag was the one the 51 Eri astrometry baseline was frozen against. With a moving ref, two installs on different dates can resolve to different `trap` commits, with only the runtime floor of 2.0.1 catching it. Given the baseline is the paper's main correctness claim, is a tag or a commit pin still workable?

Four items held. Installation, the Windows build fails while the package declares `OS Independent`. Functionality, I confirmed the database and selection claims but did not run a full reduction, which needs archive data and hours, so I have not exercised the reduction end to end. Reproducibility, the 51 Eri validation cannot be re-run by a reviewer: all six regression tests skip pending output from a prior reduction. Quality of writing, pending a closer read.

### AI usage disclosure

Per the JOSS AI policy: I ran all of the commands reported here myself. An AI assistant helped set up the review environment and with some of the installs and test-suite runs. The notebook execution and the database sync I did myself. The analysis, the writing, and all evaluative decisions are mine.
