+++ tests/regression/data/51eri_astrometry_benchmark.md
@@ -4,16 +4,6 @@ Reference document for evaluating pipeline astrometry (position + uncertainties)
 51 Eri b companion against external ground truth. Written 2026-07-27 from the IRDIS
 DB_K12 phase-6 smoke test; **designed to be reused for the simultaneous IFS dataset**.
 
-> **History (2026-07-27), read this first:** this benchmark went through two corrections.
-> (1) An early draft compared a `flat` "K1-only" detection to GRAVITY using a *stale* CSV
-> that the pipeline erroneously ingested — a real bug, now fixed
-> (`../trap/docs/llm_reference/GITHUB_ISSUE_stale_template_csv_ingestion.md`).
-> (2) After fixing that, a **clean, freshly-computed per-channel measurement** confirmed the
-> underlying point anyway: the per-channel astrometry (channels above threshold, here K1
-> only) lands on GRAVITY (Δρ ≈ −0.7 mas) while the template collapse is +8.7 mas / 2.6σ
-> off. TRAP now **measures per-channel astrometry and reports it as primary**
-> (`astrometry_source` column), collapse as fallback. The numbers below reflect that.
-
 All conversions use the instrument plate scale; see the provenance table for exact values.
 
 ---
@@ -33,18 +23,17 @@ All conversions use the instrument plate scale; see the provenance table for exa
 | TRAP search annulus | inner 31 px, outer 43 px; `yx_known_companion_position = (-35.95, -8.43)` |
 | `temporal_components_fraction` | `[0.2]` |
 | Multiwavelength regressors | **None** (WP1 baseline; regressors OFF — the baseline condition) |
-| Driver | `examples/irdis_reduction_phase6_smoketest.py` (full) or a detection-only re-run |
+| Driver | `examples/irdis_reduction_template.py` with the settings in this table (night, annulus, known-companion position), or a detection-only re-run |
 | σ_PSF (K-band) | FWHM ≈ 3.4 px → σ_PSF ≈ **1.44 px ≈ 17.7 mas** |
 
-> **Regressors must be OFF for the baseline.** Use `template_matching/` (regressors off)
-> or the frozen `template_matching_without_sdi/`. The `template_matching_{pool,occluded,
-> sdi}/` folders are WP2 regressor experiments and must **not** be used as an astrometry
-> baseline.
+> **Regressors must be OFF for the baseline.** Read `template_matching/` (regressors off).
+> The `template_matching_{pool,occluded,sdi}/` folders are WP2 regressor experiments and
+> must **not** be used as an astrometry baseline.
 >
-> **Clear stale per-template CSVs before re-running detection.** A template that finds
-> nothing in a run leaves its previous `companion_table_*.csv` on disk. The combination
-> bug that ingested those has been fixed (it now uses in-memory tables), but stale files
-> on disk are still confusing when inspecting output by hand.
+> **Stale per-template CSVs linger on disk.** A template that finds nothing in a run
+> leaves its previous `companion_table_*.csv` behind. Nothing downstream reads them —
+> the combination step uses in-memory tables — but they mislead when you inspect the
+> output folder by hand.
 
 ---
 
@@ -64,17 +53,29 @@ detection map, then detects on that one map (`template_matching_detection` →
   **diagnostic flag only** (never combined into σ — templates re-weight the same data, so
   their scatter is not an independent error).
 
-**Per-channel astrometry is now measured and is primary (2026-07-27).** Because the
-collapse is optimized for detection SNR, not astrometry — it folds in signal-free channels
[…]
