### 2026-09-10 f0fbb7a
Selecting correct number of frames for center plot (Fix #129, #130)
* Selecting correct number of frames for center plot

* Fixing the colors

* Restored plot_center_evolution.py

* Fixing the empty flux plots issue

* Use WAFFLE_MODE to check if CENTER or CORO are used for science
