### 2026-09-10 3a24379
Fix tests for #129 #132; make sure it works for IFS
