+++ tests/pipeline/test_process_centers_irdis.py
@@ -22,14 +22,25 @@ def _make_image_centers(tmp_path, n_time=100, jitter_amplitude=0.05, outliers=(5
     return centers
 
 
+def _waffle_observation(coro_frames=None):
+    """A continuous-waffle observation: CENTER frames are the science frames.
+
+    ``WAFFLE_MODE`` — not the presence of CORO frames — is what selects the
+    branch, so ``coro_frames`` is free to be non-empty (waffle sequences do
+    pick up stray CORO exposures; see the dispatch test below).
+    """
+    obs = MagicMock()
+    obs.observation = {"INSTRUMENT": ["IRDIS"], "FILTER": ["DB_K12"], "WAFFLE_MODE": [True]}
+    obs.frames = {"CORO": coro_frames}
+    return obs
+
+
 class TestIRDISWaffleCenterFit:
     def test_writes_robust_output(self, tmp_path):
         from spherical.pipeline.steps.process_centers import run_polynomial_center_fit
 
         _make_image_centers(tmp_path)
-        observation = MagicMock()
-        observation.observation = {"INSTRUMENT": ["IRDIS"], "FILTER": ["DB_K12"]}
-        observation.frames = {"CORO": None}
+        observation = _waffle_observation()
 
         run_polynomial_center_fit(
             converted_dir=str(tmp_path),
@@ -51,9 +62,7 @@ class TestIRDISWaffleCenterFit:
         from spherical.pipeline.steps.process_centers import run_polynomial_center_fit
 
         _make_image_centers(tmp_path, outliers=(50,))
-        observation = MagicMock()
-        observation.observation = {"INSTRUMENT": ["IRDIS"], "FILTER": ["DB_K12"]}
-        observation.frames = {"CORO": None}
+        observation = _waffle_observation()
 
         run_polynomial_center_fit(
             converted_dir=str(tmp_path),
@@ -70,9 +79,7 @@ class TestIRDISWaffleCenterFit:
         from spherical.pipeline.steps.process_centers import run_polynomial_center_fit
 
         _make_image_centers(tmp_path, outliers=(50, 75))
-        observation = MagicMock()
-        observation.observation = {"INSTRUMENT": ["IRDIS"], "FILTER": ["DB_K12"]}
-        observation.frames = {"CORO": None}
+        observation = _waffle_observation()
 
         run_polynomial_center_fit(
             converted_dir=str(tmp_path),
@@ -109,10 +116,86 @@ class TestIRDISWaffleCenterFit:
         assert (tmp_path / "image_centers_fitted_robust.fits").exists()
 
 
+class TestIRDISBranchDispatch:
+    """``WAFFLE_MODE`` selects the branch, not the presence of CORO frames.
+
+    DMS propagation exists to place the star in CORO frames from the header
+    dither offsets. A waffle sequence measures the star directly off the
+    satellite spots in every CENTER frame, so it must never be propagated —
+    even when the sequence also picked up a few stray CORO exposures, which
+    27 IRDIS observations in the archive do.
+    """
+
+    def test_waffle_with_stray_coro_frames_uses_center_branch(self, tmp_path):
+        """A waffle sequence with CORO frames present stays on the CENTER branch.
+
+        Dispatching on ``frames["CORO"]`` instead sent these to DMS propagation,
+        which wrote a CORO-length centers array against a CENTER-length raw
+        array (issue #129).
+        """
+        from spherical.pipeline.steps.process_centers import run_polynomial_center_fit
+
+        _make_image_centers(tmp_path, n_time=17, outliers=())
+        # No frames_info_*.csv on disk: the DMS branch would fail to find them.
+        run_polynomial_center_fit(
[…]
