+++ src/spherical/pipeline/steps/plot_center_evolution.py
@@ -7,6 +7,7 @@ converted_dir : str
     Directory where the output files are stored and written.
 """
 import os
+from dataclasses import dataclass
 
 import matplotlib.pyplot as plt
 import numpy as np
@@ -18,27 +19,170 @@ from matplotlib.lines import Line2D
 
 from spherical.pipeline.logging_utils import optional_logger
 
+# Drawn for a position array whose frame timestamps are unavailable, so it is
+# visibly not part of the time colour scale.
+_UNTIMED_COLOR = "0.5"
+
+# Draw order, back to front. Frame-major within it, which is what the step has
+# always done, so an aligned reduction (IFS, or a waffle sequence) keeps exactly
+# the stacking it had.
+_MARKER_DRAW_ORDER = ("o", "x", "+")
+
+
+@dataclass(frozen=True)
+class _CenterSeries:
+    """One set of center positions sharing a single frame grid and time base.
+
+    Attributes
+    ----------
+    label : str
+        Legend entry.
+    marker : str
+        Matplotlib marker.
+    positions : numpy.ndarray
+        ``(n_wave, n_frames, 2)`` center positions in (x, y) pixels.
+    minutes : numpy.ndarray or None
+        Elapsed minutes per frame, ``(n_frames,)``, or None when the frame
+        timestamps for this grid are not on disk.
+    alpha : float
+        Marker transparency.
+    """
+
+    label: str
+    marker: str
+    positions: np.ndarray
+    minutes: np.ndarray | None
+    alpha: float = 0.6
+
+
+def _build_center_series(raw, fitted, robust, center_minutes, coro_minutes):
+    """Pair each center array with the time base of the frames it describes.
+
+    ``image_centers.fits`` always holds CENTER-frame measurements. The two
+    fitted arrays hold either the same CENTER frames (waffle sequence, where
+    the satellite spots give the center directly) or DMS-propagated CORO
+    frames (coronagraphic sequence). Those two grids have unrelated lengths,
+    so each series carries its own timestamps and its own frame count instead
+    of borrowing the raw array's.
+
+    Position arrays that duplicate one already kept are dropped: the waffle
+    branch writes ``fitted`` as a copy of the raw centers and the DMS branch
+    writes ``robust`` as a copy of ``fitted``, and drawing either twice implies
+    a fit that never happened.
+
+    Parameters
+    ----------
+    raw, fitted, robust : numpy.ndarray
+        ``(n_wave, n_frames, 2)`` center positions.
+    center_minutes : numpy.ndarray
+        Elapsed minutes for the CENTER frames.
+    coro_minutes : numpy.ndarray or None
+        Elapsed minutes for the CORO frames, or None if unavailable.
+
+    Returns
+    -------
+    list of _CenterSeries
+    """
+    propagated = fitted.shape[1] != raw.shape[1]
+    fitted_minutes = coro_minutes if propagated else center_minutes
+
+    candidates = [
+        _CenterSeries(
+            "Measured (CENTER)" if propagated else "Original Data",
+            "+", raw, center_minutes, 0.6,
+        ),
+        _CenterSeries(
+            "DMS-propagated (CORO)" if propagated else "1st Fit (fitted)",
+            "o", fitted, fitted_minutes, 0.6,
+        ),
+        _CenterSeries(
+            "DMS-propagated, robust (CORO)" if propagated else "2nd Fit (robust)",
+            "x", robust, fitted_minutes, 0.9,
+        ),
+    ]
+
+    series: list[_CenterSeries] = []
+    for candidate in candidates:
+        if candidate.positions.shape[1] == 0:
+            continue
+        if any(np.array_equal(candidate.positions, kept.positions) for kept in series):
+            continue
+        series.append(candidate)
+    return series
+
+
+def _elapsed_minutes(time_strings, start_time):
[…]
