+++ tests/pipeline/test_plot_center_evolution.py
@@ -1,0 +1,249 @@
+"""Tests for the center-evolution plot, in particular the two frame grids.
+
+``image_centers.fits`` always holds the raw CENTER-frame measurements. What
+the two fitted files hold depends on the branch ``process_centers`` took:
+
+* waffle sequence -> CENTER-frame centers, same length as the raw array;
+* coronagraphic sequence -> DMS-propagated CORO-frame centers, a different
+  length entirely.
+
+The plot used to derive its loop bound from the raw array and index the
+fitted ones with it, which raised IndexError whenever there were more CENTER
+than CORO frames (issue #129) and silently dropped frames otherwise.
+"""
+from __future__ import annotations
+
+import numpy as np
+import pandas as pd
+import pytest
+from astropy.io import fits
+
+from spherical.pipeline.steps.plot_center_evolution import (
+    _build_center_series,
+    _time_bases,
+    run_image_center_evolution_plot,
+)
+
+
+def _positions(n_frames, value, n_wave=2):
+    """A (n_wave, n_frames, 2) center array with every entry set to ``value``."""
+    return np.full((n_wave, n_frames, 2), float(value), dtype=np.float32)
+
+
+def _times(n, start="2023-06-13T02:00:00"):
+    return pd.date_range(start, periods=n, freq="60s").strftime("%Y-%m-%dT%H:%M:%S").tolist()
+
+
+def _write_inputs(tmp_path, raw, fitted, robust, n_center, n_coro=None):
+    fits.writeto(tmp_path / "image_centers.fits", raw, overwrite=True)
+    fits.writeto(tmp_path / "image_centers_fitted.fits", fitted, overwrite=True)
+    fits.writeto(tmp_path / "image_centers_fitted_robust.fits", robust, overwrite=True)
+    pd.DataFrame({"TIME": _times(n_center)}).to_csv(tmp_path / "frames_info_center.csv", index=False)
+    if n_coro is not None:
+        pd.DataFrame({"TIME": _times(n_coro, start="2023-06-13T02:30:00")}).to_csv(
+            tmp_path / "frames_info_coro.csv", index=False
+        )
+
+
+class TestBuildCenterSeries:
+    """The series builder pairs each position array with its own time base."""
+
+    def test_aligned_arrays_all_use_center_times(self):
+        center_minutes = np.arange(5.0)
+        series = _build_center_series(
+            raw=_positions(5, 1), fitted=_positions(5, 2), robust=_positions(5, 3),
+            center_minutes=center_minutes, coro_minutes=None,
+        )
+        assert len(series) == 3
+        for entry in series:
+            np.testing.assert_array_equal(entry.minutes, center_minutes)
+
+    def test_duplicate_position_arrays_are_drawn_once(self):
+        """The waffle branch writes ``fitted`` as a copy of the raw centers.
+
+        Drawing it twice puts one marker exactly under another and implies a
+        fit that never happened.
+        """
+        raw = _positions(5, 1)
+        series = _build_center_series(
+            raw=raw, fitted=raw.copy(), robust=_positions(5, 3),
+            center_minutes=np.arange(5.0), coro_minutes=None,
+        )
+        assert [entry.marker for entry in series] == ["+", "x"]
+
+    def test_propagated_arrays_use_coro_times(self):
+        """Propagated centers describe CORO frames, so they carry CORO times."""
+        center_minutes = np.arange(2.0)
+        coro_minutes = np.arange(10.0)
+        series = _build_center_series(
+            raw=_positions(2, 1), fitted=_positions(10, 2), robust=_positions(10, 2),
+            center_minutes=center_minutes, coro_minutes=coro_minutes,
+        )
[…]
