### 2026-09-10 e6ffbe7
Fix center evolution legend order to match draw order
