### 2026-09-10 61b9851
Improve the center evolution diagnostic plot (Fix #141)
Three changes to plot_center_evolution, all found while looking at real
reductions after the frame-grid fix in 3a24379.

Marker area encodes the wavelength channel, which the legend never said,
so the two IRDIS clusters looked unexplained. A second legend labels them
by wavelength, summarised to first/middle/last so IFS does not produce 39
entries.

The duplicate-series check was exact. A float32 round trip through FITS
leaves the copies differing by ~3e-5 px, so on a real reduction it never
fired and the raw and fitted markers were still
[…]
