+++ tests/pipeline/test_plot_center_evolution.py
@@ -20,7 +20,9 @@ from astropy.io import fits
 
 from spherical.pipeline.steps.plot_center_evolution import (
     _build_center_series,
+    _residuals_from_median,
     _time_bases,
+    _wavelength_legend_entries,
     run_image_center_evolution_plot,
 )
 
@@ -34,15 +36,23 @@ def _times(n, start="2023-06-13T02:00:00"):
     return pd.date_range(start, periods=n, freq="60s").strftime("%Y-%m-%dT%H:%M:%S").tolist()
 
 
-def _write_inputs(tmp_path, raw, fitted, robust, n_center, n_coro=None):
+def _write_inputs(tmp_path, raw, fitted, robust, n_center, n_coro=None, wavelengths=None,
+                  outliers=None):
     fits.writeto(tmp_path / "image_centers.fits", raw, overwrite=True)
     fits.writeto(tmp_path / "image_centers_fitted.fits", fitted, overwrite=True)
     fits.writeto(tmp_path / "image_centers_fitted_robust.fits", robust, overwrite=True)
+    if wavelengths is None:
+        wavelengths = np.linspace(2110.0, 2251.0, raw.shape[0])
+    fits.writeto(tmp_path / "wavelengths.fits", np.asarray(wavelengths, dtype=float), overwrite=True)
     pd.DataFrame({"TIME": _times(n_center)}).to_csv(tmp_path / "frames_info_center.csv", index=False)
     if n_coro is not None:
         pd.DataFrame({"TIME": _times(n_coro, start="2023-06-13T02:30:00")}).to_csv(
             tmp_path / "frames_info_coro.csv", index=False
         )
+    if outliers is not None:
+        (tmp_path / "additional_outputs").mkdir(exist_ok=True)
+        fits.writeto(tmp_path / "additional_outputs" / "center_outlier_frames.fits",
+                     np.asarray(outliers, dtype=np.int32), overwrite=True)
 
 
 class TestBuildCenterSeries:
@@ -70,6 +80,30 @@ class TestBuildCenterSeries:
             center_minutes=np.arange(5.0), coro_minutes=None,
         )
         assert [entry.marker for entry in series] == ["+", "x"]
+
+    def test_near_identical_arrays_are_drawn_once(self):
+        """The copies differ by float32 round-trip noise on a real reduction.
+
+        bet Pic 2014-12-07 has ``image_centers_fitted.fits`` matching
+        ``image_centers.fits`` to 3e-5 px, which exact equality misses, so both
+        markers were still drawn on top of each other.
+        """
+        raw = _positions(5, 1)
+        almost = raw + 3e-5
+        series = _build_center_series(
+            raw=raw, fitted=almost, robust=_positions(5, 3),
+            center_minutes=np.arange(5.0), coro_minutes=None,
+        )
+        assert [entry.marker for entry in series] == ["+", "x"]
+
+    def test_genuinely_different_arrays_are_both_kept(self):
+        """A real fit moves the centers far more than the rounding tolerance."""
+        raw = _positions(5, 1)
+        series = _build_center_series(
+            raw=raw, fitted=raw + 0.01, robust=_positions(5, 3),
+            center_minutes=np.arange(5.0), coro_minutes=None,
+        )
+        assert [entry.marker for entry in series] == ["+", "o", "x"]
 
     def test_propagated_arrays_use_coro_times(self):
         """Propagated centers describe CORO frames, so they carry CORO times."""
@@ -173,11 +207,54 @@ class TestTimeBases:
         assert center_minutes[0] == 0.0
 
 
+class TestWavelengthLegendEntries:
+    """Marker size encodes wavelength, which the legend never explained."""
+
[…]
