+++ src/spherical/pipeline/steps/plot_center_evolution.py
@@ -28,6 +28,12 @@ _UNTIMED_COLOR = "0.5"
 # the stacking it had.
 _MARKER_DRAW_ORDER = ("o", "x", "+")
 
+# Two center arrays agreeing this closely are the same measurement written
+# twice. A float32 round trip through FITS costs ~3e-5 px, which exact equality
+# would call a difference, and nothing in SPHERE astrometry is meaningful at
+# 1e-4 px.
+_DUPLICATE_TOLERANCE_PX = 1e-4
+
 
 @dataclass(frozen=True)
 class _CenterSeries:
@@ -68,7 +74,8 @@ def _build_center_series(raw, fitted, robust, center_minutes, coro_minutes):
     Position arrays that duplicate one already kept are dropped: the waffle
     branch writes ``fitted`` as a copy of the raw centers and the DMS branch
     writes ``robust`` as a copy of ``fitted``, and drawing either twice implies
-    a fit that never happened.
+    a fit that never happened. The comparison carries a tolerance because a
+    float32 round trip through FITS leaves the two copies differing by ~3e-5 px.
 
     Parameters
     ----------
@@ -105,16 +112,179 @@ def _build_center_series(raw, fitted, robust, center_minutes, coro_minutes):
     for candidate in candidates:
         if candidate.positions.shape[1] == 0:
             continue
-        if any(np.array_equal(candidate.positions, kept.positions) for kept in series):
+        if any(_same_positions(candidate.positions, kept.positions) for kept in series):
             continue
         series.append(candidate)
     return series
+
+
+def _same_positions(a, b):
+    """True when two center arrays are the same measurement written twice."""
+    if a.shape != b.shape:
+        return False
+    return bool(np.allclose(a, b, rtol=0.0, atol=_DUPLICATE_TOLERANCE_PX, equal_nan=True))
+
+
+def _wavelength_legend_entries(wavelengths, sizes):
+    """Marker sizes paired with the wavelength each one stands for.
+
+    Marker area ramps with wavelength channel, which is the only thing telling
+    the IRDIS channels apart in the scatter plot and was never in the legend.
+    IFS has 39 channels, so only the first, middle and last are listed rather
+    than a legend taller than the figure.
+
+    Parameters
+    ----------
+    wavelengths : numpy.ndarray
+        Channel wavelengths in nanometres, as stored in ``wavelengths.fits``.
+    sizes : numpy.ndarray
+        Marker size per channel, same length as ``wavelengths``.
+
+    Returns
+    -------
+    list of (float, str)
+        Marker size and its label, in micron.
+    """
+    n_wave = len(wavelengths)
+    indices = range(n_wave) if n_wave <= 3 else (0, n_wave // 2, n_wave - 1)
+    return [(float(sizes[i]), f"{wavelengths[i] / 1000.0:.2f} µm") for i in indices]
+
+
+def _residuals_from_median(positions):
+    """Center positions with each channel's median subtracted.
+
+    The IRDIS channels sit ~13 px apart on the detector, so subtracting the
+    per-channel median is what lets both share one axis in the time series.
+
+    Parameters
+    ----------
+    positions : numpy.ndarray
+        ``(n_wave, n_frames, 2)`` center positions.
+
+    Returns
+    -------
+    numpy.ndarray
+        Same shape, in pixels relative to each channel's median.
+    """
+    return positions - np.nanmedian(positions, axis=1)[:, None, :]
 
 
 def _elapsed_minutes(time_strings, start_time):
     """Minutes elapsed since ``start_time`` for a column of timestamps."""
     times = pd.to_datetime(time_strings)
     return ((times - start_time).dt.total_seconds() / 60.0).to_numpy()
+
+
+def _optional_fits(path):
+    """Return the data in a FITS file, or None if it is not there."""
[…]
