### 2026-09-10 5147668
Add the satellite spot amplitude regression test (Fix #142)
687b026 made the mask_deviating refit carry the Const2D pedestal from the
first pass, so the fitted amplitude measures the spot rather than the spot
plus the stellar halo under it. The tests that shipped with it cover the
fit_background flag only, not the numerical outcome.

This test builds four unit-peak waffle spots on an exponential halo and
asserts the reported amplitude stays at 1.0 across halo-to-spot ratios of
0 to 3. Against 687b026^ it fails as 1 + 2 x halo, giving 1.60 at a ratio
of 0.3 and 3.00 at 1.0.

Written at the time
[…]
