### 2026-09-11 5a507aa
Re-seed the waffle centre fit and fix the spot radius
The spot radius constant is now 10*sqrt(2) exactly, and the fit re-seeds
itself from the ensemble median of the first pass when that would move the
search boxes. Also adds n_center_plots to stop plotting every frame.

Fixes #144
