### 2026-09-11 912b152
Keep measured centres on IRDIS and rename the plot series
The outlier moving-median substitution is gone, so all three IRDIS centre
products carry the measurement and rejection lives only in the outlier
mask. Plot series are now named for what they are.

Fixes #145
