### 2026-09-11 9638e5d
Disambiguate the center time series legend and overlapping channels
The legend folded colour (wavelength) and line style (series) into one
block, so its neutral style keys read as grey curves that are not in the
figure. They are now two titled blocks. The zero anchor is dotted so it
cannot be read as a flat series, channel line width ramps with wavelength
and is drawn widest first so coincident channels stay visible, and a series
whose channels are identical by construction says so. The clipped-point
count moves above each panel, out of the data.

Refs #132
