+++ src/spherical/pipeline/steps/plot_center_evolution.py
@@ -169,6 +169,27 @@ def _residuals_from_median(positions):
     return positions - np.nanmedian(positions, axis=1)[:, None, :]
 
 
+def _channels_coincide(positions):
+    """True when every channel of a series traces the same residual curve.
+
+    Not a fluke when it happens. The DMS-propagated track is
+    ``S₀[ch] + PAC[i]/18``, and `_residuals_from_median` subtracts each
+    channel's median, which cancels ``S₀[ch]`` exactly and leaves a residual
+    with no channel dependence at all. So on a coronagraphic IRDIS sequence the
+    channels are guaranteed to lie on top of each other, and a reader who takes
+    the colour legend at face value goes looking for a curve that cannot be
+    seen. Saying so in the legend is cheaper than leaving them to find out.
+    """
+    residuals = _residuals_from_median(positions)
+    if residuals.shape[0] < 2:
+        return False
+    return all(
+        np.allclose(residuals[ch], residuals[0], rtol=0.0,
+                    atol=_DUPLICATE_TOLERANCE_PX, equal_nan=True)
+        for ch in range(1, residuals.shape[0])
+    )
+
+
 def _elapsed_minutes(time_strings, start_time):
     """Minutes elapsed since ``start_time`` for a column of timestamps."""
     times = pd.to_datetime(time_strings)
@@ -210,19 +231,29 @@ def _plot_center_timeseries(series, wavelengths, outlier_frames, output_path):
 
     n_wave = timed[0].positions.shape[0]
     channel_colors = plt.cm.viridis(np.linspace(0, 0.9, n_wave))
-    # A line per series, so the marker tells you which array a point came from.
-    styles = {"+": dict(marker=".", linestyle="none", markersize=4),
-              "o": dict(marker="none", linestyle="-", linewidth=1.0),
-              "x": dict(marker="none", linestyle="--", linewidth=1.0)}
+    # Line width ramps with wavelength, the same way marker area does in the
+    # scatter plot, and channels are drawn widest first. Coincident channels
+    # then show as a narrow line inside a wider band instead of the later one
+    # erasing the earlier one. Transparency would not do: two exactly
+    # overlapping curves blend into a third colour matching neither legend key.
+    channel_widths = np.linspace(2.6, 0.9, n_wave)[::-1]
+    channel_markersizes = np.linspace(7.0, 3.0, n_wave)[::-1]
+    # A line per series, so the style tells you which array a point came from.
+    styles = {"+": dict(marker=".", linestyle="none"),
+              "o": dict(marker="none", linestyle="-"),
+              "x": dict(marker="none", linestyle="--")}
 
     fig, axes = plt.subplots(2, 1, figsize=(9, 6), sharex=True)
     for entry in timed:
         residuals = _residuals_from_median(entry.positions)
         style = styles.get(entry.marker, styles["+"])
-        for ch in range(entry.positions.shape[0]):
+        for ch in reversed(range(entry.positions.shape[0])):
+            width = dict(linewidth=channel_widths[ch % n_wave],
+                         markersize=channel_markersizes[ch % n_wave])
             for axis_idx, ax in enumerate(axes):
                 ax.plot(entry.minutes, residuals[ch, :, axis_idx],
-                        color=channel_colors[ch % n_wave], alpha=0.8, **style)
+                        color=channel_colors[ch % n_wave], alpha=0.9,
+                        **style, **width)
 
     raw = timed[0]
     if outlier_frames is not None:
@@ -257,32 +288,48 @@ def _plot_center_timeseries(series, wavelengths, outlier_frames, output_path):
             ax.set_ylim(-span, span)
[…]
