### 2026-09-11 84f143f
Skip trap/charis-dependent pipeline tests when those deps are absent
Tests that import ifs_reduction or run_trap inside the test function bypassed
the module-level importorskip guard, so the CI pipeline job -- which installs
neither charis nor trap -- failed instead of skipping. Local runs with both
installed are unchanged.
