### 2026-09-11 425148d
Fail early with a clear message when TRAP inputs are missing (#139)
Running TRAP with the preprocessing steps switched off failed on the first
`fits.getdata` call with a bare FileNotFoundError naming a single file, which
said nothing about why that file was absent. In a batch run the message then
repeated once per target with no hint that the whole queue was misconfigured.

Check the required products up front and name all the missing ones at once, so
the directory need not be probed one file at a time. The guard sits inside the
existing try block, so a misconfigured target still gets a crash 
[…]
