### 2026-09-11 26618ad
Add changelog entry for the file-descriptor leak fix (#139)
