### 2026-09-12 4f89f3e
More robust fallback implementation in mosaic plot
