+++ src/spherical/pipeline/visualize/mosaic.py
@@ -43,6 +43,7 @@ CANDIDATE_PATTERNS = {
 
 REGULAR_DETECTION_PATTERN = "norm_detection_*.fits"
 REGULAR_CANDIDATE_FILENAME = "validated_companion_table_short.csv"
+BROADBAND_OBS_MODES = frozenset({"BB_Y", "BB_J", "BB_H", "BB_Ks"})
 
 # Default color scheme for candidates (easily distinguishable colors)
 DEFAULT_CANDIDATE_COLORS = [
@@ -285,8 +286,10 @@ def get_mosaic_file_combinations(
     """Get file combinations for mosaic plotting.
 
     Template-matching products are preferred when available.
-    If an observation has no ``template_matching`` directory,
-    fall back to the regular template-free TRAP products.
+    For broadband IRDIS observations (BB_Y, BB_J, BB_H, and
+    BB_Ks), regular TRAP products are used as a fallback when
+    the requested template-matching product is unavailable.
+    Other observing modes remain missing in that case.
 
     Args:
         base_path: Root path to search
@@ -309,39 +312,36 @@ def get_mosaic_file_combinations(
         combinations = get_all_combinations(base_path)
 
     for target, obs_mode, date in combinations:
+        file_path = None
+
         obs_dir = base_path / target / obs_mode / date
-        template_dir = obs_dir / "template_matching"
         template_path = obs_dir / pattern
 
         if template_path.exists():
-            # Template-matching result from dual-band imaging
+            # template product exists -> use template product
             file_path = template_path
 
-        elif not template_dir.is_dir():
-            # Regular TRAP detection map for broad-band filters
+        elif obs_mode in BROADBAND_OBS_MODES:
+            # BB_Y/J/H/Ks -> use regular detection map
             if file_type == "fits":
-                matches = sorted(obs_dir.glob(REGULAR_DETECTION_PATTERN))
+                matches = sorted(
+                    obs_dir.glob(REGULAR_DETECTION_PATTERN),
+                    key=lambda path: path.stat().st_mtime,
+                )
 
-                if len(matches) == 1:
-                    file_path = matches[0]
-                elif len(matches) > 1:
-                    logger.warning(
-                        "Multiple regular detection maps found in %s; using %s",
-                        obs_dir,
-                        matches[0].name,
-                    )
-                    file_path = matches[0]
-                else:
-                    file_path = None
+                if matches:
+                    file_path = matches[-1]
+
+                    if len(matches) > 1:
+                        logger.warning(
+                            "Multiple regular detection maps found in %s; using newest %s",
+                            obs_dir,
+                            file_path.name,
+                        )
 
             else:
                 regular_csv = obs_dir / REGULAR_CANDIDATE_FILENAME
                 file_path = regular_csv if regular_csv.exists() else None
-
-        else:
-            # template_matching exists, but this particular product is missing.
-            # Do not silently replace it with the regular detection map.
-            file_path = None
 
         results[(target, obs_mode, date)] = file_path
 
@@ -662,30 +662,42 @@ def plot_detection_mosaic(
             f"for template type {template_type}"
         )
     
+    csv_files = get_mosaic_file_combinations(
+        base_path,
+        template_type,
+        "csv",
+        combinations=list(fits_files),
+    )
+    
     # Auto-scale color limits if requested (global scaling only)
     if auto_scale and not individual_scaling:
         all_data = []
+
         for fits_path in fits_files.values():
-            if fits_path is not None and fits_path.exists():
-                try:
-                    data = _load_detection_image(fits_path)
-                    # Only include finite values for statistics
-                    finite_data = data[np.isfinite(data)]
-                    if len(finite_data) > 0:
[…]
