### 2026-09-14 db33ed3
Fix/logging fd leak (#147)
* Fix file-descriptor leak in per-target logging (#139)

`get_pipeline_logger` builds a fresh `multiprocessing.Queue` per call — two
pipe fds plus three POSIX semaphores — and attaches it to a logger named after
the target. Python keeps every named logger alive forever in
`Logger.manager.loggerDict`, and `remove_queue_listener()` stopped only the
listener thread, so nothing ever detached the handler or closed the queue.
Batch runs leaked 5 fds and one Queue feeder thread per observation until
`OSError: [Errno 24] Too many open files` killed the run.

The reported tra
[…]
