### 2026-09-14 1bd8af0
Make `pip install ".[test]"` run the whole offline test suite (#149)
A plain pytest on a base test install failed 50 pipeline tests: the suite only
worked with the install each CI job used, and the per-module importorskip guards
for scipy/photutils had drifted because no job ran the pipeline tests without them.

Move the non-git pipeline deps (scipy, photutils, scikit-image, dill) into the
test extra and the pixi test feature, drop the now-redundant guards, and run
tests/database and tests/pipeline as separate steps of one CI job on 3.11-3.13.
Only charis/trap tests skip.

Since no test env la
[…]
