### 2026-09-14 1d19801
Update stale test-pipeline task comment
