### 2026-09-14 1417084
Scale mosaic fonts with the panel, not the whole figure
Panels in the detection and spectrum mosaics shrink as observations are added, but titles and SNR labels were scaled with the total figure size, so they grew. With ~50 data sets the titles took up more room than the detection maps (#146).

Fonts now follow the size of one grid cell, which keeps the title a constant fraction of its panel: the title/image height ratio at N=50 goes from 0.86 to 0.23, against 0.20 at N=4. SNR labels keep an 8 pt floor so they stay legible on the noisy map.
