### 2026-09-14 28d55bd
Validate run_trap force sets against the IRDIS registry for IRDIS data
run_trap_on_observation checked `force` against the default IFS step
registry, so an IRDIS config forcing an IRDIS-only step (e.g.
`preprocess_irdis`) got through execute_targets and then raised
"Unknown step name(s) in force" once TRAP started. The `_forced` and
`should_run` calls for the TRAP steps used the IFS step order as well.

Pick the registry and step order from the observation's instrument and
use them for all three calls.

Fixes #152
