### 2026-09-15 722b6c6
Overwrite cross_channel_offset.fits whenever the center fit runs
fit_centers_in_parallel only wrote additional_outputs/cross_channel_offset.fits
if it did not exist yet, so a forced re-run kept the old offset while
spot_centers.fits, image_centers.fits and the other centring products next
to it were overwritten. The offset could go stale relative to the centres it
was derived from.

Write it with overwrite=True like the rest.

Fixes #153
