### 2026-09-15 0aa84a9
Interpolate failed IRDIS center fits before they reach TRAP
Since #145 the IRDIS robust centers are the raw measurements, so a frame whose waffle fit failed went to TRAP as NaN, and TRAP skips a whole wavelength when any of its centers is NaN (m-samland/trap#40). Failed fits carry no measurement, so only those frames are interpolated in time; measured outliers stay as measured. The plot's duplicate check ignores NaN frames so the filled robust file is not drawn as a separate series.
