+++ CHANGELOG.md
@@ -8,95 +8,70 @@ This project follows [Semantic Versioning](https://semver.org/) and the [Keep a 
 
 ## [Unreleased]
 
+---
+
+## [3.1.0] - 2026-09-15 – JOSS Review and Various Improvements
+
+Minor release that concludes the JOSS review, with fixes found while running the pipeline on larger samples.
+No breaking changes, and the v3.0.0 Zenodo tables are unchanged.
+The centering fixes can move fitted star centers slightly, so reductions may differ slightly from 3.0.1.
+
 ### ✨ Added
-- **A center-position time series alongside the center evolution scatter plot** – `x` and `y`
-  against time, one panel each, relative to each channel's median, with frames flagged by the
-  center fit ringed. A slow drift reads as a slope rather than a colour gradient, and the CENTER
-  measurements and the DMS-propagated CORO track share one time axis
-  ([#141](https://github.com/m-samland/spherical/issues/141)).
+- **A center-position time series next to the center evolution plot** – `x` and `y` against time, relative to each channel's median, with flagged frames ringed.
+  Slow drifts show up as a slope
+  ([#141](https://github.com/m-samland/spherical/issues/141), [@m-samland](https://github.com/m-samland)).
 
 ### 🔧 Changed
-- **The waffle center fit re-seeds itself from what it measured** – The nominal star position is
-  calibrated on several data sets, so a realignment of the coronagraph can leave it pointing
-  several pixels off. The fit now refits once from the ensemble median of the first pass. The
-  refit is only done if the starting position would change
-  ([#144](https://github.com/m-samland/spherical/issues/144)).
-- **The center fit plots a subsample of frames instead of all of them** – Plotting was ~85% of the
-  runtime of the step, and a single IFS observation emitted over ten thousand diagnostic pages.
-  `n_center_plots` (default 10, `None` for all, `0` for none) spreads them across the sequence
-  ([#144](https://github.com/m-samland/spherical/issues/144)).
-- **IRDIS hands TRAP the measured centers, not outlier-interpolated ones** – Flagged frames were
-  replaced by a 21-frame moving median. The waffle fit is far more precise than the stellar motion
-  it measures, so that smoothed away real jitter which the planet shares with the star. Outliers
-  are still detected and still written to `center_outlier_frames.fits`, which is where frame
-  rejection belongs ([#145](https://github.com/m-samland/spherical/issues/145)).
-- **The center evolution plot explains its marker sizes** – Marker area encodes the wavelength
-  channel, which was undocumented, so the two IRDIS clusters looked unexplained. A second legend
-  now labels them by wavelength. A position array that merely duplicates one already drawn is
-  also dropped properly: the check was exact, and a float32 round trip through FITS left the
-  copies differing by ~3e-5 px, so both markers were still drawn on top of each other
-  ([#141](https://github.com/m-samland/spherical/issues/141)).
-- **CI now runs on `develop` and covers the reduction steps** – The workflow only triggered on
-  `main` and only ran the database subject, so pipeline-step breakage was invisible to it.
-- **`pip install ".[test]"` now runs the whole offline test suite** – A plain `pytest` on that
-  install failed 50 pipeline tests for lack of scipy, photutils, scikit-image and dill. These are
-  now part of the `test` extra; only tests needing the git-sourced `charis` or `trap` skip. A new
-  test keeps `spherical.database` free of pipeline-only imports. The test tiers are documented in
-  the README ([#149](https://github.com/m-samland/spherical/issues/149), reported by
[…]
