### 2026-09-15 72f2410
Add a reference driver for the 51 Eri IFS regression baseline
The IFS baseline could only be reproduced by editing the example template by hand. run_51eri_ifs_reference.py runs the full IFS chain with the benchmark settings and writes provenance.json, like the IRDIS driver. SPHERICAL_BASE_PATH and SPHERICAL_NCPU adapt it to a cluster without changing its sha256.
