### 2026-09-15 ac97ac3
Import astroquery.gaia only where query_gspphot needs it (Fix #158)
Importing astroquery.gaia queries the Gaia archive status and prints the response. As a module-level import in database_utils, every spherical import and every extraction or center-fit worker hit the network, and printed the archive's HTML maintenance page while it was down.
