### 2026-09-15 87af54e
Record the 51 Eri IFS verification run on the 3.1.0 code
The full IFS chain re-run with run_51eri_ifs_reference.py (charis 2.1.0, trap 2.0.1) agrees with the frozen baseline to 0.02 px, so the baseline stays. The benchmark doc records the provenance and explains the Y-band flux-PSF difference, which comes from spherical's flux bad-pixel threshold change in 2e72b4a rather than from charis.
