### 2026-09-15 490f1e4
Implemented suggestions
