+++ src/spherical/pipeline/visualize/mosaic.py
@@ -305,15 +305,15 @@ def get_mosaic_file_combinations(
     base_path: Path,
     template_type: TemplateType,
     file_type: Literal["fits", "csv"] = "fits",
-    combinations: Optional[List[Tuple[str, str, str]]] = None
+    combinations: Optional[List[Tuple[str, str, str]]] = None,
 ) -> Dict[Tuple[str, str, str], Optional[Path]]:
     """Get file combinations for mosaic plotting.
 
-    Template-matching products are preferred when available.
-    For broadband IRDIS observations (BB_Y, BB_J, BB_H, and
-    BB_Ks), regular TRAP products are used as a fallback when
-    the requested template-matching product is unavailable.
-    Other observing modes remain missing in that case.
+    Template-matching products are used when the template-matching FITS
+    product exists. For broadband IRDIS observations (BB_Y, BB_J, BB_H,
+    and BB_Ks), regular TRAP products are used as a fallback when the
+    template-matching FITS product is unavailable. Other observing modes
+    remain missing in that case.
 
     Args:
         base_path: Root path to search
@@ -326,12 +326,8 @@ def get_mosaic_file_combinations(
     Returns:
         Dictionary mapping (target, obs_mode, date) to file paths
     """
-    if file_type == "fits":
-        pattern = TEMPLATE_PATTERNS[template_type]
-    else:  # csv
-        pattern = CANDIDATE_PATTERNS[template_type]
-
     results = {}
+
     if combinations is None:
         combinations = get_all_combinations(base_path)
 
@@ -339,14 +335,18 @@ def get_mosaic_file_combinations(
         file_path = None
 
         obs_dir = base_path / target / obs_mode / date
-        template_path = obs_dir / pattern
 
-        if template_path.exists():
-            # template product exists -> use template product
-            file_path = template_path
+        template_fits = obs_dir / TEMPLATE_PATTERNS[template_type]
+        template_csv = obs_dir / CANDIDATE_PATTERNS[template_type]
+
+        # The template FITS determines whether template matching succeeded.
+        if template_fits.exists():
+            if file_type == "fits":
+                file_path = template_fits
+            else:
+                file_path = template_csv if template_csv.exists() else None
 
         elif obs_mode in BROADBAND_OBS_MODES:
-            # BB_Y/J/H/Ks -> use regular detection map
             if file_type == "fits":
                 matches = sorted(
                     obs_dir.glob(REGULAR_DETECTION_PATTERN),
@@ -1748,7 +1748,8 @@ def plot_detection_mosaic_batched(
     output_dir: Optional[Path] = None,
     output_format: str = "png",
     suffix: Optional[str] = None,
-    **kwargs
+    show_missing: bool = False,
+    **kwargs,
 ) -> List[Figure]:
     """Create batched detection mosaic plots.
     
@@ -1759,6 +1760,8 @@ def plot_detection_mosaic_batched(
         output_dir: Directory to save batched figures. If None, no files are saved
         output_format: File extension/format for saved batches (default: "png")
         suffix: Optional filename suffix so subset runs do not overwrite
+        show_missing: Whether to include observations without a detection map
+            as blank panels (default: False)
         **kwargs: All other arguments passed to plot_detection_mosaic()
 
     Returns:
@@ -1771,15 +1774,21 @@ def plot_detection_mosaic_batched(
         "fits",
     )
 
-    sorted_combinations = sorted(fits_files)
+    if show_missing:
+        sorted_combinations = sorted(fits_files)
+    else:
+        sorted_combinations = sorted(
+            combo
+            for combo, fits_path in fits_files.items()
+            if fits_path is not None and fits_path.exists()
+        )
 
     if not sorted_combinations:
         raise ValueError(
[…]
