### 2026-09-15 12e5bb7
Update paper for v3.1.0 / language editing and clarification
