+++ joss/paper.md
@@ -25,17 +25,17 @@ bibliography: paper.bib
 
 The Spectro-Polarimetric High-contrast Exoplanet REsearch instrument (SPHERE; @Beuzit:2019) at the Very Large Telescope (VLT) is a leading facility for coronagraphic imaging of exoplanets and circumstellar disks in the optical and near-infrared. Over the last decade, SPHERE has contributed to hundreds of publications and major legacy surveys (e.g., the SpHere INfrared survey for Exoplanets, SHINE; @Chauvin:2017; @Chomez:2025).
 
-`spherical` combines a curated, searchable database of the complete SPHERE observation history with automated reduction pipelines for two of its subsystems. Users filter observations by target properties, observing mode, or observing conditions, download the selected raw data and calibrations from the European Southern Observatory (ESO) archive, and reduce them end to end. Integral Field Spectrograph (IFS; @Claudi:2008) data are extracted into spectral cubes by a pipeline adapted from the Coronagraphic High Angular Resolution Imaging Spectrograph (CHARIS; @Brandt:2017; @Samland:2022); data from the Infra-Red Dual-band Imager and Spectrograph (IRDIS; @Dohlen:2008) in dual-band imaging (DBI) mode are calibrated and pre-processed by a pipeline that mirrors the IFS workflow step for step. Both feed post-processing with TRAP (Temporal Reference Analysis of Planets; @Samland:2021), a detection algorithm that models the temporal systematics of the residual starlight instead of subtracting a reference image, and recovers the astrometry and contrast spectra of detected companions.
+`spherical` combines a curated, searchable database of the complete SPHERE observation history with automated reduction pipelines for two of its subsystems. Users can filter observations by target star properties, observing mode, or observing conditions, download the selected raw data and calibrations from the European Southern Observatory (ESO) archive, and reduce them end to end. Integral Field Spectrograph (IFS; @Claudi:2008) data are extracted into spectral cubes by a pipeline adapted from the Coronagraphic High Angular Resolution Imaging Spectrograph (CHARIS; @Brandt:2017; @Samland:2022); data from the Infra-Red Dual-band Imager and Spectrograph (IRDIS; @Dohlen:2008) in dual-band imaging (DBI) mode are calibrated and pre-processed by a pipeline that mirrors the IFS workflow. Both feed post-processing with TRAP (Temporal Reference Analysis of Planets; @Samland:2021), a detection algorithm that identifies companions by modeling the temporal systematics of the starlight instead of subtracting a reference image, and recovers their astrometry and contrast spectra.
 
-As of the v3.0.0 release described here, the database covers observations taken between 2014 May and 2026 August: 6101 IRDIS DBI sequences, 4733 IFS sequences, 1205 IRDIS dual-beam polarimetric imaging (DPI; @deBoer:2020; @vanHolstein:2020) sequences, and 355 sparse aperture masking (SAM; @Cheetham:2016) sequences across both subsystems. DPI and SAM data are covered for discovery and download only; their reduction is left to community tools. SPHERE is undergoing the *SPHERE+* upgrade [@Boccaletti:2020; @Boccaletti:2022], including a second-stage adaptive optics system (SAXO+; @Stadler:2022), which extends its scientific lifetime and makes it a pathfinder for the Planetary Camera and Spectrograph (PCS; @Kasper:2021) of the Extremely Large Telescope (ELT).
[…]
