### 2026-09-16 0cf6150
fix: count a multi-epoch target's epochs as nights, not rows
The observation table groups on [NIGHT_START, OBS_NUMBER, mode], so one night
gives several rows when a host was observed in more than one mode or block.
Counting rows let a host with two modes on one night and a single later night
pass min_epochs=3, and overstated _MULTI_EPOCH_N for 12 of the 466 hosts the
selector keeps on the real IRDIS table. The default min_epochs=2 was protected
by the zero-span check, but a raised one was not.

Grouping stays per MAIN_ID, whatever the FILTER: TRAP's astrometry does not
depend on the band, so e
[…]
