### 2026-09-17 fb5f566
Extract stamps NaN-safely and keep the PSF cube finite (Fix #163)
