### 2026-09-17 5d95863
Share one usable-FOV footprint between run_trap and the pipeline
