### 2026-09-17 afdea49
Base the PSF guess on the shared FOV footprint and the measured coronagraph centre (#164)
