+++ src/spherical/pipeline/steps/flux_psf_calibration.py
@@ -9,6 +9,7 @@ reduction_parameters : dict
     Reduction parameters dict, must contain 'flux_combination_method', 'exclude_first_flux_frame', and 'exclude_first_flux_frame_all'.
 """
 import os
+import warnings
 from pathlib import Path
 from typing import Dict
 
@@ -23,6 +24,60 @@ from spherical.pipeline.ivar_badpixels import bad_pixel_mask_from_ivar
 from spherical.pipeline.logging_utils import optional_logger
 from spherical.pipeline.psf_repair import repair_psf_core
 from spherical.pipeline.steps.find_star import guess_position_psf, star_centers_from_PSF_img_cube
+from spherical.pipeline.steps.irdis_preprocess import nominal_star_positions
+
+
+def coronagraph_center_from_disk(converted_dir, logger, irdis_filter_comb=None):
+    """``(x, y)`` of the coronagraph centre, or ``None`` if unavailable.
+
+    The centre marks where the star sat behind the coronagraph, which is where
+    detector persistence from the CORO sequence appears in a trailing FLUX
+    block (#83). The flux step runs after ``find_centers`` and
+    ``process_extracted_centers``, so the measured centre is already on disk;
+    prefer it to any literal. On the 51 Eri OBS_H reference run it is
+    (127.38, 127.80), about 3 px from the (126, 131) this used to hard-code --
+    and that literal was being applied to IRDIS half-frames too, where it
+    means nothing.
+
+    ``irdis_filter_comb`` supplies the IRDIS fallback: the nominal star
+    position is the same CORO-frame quantity, measured per filter. Pass it
+    only for IRDIS; IFS has no nominal and falls through to ``None``.
+    """
+    for name in ("image_centers_fitted_robust.fits", "image_centers.fits"):
+        path = os.path.join(converted_dir, name)
+        if not os.path.exists(path):
+            continue
+        centers = np.asarray(fits.getdata(path), dtype=float).reshape(-1, 2)
+        with warnings.catch_warnings():
+            warnings.filterwarnings("ignore", message="All-NaN slice encountered",
+                                    category=RuntimeWarning)
+            xy = np.nanmedian(centers, axis=0)
+        if np.all(np.isfinite(xy)):
+            logger.info(
+                f"Coronagraph centre from {name}: "
+                f"(x, y) = ({xy[0]:.2f}, {xy[1]:.2f})",
+                extra={"step": "flux_psf_calibration", "status": "center_from_disk"},
+            )
+            return float(xy[0]), float(xy[1])
+        logger.warning(f"{name} holds no finite centre; trying the next source.")
+
+    if irdis_filter_comb is not None:
+        x, y = nominal_star_positions(irdis_filter_comb)[0]
+        logger.warning(
+            f"No measured coronagraph centre on disk; falling back to the "
+            f"{irdis_filter_comb} nominal star position "
+            f"(x, y) = ({x:.2f}, {y:.2f}).",
+            extra={"step": "flux_psf_calibration", "status": "center_from_nominal"},
+        )
+        return float(x), float(y)
+
+    logger.warning(
+        "No measured coronagraph centre on disk; the centre guess will run "
+        "without a coronagraph mask, so residual persistence from the CORO "
+        "sequence is not rejected (see #83).",
+        extra={"step": "flux_psf_calibration", "status": "no_coronagraph_mask"},
+    )
+    return None
 
 
 def finalize_psf_cube(cube, logger):
@@ -250,15 +305,25 @@ def run_flux_psf_calibration(
     # First, compute guess positions for all frames
     logger.debug("Computing initial guess positions for all flux frames")
[…]
