### 2026-09-17 3014105
Restrict the broadband mosaic fallback to the flat template
The regular TRAP detection map and candidate table carry no template
information, so serving them for "L-type" or "T-type" presented a
non-template map under a template-matched label. plot_trap_mosaics
defaults to all three templates, so a broadband tree produced three
figures whose panels were identical but differently titled.

Only "flat" falls back now; the other template types stay missing for
broadband and render as blank panels.

This also matters once #134 lands: with template_names=("flat",) a
broadband folder gets a flat temp
[…]
