### 2026-09-17 c1e4774
Replace healpy with astropy-healpix
healpy was pulled in for a single call, hp.ang2pix, used to index
observations by sky position so they can be grouped spatially. It is
GPL-2.0 around libhealpix_cxx and needs cfitsio, which is why there are
no Windows wheels; we used none of its spherical-harmonic, visualization
or map I/O surface.

astropy-healpix is BSD-3 like spherical, depends only on numpy and
astropy, and ships wheels for macOS, Linux and win_amd64. It takes
longitude and latitude directly, so the manual colatitude conversion
goes away.

That conversion was wrong: it computed dec + pi/
[…]
