### 2026-09-17 d8afa4a
Document that compute_healpix_indices follows the input shape
For a scalar SkyCoord astropy-healpix returns a scalar index, not a
0-d array, so the ndarray-only annotation and Returns section were
wrong for an input the docstring explicitly allows.
