### 2026-09-18 ee1573a
Add waffle-derived minimum crop size helper
The crop floor has to keep all four waffle search boxes inside the array.
Uses the full 10*sqrt(2) spot radius rather than a per-axis 10 so that '+'
orientation sequences, whose spots land on the axes at the full radius, are
also bounded.
