### 2026-09-18 209f724
Reject even and undersized crop_size
An even crop puts TRAP's N//2 centre half a pixel off the geometric centre,
which FFT rotation and scaling do not tolerate. A crop below the waffle floor
cuts into the spot search boxes. Both are rejected rather than adjusted so the
configured size is the size that is used. Default moves 512 -> 257.
