### 2026-09-18 9defaf4
Raise when a waffle search box leaves the frame
A negative slice start indexes from the far end, so an out-of-bounds search box
produced a cutout from the wrong part of the frame and a confident-looking fit
rather than an error.
