### 2026-09-18 1959d25
Compute crop origins once per observation, centred on the odd pixel
round(star) - N//2 puts the nominal star within half a pixel of the central
pixel; round(star - N/2) was half a pixel off for odd N. Computing the origins
once and reusing them for every frame type makes the shared-origin requirement
between CORO and CENTER structural rather than coincidental.
