+++ tests/pipeline/test_irdis_preprocess.py
@@ -408,24 +408,24 @@ class TestApplyCrop:
     def test_crops_square_around_star(self):
         cube = np.ones((3, 1024, 1024), dtype=np.float32)
         ivar = np.ones_like(cube)
-        cube_c, ivar_c, (x0, y0) = apply_crop(cube, ivar, star_xy=(500.0, 500.0), crop_size=200)
-        assert cube_c.shape == (3, 200, 200)
-        assert ivar_c.shape == (3, 200, 200)
+        cube_c, ivar_c, (x0, y0) = apply_crop(cube, ivar, star_xy=(500.0, 500.0), crop_size=201)
+        assert cube_c.shape == (3, 201, 201)
+        assert ivar_c.shape == (3, 201, 201)
         assert x0 == 400
         assert y0 == 400
 
     def test_offset_clamped_to_frame(self):
         cube = np.ones((2, 100, 100), dtype=np.float32)
         ivar = np.ones_like(cube)
-        cube_c, ivar_c, (x0, y0) = apply_crop(cube, ivar, star_xy=(10.0, 90.0), crop_size=50)
+        cube_c, ivar_c, (x0, y0) = apply_crop(cube, ivar, star_xy=(10.0, 90.0), crop_size=51)
         assert x0 == 0
-        assert y0 == 50
+        assert y0 == 49
 
     def test_preserves_values(self):
         cube = np.arange(1024 * 1024, dtype=np.float32).reshape(1, 1024, 1024)
         ivar = cube.copy()
-        cube_c, _, (x0, y0) = apply_crop(cube, ivar, star_xy=(500.0, 500.0), crop_size=100)
-        np.testing.assert_array_equal(cube_c[0], cube[0, y0:y0 + 100, x0:x0 + 100])
+        cube_c, _, (x0, y0) = apply_crop(cube, ivar, star_xy=(500.0, 500.0), crop_size=101)
+        np.testing.assert_array_equal(cube_c[0], cube[0, y0:y0 + 101, x0:x0 + 101])
 
 
 def _write_raw_irdis_file(path, n_dit=1, level=500.0):
@@ -983,3 +983,87 @@ class TestCropFloorEnforcement:
                 converted_outputdir=tmp_path / "converted",
                 logger=MagicMock(),
             )
+
+
+class TestCropOrigins:
+    def test_origin_is_round_star_minus_half(self):
+        from spherical.pipeline.steps.irdis_preprocess import crop_origins_for_channels
+
+        stars = np.array([[480.0, 524.7], [482.5, 511.4]])
+        origins = crop_origins_for_channels(stars, crop_size=257)
+        # round(480) - 128, round(524.7) - 128
+        np.testing.assert_array_equal(origins[0], [352, 397])
+        # round(482.5) == 482 (banker's rounding), round(511.4) == 511
+        np.testing.assert_array_equal(origins[1], [354, 383])
+
+    def test_star_lands_within_half_a_pixel_of_the_centre(self):
+        from spherical.pipeline.steps.irdis_preprocess import crop_origins_for_channels
+
+        n = 257
+        stars = np.array([[480.0, 524.7], [482.5, 511.4]])
+        origins = crop_origins_for_channels(stars, crop_size=n)
+        for ch in range(2):
+            in_crop = stars[ch] - origins[ch]
+            np.testing.assert_allclose(in_crop, [n // 2, n // 2], atol=0.5)
+
+    def test_even_crop_size_rejected(self):
+        from spherical.pipeline.steps.irdis_preprocess import crop_origins_for_channels
+
+        with pytest.raises(ValueError, match="odd"):
[…]
