### 2026-09-18 5ab7bcf
Crop IRDIS frames right after the background fit (#151)
Every per-frame operation after the background fit now runs on the crop instead
of the full detector half: a 12.6x reduction in pixels processed at a 257 px
crop. A 16 px working margin, wider than the bad-pixel fixer's 21x21 window and
the 7x7 sigma-clip box, is carried through the chunk and trimmed before the
frame is stored, so delivered pixels are bit-identical to a full-frame run.

FLUX stays full-frame: its star is deliberately offset from the coronagraph, so
a box on the coronagraph nominal is not the box the PSF is in. Its header 
[…]
