+++ tests/pipeline/test_irdis_preprocess.py
@@ -1067,3 +1067,200 @@ class TestApplyCropOrigin:
         assert cube_c.shape == (1, 257, 257)
         assert cube_c[0, 257 // 2, 257 // 2] == 1.0
         assert (x0, y0) == (480 - 128, 525 - 128)
+
+
+class TestEarlyCropBitIdentity:
+    """The margin is what makes cropped output exactly equal to full-frame output."""
+
+    def _run(self, tmp_path, crop_origins, crop_size, nsigma=0.0):
+        from astropy.io import fits
+
+        from spherical.pipeline.steps.irdis_preprocess import preprocess_frame_type
+
+        rng = np.random.default_rng(20260918)
+        dm = dead_region_mask()
+        flat = np.ones((2, 1024, 1024), dtype=np.float32)
+        flat[dm] = np.nan
+        bg = np.full((2, 1024, 1024), 100.0, dtype=np.float32)
+        bg[dm] = np.nan
+        bpm = np.zeros((2, 1024, 1024), dtype=bool)
+        # Scatter bad pixels through the crop region so the 21x21 fixer window
+        # actually reaches across the crop boundary.
+        bpm[:, 380:680, 330:640] = rng.random((2, 300, 310)) < 0.01
+
+        path = tmp_path / "raw_shared.fits"
+        if not path.exists():
+            frame = rng.normal(300.0, 20.0, size=(2, 1024, 2048)).astype(np.float32)
+            fits.writeto(path, frame, overwrite=True)
+
+        cfg = IRDISPreprocessConfig(
+            crop=crop_origins is not None,
+            crop_size=crop_size,
+            fix_badpix=True,
+            transient_nsigma=nsigma,
+        )
+        return preprocess_frame_type(
+            [str(path)], flat, bg, bpm,
+            np.array([[480.0, 524.7], [482.5, 511.4]]),
+            is_flux=False, preprocess_config=cfg, logger=MagicMock(),
+            crop_origins=crop_origins,
+        )
+
+    def test_cropped_region_is_bit_identical_to_full_frame(self, tmp_path):
+        from spherical.pipeline.steps.irdis_preprocess import crop_origins_for_channels
+
+        n = 257
+        origins = crop_origins_for_channels(
+            np.array([[480.0, 524.7], [482.5, 511.4]]), crop_size=n
+        )
+        full_cube, full_ivar, _, full_off = self._run(tmp_path, None, n)
+        crop_cube, crop_ivar, _, crop_off = self._run(tmp_path, origins, n)
+
+        assert full_off is None
+        np.testing.assert_array_equal(crop_off, origins)
+        assert crop_cube.shape == (2, 2, n, n)
+
+        for ch in range(2):
+            x0, y0 = int(origins[ch, 0]), int(origins[ch, 1])
+            np.testing.assert_array_equal(
+                crop_cube[ch], full_cube[ch][:, y0:y0 + n, x0:x0 + n]
+            )
+            np.testing.assert_array_equal(
+                crop_ivar[ch], full_ivar[ch][:, y0:y0 + n, x0:x0 + n]
+            )
+
+    def test_bit_identical_with_the_transient_clip_on(self, tmp_path):
+        from spherical.pipeline.steps.irdis_preprocess import crop_origins_for_channels
+
+        n = 257
+        origins = crop_origins_for_channels(
+            np.array([[480.0, 524.7], [482.5, 511.4]]), crop_size=n
+        )
+        full_cube, _, _, _ = self._run(tmp_path, None, n, nsigma=5.0)
+        crop_cube, _, _, _ = self._run(tmp_path, origins, n, nsigma=5.0)
+        for ch in range(2):
[…]
