+++ src/spherical/pipeline/steps/irdis_preprocess.py
@@ -722,20 +722,48 @@ def _worker_init(
     fix_badpix: bool,
     warn_threshold: int,
     transient_nsigma: float,
+    crop_origins: np.ndarray | None = None,
+    crop_size: int | None = None,
 ) -> None:
     """Initializer for both ProcessPoolExecutor workers and the serial path.
 
     Every call fully overwrites ``_WORKER_STATE`` so back-to-back invocations
     from ``run_irdis_preprocess`` (one per frame type) don't leak state.
+
+    When ``crop_origins`` is given, the per-channel working boxes and the
+    correspondingly cropped flat, bpm and dead mask are derived here once: they
+    are static per channel, so a worker pays for them at startup instead of per
+    frame. The full-resolution ``master_bg``, flat, bpm and dead mask stay
+    available — the background fit runs on the full frame.
     """
     from spherical.pipeline.steps.irdis_calibration import dead_region_mask
+
+    dead = dead_region_mask()
+    crop_boxes = None
+    flat_crop = None
+    bpm_crop = None
+    dead_crop = None
+    if crop_origins is not None:
+        frame_shape = (master_flat.shape[-2], master_flat.shape[-1])
+        crop_boxes = [
+            crop_working_box(
+                (int(crop_origins[ch, 0]), int(crop_origins[ch, 1])),
+                crop_size,
+                CROP_WORKING_MARGIN,
+                frame_shape,
+            )
+            for ch in range(2)
+        ]
+        flat_crop = [master_flat[ch][ys, xs] for ch, (ys, xs, _) in enumerate(crop_boxes)]
+        bpm_crop = [bpm_bool[ch][ys, xs] for ch, (ys, xs, _) in enumerate(crop_boxes)]
+        dead_crop = [dead[ch][ys, xs] for ch, (ys, xs, _) in enumerate(crop_boxes)]
 
     _WORKER_STATE.clear()
     _WORKER_STATE.update(
         master_flat=master_flat,
         master_bg=master_bg,
         bpm_bool=bpm_bool,
-        dead=dead_region_mask(),
+        dead=dead,
         star_positions=star_positions_xy,
         mask_radius=mask_radius,
         is_flux=is_flux,
@@ -744,6 +772,11 @@ def _worker_init(
         fix_badpix=fix_badpix,
         warn_threshold=warn_threshold,
         transient_nsigma=transient_nsigma,
+        crop_boxes=crop_boxes,
+        crop_size=crop_size,
+        flat_crop=flat_crop,
+        bpm_crop=bpm_crop,
+        dead_crop=dead_crop,
     )
 
 
@@ -761,7 +794,8 @@ def _process_chunk(
     Returns
     -------
     cube_slice : np.ndarray
-        ``(2, n_chunk, 1024, 1024)`` float32.
+        ``(2, n_chunk, ny, nx)`` float32, where ``ny = nx = crop_size`` when the
+        worker was initialized with crop origins and ``1024`` otherwise.
     ivar_slice : np.ndarray
         Same shape.
     transient_counts : list of (frame_idx, ch, n_transient)
@@ -788,9 +822,19 @@ def _process_chunk(
     do_fix_badpix = _WORKER_STATE["fix_badpix"]
     warn_threshold = _WORKER_STATE["warn_threshold"]
     transient_nsigma = _WORKER_STATE["transient_nsigma"]
+    crop_boxes = _WORKER_STATE["crop_boxes"]
+    crop_size = _WORKER_STATE["crop_size"]
+    flat_crop = _WORKER_STATE["flat_crop"]
+    bpm_crop = _WORKER_STATE["bpm_crop"]
[…]
