### 2026-09-18 19c8cee
Crop badpixel_map.fits with the science cubes
TRAP indexes bad_pixel_mask_full against the data cube, so a full-frame map
beside a cropped cube was a silent shape mismatch. Pre-existing with crop=True.
