### 2026-09-18 f9cd6fc
Keep cross_channel_offset.fits in detector coordinates
The two channels are cropped about their own stars, so a plain difference of
crop-frame centres drops the origin difference: K-band's (2.5, -13.3) collapsed
to (0.5, 0.7). The file is consumed as a detector-frame quantity.
