### 2026-09-18 7fdaade
Compare seeds in crop coordinates in process_centers
The nominal is in detector coordinates and image_centers.fits is in crop
coordinates, so the logged seed delta was reporting the crop origin. In the DMS
branch the nominal is also the all-NaN fallback for the anchor S0, where the
frame mismatch was a wrong value rather than a misleading log: every propagated
centre for that channel was offset by the crop origin.
