+++ tests/pipeline/test_process_centers_irdis.py
@@ -7,6 +7,31 @@ import numpy as np
 from astropy.io import fits
 
 
+def _write_center_cube(tmp_path, crop_origins=None, crop_size=257):
+    """Minimal center_cube.fits so the crop-frame lookup has a header to read.
+
+    `process_centers` always runs after preprocess, so the cube is always on disk
+    in the real pipeline. Pass `crop_origins` as ((x0, y0), (x0, y0)) to simulate
+    a cropped reduction.
+    """
+    header = fits.Header()
+    header["HIERARCH SPHERICAL CROP APPLIED"] = crop_origins is not None
+    n = 1024
+    if crop_origins is not None:
+        n = crop_size
+        header["HIERARCH SPHERICAL CROP SIZE"] = crop_size
+        header["HIERARCH SPHERICAL CROP X0 CH0"] = int(crop_origins[0][0])
+        header["HIERARCH SPHERICAL CROP Y0 CH0"] = int(crop_origins[0][1])
+        header["HIERARCH SPHERICAL CROP X0 CH1"] = int(crop_origins[1][0])
+        header["HIERARCH SPHERICAL CROP Y0 CH1"] = int(crop_origins[1][1])
+    fits.writeto(
+        tmp_path / "center_cube.fits",
+        np.zeros((2, 1, n, n), dtype=np.float32),
+        header=header,
+        overwrite=True,
+    )
+
+
 def _make_image_centers(tmp_path, n_time=100, jitter_amplitude=0.05, outliers=(50,)):
     rng = np.random.default_rng(0)
     centers = np.zeros((2, n_time, 2), dtype=np.float32)
@@ -19,6 +44,7 @@ def _make_image_centers(tmp_path, n_time=100, jitter_amplitude=0.05, outliers=(5
         centers[1, t, 1] += 40.0  # large outlier in ch1 y
     fits.writeto(tmp_path / "image_centers.fits", centers, overwrite=True)
     fits.writeto(tmp_path / "wavelengths.fits", np.array([2100.0, 2250.0]), overwrite=True)
+    _write_center_cube(tmp_path)
     return centers
 
 
@@ -146,6 +172,7 @@ class TestIRDISWaffleCenterFit:
         centers = np.zeros((n_wave, n_time, 2), dtype=np.float32) + 128.0
         fits.writeto(tmp_path / "image_centers.fits", centers, overwrite=True)
         fits.writeto(tmp_path / "wavelengths.fits", np.linspace(1000, 1600, n_wave), overwrite=True)
+        _write_center_cube(tmp_path)
 
         observation = MagicMock()
         observation.observation = {"INSTRUMENT": ["IFS"]}
@@ -256,6 +283,7 @@ class TestIRDISDmsPropagation:
         center_centers[1] = [[28.0, 27.0]] * 3
         fits.writeto(tmp_path / "image_centers.fits", center_centers, overwrite=True)
         fits.writeto(tmp_path / "wavelengths.fits", np.array([2100.0, 2250.0]), overwrite=True)
+        _write_center_cube(tmp_path)
 
         pd.DataFrame({
             "MJD": [0.0, 0.5, 1.0],
@@ -296,6 +324,7 @@ class TestIRDISDmsPropagation:
         center_centers[1] = [[28.0 + 1.0, 27.0 + 1.0]] * 2   # true S₀ ch1 = (28, 27)
         fits.writeto(tmp_path / "image_centers.fits", center_centers, overwrite=True)
         fits.writeto(tmp_path / "wavelengths.fits", np.array([2100.0, 2250.0]), overwrite=True)
+        _write_center_cube(tmp_path)
 
         pd.DataFrame({
[…]
