+++ src/spherical/pipeline/steps/process_centers.py
@@ -53,9 +53,45 @@ def run_polynomial_center_fit(
     _run_ifs_polynomial_center_fit(converted_dir, extraction_parameters, non_least_square_methods, logger)
 
 
-def _run_irdis_temporal_center_fit(converted_dir: str, observation, logger) -> None:
+def _nominal_in_crop_frame(converted_dir: str, filter_comb: str) -> np.ndarray:
+    """Filter nominal star positions expressed in the cube's own coordinates.
+
+    ``nominal_star_positions`` is in per-half detector coordinates, but everything
+    in ``converted/`` is in crop coordinates. Differencing the two directly makes
+    the crop origin look like a seed error — and where the nominal is used as a
+    *value* rather than for a log, it is simply the wrong number.
+
+    Parameters
+    ----------
+    converted_dir : str
+        The observation's ``converted/`` directory.
+    filter_comb : str
+        IRDIS filter combination string.
+
+    Returns
+    -------
+    np.ndarray
+        Shape ``(n_wave, 2)`` ``(x, y)`` in the cube's coordinate frame.
+    """
     from spherical.pipeline.steps.find_star import nominal_star_positions
 
+    nominal = nominal_star_positions(filter_comb).astype(np.float64)
+    header = fits.getheader(os.path.join(converted_dir, "center_cube.fits"))
+    if not bool(header.get("HIERARCH SPHERICAL CROP APPLIED", False)):
+        return nominal
+    nominal = nominal.copy()
+    nominal[:, 0] -= np.array([
+        int(header.get("HIERARCH SPHERICAL CROP X0 CH0", 0)),
+        int(header.get("HIERARCH SPHERICAL CROP X0 CH1", 0)),
+    ])
+    nominal[:, 1] -= np.array([
+        int(header.get("HIERARCH SPHERICAL CROP Y0 CH0", 0)),
+        int(header.get("HIERARCH SPHERICAL CROP Y0 CH1", 0)),
+    ])
+    return nominal
+
+
+def _run_irdis_temporal_center_fit(converted_dir: str, observation, logger) -> None:
     image_centers = np.asarray(
         fits.getdata(os.path.join(converted_dir, "image_centers.fits")),
         dtype=np.float32,
@@ -84,7 +120,7 @@ def _run_irdis_temporal_center_fit(converted_dir: str, observation, logger) -> N
     # a physical realignment, not a bug).
     n_wave = image_centers.shape[0]
     filter_comb = str(observation.observation["FILTER"][0])
-    nominal = nominal_star_positions(filter_comb)  # (n_wave, 2) in (x, y)
+    nominal = _nominal_in_crop_frame(converted_dir, filter_comb)
     measured_median = np.nanmedian(image_centers, axis=1)  # (n_wave, 2)
     for ch in range(n_wave):
         mx, my = float(measured_median[ch, 0]), float(measured_median[ch, 1])
@@ -188,8 +224,6 @@ def _run_irdis_dms_propagation(converted_dir: str, observation, logger) -> None:
     """
     import pandas as pd
 
-    from spherical.pipeline.steps.find_star import nominal_star_positions
-
     PIXEL_SCALE_UM = 18.0
 
     center_centers = np.asarray(
@@ -214,14 +248,15 @@ def _run_irdis_dms_propagation(converted_dir: str, observation, logger) -> None:
     S0_scatter = np.nanstd(per_center_S0, axis=1)                  # (n_wave, 2)
 
     filter_comb = str(observation.observation["FILTER"][0])
+    # S0 is a crop-frame anchor, so the fallback nominal has to be one too.
[…]
