### 2026-09-18 03620e6
Changelog for the early crop (#151)
