+++ CHANGELOG.md
@@ -24,12 +24,47 @@ This project follows [Semantic Versioning](https://semver.org/) and the [Keep a 
   ([#128](https://github.com/m-samland/spherical/pull/128), [@tomasstolker](https://github.com/tomasstolker)).
 
 ### 🔧 Changed
+- **IRDIS cropping moved to the start of preprocessing** – The crop used to be the last operation before writing, so flat division, inverse variance, bad-pixel repair and the transient clip all ran on the full 1024×1024 detector half and were then thrown away.
+  It now happens immediately after the background fit, which is the last step that genuinely needs the full frame (it fits on everything outside a 285 px star mask).
+  A 16 px working margin — wider than the bad-pixel fixer's 21×21 window and the 7×7 sigma-clip box — is carried through the per-frame loop and trimmed before the frame is stored, so every delivered pixel is *bit-identical* to a full-frame run rather than merely close.
+  At a 257 px crop this processes 12.6× fewer pixels
+  ([#151](https://github.com/m-samland/spherical/issues/151), [@m-samland](https://github.com/m-samland)).
+- **`crop_size` must be odd, and defaults to 257 instead of 512** – TRAP takes the image centre as `yx_dim[0] // 2`; for odd N that integer *is* the array's geometric centre, so TRAP's convention, the geometric centre and the pixel the star sits on are one point.
+  For even N they differ by half a pixel, which FFT rotation and scaling do not tolerate, and an even axis also carries an unpaired Nyquist bin that leaks ringing into a real-valued FFT shift.
+  An even value raises `ValueError` at config construction rather than being rounded, so the configured size is always the size that is used.
+  The crop origin changed from `round(star − N/2)` to `round(star) − N//2`, which puts the nominal star within half a pixel of the central pixel
+  ([#151](https://github.com/m-samland/spherical/issues/151), [@m-samland](https://github.com/m-samland)).
+- **`crop_size` is validated against a per-band floor before any data is read** – A crop smaller than the waffle-spot search boxes would make the centre fit cut into the spots.
+  The floor is `2·(10√2·λ/D + 16 + 10)` rounded up to odd, evaluated at the filter's longest wavelength: 189 px for `DB_K12`, 185 for `BB_Ks`, 153 for `DB_H23`, 151 for `BB_H`.
+  The full `10√2` spot radius is used rather than a per-axis `10` because with `'+'` waffle orientation the spots land on the axes at the full radius, and a per-axis figure would under-size the crop for exactly those sequences.
+  Available as `pipeline.steps.find_star.minimum_crop_size()`
+  ([#151](https://github.com/m-samland/spherical/issues/151), [@m-samland](https://github.com/m-samland)).
+- **FLUX frames are no longer cropped** – The FLUX star is deliberately offset from the coronagraph, so a box on the coronagraph nominal is not the box the PSF is in.
+  Cropping a handful of frames saved nothing measurable, and the 57×57 PSF stamp extraction already decouples the FLUX cube size from everything downstream.
+  `flux_cube.fits` now reports `SPHERICAL CROP APPLIED = F` rather than echoing the config flag
+  ([#151](https://github.com/m-samland/spherical/issues/151), [@m-samland](https://github.com/m-samland)).
+- **CORO and CENTER are guaranteed to share one crop origin** – The origins are computed once per observation and reused for every frame type, instead of being recomputed per frame type.
+  They must match: a CORO-derived speckle model is subtracted from CENTER frames pixel-for-pixel, and the non-waffle path propagates CENTER-measured centres onto CORO frames with no coordinate bookkeeping between them, so differing origins would offset the propagated centres by tens of pixels with no error raised
[…]
