### 2026-09-18 85332b4
Record the measured crop speedup on real data
12.6x is the pixel-count reduction for the stages after the background fit; the
whole step measures 3.7x on 51 Eri DB_K12, since the background fit is
unavoidably full-frame and now dominates.
