### 2026-09-18 392c166
Add a leaf flag to StepSpec and keep leaves out of the force cascade
A leaf step's outputs feed nothing downstream, so forcing one should re-run only
itself rather than dragging every later step (TRAP included) along with it.
Being forced by an earlier step still works.

A force name the given registry does not know counts as non-leaf, preserving the
behaviour of call sites that pass an instrument step_order with the default
registry.
