### 2026-09-18 049b8ee
Repair imutils.shift and guard _shift_fft against non-square arrays
collections.Iterable was removed in Python 3.10, so the public shift() wrapper
raised AttributeError for any sequence shift value. _shift_fft builds its ramp
with 2*pi/Nx on both axes and mixes Nx/Ny, so it is only correct for square
arrays; that is now enforced rather than assumed.
