### 2026-09-18 25b58e0
Extract science-frame helpers from the TRAP wrapper
Frame-type selection, centre normalization and the frame-axis check are needed
by frame alignment too, and neither implies a TRAP dependency.
