### 2026-09-18 e9bba1a
Add the pad-shift-trim primitive for frame alignment
Padding is what makes the border exact: an FFT shift is periodic and would wrap
flux across the frame, and a cubic spline reaches 2 px past the border. 'auto'
picks FFT on clean frames and a spline where NaN is, because FFT ringing from a
filled NaN edge is global while the spline keeps it local.

_shift_fft un-shifts its phase ramp with ifftshift rather than fftshift. The two
agree for even N, which is what the old even-width restriction quietly relied
on, but for odd N fftshift is off by one frequency bin: a unit delta shifted in
a 17x17 a
[…]
