+++ tests/pipeline/test_align_frames.py
@@ -43,3 +43,145 @@ class TestImutilsShiftRepair:
     def test_square_fft_shift_does_not_raise(self):
         frame = np.zeros((16, 16))
         imutils._shift_fft(frame, np.array([1.0, 1.0]))
+
+    @pytest.mark.parametrize("n", [16, 17, 65])
+    def test_integer_shift_is_exact_at_both_parities(self, n):
+        """The ramp must be un-shifted with ifftshift, which differs for odd n."""
+        frame = np.zeros((n, n))
+        frame[n // 2, n // 2] = 1.0
+        shifted = imutils.shift(frame, (3.0, -2.0), method="fft")
+        peak = np.unravel_index(np.argmax(shifted), shifted.shape)
+        assert peak == (n // 2 - 2, n // 2 + 3)
+        assert shifted.max() == pytest.approx(1.0, abs=1e-9)
+
+    def test_odd_square_frame_is_accepted(self):
+        frame = np.zeros((17, 17))
+        imutils.shift(frame, (1.5, 0.0), method="fft")
+
+    def test_non_square_shift_still_rejected(self):
+        with pytest.raises(ValueError, match="square"):
+            imutils.shift(np.zeros((16, 32)), (1.5, 0.0), method="fft")
+
+
+from spherical.pipeline.steps.align_frames import (  # noqa: E402
+    DEFAULT_PAD_WIDTH,
+    pad_to_odd,
+    shift_frame,
+    shift_to_target,
+)
+
+
+def _gaussian(n, cx, cy, sigma=2.0):
+    yy, xx = np.mgrid[:n, :n]
+    return np.exp(-((xx - cx) ** 2 + (yy - cy) ** 2) / (2 * sigma ** 2))
+
+
+def _centroid(frame):
+    finite = np.where(np.isfinite(frame), frame, 0.0)
+    yy, xx = np.mgrid[: frame.shape[0], : frame.shape[1]]
+    total = finite.sum()
+    return (float((xx * finite).sum() / total), float((yy * finite).sum() / total))
+
+
+class TestPadToOdd:
+    def test_even_axis_padded_at_the_high_edge(self):
+        cube = np.arange(2 * 3 * 4 * 4, dtype=np.float32).reshape(2, 3, 4, 4)
+        out = pad_to_odd(cube)
+        assert out.shape == (2, 3, 5, 5)
+        np.testing.assert_array_equal(out[..., :4, :4], cube)
+        np.testing.assert_array_equal(out[..., 4, :4], cube[..., 3, :4])
+        np.testing.assert_array_equal(out[..., :4, 4], cube[..., :4, 3])
+
+    def test_odd_axis_untouched(self):
+        cube = np.zeros((2, 3, 5, 5), dtype=np.float32)
+        out = pad_to_odd(cube)
+        assert out.shape == cube.shape
+        assert out is cube
+
+    def test_ifs_262_becomes_263_with_the_same_target_pixel(self):
+        cube = np.zeros((39, 2, 262, 262), dtype=np.float32)
+        out = pad_to_odd(cube)
+        assert out.shape[-2:] == (263, 263)
+        assert 262 // 2 == 263 // 2 == 131
+
+
+class TestShiftFrame:
+    @pytest.mark.parametrize("method", ["fft", "interp", "auto"])
+    def test_subpixel_shift_moves_the_centroid(self, method):
+        n = 65
+        frame = _gaussian(n, 30.0, 30.0)
+        shifted = shift_frame(frame, (2.4, -1.7), method=method)
+        cx, cy = _centroid(shifted)
+        assert cx == pytest.approx(32.4, abs=0.05)
+        assert cy == pytest.approx(28.3, abs=0.05)
+
+    def test_coarse_rounds_to_an_integer_shift(self):
[…]
