+++ src/spherical/pipeline/steps/align_frames.py
@@ -1,0 +1,161 @@
+"""Optional frame alignment: a science cube with the star on a fixed pixel.
+
+A leaf product. Nothing in the pipeline consumes it — TRAP forward-models the
+centre instead, spot photometry uses spot positions on the unshifted cube, and
+the flux PSF is an independently extracted stamp. It exists for external
+consumers: classical ADI/PCA, SDI, and visual inspection.
+
+No inverse variance is written alongside it. Interpolation correlates
+neighbouring pixels, so a shifted ivar is no longer a valid per-pixel weight;
+that is the same reason TRAP works on unshifted data.
+"""
+from __future__ import annotations
+
+import numpy as np
+
+# Padding carried around the frame during the shift and removed afterwards.
+# An FFT shift is periodic and would wrap flux from one edge to the other; a
+# cubic spline's support reaches 2 px past the border. Shifts are sub-pixel once
+# the crop origin puts the star within half a pixel of the centre, so 8 is
+# generous for both.
+DEFAULT_PAD_WIDTH = 8
+
+_VALID_METHODS = ("auto", "fft", "interp", "coarse")
+
+
+def pad_to_odd(cube: np.ndarray) -> np.ndarray:
+    """Pad the last two axes to odd length at the high edge, by edge replication.
+
+    IFS cubes arrive at 262x262 from the charis extraction. With an even axis
+    TRAP's ``N // 2`` centre sits half a pixel off the geometric centre, and the
+    axis carries a lone Nyquist bin with no conjugate partner that leaks ringing
+    into a real-valued FFT shift. Padding 262 to 263 fixes both without losing
+    data or resampling: the target pixel is 131 either way, and in a 263 array
+    131 is both ``N // 2`` and the exact geometric centre.
+
+    Edge replication rather than NaN, so the added row and column do not create a
+    new invalid stripe that the NaN-restore path then has to carry.
+
+    Args:
+        cube: Any array whose last two axes are spatial.
+
+    Returns:
+        The input unchanged when both axes are already odd, otherwise a padded
+        copy.
+    """
+    h, w = cube.shape[-2], cube.shape[-1]
+    pad_y, pad_x = (h % 2 == 0), (w % 2 == 0)
+    if not (pad_y or pad_x):
+        return cube
+    pad_width = [(0, 0)] * (cube.ndim - 2) + [(0, int(pad_y)), (0, int(pad_x))]
+    return np.pad(cube, pad_width, mode="edge")
+
+
+def _resolve_method(frame: np.ndarray, method: str) -> str:
+    """Resolve ``"auto"`` to ``"fft"`` on clean frames, ``"interp"`` where NaN is."""
+    if method != "auto":
+        return method
+    return "interp" if np.isnan(frame).any() else "fft"
+
+
+def shift_frame(
+    frame: np.ndarray,
+    shift_xy: tuple[float, float],
+    method: str = "auto",
+    pad: int = DEFAULT_PAD_WIDTH,
+) -> np.ndarray:
+    """Shift one square frame by ``(dx, dy)`` pixels, preserving its shape.
+
+    ``"auto"`` picks FFT when the frame has no NaN and a cubic spline when it
+    does. FFT avoids the spline's photometric smoothing, but it is global, so
+    ringing from a filled NaN edge would spread across the whole frame; the
+    spline keeps it local. ``"coarse"`` rounds the shift to an integer and does
+    no interpolation at all.
+
+    NaN is filled before the shift and restored afterwards from a
+    nearest-neighbour-shifted validity mask, so invalid pixels move with the data
+    instead of growing.
+
[…]
