### 2026-09-18 82b8958
Add AlignmentConfig and the opt-in align_frames step flag
