### 2026-09-18 1b0b54a
Add the align_frames step (#161)
Writes a copy of the science cube with the star on N//2. Derives instrument,
crop offsets and waffle mode from files on disk, so it re-runs standalone on an
already-reduced dataset. No shifted inverse variance: interpolation correlates
neighbouring pixels, so it would not be a valid per-pixel weight. The IFS
bad-pixel repair is a logged placeholder.
