+++ src/spherical/pipeline/steps/align_frames.py
@@ -159,3 +159,186 @@ def shift_to_target(
         method=method,
         pad=pad,
     )
+
+
+def infer_continuous_satellite_spots(converted_dir) -> bool:
+    """Decide from files on disk whether CENTER frames are the science frames.
+
+    Two signals, in order:
+
+    1. ``coro_cube.fits`` is a symlink. The IRDIS orchestrator links CENTER to
+       CORO only when the observation has no CORO frames at all, so the CENTER
+       frames must be the science frames.
+    2. The centre array's frame axis matches exactly one of the two frame tables.
+
+    Args:
+        converted_dir: The observation's ``converted/`` directory.
+
+    Returns:
+        ``True`` when the CENTER frames carry the science.
+
+    Raises:
+        ValueError: When both tables have the same row count, so the frame axis
+            cannot distinguish them. Pass ``continuous_satellite_spots``
+            explicitly.
+    """
+    from pathlib import Path
+
+    import pandas as pd
+    from astropy.io import fits
+
+    converted_dir = Path(converted_dir)
+    if (converted_dir / "coro_cube.fits").is_symlink():
+        return True
+
+    n_centers = fits.getdata(
+        converted_dir / "image_centers_fitted_robust.fits"
+    ).shape[1]
+
+    def _rows(identifier: str) -> int | None:
+        path = converted_dir / f"frames_info_{identifier}.csv"
+        return len(pd.read_csv(path)) if path.exists() else None
+
+    n_center, n_coro = _rows("center"), _rows("coro")
+    if n_coro is None:
+        return True
+    if n_center is None:
+        return False
+    if n_centers == n_coro and n_centers != n_center:
+        return False
+    if n_centers == n_center and n_centers != n_coro:
+        return True
+    raise ValueError(
+        f"Cannot infer the science frame type: image_centers_fitted_robust has "
+        f"{n_centers} frames and both frames_info_center.csv ({n_center}) and "
+        f"frames_info_coro.csv ({n_coro}) are ambiguous. Pass "
+        "continuous_satellite_spots explicitly."
+    )
+
+
+def _run_bad_pixel_repair(instrument: str, alignment_config, logger) -> bool:
+    """Bad-pixel repair gate. Returns whether repaired data reaches the shift.
+
+    The scaffolding is deliberately identical for both instruments so the gap is
+    explicit in the log rather than implicit in the structure. On IRDIS the
+    requirement is already met by ``fix_badpix`` in preprocess. On IFS the
+    interpolation is not implemented: charis marks bad lenslets as ``ivar == 0``
+    and the repair has to work on the extracted spaxel grid rather than the
+    detector, which is its own design.
+    """
+    if not alignment_config.repair_bad_pixels:
+        return False
+    if instrument == "IRDIS":
+        return True
+    logger.warning(
+        "IFS bad-pixel interpolation not yet implemented. Skipped.",
+        extra={"step": "frame_alignment", "status": "repair_not_implemented"},
+    )
+    return False
+
+
+def run_frame_alignment(
+    converted_dir,
+    alignment_config,
+    logger,
+    continuous_satellite_spots: bool | None = None,
+):
+    """Write a copy of the science cube with the star on the centre pixel.
+
+    Everything else is derived from files on disk — the instrument from the
+    wavelength axis, the crop offsets from the header, the science frame type
+    from the centres' frame axis — so this runs standalone on an already-reduced
+    dataset without an observation object.
+
+    Args:
+        converted_dir: The observation's ``converted/`` directory.
[…]
