### 2026-09-18 ee36f1f
Wire align_frames into both orchestrators as a leaf step
Gated on a marker because the output filename depends on the science frame type,
which StepDirs cannot know. Excluded from check_output so an opt-in step nobody
enabled does not make a finished reduction look incomplete.
