+++ src/spherical/pipeline/irdis_reduction.py
@@ -42,7 +42,9 @@ from spherical.pipeline.step_registry import (
     StepDirs,
     _forced,
     should_run,
+    write_marker,
 )
+from spherical.pipeline.steps.align_frames import run_frame_alignment
 from spherical.pipeline.steps.cube_header_update import run_cube_header_update
 from spherical.pipeline.steps.download_data import (
     download_data_for_observation,
@@ -303,6 +305,20 @@ def execute_irdis_target(
         ):
             run_spot_to_flux_normalization(converted_dir_str, reduction_parameters, logger=logger)
 
+        if should_run(
+            "align_frames", steps.align_frames, dirs, steps.force, logger,
+            step_order=IRDIS_STEP_ORDER, registry=IRDIS_STEP_REGISTRY,
+        ):
+            run_frame_alignment(
+                converted_dir=converted_dir,
+                alignment_config=config.alignment,
+                logger=logger,
+                continuous_satellite_spots=bool(
+                    observation.observation["WAFFLE_MODE"][0]
+                ),
+            )
+            write_marker("align_frames", converted_dir)
+
         elapsed_min = (time.time() - start) / 60.0
         logger.info(f"IRDIS pipeline finished in {elapsed_min:.2f} minutes.")
         return None
@@ -385,7 +401,7 @@ def check_output(
         )
         missing_files: list[str] = []
         for step, spec in IRDIS_STEP_REGISTRY.items():
-            if spec.internal_guard or spec.is_trap:
+            if spec.internal_guard or spec.is_trap or spec.leaf:
                 continue
             for p in expected_outputs(step, dirs, registry=IRDIS_STEP_REGISTRY):
                 if not p.exists():
