### 2026-09-18 3fd95ca
Changelog for optional frame alignment (#161)
