### 2026-09-21 f7e6343
Pin reduction_status behaviour for leaf steps after the final step
align_frames is the first step that can appear in a reduction log after the
completion gate. Completion is a sticky flag on spot_to_flux, so a leaf that
runs or fails afterwards cannot reopen it — cover both cases and correct the
comment that claimed spot_to_flux was the last pre-TRAP step.
