### 2026-09-22 482d127
refactor(pipeline): derive the science frame type from WAFFLE_MODE, not symlinks
Remove _link_center_as_coro_if_missing and the file-based inference that
consumed it. Which frame type carries the science is a property of the
observation, already encoded in WAFFLE_MODE, so nothing needs to guess it
from files on disk.

The symlink aliased coro_cube.fits onto center_cube.fits, which made an
in-place header update on one silently rewrite the other (cube_header_update
opens mode='update'). What it actually propped up was the resume check:
should_run requires every declared output to exist, and pre
[…]
