### 2026-09-22 b7be514
fix(pipeline): gate resume on the frame types present, not the science type
WAFFLE_MODE is a majority-exposure-time test, not an existence test.
select_primary_science_frames picks the frame type with the most total
exposure time and observation_table sets WAFFLE_MODE from that, so a waffle
sequence can carry CORO frames alongside the CENTER frames that hold its
science. check_frames demands CORO only when WAFFLE_MODE is False.

Declaring outputs from the science frame type therefore under-declared such
a target: preprocess writes coro_cube.fits and frame_info writes
frames_info_coro.csv whene
[…]
