+++ src/spherical/pipeline/steps/cube_header_update.py
@@ -23,8 +23,13 @@ def _stamp_waffle_mode(converted_dir, frame_types, continuous_satellite_spots, l
     ``science_frames.science_frame_type``), and the cubes are the only place a
     consumer outside the pipeline can learn it. Written after the metadata pass
     so nothing can clobber it.
+
+    Returns:
+        True when every cube that exists was stamped. A cube that is not there
+        is not a failure; a cube that could not be written is.
     """
     value = bool(continuous_satellite_spots)
+    stamped_all = True
     for frame_type in frame_types:
         path = os.path.join(converted_dir, f"{frame_type.lower()}_cube.fits")
         if not os.path.exists(path):
@@ -35,10 +40,12 @@ def _stamp_waffle_mode(converted_dir, frame_types, continuous_satellite_spots, l
                 # and astropy truncates anything appended to it.
                 hdulist[0].header[WAFFLE_KEYWORD] = value
         except Exception:
+            stamped_all = False
             logger.exception(
                 f"Failed to stamp {WAFFLE_KEYWORD} on {path}.",
                 extra={"step": "cube_header_update", "status": "failed"},
             )
+    return stamped_all
 
 
 @optional_logger
@@ -137,10 +144,22 @@ def run_cube_header_update(
         except Exception:
             logger.exception(f"Failed to update FITS header for {frame_type}.", extra={"step": f"cube_header_update_{frame_type.lower()}", "status": "failed"})
     
+    stamped_all = True
     if continuous_satellite_spots is not None:
-        _stamp_waffle_mode(
+        stamped_all = _stamp_waffle_mode(
             converted_dir, frame_types_to_extract, continuous_satellite_spots, logger
         )
 
     logger.info("FITS headers updated for all processed frame types.")
+    if not stamped_all:
+        # Reporting success here would send a later standalone frame-alignment
+        # re-run to the "re-run cube_header_update" message, naming the step
+        # that had just claimed to succeed.
+        logger.error(
+            f"Finished cube header update step, but {WAFFLE_KEYWORD} is missing from at "
+            "least one cube. A standalone frame-alignment re-run cannot resolve the "
+            "science frame type from these cubes.",
+            extra={"step": "cube_header_update", "status": "failed"},
+        )
+        return
     logger.info("Finished cube header update step.", extra={"step": "cube_header_update", "status": "success"})
