+++ tests/pipeline/test_resume_skip.py
@@ -33,9 +33,9 @@ def test_step_order_covers_config_steps_and_trap_last():
 def test_expected_outputs_locations(tmp_path):
     d = _dirs(tmp_path)
     assert sr.expected_outputs("bundle_output", d) == [
+        d.converted_dir / "wavelengths.fits",
         d.converted_dir / "coro_cube.fits",
         d.converted_dir / "center_cube.fits",
-        d.converted_dir / "wavelengths.fits",
     ]
     # additional_outputs lives INSIDE converted_dir (matches the step modules).
     assert sr.expected_outputs("calibrate_spot_photometry", d) == [
@@ -133,8 +133,12 @@ def test_check_output_uses_registry_and_real_additional_dir(tmp_path, monkeypatc
     converted.mkdir(parents=True)
     monkeypatch.setattr(ir, "output_directory_path", lambda rd, obs, method='optext': str(converted) + "/")
 
-    # check_output reads WAFFLE_MODE to know which frame type is the science one.
-    observation = SimpleNamespace(observation={"WAFFLE_MODE": [False]})
+    # check_output reads the frame types the observation carries, which decide
+    # which per-frame-type products it expects on disk.
+    observation = SimpleNamespace(
+        observation={"WAFFLE_MODE": [False]},
+        frames={"CORO": [1], "CENTER": [1], "FLUX": [1]},
+    )
 
     # Create every registry output for a "complete" target.
     dirs = sr.StepDirs(converted_dir=converted, cube_outputdir=converted.parent, wavecal_outputdir=tmp_path)
@@ -150,40 +154,43 @@ def test_check_output_uses_registry_and_real_additional_dir(tmp_path, monkeypatc
     assert missing == [[]]
 
 
-class TestScienceIdentifierOutputs:
-    """Frame-type-dependent outputs follow WAFFLE_MODE, not a fixed 'coro'.
+class TestFrameTypeDependentOutputs:
+    """Outputs that exist once per frame type follow the frame types present.
 
-    A continuous-waffle observation has no CORO frames, so preprocess never
-    writes coro_cube.fits and frame_info never writes frames_info_coro.csv.
-    Declaring those unconditionally kept such a target permanently 'incomplete'
-    and re-ran the step on every invocation.
+    WAFFLE_MODE is a majority-exposure-time test, not an existence test, so it
+    cannot stand in for this. A waffle sequence with no CORO frames must not be
+    asked for CORO products, which is what kept such a target permanently
+    'incomplete' and re-ran the step on every invocation. A waffle sequence that
+    does carry CORO frames writes CORO products and must still be asked for them.
     """
 
-    def test_default_identifier_is_coro(self):
-        assert sr.StepDirs().science_identifier == "coro"
+    WAFFLE_NO_CORO = ("CENTER", "FLUX")
 
-    def test_bundle_output_follows_the_identifier(self, tmp_path):
-        d = sr.StepDirs(converted_dir=tmp_path, science_identifier="center")
-        assert sr.expected_outputs("bundle_output", d) == [
-            tmp_path / "center_cube.fits",
-            tmp_path / "wavelengths.fits",
+    def test_default_covers_every_frame_type(self):
+        assert sr.StepDirs().available_frame_types == ("CORO", "CENTER", "FLUX")
+
+    def test_bundle_output_drops_absent_frame_types(self, tmp_path):
+        d = sr.StepDirs(converted_dir=tmp_path, available_frame_types=self.WAFFLE_NO_CORO)
+        assert [p.name for p in sr.expected_outputs("bundle_output", d)] == [
+            "wavelengths.fits",
+            "center_cube.fits",
         ]
 
[…]
