### 2026-09-22 5099c71
fix(pipeline): declare the IFS flux and inverse-variance cubes
bundle_output writes a data cube and its inverse-variance sibling together,
unconditionally, for every frame type it bundles, plus wavelengths.fits. The
registry declared only the CORO and CENTER data cubes, so flux_cube.fits and
all three ivar cubes were written on every run and gated on by nothing.

A run that died partway through bundling therefore left enough behind for
should_run to skip the step on the next invocation, and for check_output to
call the target complete. The missing products are all consumed later:
flux_psf_cali
[…]
