### 2026-09-23 8f61eba
Changelog for the center->coro symlink purge (#175)
Updates the pipeline.science_frames entry in place rather than adding a
second one, and collapses the duplicate "Fixed" heading the develop merge
left in [Unreleased].
