+++ CHANGELOG.md
@@ -19,9 +19,9 @@ This project follows [Semantic Versioning](https://semver.org/) and the [Keep a 
 - **`StepSpec.leaf` marks steps whose outputs feed nothing downstream** – Forcing a leaf re-runs only itself instead of cascading into TRAP, and leaf steps are excluded from `check_output`'s completeness sweep, so an opt-in step nobody enabled does not make a finished reduction look incomplete.
   Being forced *by* an earlier step still works
   ([#161](https://github.com/m-samland/spherical/issues/161), [@m-samland](https://github.com/m-samland)).
-- **`pipeline.science_frames`** holds the science frame type, the centre-to-frame-axis normalization and the frame-axis consistency check, previously inline in the TRAP wrapper.
+- **`pipeline.science_frames`** holds the science frame type, the frame types an observation actually carries (`frame_types_present`), the centre-to-frame-axis normalization, the frame-axis consistency check and the name of the `HIERARCH SPHERICAL WAFFLE MODE` card, previously inline in the TRAP wrapper.
   It imports numpy only, so using it implies no TRAP dependency
-  ([#161](https://github.com/m-samland/spherical/issues/161), [@m-samland](https://github.com/m-samland)).
+  ([#161](https://github.com/m-samland/spherical/issues/161), [#175](https://github.com/m-samland/spherical/pull/175), [@m-samland](https://github.com/m-samland)).
 - **Multi-epoch target selection** – `database.multi_epoch_filter.select_multi_epoch_targets()` keeps hosts observed on at least two nights, in any combination of modes, whose proper motion moves a stationary background source by at least one pixel between the first and last epoch, the precondition for telling a comoving companion from a background star.
   Surviving rows get the number of nights, span and predicted background motion. Apply quality cuts first, since the span is measured over the rows passed in.
   `read_host_list()` reads a name-per-line file for `SphereDatabase.filter(exclude_targets=...)`
@@ -63,6 +63,11 @@ This project follows [Semantic Versioning](https://semver.org/) and the [Keep a 
   `astropy-healpix` is BSD-3 like spherical, depends only on numpy and astropy, and has wheels for macOS, Linux and Windows.
   The base install now has no blocker on Windows; the `pipeline` extra still does, since `charis` and `trap` come from git with no Windows story, and there is no Windows CI
   ([#140](https://github.com/m-samland/spherical/issues/140), [@m-samland](https://github.com/m-samland)).
+- **The science frame type is derived from `WAFFLE_MODE` instead of from a symlink** – `_link_center_as_coro_if_missing` pointed `coro_cube.fits` at `center_cube.fits` for continuous-waffle sequences with no CORO frames.
+  Since `cube_header_update` opens cubes with `mode='update'`, a header write to one silently rewrote the other. What the symlink actually propped up was the resume check, now fixed at its source.
+  `cube_header_update` stamps `HIERARCH SPHERICAL WAFFLE MODE` onto every cube, so a standalone `align_frames` re-run can resolve the science frame type without an observation object; a reduction predating the card reports the missing keyword rather than guessing.
+  `align_frames.infer_continuous_satellite_spots` is replaced by `resolve_waffle_mode`. The old heuristic read the centre array's frame axis, which is the CENTER axis for IFS either way, so every IFS observation took the waffle branch and would have aligned the calibration frames without raising
+  ([#175](https://github.com/m-samland/spherical/pull/175), [@m-samland](https://github.com/m-samland)).
 
 ### 🐛 Fixed
 - **`imutils.shift` raised `AttributeError` for any sequence shift value** – `collections.Iterable` was removed in Python 3.10, so the public wrapper had been dead for every supported interpreter
[…]
