### 2026-09-23 2d451c7
fix(pipeline): stop claiming a bad-pixel repair that may not have happened
_run_bad_pixel_repair returned True for IRDIS whenever repair_bad_pixels was
set, on the reasoning that preprocess already did the work. But that work is
gated by fix_badpix, a separate flag the alignment step never consulted, so
HIERARCH SPHERICAL ALIGN REPAIRED read True even with fix_badpix=False. The
keyword exists to be checked, which makes a confident wrong answer the worst
of the three possible ones.

run_frame_alignment takes fix_badpix and the IRDIS orchestrator passes it.
A standalone re-run has no preprocess 
[…]
