### 2026-09-23 00ca9ae
Changelog for the align_frames repair-state fix
