### 2026-09-23 3968444
fix(pipeline): check IFS completeness against the configured frame types
The driver narrows its declared outputs by preprocessing.frame_types_to_extract
intersected with what has files, but check_output read frame_types_present
alone. A reduction configured without FLUX therefore ran correctly, resumed
correctly, and was reported incomplete forever, naming products the pipeline
had been told not to make.

check_output takes frame_types_to_extract and intersects the two. Defaults to
all three, so existing callers are unaffected. The narrowing lives in
science_frames.configured_frame_types, next
[…]
