### 2026-09-24 8b0688d
Tidying up changelog entries
