### 2026-09-24 0a9af88
Address #174 review: check_output wiring, align_frames cleanup, docs
execute_targets called check_output without frame_types_to_extract, so the
narrowing added for #175 never reached the pipeline's own completeness check
and a reduction configured without FLUX still reported incomplete.

align_frames:
- Drop AlignmentConfig.repair_bad_pixels. It repaired nothing: on IRDIS it
  echoed fix_badpix, on IFS it logged a warning on every default run. ALIGN
  REPAIRED is now written directly (False on IFS, fix_badpix on IRDIS,
  UNKNOWN when not passed).
- Raise when an fft or coarse shift exceeds pad
[…]
