+++ src/spherical/pipeline/pipeline_config.py
@@ -324,12 +324,27 @@ class PipelineStepsConfig:
 
 # --- Composite reduction config --------------------------------------------
 
+ALIGN_SHIFT_METHODS = ("auto", "fft", "interp", "coarse")
+
+# Padding carried around the frame during the shift and removed afterwards.
+# An FFT shift is periodic and would wrap flux from one edge to the other; a
+# cubic spline's support reaches 2 px past the border. Shifts are sub-pixel once
+# the crop origin puts the star within half a pixel of the centre, so 8 is
+# generous for both.
+DEFAULT_ALIGN_PAD_WIDTH = 8
+
+
 @dataclass(slots=True)
 class AlignmentConfig:
     """Parameters for the optional ``align_frames`` step.
 
     The aligned cube is a leaf product for external consumers (classical
     ADI/PCA, SDI, inspection). Nothing in the pipeline reads it back.
+
+    The step is gated on a marker file rather than on the aligned cube, so
+    deleting the cube does not regenerate it and changing these parameters does
+    not re-run it. Use ``config.steps.force = {"align_frames"}`` for either
+    (`#177 <https://github.com/m-samland/spherical/issues/177>`_).
     """
 
     # "auto" uses FFT on frames with no NaN (cropped IRDIS) and a cubic spline
@@ -338,17 +353,14 @@ class AlignmentConfig:
     # would spread across the whole frame. "coarse" rounds to an integer shift
     # and does not interpolate at all.
     shift_method: str = "auto"
-    pad_width: int = 8
-    # Gate for the pre-shift bad-pixel repair stage. Satisfied on IRDIS by
-    # `fix_badpix` in preprocess; on IFS the interpolation is not implemented yet
-    # and the step logs a warning and passes the cube through.
-    repair_bad_pixels: bool = True
+    # Must exceed the largest shift for "fft" and "coarse"; the step raises
+    # rather than wrap flux across the frame.
+    pad_width: int = DEFAULT_ALIGN_PAD_WIDTH
 
     def __post_init__(self) -> None:
-        valid = ("auto", "fft", "interp", "coarse")
-        if self.shift_method not in valid:
+        if self.shift_method not in ALIGN_SHIFT_METHODS:
             raise ValueError(
-                f"shift_method must be one of {valid}, got {self.shift_method!r}."
+                f"shift_method must be one of {ALIGN_SHIFT_METHODS}, got {self.shift_method!r}."
             )
         if self.pad_width < 0:
             raise ValueError(f"pad_width must be >= 0, got {self.pad_width}.")
