+++ src/spherical/pipeline/steps/align_frames.py
@@ -11,18 +11,21 @@ that is the same reason TRAP works on unshifted data.
 """
 from __future__ import annotations
 
+from pathlib import Path
+
 import numpy as np
+import pandas as pd
+from astropy.io import fits
+from scipy import ndimage
 
-from spherical.pipeline.science_frames import WAFFLE_KEYWORD
-
-# Padding carried around the frame during the shift and removed afterwards.
-# An FFT shift is periodic and would wrap flux from one edge to the other; a
-# cubic spline's support reaches 2 px past the border. Shifts are sub-pixel once
-# the crop origin puts the star within half a pixel of the centre, so 8 is
-# generous for both.
-DEFAULT_PAD_WIDTH = 8
-
-_VALID_METHODS = ("auto", "fft", "interp", "coarse")
+from spherical.pipeline import imutils
+from spherical.pipeline.pipeline_config import ALIGN_SHIFT_METHODS, DEFAULT_ALIGN_PAD_WIDTH
+from spherical.pipeline.science_frames import (
+    WAFFLE_KEYWORD,
+    normalize_centers_to_frames,
+    science_frame_type,
+    verify_frame_axis,
+)
 
 
 def pad_to_odd(cube: np.ndarray) -> np.ndarray:
@@ -64,7 +67,7 @@ def shift_frame(
     frame: np.ndarray,
     shift_xy: tuple[float, float],
     method: str = "auto",
-    pad: int = DEFAULT_PAD_WIDTH,
+    pad: int = DEFAULT_ALIGN_PAD_WIDTH,
 ) -> np.ndarray:
     """Shift one square frame by ``(dx, dy)`` pixels, preserving its shape.
 
@@ -89,21 +92,26 @@ def shift_frame(
         Shifted frame, same shape as the input, as float32.
 
     Raises:
-        ValueError: If ``method`` is unknown, or the frame is not square.
+        ValueError: If ``method`` is unknown, the frame is not square, or an
+            ``"fft"`` or ``"coarse"`` shift is larger than ``pad``.
     """
-    from scipy import ndimage
-
-    from spherical.pipeline import imutils
-
-    if method not in _VALID_METHODS:
+    if method not in ALIGN_SHIFT_METHODS:
         raise ValueError(
-            f"Unknown shift method {method!r}; expected one of {_VALID_METHODS}."
+            f"Unknown shift method {method!r}; expected one of {ALIGN_SHIFT_METHODS}."
         )
     if frame.ndim != 2 or frame.shape[0] != frame.shape[1]:
         raise ValueError(f"shift_frame requires a square 2-D frame, got {frame.shape}.")
 
     dx, dy = float(shift_xy[0]), float(shift_xy[1])
     resolved = _resolve_method(frame, method)
+    # Both shifts are periodic, so anything beyond the padding wraps flux from
+    # the opposite edge into the frame at full amplitude. The spline clamps to
+    # the nearest edge value instead and needs no such limit.
+    if resolved in ("fft", "coarse") and max(abs(dx), abs(dy)) > pad:
+        raise ValueError(
+            f"Shift ({dx:.2f}, {dy:.2f}) px exceeds the {pad} px padding, so the "
+            f"{resolved!r} shift would wrap flux across the frame. Increase pad_width."
+        )
 
     invalid = np.isnan(frame)
     has_nan = bool(invalid.any())
@@ -137,7 +145,7 @@ def shift_to_target(
     frame: np.ndarray,
     center_xy: tuple[float, float],
     method: str = "auto",
-    pad: int = DEFAULT_PAD_WIDTH,
+    pad: int = DEFAULT_ALIGN_PAD_WIDTH,
 ) -> np.ndarray:
[…]
