### 2026-09-24 579ad78
Record the crop on every product that carries crop coordinates
With irdis_preprocessing.crop on, badpixel_map.fits and the image_centers*
files are in crop coordinates, but only the science cubes said so. Which
frame a file is in cannot be recovered from the data, and a consumer that
assumes detector coordinates is off by the crop origin, hundreds of pixels,
with nothing to raise on. TRAP indexes the bad-pixel map against the cube,
and when both are cropped the shapes agree, so a wrong origin is invisible
without the cards.

The card names were spelled out in three modules. crop_provenance now
[…]
