### 2026-09-24 41128f0
Fold the crop cards into irdis_preprocess
crop_provenance was a new module with no import-graph reason to exist:
irdis_preprocess makes the crop and already holds the nominal star
positions, and find_star and process_centers already import from it.

Trimmed on the way:
- Three helpers instead of four. crop_cards() takes a header, so the
  cube lookup with its CORO fallback is gone; the IRDIS centre steps
  always have center_cube.fits, since find_star reads it.
- The IFS centre fit no longer copies cards. IFS is never cropped, so the
  call could only ever return an empty header.
- No header .
[…]
