### 2026-09-24 1455553
Decide primary science between CORO and CENTER only
FLUX frames are never science, but select_primary_science_frames let them
win the exposure-time contest. A sequence aborted before its coronagraphic
frames was then labelled FLUX, got WAFFLE_MODE=False with no CORO frames,
and had TOTAL_EXPTIME_SCI and the rest of its metadata computed from the
flux frames. FLUX is now only the label of a flux-only sequence, and a
group with no exposure at all returns no rows so the caller skips it.

check_frames() treated a None FLAT as present and raised TypeError; it now
reports it as missing.

This change
[…]
