+++ tests/database/test_observation_aborted_sequences.py
@@ -15,6 +15,8 @@ from astropy.table import Table
 
 from spherical.database.ifs_observation import IFSObservation
 from spherical.database.irdis_observation import IRDISObservation
+from spherical.database.observation_table import select_primary_science_frames
+from spherical.database.sphere_database import USABLE_MIN_EXPTIME_SCI
 
 NIGHT = "2026-06-16"
 
@@ -102,3 +104,95 @@ def test_coronagraphic_sequence_still_splits_centers(cls, mode_key, mode_value):
     assert len(obs.frames["CENTER_AFTER"]) == 1
     assert obs.frames["CENTER_BEFORE"]["MJD_OBS"][0] < obs.frames["CORO"]["MJD_OBS"][0]
     assert obs.frames["CENTER_AFTER"]["MJD_OBS"][0] > obs.frames["CORO"]["MJD_OBS"][-1]
+
+
+# ---------------------------------------------------------------------------
+# PRIMARY_SCIENCE contest (#179): FLUX frames are never science, so the contest
+# runs between CORO and CENTER, and FLUX is only the label of last resort.
+# ---------------------------------------------------------------------------
+
+
+def _group(frames):
+    """Observation group from ``(dpr_type, exptime, ndit)`` tuples."""
+    return Table(
+        {
+            "DPR_TYPE": [f[0] for f in frames],
+            "EXPTIME": [float(f[1]) for f in frames],
+            "NDIT": [int(f[2]) for f in frames],
+        }
+    )
+
+
+def _science_minutes(rows):
+    """``TOTAL_EXPTIME_SCI`` as ``compute_basic_metadata`` derives it from the primary rows."""
+    return round(float(np.sum(rows["EXPTIME"] * rows["NDIT"])) / 60, 3)
+
+
+CORO, CENTER, FLUX = "OBJECT", "OBJECT,CENTER", "OBJECT,FLUX"
+
+
+def test_flux_only_sequence_keeps_flux_label():
+    kind, rows, _ = select_primary_science_frames(_group([(FLUX, 2.0, 1)]), "NDIT")
+    assert kind == "FLUX"
+    assert len(rows) == 1
+
+
+def test_center_beats_more_flux_exposure_without_coro():
+    """HD 106036: 2 FLUX frames outweigh 1 CENTER cube, but only CENTER is science."""
+    group = _group([(FLUX, 2.0, 60), (FLUX, 2.0, 60), (CENTER, 4.0, 15)])
+    kind, rows, exptimes = select_primary_science_frames(group, "NDIT")
+    assert kind == "CENTER"
+    assert list(rows["DPR_TYPE"]) == [CENTER]
+    assert _science_minutes(rows) == 1.0
+    assert exptimes["TOTAL_EXPTIME_FLUX"] == 4.0
+
+
+def test_single_cube_waffle_counts_its_full_exposure():
+    """NCENTER=1 with a large NDIT is a real sequence, and must clear the usable cut."""
+    group = _group([(FLUX, 2.0, 10), (CENTER, 4.0, 100)])
+    kind, rows, _ = select_primary_science_frames(group, "NDIT")
+    assert kind == "CENTER"
+    assert _science_minutes(rows) >= USABLE_MIN_EXPTIME_SCI
+
+
+def test_coronagraphic_sequence_with_one_center_stays_coro():
+    group = _group([(CENTER, 4.0, 1)] + [(CORO, 32.0, 4)] * 10 + [(FLUX, 2.0, 1)])
+    kind, _, _ = select_primary_science_frames(group, "NDIT")
+    assert kind == "CORO"
+
+
+def test_center_outweighing_stray_coro_frames_is_waffle():
[…]
