### 2026-09-24 0238d59
Keep a NaN exposure from winning the primary-science contest
totals['CORO'] >= nan is False, so a NaN CENTER total would hand the
win to CENTER and mark a coronagraphic sequence as waffle. Sum with
nansum instead. No row in the v3.0.0 tables has a NaN exposure.
