### 2026-09-24 df01449
Wrap the hour angle so the derotation angle survives the LST wrap
compute_angles took the hour angle as LST - RA without wrapping it, so it
spans (-24h, 24h). parallactic_angle adds 360 deg for ha < 0 when the
target is north of the zenith, which treated +0.9h and -23.1h differently.
A sequence during which the sidereal time wrapped past 0h therefore jumped
by 360 deg, and the observation table's ROTATION reported its complement
(343 deg instead of 17 deg for HR 8799 on 2016-11-17).

Wrap to [-12h, 12h) where the hour angle is computed and inside
parallactic_angle. Rebuilding the v3.0.0 observ
[…]
