### 2026-09-24 fcfb78d
Guess the flux PSF position per IRDIS channel instead of on a wavelength median (#170)
IRDIS FLUX cubes are never cropped, so the two channels keep the
cross-channel offset and a single guess on their median is wrong for at
least one of them. IFS keeps the median guess.

star_centers_from_PSF_img_cube now accepts one guess per channel.
