### 2026-09-24 4427a05
Shorten the #170 changelog entry
