### 2026-09-25 2b9b795
Order and pair flux calibration frames by MJD, not LST
LST wraps at 24h, so for a sequence crossing LST 0h the flux blocks were
merged, sorted out of order (leaving an empty, zeroed PSF block) and paired
with the wrong CENTER frames. The LST columns stay in
flux_calibration_indices.csv for labelling, so spot_to_flux still reads
indices written before this change. Fixes #193.
