+++ src/spherical/pipeline/flux_calibration.py
@@ -204,24 +204,30 @@ def get_aperture_photometry(flux_stamps, aperture_radius_range=[1, 15],
 
 
 def get_flux_calibration_indices(frames_info_center, frames_info_flux, number_excluded_after_coro=None):
+    # Order, pair and split on MJD: LST wraps at 24h, which reverses the order
+    # and breaks the pairing for a sequence crossing LST 0h (#193). The LST
+    # columns are kept for labelling.
+    def calibration_row(flux_idx):
+        science_idx = find_nearest(
+            array=frames_info_center['MJD'], value=frames_info_flux['MJD'].iloc[flux_idx])
+        return {'flux_idx': flux_idx,
+                'flux_lst': frames_info_flux['LST'].iloc[flux_idx],
+                'flux_mjd': frames_info_flux['MJD'].iloc[flux_idx],
+                'science_idx': science_idx,
+                'science_lst': frames_info_center['LST'].iloc[science_idx],
+                'science_mjd': frames_info_center['MJD'].iloc[science_idx]}
+
     flux_calibration_indices = []
     number_of_flux_frames = len(frames_info_flux)
     if number_of_flux_frames == 1:
-        science_idx = find_nearest(
-            array=frames_info_center['LST'], value=frames_info_flux['LST'].iloc[0])
-        flux_calibration_indices.append({
-            'flux_idx': 0, 
-            'flux_lst': frames_info_flux['LST'].iloc[0],
-            'science_idx': science_idx,
-            'science_lst': frames_info_center['LST'].iloc[science_idx]
-        })
+        flux_calibration_indices.append(calibration_row(0))
         indices_of_discontinuity = np.array([], dtype=int)  # Empty array for single frame case
     else:
         # assert np.all(
         #     frames_info_flux['EXPTIME'] == frames_info_flux['EXPTIME'].iloc[0]), "Different exposure times for flux frames"
         # flux_time_diff = np.diff(frames_info_flux['LST'])
         flux_time_diff_in_dit = np.diff(
-            frames_info_flux['LST'] * 60 * 60) / np.max(frames_info_flux['EXPTIME'].iloc[0])
+            frames_info_flux['MJD'] * 24 * 60 * 60) / np.max(frames_info_flux['EXPTIME'].iloc[0])
 
         # center_time_diff = np.diff(frames_info_center['LST'])
         # center_time_diff_in_dit = np.diff(
@@ -235,42 +241,21 @@ def get_flux_calibration_indices(frames_info_center, frames_info_flux, number_ex
         #             f"Excluded {number_excluded_after_coro} flux frames after science seqeunce from normalization.")
         #         indices_of_discontinuity[i] += number_excluded_after_coro
 
-        science_idx_first = find_nearest(
-            array=frames_info_center['LST'], value=frames_info_flux['LST'].iloc[0])
-        science_idx_last = find_nearest(
-            array=frames_info_center['LST'], value=frames_info_flux['LST'].iloc[-1])
-        flux_calibration_indices.append(
-            {'flux_idx': 0,
-             'flux_lst': frames_info_flux['LST'].iloc[0],
-             'science_idx': science_idx_first,
-             'science_lst': frames_info_center['LST'].iloc[science_idx_first]})
-        index_of_last_flux_frame = number_of_flux_frames - 1
-        flux_calibration_indices.append(
-            {'flux_idx': index_of_last_flux_frame,
-             'flux_lst': frames_info_flux['LST'].iloc[index_of_last_flux_frame],
-             'science_idx': science_idx_last,
-             'science_lst': frames_info_center['LST'].iloc[science_idx_last]})
[…]
