### 2026-09-25 0818c75
Add cube-level flux selection that drops saturated cubes
Flux cubes are summarised per raw file, measured by their median core peak
in a box around the PSF guess, and every cube below flux_saturation_adu is
kept whatever its DIT and ND, since those are scaled per frame. Users can
restrict to the cubes before or after the science sequence, or force all
cubes, one index or one file. Part of #172.
