+++ tests/pipeline/test_cube_selection.py
@@ -1,0 +1,171 @@
+"""Flux cube selection (#172)."""
+import numpy as np
+import pandas as pd
+import pytest
+
+from spherical.pipeline.cube_selection import (
+    core_peaks,
+    resolve_cube_selection,
+    select_flux_cubes,
+    summarise_cubes,
+)
+
+
+def _frames(origfiles, dits, nds, mjds):
+    return pd.DataFrame({
+        "ORIGFILE": origfiles, "DET SEQ1 DIT": dits,
+        "INS4 FILT2 NAME": nds, "MJD": mjds,
+    })
+
+
+def _cubes(frames, peak_per_frame, exposure_level=None):
+    """Summarise with a (n_wave=2, n_frames) peak array built from one value per frame."""
+    peaks = np.tile(np.asarray(peak_per_frame, dtype=float), (2, 1))
+    if exposure_level is None:
+        exposure_level = np.asarray(frames["DET SEQ1 DIT"], dtype=float)
+    return summarise_cubes(frames, np.asarray(exposure_level, dtype=float), peaks)
+
+
+def _pi_men():
+    """2 cubes, 90 + 8 DITs, as expand_frames_info produces them."""
+    frames = _frames(
+        origfiles=["a.fits"] * 90 + ["b.fits"] * 8,
+        dits=[0.837464] * 90 + [8.0] * 8,
+        nds=["ND_2.0"] * 90 + ["ND_1.0"] * 8,
+        mjds=list(np.linspace(57376.10, 57376.11, 90)) + list(np.linspace(57376.18, 57376.19, 8)),
+    )
+    return _cubes(frames, [8598.0] * 90 + [47029.0] * 8)
+
+
+def _saturated(cubes, threshold=40000.0):
+    return cubes["peak_adu"].to_numpy() >= threshold
+
+
+# --- core peaks -------------------------------------------------------------
+
+def test_core_peaks_ignores_a_hot_pixel_outside_the_box():
+    cube = np.zeros((2, 1, 200, 200))
+    cube[:, 0, 100, 100] = 9000.0
+    cube[:, 0, 10, 10] = 65000.0
+    peaks = core_peaks(cube, guesses=[[(100, 100), (100, 100)]], box=15)
+    assert peaks.shape == (2, 1)
+    assert np.allclose(peaks[:, 0], 9000.0)
+
+
+def test_core_peaks_uses_the_per_channel_guess():
+    cube = np.zeros((2, 1, 200, 200))
+    cube[0, 0, 100, 100] = 9000.0
+    cube[1, 0, 130, 100] = 47000.0
+    peaks = core_peaks(cube, guesses=[[(100, 100), (130, 100)]], box=15)
+    assert peaks[0, 0] == pytest.approx(9000.0)
+    assert peaks[1, 0] == pytest.approx(47000.0)
+
+
+def test_core_peaks_accepts_one_guess_for_all_channels():
+    """IFS passes one (cy, cx) per frame, not one per channel."""
+    cube = np.zeros((3, 1, 100, 100))
+    cube[:, 0, 50, 50] = [100.0, 200.0, 300.0]
+    peaks = core_peaks(cube, guesses=[(50, 50)], box=5)
+    assert peaks[:, 0].tolist() == [100.0, 200.0, 300.0]
+
+
+# --- cube summary -----------------------------------------------------------
+
+def test_summarise_cubes_collapses_dits_into_cubes():
+    cubes = _pi_men()
+    assert cubes["origfile"].tolist() == ["a.fits", "b.fits"]
+    assert cubes["n_dits"].tolist() == [90, 8]
+    assert cubes["dit"].tolist() == [0.837464, 8.0]
+    assert cubes["peak_adu"].tolist() == [8598.0, 47029.0]
+
+
[…]
