### 2026-09-25 0081c95
Apply flux cube selection before PSF centring, subsetting all co-indexed arrays
Selection runs after the guess loop and before centring and Moffat repair,
and subsets the flux cube, ivar, bad-pixel mask, frame table and guesses in
one place. The full extraction products stay untouched; the selected frame
table and a per-cube report are written alongside. Kept cubes whose core
reaches flux_nonlinearity_adu are flagged and warned about. The flux block
split now measures each gap in units of the longer neighbouring DIT, so
mixed DITs no longer split a long-DIT cube. Part of #172.
