### 2026-09-25 acb8831
Normalise flux frames to their own block's median (#159)
Each flux frame was scaled to its block's mean aperture flux before the
frames were combined, so one frame with bad photometry rescaled the whole
block's PSF. The median is robust to that. For every block after the first
the reference was also sliced from the full frame axis (a block-relative
start used as an absolute index), so later blocks took in the earlier
blocks' flux; the reference now comes from the block alone, which changes
the later PSF blocks whenever the blocks differ in flux. Combined frames
deviating by more than 20% from 
[…]
