### 2026-09-25 5d1e7e3
Read the selected flux frame table in spot_to_flux
After flux cube selection, flux_amplitude_calibrated.fits carries only the
selected frames, so spot_to_flux reads frames_info_flux_selected.csv and
falls back to frames_info_flux.csv for older reductions. A length mismatch
now fails with a clear message instead of inside matplotlib. Part of #172.
